% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_AddLine.R
\name{AddLine}
\alias{AddLine}
\title{Add a line to the current plot}
\usage{
AddLine(at = NULL, Outer = FALSE, H = TRUE, ...)
}
\arguments{
\item{at}{relative location of where the line should be plotted}

\item{Outer}{plot the line out of plotting area. Default: \code{FALSE}}

\item{H}{Horizontal line? if H == FALSE, vertical line will be added. Default: \code{TRUE}}

\item{...}{Other arguments}
}
\description{
Add a line to the current plot
}
\examples{
# Horizontal line
par(oma = c(1, 1, 1, 1), mar = c(3, 3, 1, 1))
plot(1:100)
AddLine(at = 0.75)
AddLine(at = 0.25, Outer = TRUE, lwd = 2)
AddLine(at = 0.5, Outer = TRUE, lwd = 2, col = "red")

# ---------------------------------------------

# Vertical line
plot(1:100)
AddLine(H = FALSE, at = 0.75)
AddLine(H = FALSE, at = 0.25, Outer = TRUE, lwd = 2)
AddLine(H = FALSE, at = 0.5, Outer = TRUE, lwd = 2, col = "red")
}
\references{
\href{https://stackoverflow.com/questions/27800307/}{Click here}
}
