% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{AddMissingCols}
\alias{AddMissingCols}
\title{Add missing columns to data frame}
\usage{
AddMissingCols(DT, FillVal = NA_character_, ...)
}
\arguments{
\item{DT}{data frame}

\item{FillVal}{value to be used: default: NA_character}

\item{...}{list of column names to add}
}
\description{
Add missing columns to data frame
}
\examples{
mtcars \%>\%
 dplyr::select(1:3) \%>\%
 AddMissingCols(FillVal = NA_character_, A, B, C) \%>\%
 AddMissingCols(FillVal = as.integer(10), D)

# -------------------------------------------

AddCols <- c("Add1", "Add2")
mtcars \%>\%
 dplyr::select(1:3) \%>\%
 AddMissingCols(FillVal = NA_real_, AddCols)
}
\author{
Ahmed El-Gabbas
}
