% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_AssignIfNotExist.R
\name{AssignIfNotExist}
\alias{AssignIfNotExist}
\title{Assign a value to a variable if it does not already exist in the specified environment}
\usage{
AssignIfNotExist(Variable, Value, Env = globalenv())
}
\arguments{
\item{Variable}{Character; the name of the variable to be checked and potentially assigned a value.}

\item{Value}{any; the value to be assigned to the variable if it does not already exist.}

\item{Env}{environment; the environment in which to check for the existence of the variable and potentially assign the value. Defaults to the global environment.}
}
\value{
The function explicitly returns \code{NULL}, but its primary effect is the side-effect of assigning a value to a variable in an environment or printing the current value of an existing variable.
}
\description{
This function checks if a given variable exists in the specified environment (global environment by default). If the variable does not exist, it assigns a given value to it. If the variable already exists, it prints the current value of the variable. The function is designed to prevent overwriting existing variables unintentionally.
}
\examples{
AssignIfNotExist(Variable = "x", Value = TRUE)
print(x)

# --------------------------------------------------

y <- 10
# y exists and thus its value was not changed
AssignIfNotExist(Variable = "y", Value = TRUE)
print(y)
}
\author{
Ahmed El-Gabbas
}
