% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Summary.R
\name{BetaSummary}
\alias{BetaSummary}
\alias{Mod_Summary}
\title{Summary of the Hmsc model parameters}
\usage{
Mod_Summary(Path_Coda = NULL, EnvFile = ".env", ReturnData = FALSE)
}
\arguments{
\item{Path_Coda}{String. Path to .RData file containing coda object}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{ReturnData}{Logical. Should the response curve data be returned as an R object? Default: \code{FALSE}}
}
\description{
Summary of the Hmsc model parameters
}
\author{
Ahmed El-Gabbas
}
