% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Bioreg.R
\name{BioReg_Process}
\alias{BioReg_Process}
\title{Process biogeographical regions dataset}
\usage{
BioReg_Process(FromHPC = TRUE, EnvFile = ".env")
}
\arguments{
\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}
}
\value{
Invisible \code{NULL}. Processed data is saved to disk as raster, vector,
and RData files.
}
\description{
Downloads and processes the Biogeographical Regions dataset (Europe 2016, v1)
from the \href{https://www.eea.europa.eu/en/datahub/datahubitem-view/11db8d14-f167-4cd5-9205-95638dfd9618}{European Environment Agency}.
This function extracts biogeographical region names per reference grid cell
for use in counting species presence across biogeographical regions.
}
\details{
\itemize{
\item \emph{Temporal coverage}: 2011-2015
\item \emph{Spatial coverage}: 28째E to 81째E, 31.27째W to 62째E
\item \emph{CRS}: EPSG:3035
\item \emph{file format}: shapefile (compressed in zip file)
\item \emph{Requirements}: \code{curl} (download) and \code{unzip} (extraction)
}
}
\author{
Ahmed El-Gabbas
}
