% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cat_sep.R
\name{cat_sep}
\alias{cat_sep}
\title{Print separator(s) to the console}
\usage{
cat_sep(
  n_separators = 1L,
  lines_before = 0L,
  lines_after = 1L,
  line_char = "-",
  repetitions = 50L,
  bold = FALSE,
  red = FALSE,
  ...
)
}
\arguments{
\item{n_separators}{integer; the number of separator lines to print. Default
is \code{1}.}

\item{lines_before, lines_after}{integer; the number of extra empty lines to
print before and after the separator lines. Default is \code{0} and \code{1},
respectively.}

\item{line_char}{character; the character used to construct the separator
line. Default is \code{"-"}.}

\item{repetitions}{integer; the number of times the character is repeated to
form a separator line. Default is \code{50}.}

\item{bold}{logical; whether to print the text in bold. Default is \code{FALSE}.}

\item{red}{logical; whether to print the text in red. Default is \code{FALSE}.}

\item{...}{additional arguments to be passed to \code{cat()}.}
}
\value{
The function is called for its side effect (printing to the console)
and does not return a meaningful value.
}
\description{
This function prints customizable separator lines to the console, optionally
preceded and followed by empty lines. It is useful for improving the
readability of console output in R scripts or during interactive sessions.
}
\examples{
cat_sep()

cat_sep(2)

cat_sep(2,2,3)

cat_sep(2,2,3, line_char = "*")

cat_sep(2,2,3, line_char = "*", repetitions = 20)
}
\author{
Ahmed El-Gabbas
}
