% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cat_time.R
\name{cat_time}
\alias{cat_time}
\title{Print text with time stamp}
\usage{
cat_time(
  text = "",
  n_lines = 1L,
  time = TRUE,
  bold = FALSE,
  red = FALSE,
  date = FALSE,
  time_zone = "CET",
  level = 0L,
  ...
)
}
\arguments{
\item{text}{character; the text to print before the timestamp. If empty
(default), only the timestamp is printed.}

\item{n_lines}{integer; the number of newline characters to print after the
message. Default is 1.}

\item{time}{logical; whether to include the time in the timestamp. Default is
\code{TRUE}. If \code{FALSE}, only the text is printed.}

\item{bold}{logical; whether to print the text in bold. Default is \code{FALSE}.}

\item{red}{logical; whether to print the text in red. Default is \code{FALSE}.}

\item{date}{logical; whether to include the date in the timestamp. Only
effective if \code{time} is \code{TRUE}. Default is \code{FALSE}, meaning only the time is
printed. If \code{TRUE}, the date is printed in the format "\%d/\%m/\%Y \%X".}

\item{time_zone}{character; the time zone to use for the timestamp. Default
is \code{CET}.}

\item{level}{integer; the level at which the message will be printed. If e.g.
level = 1, the following string will be printed at the beginning of the
message: "   >>>   ". Default is \code{0}.}

\item{...}{additional arguments passed to \code{cat}.}
}
\value{
The function is called for its side effect of printing to the
console.
}
\description{
This function prints a given text followed by the current time (and
optionally the date) to the console. It allows for customization of the time
zone, the inclusion of the date, and the number of newline characters to
print after the message.
}
\examples{
cat_time()

cat_time(date = TRUE)

cat_time("time now")

cat_time("\n\nTime now", n_lines = 2L, level = 1L)

cat_time(
  "\ntime now", date = TRUE, bold = TRUE, red = TRUE,
  n_lines = 2L, level = 1L)

# The use of levels
{
  cat_time("Task 1")
  cat_time("subtask L1", level = 1L)
  cat_time("subtask L2", level = 2L)
  cat_time("subtask L3", level = 3L)
}
}
\author{
Ahmed El-Gabbas
}
