% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CatFunArgs.R
\name{CatFunArgs}
\alias{CatFunArgs}
\title{Print function Arguments}
\usage{
CatFunArgs(Function, Assign = FALSE, Package = NULL)
}
\arguments{
\item{Function}{A function whose arguments you want to print. Must be a valid
R function.}

\item{Assign}{Logical. Whether to assign the arguments as variables in the
global environment. Defaults to \code{FALSE}.}

\item{Package}{Character string. The name of the R package to be loaded
before processing the function. Default is NULL.}
}
\value{
The function prints the formatted arguments to the console. If
\code{Assign} is TRUE, it also assigns arguments to the global environment.
}
\description{
This function takes another function as input and prints its arguments in the
format \code{ArgumentName = DefaultValue}. Each argument is printed on a new line.
The function can optionally assign the formatted arguments to the global
environment and can load a specified package before processing.
}
\examples{
formals(stats::setNames)
CatFunArgs(stats::setNames)
}
\author{
Ahmed El-Gabbas
}
