% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{CatPipe}
\alias{CatPipe}
\title{print a message with time in the middle of the pipe}
\usage{
CatPipe(x, message = NULL)
}
\arguments{
\item{x}{input object to pipe}

\item{message}{printed message}
}
\description{
print a message with time in the middle of the pipe
}
\author{
Ahmed El-Gabbas
}
