% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CatSep.R
\name{CatSep}
\alias{CatSep}
\title{Print separator(s) to the console}
\usage{
CatSep(Rep = 1, Extra1 = 0, Extra2 = 0, Char = "-", CharReps = 50)
}
\arguments{
\item{Rep}{integer; the number of separator lines to print. Default is \code{1}.}

\item{Extra1, Extra2}{integer; the number of extra empty lines to print before
and after the separator lines. Default is \code{0}.}

\item{Char}{character; the character used to construct the separator line.
Default is \code{"-"}.}

\item{CharReps}{integer; the number of times the character is repeated to
form a separator line. Default is \code{50}.}
}
\value{
The function is called for its side effect (printing to the console)
and does not return a meaningful value.
}
\description{
This function prints customizable separator lines to the console, optionally
preceded and followed by empty lines. It is useful for improving the
readability of console output in R scripts or during interactive sessions.
}
\examples{
CatSep()

CatSep(2)

CatSep(2,2,3)

CatSep(2,2,3, Char = "*")

CatSep(2,2,3, Char = "*", CharReps = 20)
}
\author{
Ahmed El-Gabbas
}
