% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{CatTime}
\alias{CatTime}
\title{Print text and a time stamp}
\usage{
CatTime(Text = "", NLines = 1, Date = FALSE, ...)
}
\arguments{
\item{Text}{the text to print: default empty string (print time only)}

\item{NLines}{number of empty lines after the printing; default: 1}

\item{Date}{Also print date? Default value \code{FALSE}}

\item{...}{other arguments passed to \code{cat}}
}
\description{
Print text and a time stamp
}
\examples{
CatTime()
CatTime(Date = TRUE)

CatTime("Time now")
CatTime("Time now", Date = TRUE)
}
\author{
Ahmed El-Gabbas
}
