% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{CatTimeSince}
\alias{CatTimeSince}
\title{Print time since a time event}
\usage{
CatTimeSince(InitTime, Prefix = "Completed  in ", CatInfoChunk = TRUE)
}
\arguments{
\item{InitTime}{day and time of the time event}

\item{Prefix}{Prefix character to print}

\item{CatInfoChunk}{logical; also print \code{IASDT.R::InfoChunk("Session summary")}}
}
\description{
Print time since a time event
}
\author{
Ahmed El-Gabbas
}
