% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{cc}
\alias{cc}
\title{Concatenate without quotes}
\usage{
cc(...)
}
\arguments{
\item{...}{one or more string to concatenate}
}
\description{
Concatenate without quotes
}
\examples{
cc(A, B, C)

# -------------------------------------------

cc(A, B, "A and B")
}
\author{
Ahmed El-Gabbas
}
