% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{ChangeClass}
\alias{ChangeClass}
\title{Change the class of some of GBIF columns}
\usage{
ChangeClass(DF)
}
\arguments{
\item{DF}{Data frame}
}
\description{
Change the class of some of GBIF columns
}
\details{
A function to extract EASIN data for a given EASIN_ID
}
\author{
Ahmed El-Gabbas
}
