% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckData.R
\name{Check_data}
\alias{Check_data}
\alias{CheckData}
\alias{CheckRData}
\alias{CheckQs}
\alias{CheckRDS}
\alias{CheckFeather}
\title{Check the integrity of \code{RData} / \code{qs2} / \code{rds} / \code{feather} files}
\usage{
CheckData(File, warning = TRUE, nthreads = 5)

CheckRData(File, warning = TRUE)

CheckQs(File, warning = TRUE, nthreads = 5)

CheckRDS(File, warning = TRUE)

CheckFeather(File, warning = TRUE)
}
\arguments{
\item{File}{Character. The path to the file to be checked. Cannot be empty.}

\item{warning}{logical. If \code{TRUE} (default), warnings will be printed if the
file does not exist.}

\item{nthreads}{Integer. The number of threads to use when reading qs2 files.
Default is 5.}
}
\value{
Logical. \code{TRUE} if the file is valid, otherwise \code{FALSE}.
}
\description{
These functions validate a given file by checking its extension and
attempting to load its contents. A file is considered valid if it loads
successfully and contains a non-null object, returning \code{TRUE}. Otherwise, it
returns \code{FALSE}.
}
\details{
The \code{CheckData()} function determines the file type based on its
extension. If the extension is unrecognized, it returns \code{FALSE}. Supported
file types:
\itemize{
\item \strong{RData}: Checked with \code{CheckRData()}, read using \link{LoadAs}
\item \strong{qs2}: Checked with \code{CheckQs()}, read using \link[qs2:qs_read]{qs2::qs_read}
\item \strong{rds}: Checked with \code{CheckRDS()}, read using \link{readRDS}
\item \strong{feather}: Checked with \code{CheckFeather()}, read using
\link[arrow:read_feather]{arrow::read_feather}
}
}
\author{
Ahmed El-Gabbas
}
