% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_CheckGBIF.R
\name{Check_GBIF}
\alias{Check_GBIF}
\title{Check if GBIF login credentials exist}
\usage{
Check_GBIF(Renviron = ".Renviron")
}
\arguments{
\item{Renviron}{Character. The path to the \code{.Renviron} file containing GBIF
login credentials: \code{GBIF_EMAIL}, \code{GBIF_USER}, and \code{GBIF_PWD}. Defaults to
the current working directory's \code{.Renviron} file. The credentials must be
in the format:
\itemize{
\item \code{GBIF_EMAIL=your_email}
\item \code{GBIF_USER=your_username}
\item \code{GBIF_PWD=your_password}
}}
}
\description{
Check if GBIF login credentials exist
}
\author{
Ahmed El-Gabbas
}
