% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckArgs.R
\name{CheckArgs}
\alias{CheckArgs}
\title{Check function arguments}
\usage{
CheckArgs(AllArgs, Args, Type)
}
\arguments{
\item{AllArgs}{String. Vector for the input parameters of the function. Usually as a result of \code{formals()} function}

\item{Args}{Arguments to check}

\item{Type}{Class of the arguments to check. This has to be one of "character", "logical", or "numeric".}
}
\description{
Check function arguments
}
\author{
Ahmed El-Gabbas
}
