% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckFeather.R
\name{CheckFeather}
\alias{CheckFeather}
\title{Check the Integrity of an \code{feather} File}
\usage{
CheckFeather(File, warning = TRUE)
}
\arguments{
\item{File}{A character string specifying the path to the file to be checked.
This can not be empty. The function will attempt to determine the file type
based on the file extension. If the file extension is not \code{feather}, the
function will return \code{FALSE}.}

\item{warning}{logical. If \code{TRUE} (default), warnings will be printed if the
file does not exist.}
}
\value{
A logical value: \code{TRUE} if the file is an \code{feather} file and contains
a non-null object, \code{FALSE} otherwise.
}
\description{
This function checks if the specified file is an \code{feather} file and attempts
to read it using \link[arrow:read_feather]{arrow::read_feather}. It returns \code{TRUE} if the file is an
\code{qs} file and contains a non-null object, otherwise \code{FALSE.}
}
\author{
Ahmed El-Gabbas
}
