% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckRStudioVersion.R
\name{CheckRStudioVersion}
\alias{CheckRStudioVersion}
\title{Check if \code{RStudio} should be updated}
\usage{
CheckRStudioVersion()
}
\description{
Check if \code{RStudio} should be updated
}
\examples{
\dontrun{
CheckRStudioVersion()
}
}
\author{
Ahmed El-Gabbas
}
