% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_spatial.R
\name{CheckStackInMemory}
\alias{CheckStackInMemory}
\title{Check if the raster stack reads from disk or memory}
\usage{
CheckStackInMemory(Stack = NULL)
}
\arguments{
\item{Stack}{Stack}
}
\description{
Check if the raster stack reads from disk or memory
}
\examples{
require(raster, warn.conflicts = F, quietly = T)
logo <- raster(system.file("external/rlogo.grd", package = "raster"))
logo@data@inmemory
logo@data@fromdisk
logo@file@name

# -------------------------------------------

# A raster stack reading from files
ST2 <- raster::stack(logo, logo)
CheckStackInMemory(ST2)
c(ST2[[1]]@data@inmemory, ST2[[2]]@data@inmemory)
c(ST2[[1]]@data@fromdisk, ST2[[2]]@data@fromdisk)
c(ST2[[1]]@file@name, ST2[[2]]@file@name)

# -------------------------------------------

logo2 <- raster::readAll(logo)
ST3 <- raster::stack(logo, logo2)
CheckStackInMemory(ST3)
c(ST3[[1]]@data@inmemory, ST3[[2]]@data@inmemory)
c(ST3[[1]]@data@fromdisk, ST3[[2]]@data@fromdisk)
c(ST3[[1]]@file@name, ST3[[2]]@file@name)
}
\author{
Ahmed El-Gabbas
}
