% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_CheckTiff.R
\name{CheckTiff}
\alias{CheckTiff}
\title{Check if tiff file corrupted}
\usage{
CheckTiff(x)
}
\arguments{
\item{x}{tiff file path}
}
\value{
logical: \code{TRUE} when the tiff file is okay; \code{FALSE} when corrupted
}
\description{
Check if tiff file corrupted
}
\examples{
(f <- system.file("ex/elev.tif", package="terra"))

CheckTiff(f)
}
\author{
Ahmed El-Gabbas
}
