% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckURL.R
\name{CheckURL}
\alias{CheckURL}
\title{Check the validity of a URL}
\source{
The source code of this function was taken from this
\href{https://stackoverflow.com/q/52911812}{stackoverflow} discussion.
}
\usage{
CheckURL(url_in, t = 2)
}
\arguments{
\item{url_in}{Character. The URL to be checked.}

\item{t}{Numeric. Timeout in seconds for the connection attempt. Default is
2 seconds.}
}
\value{
A logical value: \code{TRUE} if the URL is valid, \code{FALSE} if not.
}
\description{
This function opens a connection to the specified URL to check its validity.
It returns \code{TRUE} if the URL is valid (i.e., the connection can be opened),
and \code{FALSE} otherwise.
}
\examples{
urls <- c(
     "http://www.amazon.com", "http://this.isafakelink.biz",
     "https://stackoverflow.com", "https://stack-overflow.com")
sapply(urls, CheckURL)
}
