% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_CHELSA_Prepare.R
\name{CHELSA_Prepare}
\alias{CHELSA_Prepare}
\title{Prepare and Download CHELSA Climate Data}
\usage{
CHELSA_Prepare(
  EnvFile = ".env",
  FromHPC = TRUE,
  Download = FALSE,
  NCores = 4,
  Overwrite = FALSE,
  Download_Attempts = 10,
  Sleep = 5,
  BaseURL = "https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/"
)
}
\arguments{
\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{Download}{Logical, whether to download the CHELSA files. Defaults to
\code{FALSE}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Defaults to 4.}

\item{Overwrite}{Logical, whether to re-download files that already exist.
Defaults to \code{FALSE}.}

\item{Download_Attempts}{Integer. The maximum number of download attempts.
Defaults to 10.}

\item{Sleep}{Integer. Time in seconds to wait after each download attempt.
Defaults to 5.}

\item{BaseURL}{String, the base URL for downloading CHELSA climate data.}
}
\description{
This function reads the contents of local text files containing URLs for
\code{CHELSA} climate data and extracts and saves their metadata (e.g., variable
name, climate model & scenario, time, and download URLs). It can optionally
download these links to desk, if requested. Only bioclimatic variables are
considered.
}
\details{
The function returns information on 874 tif files, representing 19
bioclimatic variables at 46 climate options (current and 45 future
scenarios)
}
\author{
Ahmed El-Gabbas
}
