% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_Chelsa_Prepare_List.R
\name{Chelsa_Prepare_List}
\alias{Chelsa_Prepare_List}
\title{Prepare list of potential variables}
\usage{
Chelsa_Prepare_List(
  Down = FALSE,
  DownParallel = TRUE,
  DwnPath = NULL,
  OutPath = NULL,
  UpdateExisting = FALSE,
  Path_Chelsa = "Data/Chelsa"
)
}
\arguments{
\item{Down}{Logical. Download Chelsa files?}

\item{DownParallel}{Logical. Download input files on parallel (if \code{Down} = \code{TRUE})}

\item{DwnPath}{String. Download path}

\item{OutPath}{String. Output path}

\item{UpdateExisting}{Logical. Reprocess existing file?}

\item{Path_Chelsa}{String. Path for Chelsa analyses}
}
\description{
Prepare list of potential variables
}
\details{
list of variables exist under current and future climates.

46 variables available at 46 options (current and 45 future scenarios)
}
\author{
Ahmed El-Gabbas
}
