% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_CHELSA_Project.R
\name{CHELSA_Project}
\alias{CHELSA_Project}
\title{Project CHELSA data to the study area}
\usage{
CHELSA_Project(
  Metadata = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  CompressLevel = 5,
  ReturnMap = FALSE
)
}
\arguments{
\item{Metadata}{\code{tibble}. Single row tibble for the metadata of the input
file. This should be prepared in the \link{CHELSA_Prepare} function and provided
via the the \link{CHELSA_Process} function.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{CompressLevel}{Integer. The compression level for the exported NetCDF
file, ranging from 1 (least compression) to 9 (most compression). Defaults
to 5.}

\item{ReturnMap}{Logical. If \code{TRUE}, the processed map (as a
\code{PackedSpatRaster} object) is returned. Defaults to \code{FALSE}.}
}
\value{
Depending on the \code{ReturnMap} parameter, either a \code{PackedSpatRaster}
object is returned, or nothing is returned. The function always writes
NetCDF and TIFF files to disk.
}
\description{
This function processes CHELSA climate data, projects it to a specified
reference grid, and optionally saves the output in NetCDF or TIFF format. It
supports downloading data from a URL, applying a land mask, and adjusting
data with scale and offset values.
}
\author{
Ahmed El-Gabbas
}
