% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwf_chelsa_variables.R
\docType{data}
\name{chelsa_variables}
\alias{chelsa_variables}
\title{Detailed information on CHELSA climate variables}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 6 columns.
}
\usage{
chelsa_variables
}
\value{
A tibble containing detailed information about various climate
variables in \code{CHELSA} climate data. Each row in the tibble represents a
different climate variable, with columns providing additional details
such as the long name, unit, scale, offset, and an explanation of the
variable
}
\description{
Detailed information on CHELSA climate variables
}
\examples{
ecokit::load_packages(dplyr, gt)

data(chelsa_variables)
gt::gt(chelsa_variables)
}
\keyword{datasets}
