% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_Chelsa_Vars.R
\name{Chelsa_Vars}
\alias{Chelsa_Vars}
\title{Detailed information on Chelsa data}
\usage{
Chelsa_Vars()
}
\description{
Detailed information on Chelsa data
}
\details{
https://chelsa-climate.org/bioclim/
}
\examples{
print(Chelsa_Vars(), n = Inf)
}
\author{
Ahmed El-Gabbas
}
