% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_CLC_Process.R
\name{CLC_Process}
\alias{CLC_Process}
\title{Process Corine land cover (CLC) data}
\usage{
CLC_Process(EnvFile = ".env", FromHPC = TRUE, MinLandPerc = 15, PlotCLC = TRUE)
}
\arguments{
\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{MinLandPerc}{A numeric value indicating the minimum percentage of land
per grid cell to be used in the reference grid cell. Defaults to 15.}

\item{PlotCLC}{Logical indicating whether to plot the percentage coverage of
different levels of CLC and custom habitat types. Defaults to \code{TRUE}.}
}
\value{
The function does not return a value but produces side effects such
as saving processed data and plots to specified directories.
}
\description{
Processes Corine Land Cover (CLC) data for environmental modeling purposes.
The function calculates the percentage coverage of CLC per grid cell. It
estimate the percent coverage at the 3 CLC levels, and cross-walk for
\code{EUNIS_19} and \code{SynHab} habitat types. Similarly, the function estimates the
most common class per grid cell (3 CLC levels, EUNIS, and SynHab habitat
types) and prepares reference grid for models. The function optionally plots
\% coverage maps.
}
\details{
\itemize{
\item \href{https://land.copernicus.eu/pan-european/corine-land-cover/clc2018}{Data source}
\item \href{https://doi.org/10.2909/960998c1-1870-4e82-8051-6485205ebbac}{Citation}
}
}
\author{
Ahmed El-Gabbas
}
