% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_spatial.R
\name{ClipRasterByPolygon}
\alias{ClipRasterByPolygon}
\title{Clip raster by a spatial polygon}
\usage{
ClipRasterByPolygon(raster = NULL, shape = NULL)
}
\arguments{
\item{raster}{raster layer}

\item{shape}{Polygon}
}
\description{
Clip raster by a spatial polygon
}
\examples{
require(sp)
require(raster)

# Example Polygon
data(wrld_simpl, package = "maptools")
SPDF <- subset(wrld_simpl, NAME == "Germany")

# Example RasterLayer
r <- raster::raster(nrow = 1e3, ncol = 1e3, crs = sp::proj4string(SPDF))
r[] <- 1:length(r)
plot(r)
plot(SPDF, add = T)

# ----------------------------------

SPDF_DE <- ClipRasterByPolygon(r, SPDF)
plot(raster::extent(SPDF_DE), axes = F, xlab = "", ylab = "")
plot(SPDF_DE, add = T)
plot(SPDF, add = T)
}
