% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Coda_to_tibble.R
\name{Coda_to_tibble}
\alias{Coda_to_tibble}
\title{Convert a Coda object to a tibble with specified parameter transformations}
\usage{
Coda_to_tibble(
  CodaObj = NULL,
  Type = NULL,
  EnvFile = ".env",
  NOmega = 100,
  FromHPC = TRUE
)
}
\arguments{
\item{CodaObj}{An object of class \code{mcmc.list} or \code{mcmc}, representing the
MCMC output.}

\item{Type}{Character. The parameter type to transform and extract. Must be
one of \code{rho}, \code{alpha}, \code{omega}, or \code{beta}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{NOmega}{Integer. The number of species to be sampled for the \code{Omega}
parameter transformation. Defaults to 100.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}
}
\value{
A tibble containing the transformed parameters based on the specified
\code{Type}. The structure of the returned tibble varies depending on the \code{Type}
parameter.
}
\description{
This function converts a Coda object (\code{mcmc.list} or \code{mcmc}) into a tibble
format, facilitating further analysis and visualization. It supports
transformation for specific parameter types: \code{rho}, \code{alpha}, \code{omega}, and
\code{beta}.
}
\examples{
library(Hmsc)
library(coda)
library(dplyr)
Coda <- Hmsc::convertToCodaObject(Hmsc::TD$m)
IASDT.R::Coda_to_tibble(CodaObj = Coda$Alpha[[1]], Type = "Alpha")
}
\author{
Ahmed El-Gabbas
}
