% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Convergence_Plot.R,
%   R/Mod_Convergence_Alpha.R, R/Mod_Convergence_Rho.R
\name{Convergence_plots}
\alias{Convergence_plots}
\alias{Convergence_Plot}
\alias{Convergence_Alpha}
\alias{Convergence_Rho}
\title{Plot model convergence of a selected model}
\usage{
Convergence_Plot(
  Path_Coda = NULL,
  Path_Model = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  Title = " ",
  NOmega = 1000L,
  NCores = 8L,
  NRC = c(2L, 2L),
  Beta_NRC = c(3L, 3L),
  SavePlotData = TRUE,
  PagePerFile = 20L,
  Cols = NULL,
  MarginType = "histogram"
)

Convergence_Alpha(
  Post = NULL,
  Model = NULL,
  Title = NULL,
  NRC = NULL,
  AddFooter = TRUE,
  AddTitle = TRUE,
  Cols = NULL,
  FromHPC = TRUE,
  MarginType = "histogram"
)

Convergence_Rho(
  Post = NULL,
  Model = NULL,
  Title = NULL,
  Cols = NULL,
  MarginType = "histogram"
)
}
\arguments{
\item{Path_Coda}{Character. Path to a saved coda object containing MCMC
samples.}

\item{Path_Model}{Character. Path to a saved Hmsc model object.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{Title}{Character. Title for \strong{rho} and \strong{alpha} convergence plots.
Default: " "}

\item{NOmega}{Integer. Number of species interactions sampled for Omega
parameter diagnostics. Default: 1000L}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{NRC}{Numeric vector. Grid layout (rows×columns) for arranging
alpha parameter plots. Default: \code{c(2, 2)}. If \code{NULL}, the layout is
automatically determined based on the number of alpha levels.}

\item{Beta_NRC}{Numeric vector. The grid layout (rows×columns) for
arranging beta parameter plots. Default: \code{c(3, 3)}.}

\item{SavePlotData}{Logical. If \code{TRUE} (default), saves the plot data as
\code{.RData} files.}

\item{PagePerFile}{Integer. Number of plots per page in the Omega parameter
output. Default: 20L.}

\item{Cols}{Character vector. MCMC chain colors (optional). Default: \code{NULL}.}

\item{MarginType}{Character. The type of marginal plot to add to
the main plot. Valid options are "histogram" (default) or "density".}

\item{Post}{\code{mcmc.list} or character. Either an MCMC object (\code{mcmc.list})
containing posterior samples, or a file path to a saved coda object.}

\item{Model}{\code{Hmsc} object or character. Either a fitted Hmsc model
object, or a file path to a saved Hmsc model.}

\item{AddFooter}{Logical. If \code{TRUE} (default), adds a footer with page
numbers to each plot.}

\item{AddTitle}{Logical. If \code{TRUE} (default), adds the main title (\code{Title})
to the plot.}
}
\description{
The \code{Convergence_Plot()} function generates and saves convergence diagnostics
plots for the \code{rho}, \code{alpha}, \code{omega}, and \code{beta} parameters in an Hmsc
model. These plots help assess whether the MCMC chains have reached
stationarity. It supports parallel processing and can work with models fitted
on HPC environments.
}
\details{
\code{Convergence_Alpha()} and \code{Convergence_Rho()} are internal functions
and should not be called directly.
}
\author{
Ahmed El-Gabbas
}
