% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CurrOS.R
\name{CurrOS}
\alias{CurrOS}
\title{Current operating system}
\usage{
CurrOS()
}
\value{
A character string representing the name of the operating system.
}
\description{
This function returns the name of the current operating system the R session
is running on.
}
\examples{
CurrOS()

}
\author{
Ahmed El-Gabbas
}
