% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_DetectAlias.R
\name{DetectAlias}
\alias{DetectAlias}
\title{Detect aliased variables}
\usage{
DetectAlias(DT, Verbose = FALSE)
}
\arguments{
\item{DT}{input data as tibble or data frame}

\item{Verbose}{print message for the output of the function}
}
\description{
Detect aliased variables
}
\examples{
require(car)
x1 <- rnorm(100)
x2 <- 2 * x1
x3 <- rnorm(100)
y <- rnorm( 100 )
model <- lm(y ~ x1 + x2 + x3)
summary(model)
try(car::vif(model))
DetectAlias(DT = cbind(x1, x2, x3))
DetectAlias(DT = cbind(x1, x2, x3), Verbose = TRUE)

# excluding x2
model <- lm(y ~ x1 + x3)
summary(model)
try(car::vif(model))
}
