% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{DirCreate}
\alias{DirCreate}
\title{Create directory if not existed}
\usage{
DirCreate(Path)
}
\arguments{
\item{Path}{Path of the folder}
}
\description{
Create directory if not existed
}
\examples{
# create new folder (random name) in the temporary folder
Path2Create <- file.path(tempdir(), stringi::stri_rand_strings(1, 5))
file.exists(Path2Create)

DirCreate(Path2Create)
file.exists(Path2Create)

}
\author{
Ahmed El-Gabbas
}
