% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_EASIN_Down.R
\name{EASIN_Down}
\alias{EASIN_Down}
\title{Extract EASIN data using EASIN ID}
\usage{
EASIN_Down(
  SpKey,
  Timeout = 200,
  Verbose = FALSE,
  BaseURL = "https://easin.jrc.ec.europa.eu/apixg/geoxg",
  NSearch = 1000,
  Attempts = 10,
  SleepTime = 5,
  Path_Raw = "datasets/interim/EASIN",
  DeleteChunks = TRUE,
  ReturnData = FALSE
)
}
\arguments{
\item{SpKey}{character; the EASIN taxon ID for which data is to be retrieved.
This parameter cannot be \code{NULL}.}

\item{Timeout}{integer; time in seconds before a download attempt times out.
Default is 200.}

\item{Verbose}{logical. Indicating whether to print messages for the progress
of the function}

\item{BaseURL}{character; the base URL for downloading EASIN data. Default
is "https://easin.jrc.ec.europa.eu/apixg/geoxg".}

\item{NSearch}{the number of records to attempt to retrieve per request.
Default is 1000, which is the current maximum allowed by the API.}

\item{Attempts}{integer; maximum number of download attempts per data
chunk. Defaults to 10.}

\item{SleepTime}{integer; the number of seconds to pause between each data
retrieval request to prevent overloading the server. Default is 5 second.}

\item{Path_Raw}{character; the path where the raw data files and temporary
parts will be stored. Default is \code{datasets/interim/EASIN}.}

\item{DeleteChunks}{logical, indicating whether to delete temporary files for
data chunks from the \code{FileParts} subdirectory. Defaults to \code{TRUE}.}

\item{ReturnData}{logical; if \code{TRUE}, the function will return the combined
data as a dataframe. If \code{FALSE} (default), the data will only be saved to
disk and the function will invisibly return \code{NULL}.}
}
\value{
If ReturnData is \code{TRUE}, returns a dataframe containing all the data
retrieved for the specified EASIN ID. If ReturnData is \code{FALSE}, returns
\code{NULL}.
}
\description{
This function downloads data for a specified species from the EASIN database.
It handles pagination automatically, downloading data in chunks until all
available data for the species is retrieved. The function also supports
pausing between requests to avoid overloading the server.
}
\details{
This function extracts EASIN data for a given EASIN_ID, handles
pagination, and ensures that data retrieval is efficient by managing
retries and pauses.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only called from the \link{EASIN_Processing} function.
}
\author{
Ahmed El-Gabbas
}
