% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_EASIN_Plot.R
\name{EASIN_Plot}
\alias{EASIN_Plot}
\title{Plotting cleaned EASIN data}
\usage{
EASIN_Plot(EnvFile = ".env", FromHPC = TRUE)
}
\arguments{
\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}
}
\value{
The function returns NULL invisibly. Plots are saved as JPEG files in
the specified directory.
}
\description{
Generates a series of plots for cleaned EASIN data, including the number of
observations, number of species, and the distribution of these metrics per
data partner. The function requires environment variables for paths to data
and expects specific files to be present at these paths.
}
\details{
The function first loads environment variables either from a
high-performance computing (HPC) environment or a local environment. Then,
it loads required data files, checks for missing files, and proceeds to
generate and save plots as JPEG files. The plots include the number of
observations and species and their distribution per data partner.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only used inside the \link{EASIN_Process} function.
}
\author{
Ahmed El-Gabbas
}
