% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_EASIN_Process.R
\name{EASIN_Process}
\alias{EASIN_Process}
\title{Extracts and processes EASIN data}
\usage{
EASIN_Process(
  ExtractTaxa = TRUE,
  ExtractData = TRUE,
  NDownTries = 10,
  NCores = 6,
  SleepTime = 10,
  NSearch = 1000,
  FromHPC = TRUE,
  EnvFile = ".env",
  DeleteChunks = FALSE,
  StartYear = 1981,
  Plot = TRUE
)
}
\arguments{
\item{ExtractTaxa}{Logical. If \code{TRUE}, the function will extract the EASIN
taxonomy list using \link{EASIN_Taxonomy}. Default is \code{TRUE}.}

\item{ExtractData}{Logical. If \code{TRUE}, the function will download EASIN
species occurrence data using \link{EASIN_Down}. Default is \code{TRUE}.}

\item{NDownTries}{Integer. Number of attempts to retry downloading data in
case of failure. Default is 10.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default is 6.}

\item{SleepTime}{Numeric. Time in seconds to wait between download attempts
and between chunks. Default is 10 seconds.}

\item{NSearch}{Integer. Number of observations or species to download during
EASIN taxonomy or data extraction, respectively. Default is 1000.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default is ".env".}

\item{DeleteChunks}{Logical. If \code{TRUE}, the function will delete intermediate
files after processing. Default is \code{FALSE}.}

\item{StartYear}{Integer. Minimum year for filtering species occurrence data.
Records before this year will be excluded. Default is \code{1981}, which matches
the year ranges of CHELSA current climate data.}

\item{Plot}{Logical. If \code{TRUE}, the function will generate summary plots of
the processed data using \link{EASIN_Plot}. Default is \code{TRUE}.}
}
\value{
The function Returns \code{NULL} invisibly after completing the data
extraction, processing, and optional plotting. The function saves multiple
outputs to disk, including the extracted and processed EASIN data,
species-specific data files, and summary statistics. The main outputs are:
\itemize{
\item \code{EASIN_Taxa.RData}: A dataset containing the standardized EASIN taxonomy.
\item \code{EASIN_Data.RData}: A cleaned and merged dataset of species occurrence
data.
\item \code{EASIN_NObs.RData}: A rasterized dataset showing the number of
observations per grid cell.
\item \code{EASIN_NObs_PerPartner.RData}: A rasterized dataset showing the number of
observations per data partner.
\item \code{EASIN_NSp.RData}: A rasterized dataset showing the number of species per
grid cell.
\item \code{EASIN_NSp_PerPartner.RData}: A rasterized dataset showing the number of
species per data partner.
\item Species-specific data files, saved as both sf and raster objects.
}
}
\description{
This function extracts and processes data from the European Alien Species
Information Network (\href{https://easin.jrc.ec.europa.eu/}{EASIN}) for vascular
plants. This function extracts plant species data from the EASIN database,
matches them with a pre-processed standardized list of taxa, and prepares
species-specific maps and summary maps. It also supports downloading data in
chunks, handling pagination, and retrying failed downloads.
}
\note{
\itemize{
\item The function assumes that the necessary environment variables are
correctly set up in the specified \code{.env} file. Users should ensure that all
required files and directories are accessible before running the function.
\item The function skips processing (i.e. reuse) species data or data chunks if
the data already exist on the raw directory. The function assumes that the
contents of this folder should be removed as part of the data workflow.
Skipping processing available data can help not to re-download already
available data from the EASIN server.
\item This function depends on the following functions: \link{EASIN_Taxonomy} for
getting the most recent EASIN taxonomy; \link{EASIN_Down} for processing EASIN
dataset; and \link{EASIN_Plot} for plotting.
}
}
\author{
Ahmed El-Gabbas
}
