% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_EASIN_Taxonomy.R
\name{EASIN_Taxonomy}
\alias{EASIN_Taxonomy}
\title{Extract EASIN Taxonomy Data}
\usage{
EASIN_Taxonomy(
  BaseURL = "https://easin.jrc.ec.europa.eu/apixg/catxg",
  Kingdom = "Plantae",
  Phylum = "Tracheophyta",
  NSearch = 100
)
}
\arguments{
\item{BaseURL}{character; the base URL for accessing the EASIN taxonomy API.
Default is "https://easin.jrc.ec.europa.eu/apixg/catxg".}

\item{Kingdom}{character; the taxonomic kingdom to search within. Default is
"Plantae".}

\item{Phylum}{character; the taxonomic phylum to filter within the specified
kingdom. Default is "Tracheophyta".}

\item{NSearch}{integer; the number of records to attempt to retrieve per
request. Default is 100.}
}
\value{
A tibble containing the retrieved taxonomy data for the specified
kingdom and phylum.
}
\description{
This function retrieves taxonomy data from the EASIN database for vascular
plants. It is also possible to get similar data for other taxonomic groups.
It handles pagination automatically, downloading data in chunks until all
available data for the specified taxa are retrieved.
}
\details{
This function loops through the EASIN API, retrieving data in chunks
until all available data for the specified kingdom and phylum are
collected. The results are returned as a tibble after filtering for the
specified taxa.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only called from the \link{EASIN_Processing} function.
}
\author{
Ahmed El-Gabbas and Marina Golivets
}
