% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Efforts_Process.R, R/DWF_Efforts_Request.R,
%   R/DWF_Efforts_Download.R, R/DWF_Efforts_Summarize.R, R/DWF_Efforts_Split.R,
%   R/DWF_Efforts_Plot.R
\name{Efforts_data}
\alias{Efforts_data}
\alias{Efforts_Process}
\alias{Efforts_Request}
\alias{Efforts_Download}
\alias{Efforts_Summarize}
\alias{Efforts_Split}
\alias{Efforts_Plot}
\title{Process GBIF sampling effort data for the \code{IAS-pDT}}
\usage{
Efforts_Process(
  EnvFile = ".env",
  Renviron = ".Renviron",
  Request = TRUE,
  Download = TRUE,
  NCores = 6L,
  StartYear = 1981L,
  Boundaries = c(-30, 50, 25, 75),
  ChunkSize = 100000L,
  DeleteChunks = TRUE,
  DeleteProcessed = TRUE
)

Efforts_Request(
  EnvFile = ".env",
  NCores = 3L,
  StartYear = 1981L,
  Renviron = ".Renviron",
  Boundaries = c(-30, 50, 25, 75)
)

Efforts_Download(NCores = 6L, EnvFile = ".env")

Efforts_Summarize(
  EnvFile = ".env",
  NCores = 6L,
  ChunkSize = 100000L,
  DeleteChunks = TRUE
)

Efforts_Split(Path_Zip = NULL, EnvFile = ".env", ChunkSize = 100000L)

Efforts_Plot(EnvFile = ".env")
}
\arguments{
\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{Renviron}{Character. Path to \code{.Renviron} file with GBIF credentials
(\code{GBIF_EMAIL}, \code{GBIF_USER}, \code{GBIF_PWD}). Default: \code{".Renviron"}. The
credentials must be in the format:
\itemize{
\item \code{GBIF_EMAIL=your_email}
\item \code{GBIF_USER=your_username}
\item \code{GBIF_PWD=your_password}
}}

\item{Request}{Logical. If \code{TRUE} (default), requests GBIF data; otherwise,
loads existing data.}

\item{Download}{Logical. If \code{TRUE} (default), downloads and saves GBIF data;
otherwise, skips download. Default: \code{TRUE}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6, except for \code{Efforts_Request}, which defaults to 3 with a
maximum of 3.}

\item{StartYear}{Integer. Earliest year for GBIF records (matches CHELSA
climate data). Default: \code{1981}.}

\item{Boundaries}{Numeric vector (length 4). GBIF data bounds (Left, Right,
Bottom, Top). Default: \code{c(-30, 50, 25, 75)}.}

\item{ChunkSize}{Integer. Rows per chunk file. Default: \code{100000}.}

\item{DeleteChunks}{Logical. If \code{TRUE} (default), deletes chunk files
post-processing.}

\item{DeleteProcessed}{Logical. If \code{TRUE} (default), removes raw GBIF files
after processing (>22 GB).}

\item{Path_Zip}{Character. Path to zip file with CSV for splitting.}
}
\description{
Downloads and processes GBIF sampling effort data for vascular plants in
Europe, supporting the Invasive Alien Species prototype Digital Twin
(\code{IAS-pDT}). Orchestrated by \code{Efforts_Process()}, it uses helper functions to
request, download, split, summarize, and visualize data at the Order level.
The functions prepares raster maps for the number of vascular plant
observations and species per grid cell.
}
\note{
\itemize{
\item \code{Efforts_Process()} is the main entry point for processing sampling effort
data.
\item Time-intensive (>9 hours on 6-core Windows PC; GBIF request ~5 hours).
\item Detects and processes only new/missing data by order.
}
}
\section{Functions details}{

\itemize{
\item \strong{\code{Efforts_Process()}}: Manages the workflow for requesting, downloading,
processing, and plotting GBIF vascular plant data.
\item \strong{\code{Efforts_Request()}}: Requests GBIF data by order in parallel. Stores
results to disk.
\item \strong{\code{Efforts_Download()}}: Downloads GBIF data, validates files, and loads
existing data if available. Returns a dataframe (\code{Efforts_AllRequests})
with paths.
\item \strong{\code{Efforts_Split()}}: Splits zipped CSV data by order into chunks, saving
each separately.
\item \strong{\code{Efforts_Summarize()}}: Processes and summarizes data into \code{RData} and
TIFF rasters.
\item \strong{\code{Efforts_Plot()}}: Plots observation efforts (raw and log10 scales).
}
}

\references{
Data source: \url{https://www.gbif.org}
}
\author{
Ahmed El-Gabbas
}
