% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Efforts_Plot.R
\name{Efforts_Plot}
\alias{Efforts_Plot}
\title{Plot the output of efforts maps}
\usage{
Efforts_Plot(Path_Efforts, EU_Bound)
}
\arguments{
\item{Path_Efforts}{Character. Path to the directory where the generated
plots will be saved. The directory must exist.}

\item{EU_Bound}{Character. Path to \code{RData} file containing country
boundaries.}
}
\value{
The function saves generated plots as JPEG files in the specified
directory and returns NULL invisibly.
}
\description{
This function generates and saves multiple plots of plant observation
efforts, both in raw and log10 scales, using provided spatial boundary and
summary data.
}
\details{
This function generates and saves effort maps visualizing the number
of plant observations and species, including both native and non-native
species, within Europe. It produces both standard and log10-scaled plots.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only used inside the \link{Efforts_Process} function.
}
\author{
Ahmed El-Gabbas
}
