% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Efforts_Process.R
\name{Efforts_Process}
\alias{Efforts_Process}
\title{Download and Process sampling efforts data}
\usage{
Efforts_Process(
  FromHPC = TRUE,
  EnvFile = ".env",
  Renviron = ".Renviron",
  RequestData = TRUE,
  DownloadData = TRUE,
  NCores = 6L,
  StartYear = 1981L,
  Boundaries = c(-30, 50, 25, 75),
  ChunkSize = 100000L,
  DeleteChunks = TRUE,
  DeleteProcessed = TRUE
)
}
\arguments{
\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{Renviron}{Character. The path to the \code{.Renviron} file containing GBIF
login credentials (email, user, password). Default: \code{.Renviron}.}

\item{RequestData}{Logical. If \code{TRUE}, the function requests data from GBIF.
If \code{FALSE}, previously requested data is loaded from disk. Defaults to
\code{TRUE}.}

\item{DownloadData}{Logical. If \code{TRUE}, the function downloads data and
stores it on disk. If \code{FALSE}, it skips the download step. Defaults to
\code{TRUE}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}

\item{StartYear}{Numeric. The starting year for the occurrence data. Only
records from this year onward will be requested from GBIF. Default is
\code{1981}, which matches the year ranges of CHELSA current climate data.}

\item{Boundaries}{Numeric vector of length 4. The boundaries of the
requested GBIF data in the order: Left, Right, Bottom, Top. Defaults to
\code{c(-30, 50, 25, 75)}.}

\item{ChunkSize}{Integer. The number of rows per chunk file. Default:
\verb{100,000}. See \link{Efforts_Split} and \link{Efforts_Summarize} for more details.}

\item{DeleteChunks}{Logical. Whether to remove file chunks after processing
the data. Defaults to \code{TRUE}.}

\item{DeleteProcessed}{Logical. Whether to delete the raw downloaded GBIF
data after processing them. This helps to free large unnecessary file space
(> 22 GB). Defaults to \code{TRUE}.}
}
\value{
Returns \code{NULL} invisibly. The function generates various output
files, maps, and logs, and it is designed to be used for its side effects.
}
\description{
This function downloads GBIF data for all vascular plants within a specified
geographical area (Europe), grouped by order, and converts the data to raster
format to represent the number of vascular plant observations and species per
grid cell.
}
\note{
\itemize{
\item This function is expected to take a substantial amount of time (>9
hours on a Windows PC with 6 cores). The data request from GBIF may take
around 5 hours to be ready. The function requests GBIF data for each vascular
plant order and waits for the data to be ready before processing them.
\item This function should be the only function to be called to prepare sampling
efforts data. It calls other functions \link{Efforts_Request} to request data from
GBIF, \link{Efforts_Download} to download zipped archive for each vascular plant
order, \link{Efforts_Summarize} and \link{Efforts_Split} to process data in small
chunks, and \link{Efforts_Plot} for plotting.
}
}
\author{
Ahmed El-Gabbas
}
