% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Sampling_Efforts.R
\name{Efforts_Request}
\alias{Efforts_Request}
\title{Request and Manage GBIF Data for Vascular Plant Orders}
\usage{
Efforts_Request(NCores, Path_Requests, Path_Efforts, StartYear, Boundaries)
}
\arguments{
\item{NCores}{Integer. The number of cores to use for parallel processing (1 to 3). Must be a positive integer, up to a maximum of 3.}

\item{Path_Requests}{Character. The directory path to save individual request
files.}

\item{Path_Efforts}{Character. The directory path to save the final compiled
data.}

\item{StartYear}{Numeric. The starting year for the occurrence data. Only
records from this year onward will be requested from GBIF. Defaults to
1980.}

\item{Boundaries}{Numeric vector of length 4. Specifies geographical
boundaries for the requested GBIF data in the order: Left, Right, Bottom,
Top. Defaults to c(-30, 50, 25, 75).}
}
\value{
The function returns the GBIF data requests processed and stored in
the specified directories.
}
\description{
This function requests GBIF data for each vascular plant order, processes it
in parallel, and manages the data download and storage. If data is already
available, it loads the data instead of making new requests.
}
\author{
Ahmed El-Gabbas
}
