% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Efforts_Split.R
\name{Efforts_Split}
\alias{Efforts_Split}
\title{Split Order data into smaller chunks}
\usage{
Efforts_Split(Path_Zip, Path_Output, ChunkSize = 1e+05)
}
\arguments{
\item{Path_Zip}{Character. The file path to the zip archive containing the
CSV file. The file must be a ZIP archive containing a single CSV file.}

\item{Path_Output}{Character. The directory where the split files will be
saved. The directory must exist.}

\item{ChunkSize}{Integer. The number of rows per chunk file. Default:
\verb{100,000}. Note: Larger chunk sizes may require significant memory and
processing power.}
}
\value{
A character vector of file path(s) to the created chunk files.
}
\description{
This function extracts Order data without extracting the zipped archive. The
function reads CSV files inside the zipped file, selects specified columns,
and splits the data into smaller chunks of specified row size, saving each
chunk as a separate file.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only used inside the \link{Sampling_Efforts} function.
}
\author{
Ahmed El-Gabbas
}
