% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Efforts_Summarize.R
\name{Efforts_Summarize}
\alias{Efforts_Summarize}
\title{Process and Summarize GBIF Data for vascular plants}
\usage{
Efforts_Summarize(
  NCores,
  Path_Efforts,
  Path_Efforts_Interim,
  Path_Efforts_Data,
  Path_Grid,
  IAS_List,
  Efforts_AllRequests,
  ChunkSize = 1e+05,
  DeleteChunks = TRUE
)
}
\arguments{
\item{NCores}{Numeric. The number of cores to use for parallel processing.}

\item{Path_Efforts}{Character. Path where the final processed data will be
saved.}

\item{Path_Efforts_Interim}{Character. The directory path to save interim
data.}

\item{Path_Efforts_Data}{Character. The directory path to save detailed
effort data as \code{RData}.}

\item{Path_Grid}{Character. The directory path to load the grid data.}

\item{IAS_List}{A list of invasive alien species keys.}

\item{Efforts_AllRequests}{A data frame containing the details of the GBIF
download, including paths to CSV files, zip files, order, class, and total
records.}

\item{ChunkSize}{Integer. The number of rows per chunk file. Default:
\verb{100,000}. See \link{Efforts_Split} for more details.}

\item{DeleteChunks}{logical, indicating whether to remove file chunks after
processing the data. Defaults to \code{TRUE}.}
}
\value{
The function returns and saves the GBIF data summary.
}
\description{
This function processes GBIF data for vascular plants by extracting,
summarizing, and saving the data. It also creates summary maps and saves them
in both \code{RData} and \code{TIFF} formats.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only used inside the \link{Efforts_Process} function.
}
\author{
Ahmed El-Gabbas
}
