% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Efforts_SummarizeMaps.R
\name{Efforts_SummarizeMaps}
\alias{Efforts_SummarizeMaps}
\title{Summarize maps for efforts data}
\usage{
Efforts_SummarizeMaps(Data, NSp, Name, ClassOrder, Grid_SF, Grid_R)
}
\arguments{
\item{Data}{An \code{sf} object containing spatial data, with a column named
\code{CellCode}.}

\item{NSp}{A logical value indicating whether to generate distinct species
counts (\code{TRUE}) or total observation counts (\code{FALSE}).}

\item{Name}{A character string that represents the name of the count field
and the prefix for the final raster layer's name.}

\item{ClassOrder}{A character string specifying the class and order
combination (separated by an underscore) represented in the \code{Data}.}
}
\value{
A processed \code{terra} raster object representing the summarized data.
}
\description{
This function processes spatial data (as an \code{sf} object), summarizes it based
on the number of observations or distinct species, and generates a raster
layer.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only used inside the \link{Efforts_Process} and \link{Efforts_Summarize}
functions.
}
\author{
Ahmed El-Gabbas
}
