% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{ElapsedTime}
\alias{ElapsedTime}
\title{Elapsed time for running an R command}
\usage{
ElapsedTime(x, Unit = c("M", "S", "H"))
}
\arguments{
\item{x}{R command}
}
\value{
nothing is returned; the elapsed time is printed to the console
}
\description{
Elapsed time for running an R command
}
\examples{
# unit in minutes by default
ElapsedTime(Sys.sleep(3))

# elapsed time in minutes
ElapsedTime(Sys.sleep(3), "S")

ElapsedTime(Sys.sleep(3), "Sec")     # the same (only first letter is used)

ElapsedTime(Sys.sleep(3), "s")     # the same (case insensitive)

try(ElapsedTime(Sys.sleep(3), "0s")) # this will give an error
}
\author{
Ahmed El-Gabbas
}
