% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_elTER_Process.R
\name{elTER_Process}
\alias{elTER_Process}
\title{eLTER data processing function}
\usage{
elTER_Process(FromHPC = TRUE, EnvFile = ".env", StartYear = 1981)
}
\arguments{
\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{StartYear}{Numeric. The starting year for the occurrence data. Only
records from this year onward will be processed. Default is \code{1981}, which
matches the year ranges of CHELSA current climate data.}
}
\value{
Returns \code{NULL} invisibly after saving the processed data.
}
\description{
This function processes pre-cleaned and pre-standardized eLTER data.
}
\author{
Ahmed El-Gabbas & Marina Golivets
}
