% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_elTER_Processing.R
\name{elTER_Processing}
\alias{elTER_Processing}
\title{eLTER Data Processing Function}
\usage{
elTER_Processing(FromHPC = TRUE, EnvFile = ".env", StartYear = 1980)
}
\arguments{
\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default is ".env".}

\item{StartYear}{Numeric. The starting year for the occurrence data. Only
records from this year onward will be processed. Defaults to 1980.}
}
\value{
Returns \code{NULL} invisibly after saving the processed data.
}
\description{
This function processes pre-cleaned and pre-standardized eLTER data.
}
\author{
Ahmed El-Gabbas & Marina Golivets
}
