% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exp_neg_div}
\alias{exp_neg_div}
\title{exp_neg_div: Exponential of Negative Division}
\usage{
exp_neg_div(D11, x)
}
\arguments{
\item{D11}{A matrix of values to be divided and exponentiated.}

\item{x}{A scalar divisor.}
}
\value{
A matrix of the same dimensions as \code{D11}, containing the computed
values.
}
\description{
Computes the element-wise exponential of the negative division of each
element in a matrix by a scalar value.
}
