% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Extract_BB}
\alias{Extract_BB}
\title{Extract a specific column from the output of GBIF standardization}
\usage{
Extract_BB(x = ., var)
}
\arguments{
\item{x}{GBIF results of GBIF standardization \code{rgbif::name_backbone}}

\item{var}{column name to extract}
}
\value{
Scientific name of the accepted taxa
}
\description{
Extract a specific column from the output of GBIF standardization
}
\details{
Extract a specific column from the output of GBIF
}
\examples{
rgbif::name_backbone(name = "Helianthus annuus", kingdom = "plants") \%>\%
   Extract_BB(status)

# ------------------------

c("Helianthus annuus", "Tagetes patula L.") \%>\%
   tibble::tibble(Taxa = .) \%>\%
   dplyr::mutate(
      BB = purrr::map(Taxa, rgbif::name_backbone),
      status = purrr::map_chr(BB, Extract_BB, status),
      SpKey = purrr::map_int(BB, Extract_BB, speciesKey))
}
\author{
Ahmed El-Gabbas
}
