% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_pnorm}
\alias{fast_pnorm}
\title{fast_pnorm: Fast Normal CDF}
\usage{
fast_pnorm(x)
}
\arguments{
\item{x}{A numeric vector for which the CDF should be computed.}
}
\value{
A numeric vector of CDF values for each element in \code{x}.
}
\description{
Computes the cumulative distribution function of the normal distribution
for a numeric vector using a fast approximation.
}
