% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{FileExt}
\alias{FileExt}
\title{Get file extension}
\usage{
FileExt(Path)
}
\arguments{
\item{Path}{File path}
}
\description{
Get file extension
}
\examples{
FileExt("File.doc")
FileExt("D:/File.doc")

}
\author{
Ahmed El-Gabbas
}
