% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{FileSize}
\alias{FileSize}
\title{File size in a Human Readable format}
\usage{
FileSize(File, ...)
}
\arguments{
\item{File}{character; file path}

\item{...}{additional arguments for \code{gdata::humanReadable} function}
}
\description{
File size in a Human Readable format
}
