% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_FileType.R
\name{FileType}
\alias{FileType}
\title{Determines the file type of a given file path}
\usage{
FileType(Path)
}
\arguments{
\item{Path}{A character string specifying the path to the file whose type is to be determined. The path must not be NULL, and the file must exist.}
}
\value{
A character string describing the file type.
}
\description{
This function uses the system's \code{file} command to determine the type of the file specified by the \code{Path} parameter. It returns a character string describing the file type.
}
\note{
This function relies on the system's \code{file} command and therefore might produce different outputs on different platforms.
}
\examples{
(f <- system.file("ex/elev.tif", package="terra"))
FileType(Path = f)
}
