% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_FunctionsInPackage.R
\name{FunctionsInPackage}
\alias{FunctionsInPackage}
\title{List of functions in a package}
\usage{
FunctionsInPackage(Package)
}
\arguments{
\item{Package}{Character. Package name.}
}
\value{
A character vector containing the names of all functions in the
specified package.
}
\description{
This function returns a character vector listing all the functions available
in the specified R package. It first checks if the package is installed and
can be loaded; if not, it raises an error.
}
\examples{
str(FunctionsInPackage(Package = "raster"))

str(FunctionsInPackage(Package = "sf"))
}
\author{
Ahmed El-Gabbas
}
