% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GBIF_Process.R, R/DWF_GBIF_Check.R,
%   R/DWF_GBIF_Download.R, R/DWF_GBIF_ReadChunk.R, R/DWF_GBIF_SpData.R
\name{GBIF_data}
\alias{GBIF_data}
\alias{GBIF_Process}
\alias{GBIF_Check}
\alias{GBIF_Download}
\alias{GBIF_ReadChunk}
\alias{GBIF_SpData}
\title{Process GBIF occurrence data for the \code{IAS-pDT}}
\usage{
GBIF_Process(
  EnvFile = ".env",
  Renviron = ".Renviron",
  NCores = 6L,
  Request = TRUE,
  Download = TRUE,
  SplitChunks = TRUE,
  Overwrite = FALSE,
  DeleteChunks = TRUE,
  ChunkSize = 50000L,
  Boundaries = c(-30, 50, 25, 75),
  StartYear = 1981L
)

GBIF_Check(Renviron = ".Renviron")

GBIF_Download(
  EnvFile = ".env",
  Renviron = ".Renviron",
  Request = TRUE,
  Download = TRUE,
  SplitChunks = TRUE,
  ChunkSize = 50000L,
  Boundaries = c(-30, 50, 25, 75),
  StartYear = 1981L
)

GBIF_ReadChunk(
  ChunkFile,
  EnvFile = ".env",
  MaxUncert = 10L,
  StartYear = 1981L,
  SaveRData = TRUE,
  ReturnData = FALSE,
  Overwrite = FALSE
)

GBIF_SpData(Species = NULL, EnvFile = ".env", Verbose = TRUE, PlotTag = NULL)
}
\arguments{
\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{Renviron}{Character. Path to \code{.Renviron} file with GBIF credentials
(\code{GBIF_EMAIL}, \code{GBIF_USER}, \code{GBIF_PWD}). Default: \code{".Renviron"}. The
credentials must be in the format:
\itemize{
\item \code{GBIF_EMAIL=your_email}
\item \code{GBIF_USER=your_username}
\item \code{GBIF_PWD=your_password}
}}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}

\item{Request}{Logical. If \code{TRUE} (default), requests GBIF data; otherwise,
loads from disk.}

\item{Download}{Logical. If \code{TRUE} (default), downloads and saves GBIF data.}

\item{SplitChunks}{Logical. If \code{TRUE} (default), splits data into chunks for
easier processing.}

\item{Overwrite}{Logical. If \code{TRUE}, reprocesses existing \code{.RData} chunks.
Default: \code{FALSE}. This helps to continue working on previously processed
chunks if the previous try failed, e.g. due to memory issue.}

\item{DeleteChunks}{Logical. If \code{TRUE} (default), deletes chunk files.}

\item{ChunkSize}{Integer. Records per data chunk. Default: \code{50000}.}

\item{Boundaries}{Numeric vector (length 4). GBIF data bounds (Left, Right,
Bottom, Top). Default: \code{c(-30, 50, 25, 75)}.}

\item{StartYear}{Integer. Earliest collection year to be included. Default is
1981.}

\item{ChunkFile}{Character. Path of chunk file for processing.}

\item{MaxUncert}{Numeric. Maximum spatial uncertainty in kilometers. Default:
\code{10}.}

\item{SaveRData}{Logical. If \code{TRUE} (default), saves chunk data as \code{.RData}.}

\item{ReturnData}{If \code{TRUE}, returns chunk data; otherwise,
\code{invisible(NULL)}. Default: \code{FALSE}.}

\item{Species}{Character. Species name for processing.}

\item{Verbose}{Logical. If \code{TRUE} (default), prints progress messages.}

\item{PlotTag}{Character. Tag for plot titles.}
}
\description{
Extracts, processes, and visualizes occurrence data from the \href{https://www.gbif.org}{Global Biodiversity Information Facility (GBIF)} for the
Invasive Alien Species prototype Digital Twin (\code{IAS-pDT}). Orchestrated by
\code{GBIF_Process()}, it requests, downloads, cleans, chunks, and maps species
data using helper functions.
}
\note{
Relies on a static RDS file listing IAS species, GBIF keys, and
metadata, standardized by Marina Golivets (Feb 2024).
}
\section{Functions details}{

\itemize{
\item \strong{\code{GBIF_Process()}}: Orchestrates GBIF data requests, downloads,
processing, and mapping. Saves \code{RData}, Excel, and JPEG summary files.
\item \strong{\code{GBIF_Check()}}: Verifies GBIF credentials in environment or
\code{.Renviron}. Returns \code{TRUE} if valid, else \code{FALSE}.
\item \strong{\code{GBIF_Download()}}: Requests and downloads GBIF data (if \code{Download = TRUE}),  using the specified criteria (taxa, coordinates, time period, and
boundaries), splits into small chunks (if \code{SplitChunks = TRUE}), and saves
metadata. Returns \code{invisible(NULL)}.
\item \strong{\code{GBIF_ReadChunk()}}: Filters chunk data (spatial/temporal, e.g.,
spatial uncertainty, collection year, coordinate precision, and taxonomic
rank), select relevant columns, and saves as \code{.RData} (if \code{SaveRData = TRUE}) or returns it (if \code{ReturnData = TRUE}). Skips if \code{.RData} exists and
\code{Overwrite = FALSE}.
\item \strong{\code{GBIF_SpData()}}: Converts species-specific data to \code{sf} and raster
formats, generating distribution maps.
}
}

\author{
Ahmed El-Gabbas
}
