% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GBIF_Process.R
\name{GBIF_Process}
\alias{GBIF_Process}
\title{GBIF Data Processing Function}
\usage{
GBIF_Process(
  FromHPC = TRUE,
  EnvFile = ".env",
  Renviron = ".Renviron",
  NCores = 6L,
  RequestData = TRUE,
  DownloadData = TRUE,
  SplitChunks = TRUE,
  Overwrite = FALSE,
  DeleteChunks = TRUE,
  ChunkSize = 50000L,
  Boundaries = c(-30, 50, 25, 75),
  StartYear = 1981L
)
}
\arguments{
\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{Renviron}{Character. The path to the \code{.Renviron} file containing GBIF
login credentials (email, user, password).}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}

\item{RequestData}{Logical. If \code{TRUE}, requests data from GBIF. If \code{FALSE},
loads a previously requested data set from \code{GBIF_Request.RData} and
\code{StatusDetailed.RData} files. Defaults to \code{TRUE}.}

\item{DownloadData}{Logical. If \code{TRUE}, downloaded data is stored on disk.
Defaults to \code{TRUE}.}

\item{SplitChunks}{Logical. If \code{TRUE}, splits the downloaded data into
smaller chunks for easier processing.}

\item{Overwrite}{Logical. Whether to process the current chunk
file if it has already processed and saved as \verb{*.RData} file. This helps to
continue working on previously processed chunks if the previous try failed,
e.g. due to memory issue.}

\item{DeleteChunks}{Logical. If \code{TRUE}, delete the chunk files. Defaults to
\code{TRUE}.}

\item{ChunkSize}{Integer. The number of records per chunk when splitting the
data. Default is 50,000.}

\item{Boundaries}{Numeric vector of length 4. The boundaries of the
requested GBIF data in the order: Left, Right, Bottom, Top. Defaults to
\code{c(-30, 50, 25, 75)}.}

\item{StartYear}{Numeric. The starting year for the occurrence data. Only
records from this year onward will be requested from GBIF. Default is
\code{1981}, which matches the year ranges of CHELSA current climate data.}
}
\value{
Saves multiple RData and Excel files containing the processed data
and summary maps. Also saves a JPEG file with summary plots.
}
\description{
This function processes GBIF (Global Biodiversity Information Facility) data,
which includes downloading, cleaning, and saving the data in various formats.
The function handles data chunking, merging, and produces summary maps as
well as species-specific data.
}
\note{
This function depends on the following functions: \link{GBIF_Download}  for
requesting, downloading and splitting data into chunks; \link{GBIF_ReadChunk} to
process chunk files; and \link{GBIF_SpData} to prepare species-specific data.
}
\author{
Ahmed El-Gabbas
}
