% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GBIF_SpData.R
\name{GBIF_SpData}
\alias{GBIF_SpData}
\title{Extract species data as sf and Raster and plot species distribution}
\usage{
GBIF_SpData(
  Species = NULL,
  FromHPC = TRUE,
  EnvFile = ".env",
  Verbose = TRUE,
  LastUpdate = NULL
)
}
\arguments{
\item{Species}{A character string specifying the species name.}

\item{FromHPC}{A logical value indicating whether the environment variables
should be read from an HPC (High-Performance Computing) environment.
Default is \code{TRUE}.}

\item{EnvFile}{A string specifying the path of the environment file. Default
is ".env".}

\item{Verbose}{A logical value indicating whether to print progress messages.
Default is \code{TRUE}.}

\item{LastUpdate}{A character string specifying the last update date to be
displayed on the plots.}
}
\description{
Extracts and processes species data from processed GBIF data, generating
grids, rasters, and visual maps.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only called from the \link{GBIF_Processing} function.
}
\author{
Ahmed El-Gabbas
}
