% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_gelman.preplot.R
\name{gelman.preplot}
\alias{gelman.preplot}
\title{gelman.preplot}
\source{
This function is copied from the unexported `coda:::gelman.preplot` function. This function is used internally in the [coda::gelman.plot] function. I make this function available in the package without exporting it. See: https://svn.r-project.org/R-packages/trunk/coda/R/gelman.R
}
\usage{
gelman.preplot(
  x,
  bin.width = bin.width,
  max.bins = max.bins,
  confidence = confidence,
  transform = transform,
  autoburnin = autoburnin
)
}
\arguments{
\item{x}{x}

\item{bin.width}{bin.width}

\item{max.bins}{max.bins}

\item{confidence}{confidence}

\item{transform}{transform}

\item{autoburnin}{autoburnin}
}
\description{
gelman.preplot
}
