% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Gelman_Beta.R
\name{Gelman_Beta}
\alias{Gelman_Beta}
\title{Gelman-Rubin-Brooks plot for all beta parameters}
\usage{
Gelman_Beta(CodaObj = NULL, NCores = NULL, PlottingAlpha = 0.25)
}
\arguments{
\item{CodaObj}{an mcmc object}

\item{NCores}{Integer. Number of parallel processes.}

\item{PlottingAlpha}{Double. Plotting alpha for line transparency}
}
\description{
Gelman-Rubin-Brooks plot for all beta parameters
}
\author{
Ahmed El-Gabbas
}
