% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Gelman_Plot.R
\name{Gelman_Plot}
\alias{Gelman_Plot}
\title{Plot Gelman-Rubin-Brooks plots}
\usage{
Gelman_Plot(
  CodaPath = NULL,
  Beta = TRUE,
  Rho = TRUE,
  Omega = TRUE,
  Alpha = TRUE,
  NCores = NULL,
  NOmega = 1000,
  PlotAlpha = 0.25,
  SavePlot = NULL,
  OutPath = NULL,
  Beta_File = NULL,
  Rho_File = NULL,
  Omega_File = NULL,
  Alpha_File = NULL
)
}
\arguments{
\item{CodaPath}{Path to RData file containing the coda object}

\item{Beta}{Logical. Run \code{IASDT.R::Gelman_Beta}?}

\item{Rho}{Logical. Run \code{IASDT.R::Gelman_Rho}?}

\item{Omega}{Logical. Run \code{IASDT.R::Gelman_Omega}?}

\item{Alpha}{Logical. Run \code{IASDT.R::Gelman_Alpha} (not yet available)?}

\item{NCores}{Integer. Number of parallel processes.}

\item{NOmega}{Integer. Number of species to be sampled for the Omega parameter}

\item{PlotAlpha}{Double. Plotting alpha for line transparency}

\item{SavePlot}{Logical. Save the outputs as JPEG file}

\item{OutPath}{String. Folder path to save the output figures}

\item{Beta_File}{String. File name (with extension) for saving the results of \code{IASDT.R::Gelman_Beta}}

\item{Rho_File}{String. File name (with extension) for saving the results of \code{IASDT.R::Gelman_Rho}}

\item{Omega_File}{String. File name (with extension) for saving the results of \code{IASDT.R::Gelman_Omega}}

\item{Alpha_File}{String. File name (with extension) for saving the results of \code{IASDT.R::Gelman_Alpha}}
}
\description{
Plot Gelman-Rubin-Brooks plots
}
\author{
Ahmed El-Gabbas
}
