% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_predictHmsc.R
\name{get1prediction}
\alias{get1prediction}
\title{get1prediction}
\usage{
get1prediction(
  object,
  X,
  XRRR,
  Yc,
  Loff,
  rL,
  rLPar,
  sam,
  predPostEta,
  PiNew,
  dfPiNew,
  nyNew,
  expected,
  mcmcStep,
  seed = NULL
)
}
\description{
get1prediction
}
