% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Get_EASIN_Data}
\alias{Get_EASIN_Data}
\title{Extract EASIN data using EASIN ID}
\usage{
Get_EASIN_Data(
  SpKey,
  NSearch = 500,
  Grid_sf = Grid_10_sf_s,
  Grid_R = terra::unwrap(Grid_10_Land_Crop),
  TaxaList = EASIN_Taxa,
  MinYear = 1981
)
}
\arguments{
\item{SpKey}{EASIN ID}

\item{NSearch}{number of grid cells to extract each time}

\item{Grid_sf}{Reference grid as simple feature object}

\item{Grid_R}{Reference grid as \code{SpatRaster} object}

\item{MinYear}{exclude points before this year}
}
\description{
Extract EASIN data using EASIN ID
}
\details{
A function to extract EASIN data for a given EASIN_ID
}
\examples{
\dontrun{
c("R00006", "R18689", "R00008") \%>\%
  purrr::map(Get_EASIN_Data)
}
}
\author{
Ahmed El-Gabbas
}
