% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_get_species_name.R
\name{get_species_name}
\alias{get_species_name}
\title{Get species name or information of an \code{IAS-pDT} species ID}
\usage{
get_species_name(species_ID = NULL, env_file = ".env")
}
\arguments{
\item{species_ID}{optional IASDT species ID for which detailed information is
required. If not provided, the function returns the entire species list.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}
}
\value{
A data frame containing species information. If a species ID
\code{species_ID} is provided, it only returns species information for the
listed species, otherwise return the full list of IAS.
}
\description{
This function retrieves detailed information on \code{IAS-pDT} species list,
optionally filtered by a specific IAS-pDT species ID (\code{species_ID}).
}
\author{
Ahmed El-Gabbas
}
