% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_GetAcceptedName.R
\name{GetAcceptedName}
\alias{GetAcceptedName}
\title{Get accepted name of a taxa}
\usage{
GetAcceptedName(ID)
}
\arguments{
\item{ID}{GBIF id}
}
\value{
Scientific name of the accepted taxa
}
\description{
Get accepted name of a taxa
}
\details{
Get accepted name of a taxa
}
\examples{
\dontrun{
GetAcceptedName(5372559) ### https://www.gbif.org/species/5372559
}
}
\author{
Ahmed El-Gabbas
}
