% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_GetBashVars.R
\name{GetBashVars}
\alias{GetBashVars}
\title{Loading arguments passed to Rscript bash command}
\usage{
GetBashVars()
}
\description{
Loading arguments passed to Rscript bash command
}
\author{
Ahmed El-Gabbas
}
