% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_GetCV.R
\name{GetCV}
\alias{GetCV}
\title{GetCV}
\usage{
GetCV(DT, NR = 4, NC = 3, Path_Grid)
}
\arguments{
\item{DT}{DT}

\item{NR}{NR}

\item{NC}{NC}

\item{Path_Grid}{Path_Grid}
}
\description{
GetCV
}
\author{
Ahmed El-Gabbas
}
