% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_git_log.R
\name{git_log}
\alias{git_log}
\title{print detailed \verb{git log} of the git repo located in the current working directory}
\usage{
git_log()
}
\description{
print detailed \verb{git log} of the git repo located in the current working directory
}
\examples{
git_log()
}
