% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_GridCross.R
\name{GridCross}
\alias{GridCross}
\title{Create a \code{multilinestring} (cross in the middle of the grid) sf object from each grid cell}
\usage{
GridCross(DT)
}
\arguments{
\item{DT}{input sf tibble}
}
\description{
Create a \code{multilinestring} (cross in the middle of the grid) sf object from each grid cell
}
\examples{
IASDT.R::LoadPackages(dplyr, sf, raster, ggplot2)

Grid <- raster::raster(nrows = 10, ncols = 10, xmn = 0, xmx = 10, ymn = 0, ymx = 10, crs = 4326) \%>\%
  setNames("Grid") \%>\%
  raster::setValues(1) \%>\%
  raster::rasterToPolygons() \%>\%
  sf::st_as_sf()

ggplot2::ggplot() +
  ggplot2::geom_sf(Grid, mapping = ggplot2::aes(), color = "black",
                   linewidth = 0.5, fill = "transparent") +
  ggplot2::scale_x_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::scale_y_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::theme_minimal()

Grid_X <- GridCross(Grid)

ggplot2::ggplot() +
  ggplot2::geom_sf(Grid, mapping = ggplot2::aes(), color = "black",
                   linewidth = 0.5, fill = "transparent") +
  ggplot2::geom_sf(Grid_X, mapping = ggplot2::aes(), color = "red",
                   linewidth = 0.5, inherit.aes = TRUE) +
  ggplot2::scale_x_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::scale_y_continuous(expand = c(0, 0, 0, 0), limits = c(0, 10)) +
  ggplot2::theme_minimal()
}
\author{
Ahmed El-Gabbas
}
