% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_ht.R
\name{ht}
\alias{ht}
\title{Print head and tail of data frame}
\usage{
ht(DF, NRows = 5)
}
\arguments{
\item{DF}{data frame to print}

\item{NRows}{N umber of rows to print at the top and bottom of data frame}
}
\description{
Print head and tail of data frame
}
\examples{
ht(mtcars)

# -------------------------------------------

ht(mtcars, 2)

# -------------------------------------------

ht(mtcars, 6)
}
\author{
Ahmed El-Gabbas
}
