% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_IAS_process.R, R/DWF_IAS_distribution.R,
%   R/DWF_IAS_plot.R
\name{IAS_data}
\alias{IAS_data}
\alias{IAS_process}
\alias{IAS_distribution}
\alias{IAS_plot}
\title{Process and map Invasive Alien Species (IAS) data for the \code{IAS-pDT}}
\usage{
IAS_process(env_file = ".env", n_cores = 6L, overwrite = TRUE)

IAS_distribution(
  species = NULL,
  env_file = ".env",
  verbose = FALSE,
  overwrite = FALSE
)

IAS_plot(species = NULL, env_file = ".env", overwrite = FALSE)
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}

\item{overwrite}{Logical. If \code{TRUE}, overwrites existing maps; otherwise,
skips if maps exist. Default: \code{FALSE}.}

\item{species}{Character. Species name for distribution mapping.}

\item{verbose}{Logical. If \code{TRUE}, prints progress messages. Default:
\code{FALSE}.}
}
\description{
Processes and visualizes Invasive Alien Species (IAS) distribution data from
GBIF, EASIN, and eLTER for the Invasive Alien Species prototype Digital Twin
(\code{IAS-pDT}). Merges pre-processed data, creates presence-absence rasters,
summarizes distributions, and generates maps using helper functions.
}
\section{Functions details}{

\itemize{
\item \strong{\code{IAS_process()}}: Merges pre-processed GBIF (\code{GBIF_process()}), EASIN
(\code{EASIN_process()}), and eLTER (\code{eLTER_process()}) data (run these first).
Outputs \code{SpatRaster} distribution rasters, summary tables, and JPEG maps
using \code{IAS_distribution()} and \code{IAS_plot()}.
\item \strong{\code{IAS_distribution()}}: Generates presence-absence maps (\code{.RData},
\code{.tif})  for a species, including all grid cells in the study area and a set
excluding cultivated/casual-only countries. Returns a tibble with presence
counts (total, by source) and summary statistics for biogeographical
regions
\item \strong{\code{IAS_plot()}}: Creates JPEG distribution maps from GBIF, EASIN, and
eLTER data using \code{ggplot2}.
}
}

\author{
Ahmed El-Gabbas
}
