% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_IAS_Distribution.R
\name{IAS_Distribution}
\alias{IAS_Distribution}
\title{Prepare distribution maps and summary for Invasive Alien Species (IAS)}
\usage{
IAS_Distribution(Species, FromHPC = TRUE, EnvFile = ".env", Verbose = FALSE)
}
\arguments{
\item{Species}{Character. Name of the species to analyze.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default is ".env".}

\item{Verbose}{Logical. If TRUE, detailed output is printed. Default is
\code{FALSE}.}
}
\value{
A tibble containing species distribution information, including the
number of presence grid cells, presence by data provider, and summary
statistics for biogeographical regions.
}
\description{
This function processes and analyzes distribution data for Invasive Alien
Species (IAS) from multiple sources (GBIF, EASIN, and eLTER). It generates
presence-absence maps, summarizes distribution data per country and
biogeographical region, and outputs results as tif, RData, and tibble
formats.
}
\note{
\itemize{
\item This function is not intended to be used directly by the user or in the
IAS-pDT, but only used inside the \link{IAS_Process} function.
\item The function returns two sets of presence-absence maps: 1) at the European
scale giving the data in the three data sources; and 2) the same maps, but
with assigning grid cells from countries with only cultivated or casual
observations.
}
}
\author{
Ahmed El-Gabbas
}
