% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_IAS_Plotting.R
\name{IAS_Plot}
\alias{IAS_Plot}
\title{Plot IAS distribution map}
\usage{
IAS_Plot(Species, FromHPC = TRUE, EnvFile = ".env", Overwrite = FALSE)
}
\arguments{
\item{Species}{Character. The name of the species for which the distribution
plot will be generated.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default is ".env".}

\item{Overwrite}{Logical. If \code{TRUE}, the function will overwrite existing
maps; otherwise, it will skip generating a new map if one exists (default:
\code{FALSE}).}
}
\value{
The function uses \code{ggplot2} to generate a JPEG file containing the
species distribution map. No object is returned.
}
\description{
This function prepares species distribution maps for IAS, showing the
distribution of the species at the three data sources used (GBIF, EASIN, and
eLTER).
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only called from the \link{IAS_Processing} function.
}
\author{
Ahmed El-Gabbas
}
