% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_IAS_Process.R
\name{IAS_Process}
\alias{IAS_Process}
\title{Process IAS data}
\usage{
IAS_Process(FromHPC = TRUE, EnvFile = ".env", NCores = 6, Overwrite = TRUE)
}
\arguments{
\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default is ".env".}

\item{NCores}{Numeric. Number of cores to use for parallel processing.}

\item{Overwrite}{Logical. If \code{TRUE}, existing JPEG files will be overwritten
during processing. See \link{IAS_Plot}}
}
\value{
The function returns \code{NULL} invisibly.
}
\description{
This function processes Invasive Alien Species (IAS) data. The function
merges pre-processed distribution data from 3 data sources: GBIF
(\link{GBIF_Process}), EASIN (\link{EASIN_Process}), eLTER (\link{elTER_Process}).
The function prepares final species outputs in the form of 1) species
distribution as SpatRaster (\code{.RData}) and \code{.tif} using \link{IAS_Distribution}; 2)
summary table on the distribution of the species; and 3) JPEG files for the
species distribution showing the the sources of presence data \link{IAS_Plot}. It
further generates summary results in the form of maps and figures.
}
\note{
\itemize{
\item The function should be used only after data from the three data sources
were processed: GBIF (\link{GBIF_Process}), EASIN (\link{EASIN_Process}), eLTER
(\link{elTER_Process}).
\item The function depends on the following functions: \link{IAS_Distribution} for
prepare species-specific final maps and \link{IAS_Plot} for plotting species
distribution as JPEG.
}
}
\author{
Ahmed El-Gabbas
}
