% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Install_Hmsc.R
\name{Install_Hmsc}
\alias{Install_Hmsc}
\title{Install Hmsc-HPC in a Python Virtual Environment on Windows}
\usage{
Install_Hmsc(Path_Python, Path_VE, GitURL)
}
\arguments{
\item{Path_Python}{Character string. Path to the Python executable.}

\item{Path_VE}{Character string. Path where the virtual environment will be
created. This can not be an existing folder.}

\item{GitURL}{A character string containing the URL of the Git repository
(and branch name) for the HMSC-HPC package.}
}
\value{
The function performs installation steps and returns NULL invisibly.
}
\description{
This function sets up a Python virtual environment and installs the
\href{https://github.com/aniskhan25/hmsc-hpc}{\code{Hmsc-HPC}} package from a specified
Git repository. It also installs \code{TensorFlow} and performs checks to to
verify the installation of Python, the virtual environment, and the packages.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if the virtual environment directory already exists and stops with an error if it does.
\item Verifies the Python version and installation.
\item Creates a new Python virtual environment.
\item Upgrades \code{pip} in the virtual environment.
\item Installs the \code{Hmsc-HPC} package from the provided Git URL.
\item Installs \code{TensorFlow} version 2.15.
\item Checks \code{TensorFlow} and \code{Hmsc-HPC} package installations.
}
}
\author{
Ahmed El-Gabbas
}
