% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_install_hmsc_windows.R
\name{install_hmsc_windows}
\alias{install_hmsc_windows}
\title{Install Hmsc-HPC in a python virtual environment on Windows}
\usage{
install_hmsc_windows(
  path_python,
  path_VE,
  URL_Hmsc = "https://github.com/trossi/hmsc-hpc.git@simplify-io",
  URL_rdata = "https://github.com/trossi/rdata.git@test-for-hmsc-v2",
  force = FALSE
)
}
\arguments{
\item{path_python}{Character. Path to the Python executable.}

\item{path_VE}{Character. Path where the virtual environment will be created.
This can not be an existing folder.}

\item{URL_Hmsc}{Character. URL of the Git repository (and branch name) of the
\code{Hmsc-HPC} package.}

\item{URL_rdata}{Character. URL of the Git repository (and branch name) of
the \code{rdata} package. This is temporary to allow \code{rdata} package to write
rds file. In the near future, this functionality will be pushed to the
main branch of \code{rdata}.}

\item{force}{Logical. Whether to force the installation of \code{Hmsc-HPC}; i.e.
using \code{--force-reinstall} as a suffix to the \verb{pip install} command}
}
\value{
The function performs installation steps and returns NULL invisibly.
}
\description{
This function sets up a Python virtual environment and installs the
\href{https://github.com/aniskhan25/hmsc-hpc}{\code{Hmsc-HPC}} package from a specified
Git repository. It also installs \code{TensorFlow} and performs checks to to
verify the installation of Python, the virtual environment, and the packages.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if the virtual environment directory already exists and stops with
an error if it does
\item Verifies the Python version and installation
\item Creates a new Python virtual environment
\item Upgrades \code{pip} in the virtual environment
\item Installs the \code{Hmsc-HPC} package from the provided Git URL
\item Installs \code{TensorFlow} version 2.15
\item Installs \code{rdata}
\item Checks \code{TensorFlow} and \code{Hmsc-HPC} package installations
}
}
\examples{
\dontrun{
install_hmsc_windows(
   path_python = "C:/Python/Python310/python.exe",
   path_VE = "D:/Hmsc-HPC")
}
}
\author{
Ahmed El-Gabbas
}
