% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_integer_breaks.R
\name{integer_breaks}
\alias{integer_breaks}
\title{integer axis values for ggplot}
\source{
https://joshuacook.netlify.app/post/integer-values-ggplot-axis/
}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer breaks that requested.}

\item{...}{other arguments passed on to \code{scales::pretty()}}
}
\description{
integer axis values for ggplot
}
