% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{KeepOnly}
\alias{KeepOnly}
\title{Keep only certain objects in memory, all other objects will be removed}
\usage{
KeepOnly(Obj = NULL, Verbose = TRUE)
}
\arguments{
\item{Obj}{character vector for objects to be kept in memory}

\item{Verbose}{Should the names of kept and removed variables printed? Default: \code{TRUE}.}
}
\description{
Keep only certain objects in memory, all other objects will be removed
}
\examples{
A <- B <- C <- 15
ls()

KeepOnly("A")

ls()
rm(list = ls())


A <- B <- C <- 15
KeepOnly(c("A","B"))
ls()
}
\author{
Ahmed El-Gabbas
}
