% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{LoadAs}
\alias{LoadAs}
\title{Load RData file as specific object; i.e. rename loaded object}
\usage{
LoadAs(File = NA)
}
\arguments{
\item{File}{path of file}
}
\description{
Load RData file as specific object; i.e. rename loaded object
}
\author{
Ahmed El-Gabbas
}
