% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_LoadedPackages.R
\name{LoadedPackages}
\alias{LoadedPackages}
\title{List of loaded packages}
\usage{
LoadedPackages()
}
\value{
A character vector containing the names of all loaded packages.
}
\description{
This function returns a character vector listing all the packages that are currently loaded in the R session.
}
\examples{
LoadedPackages()

LoadPackages(sf)

LoadedPackages()

}
