% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{LoadMultiple}
\alias{LoadMultiple}
\title{Load multiple RData files together}
\usage{
LoadMultiple(
  Files = NULL,
  Verbose = TRUE,
  OneObject = TRUE,
  ReturnNames = TRUE
)
}
\arguments{
\item{Files}{vector containing the list of \code{.RData} files to be loaded}

\item{Verbose}{Verbose? Only valid when \code{OneObject = FALSE}. Default = \code{TRUE}}

\item{OneObject}{Should the objects loaded as one object, each as a list item? If \code{FALSE}, then all original object names will be loaded to the global environment. Default = \code{TRUE}.}

\item{ReturnNames}{Should the name of the objects loaded be returned? Only valid when \code{OneObject = FALSE}. Default = \code{TRUE}.}
}
\description{
Load multiple RData files together
}
\examples{
\dontrun{
(Files <- system.file("testdata", c("culcita_dat.RData", "gopherdat2.RData"), package = "lme4"))

# ---------------------------------------------------

# Load multiple *.rd files to one list object
print(Files)
file.exists(Files)
MultiObj <- LoadMultiple(Files = Files, OneObject = TRUE)
MultiObj

# ---------------------------------------------------

# Load multiple *.rd files current environment
rm("MultiObj")
LoadMultiple(Files = Files, OneObject = FALSE)

print(c("culcita_dat", "Gdat") \%in\% ls(globalenv()))
print(Gdat)
print(culcita_dat)

# ---------------------------------------------------

# Load multiple *.rd files, one object already exists
rm("culcita_dat")
ls()
print(c("culcita_dat", "Gdat") \%in\% ls(globalenv()))
LoadMultiple(Files = Files, OneObject = FALSE)
}
}
\author{
Ahmed El-Gabbas
}
