% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_LoadPackages.R
\name{LoadPackages}
\alias{LoadPackages}
\title{Load or Install Multiple R Packages}
\usage{
LoadPackages(..., List = NULL, Verbose = FALSE, InstallMissing = TRUE)
}
\arguments{
\item{...}{character. Names of the packages to be loaded or installed.}

\item{List}{character vector. An alternative or additional way to specify
package names as a vector.}

\item{Verbose}{logical. If \code{TRUE}, prints the names and versions of the
loaded packages. Defaults to \code{FALSE}.}

\item{InstallMissing}{logical. If \code{TRUE} (default), missing packages are
automatically installed and then loaded.}
}
\value{
This function is used for its side effects (loading/installing
packages) and does not return any value.
}
\description{
This function attempts to load multiple R packages specified by the user. If
a package is not installed, the function can optionally install it before
loading. It also provides an option to print the names and versions of the
loaded packages.
}
\examples{
# Currently loaded packages
(P1 <- IASDT.R::LoadedPackages())

LoadPackages(Hmsc)
# Loaded packages after implementing the function
(P2 <- IASDT.R::LoadedPackages())

# Which packages were loaded?
setdiff(P2, P1)

LoadPackages(terra, Verbose = TRUE)
}
\author{
Ahmed El-Gabbas
}
