% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{match_to_gbif.fn}
\alias{match_to_gbif.fn}
\title{Match taxonomy with GBIF; may return >1 match}
\usage{
match_to_gbif.fn(taxon_name, taxon_id, include_genus = FALSE)
}
\arguments{
\item{taxon_name}{taxonomy name}

\item{taxon_id}{taxonomy ID}

\item{include_genus}{include matches at genus level; default: \code{FALSE}}
}
\description{
Match taxonomy with GBIF; may return >1 match
}
\details{
as input, provide a vector of verbatim taxon names (preferably with authorship) and a vector of existing local identifiers for those names
}
\author{
Marina Golivets
}
