% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_Match_to_GBIF.R
\name{Match_to_GBIF}
\alias{Match_to_GBIF}
\title{Match taxonomy with GBIF; may return >1 match}
\usage{
Match_to_GBIF(
  taxon_name,
  taxon_id = NULL,
  include_genus = FALSE,
  Parallel = FALSE,
  Progress = FALSE
)
}
\arguments{
\item{taxon_name}{Taxonomy name}

\item{taxon_id}{Taxonomy ID}

\item{include_genus}{Include matches at genus level; default: \code{FALSE}}

\item{Parallel}{Logical. Whether to implement standardization on parallel; default: \code{FALSE}}

\item{Progress}{Logical. Whether to print progress bar; default: \code{FALSE}}
}
\value{
a tibble for the standardization results
}
\description{
Match taxonomy with GBIF; may return >1 match
}
\details{
as input, provide a vector of verbatim taxon names (preferably with authorship) and a vector of existing local identifiers for those names
}
\author{
Marina Golivets
}
