% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_CV_Eval.R
\name{Mod_CV_Eval}
\alias{Mod_CV_Eval}
\title{Evaluate cross-validated Hmsc Model}
\usage{
Mod_CV_Eval(Path_CV = NULL, predictEtaMean = TRUE, NCores = 8L)
}
\arguments{
\item{Path_CV}{Character. The directory path where cross-validation models
and outputs are stored.}

\item{predictEtaMean}{boolean flag indicating whether to use the estimated
mean values of posterior predictive distribution for random effects
corresponding for the new units. See \code{Hmsc:::predict.Hmsc} for more
details.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}
}
\description{
This function evaluates Hmsc models trained via cross-validation using. It
merges posterior chains, makes predictions, and computes 4 evaluation
metrics: AUC, RMSE, Tjur's R2, and Boyce Index.
}
\author{
Ahmed El-Gabbas
}
