% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cv_evaluate.R
\name{mod_cv_evaluate}
\alias{mod_cv_evaluate}
\title{Cross-validation Model Evaluation and Plotting}
\usage{
mod_cv_evaluate(model_dir = NULL, cv_type = "cv_dist")
}
\arguments{
\item{model_dir}{Character. Path to the root directory of the fitted model.}

\item{cv_type}{Character. Cross-validation type. One of \code{cv_dist} (default)
or \code{cv_large}. See \code{\link[=mod_cv_fit]{mod_cv_fit()}} for more details.}
}
\value{
Invisibly returns the path to the saved evaluation data file.
}
\description{
performs model evaluation using cross-validation results, calculates multiple
metrics (AUC, Tjur R\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}, Boyce index, RMSE), and generates summary
plots for explanatory and predictive power.
}
\author{
Ahmed El-Gabbas
}
