% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cv_fit.R
\name{mod_cv_fit}
\alias{mod_cv_fit}
\title{Prepare cross-validated Hmsc models for HPC fitting}
\usage{
mod_cv_fit(
  path_model = NULL,
  cv_name = c("cv_dist", "cv_large"),
  partitions = NULL,
  env_file = ".env",
  init_par = NULL,
  job_name = "cv_models",
  updater = list(Gamma2 = FALSE, GammaEta = FALSE),
  align_posterior = TRUE,
  to_json = FALSE,
  slurm_prepare = TRUE,
  memory_per_cpu = NULL,
  job_runtime = NULL,
  path_hmsc = NULL,
  precision = 64,
  ...
)
}
\arguments{
\item{path_model}{Character. Path to a saved model file (\verb{*.qs2}).}

\item{cv_name}{Character vector. Column name(s) in the model input data to be
used to cross-validate the models (see \link{mod_prepare_data} and
\link{mod_cv_prepare}). The function allows the possibility of using more than
one way of assigning grid cells into cross-validation folders. If multiple
names are provided, separate cross-validation models will be fitted for
each cross-validation type. Currently, there are three cross-validation
strategies: \code{cv_sac}, \code{cv_dist}, and \code{cv_large}. Defaults to \code{c("cv_dist", "cv_large")}.}

\item{partitions}{A vector for cross-validation created by
\link[Hmsc:createPartition]{Hmsc::createPartition} or similar. Defaults to \code{NULL}, which means to use
column name(s) provided in the \code{cv_name} argument. If the \code{partitions}
vector is provided, the label used in the output files will be \code{cv_custom}.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{init_par}{a named list of parameter values used for initialisation of
MCMC states. See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information.
Default: \code{NULL}.}

\item{job_name}{Character. Name of the submitted job(s) for SLURM. Default:
\code{cv_models}.}

\item{updater}{named \code{list}. Which conditional updaters should be omitted?
See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information. Defaults to
\code{list(Gamma2 = FALSE, GammaEta = FALSE)} to disable the following warnings:
\verb{setting updater$Gamma2=FALSE due to specified phylogeny matrix} and
\verb{setting updater$GammaEta=FALSE: not implemented for spatial methods 'GPP' and 'NNGP'}.}

\item{align_posterior}{Logical. Whether the posterior of each chains should
be aligned. See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information.
Default: \code{TRUE}.}

\item{to_json}{Logical. Whether to convert unfitted models to JSON before
saving to RDS file. Default: \code{FALSE}.}

\item{slurm_prepare}{Logical. Whether to prepare SLURM command files. If
\code{TRUE} (default), the SLURM commands will be saved to disk using the
\link{mod_slurm} function.}

\item{memory_per_cpu}{Character. Memory allocation per CPU core. Example:
"32G" for 32 gigabytes. Defaults to "64G".}

\item{job_runtime}{Character. Maximum allowed runtime for the job. Example:
"01:00:00" for one hour. Required --- if not provided, the function throws
an error.}

\item{path_hmsc}{Character. Path to the Hmsc-HPC installation.}

\item{precision}{Integer. Must be either 32 or 64 (default). Defines the
floating-point precision mode for \code{Hmsc-HPC} sampling (--fp 32 or --fp 64).}

\item{...}{Additional arguments passed to the \link{mod_slurm} function.}
}
\description{
This function prepares cross-validated Hmsc models for fitting using HPC. It
handles data preparation, model initialisation, and generation of SLURM
commands.
}
\details{
The function copies part of the \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues}
function, which currently does not support performing cross-validation
using Hmsc-HPC.
}
\author{
Ahmed El-Gabbas
}
