% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CV_prepare.R
\name{mod_CV_prepare}
\alias{mod_CV_prepare}
\title{Prepare spatial-block cross-validation folds for spatial analysis}
\usage{
mod_CV_prepare(
  input_data = NULL,
  env_file = ".env",
  x_vars = NULL,
  CV_n_folds = 4L,
  CV_n_grids = 20L,
  CV_n_rows = 2L,
  CV_n_columns = 2L,
  CV_SAC = FALSE,
  out_path = NULL,
  CV_plot = TRUE
)
}
\arguments{
\item{input_data}{\code{data.frame}. A data frame or tibble containing the input
dataset. This data frame should include two columns for \code{x} and \code{y}
coordinates as long as other columns matching the names of predictors
listed in \code{x_vars} argument. This argument is mandatory and can not be
empty.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{x_vars}{Character vector. Variables to be used in the model. This
argument is mandatory and can not be empty.}

\item{CV_n_folds}{Integer. Number of cross-validation folds. Default: 4L.}

\item{CV_n_grids}{Integer. Number of grid cells in both directions used in
the \code{CV_Dist} cross-validation strategy (see below). Default: 20L.}

\item{CV_n_rows, CV_n_columns}{Integer. Number of rows and columns used in the
\code{CV_Large} cross-validation strategy (see below), in which the study area
is divided into large blocks given the provided \code{CV_n_rows} and
\code{CV_n_columns} values. Both default to 2L which means to split the study
area into four large blocks at the median latitude and longitude.}

\item{CV_SAC}{Logical. Whether to use the spatial autocorrelation to
determine the block size. Defaults to \code{FALSE},}

\item{out_path}{Character. Path for directory to save the cross-validation
results. This argument is mandatory and can not be empty.}

\item{CV_plot}{Logical. Indicating whether to plot the block cross-validation
folds.}
}
\value{
The function returns a modified version of the input dataset with
additional numeric columns (integer) indicating the cross-validation
strategy used.
}
\description{
This function assign modelling input data into spatial-block cross-validation
folds using three strategies (see below) using \link[blockCV:cv_spatial]{blockCV::cv_spatial}. The
function is planned to be used inside the \link{mod_prepare_HPC}
function.
}
\note{
The function uses the following cross-validation strategies:
\itemize{
\item \code{CV_Dist} in which the size of spatial cross-validation blocks is
determined by the \code{CV_n_grids} argument. The default \code{CV_n_grids} value is
20L, which means blocks of 20×20 grid cell each.
\item \code{CV_Large} which splits the study area into large blocks, as determined
by the  \code{CV_n_rows} and \code{CV_n_columns} arguments. if \code{CV_n_rows = CV_n_columns} = 2L (default), four large blocks will be used, split the
study area at the median coordinates.
\item \code{CV_SAC} in which the size of the blocks is determined by the median
spatial autocorrelation range in the predictor data (estimated using
\link[blockCV:cv_spatial_autocor]{blockCV::cv_spatial_autocor}). This requires the availability of the
\code{automap} R package. This strategy is currently skipped by default.
}
}
\author{
Ahmed El-Gabbas
}
