% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Evaluation_Plot.R
\name{Mod_Eval_Plot}
\alias{Mod_Eval_Plot}
\title{Generate plots for the explanatory power of Hmsc models}
\usage{
Mod_Eval_Plot(ModelDir, EnvFile = ".env", FromHPC = TRUE)
}
\arguments{
\item{ModelDir}{Character. Path to the model directory containing
predictions.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}
}
\description{
This function generates four diagnostic plots (RMSE, AUC, Continuous Boyce
Index, and Tjur-R²) to evaluate the performance (explanatory power without
cross-validation) of Hmsc models.
}
\author{
Ahmed El-Gabbas
}
