% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Evaluate.R
\name{Mod_Evaluate}
\alias{Mod_Evaluate}
\title{Model Evaluation}
\usage{
Mod_Evaluate(Path_CV = NULL, predictEtaMean = TRUE, NCores = 8)
}
\arguments{
\item{Path_CV}{The path to the cross-validation folder where models and data
are stored.}

\item{predictEtaMean}{Logical; whether to predict the mean value of the
latent variable.}

\item{NCores}{Integer specifying the number of parallel cores for
parallelization. Default: 8 cores.}
}
\description{
This function performs cross-validated model prediction and evaluation. It
merges posterior chains, and makes predictions for cross-validation folds. It
evaluates the model fit either for internal evaluation or on
cross-validation.
}
