% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_evaluate.R
\name{mod_evaluate_cv}
\alias{mod_evaluate_cv}
\title{Cross-validation Model Evaluation and Plotting}
\usage{
mod_evaluate_cv(model_dir = NULL, cv_type = "CV_Dist")
}
\arguments{
\item{model_dir}{Character. Path to the root directory of the fitted model.}

\item{cv_type}{Character. Cross-validation type. One of \code{CV_Dist} (default)
or \code{CV_Large}. See \code{\link[=mod_CV_fit]{mod_CV_fit()}} for more details.}
}
\value{
Invisibly returns the path to the saved evaluation data file.
}
\description{
performs model evaluation using cross-validation results, calculates multiple
metrics (AUC, Tjur R\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}, Boyce index, RMSE), and generates summary
plots for explanatory and predictive power.
}
\author{
Ahmed El-Gabbas
}
