% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Fit_WS.R
\name{Mod_Fit_WS}
\alias{Mod_Fit_WS}
\title{Fit Hmsc-HPC models on UFZ Windows Server}
\usage{
Mod_Fit_WS(Path_Model, EnvFile = ".env", NCores = NULL)
}
\arguments{
\item{Path_Model}{Character. Path to the model files.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.}
}
\value{
The function does not return anything but prints messages to the
console regarding the progress and completion of model fitting.
}
\description{
This function fits Hmsc models on a UFZ Windows Server. It reads model
configurations from a specified path, loads environment variables, checks
input arguments for validity, and executes model fitting in parallel if
required.
}
\author{
Ahmed El-Gabbas
}
