% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_posteriors.R
\name{mod_get_posteriors}
\alias{mod_get_posteriors}
\title{Combines posteriors exported by \code{Hmsc-HPC} into an Hmsc object}
\usage{
mod_get_posteriors(path_posterior = NULL, from_JSON = FALSE)
}
\arguments{
\item{path_posterior}{Character vector. Path to the RDS files containing the
exported posterior files. This argument is mandatory and cannot be empty.}

\item{from_JSON}{Logical. Whether the loaded models should be converted from
\code{JSON} format. Defaults to \code{FALSE}, meaning the data will be read directly
from RDS files without conversion.}
}
\value{
Depending on the \code{from_JSON} parameter, returns an Hmsc object either
directly from the RDS files or after converting it from JSON format.
}
\description{
This function converts posterior files exported by \code{Hmsc-HPC} into an Hmsc
object. It can either read the data directly from RDS files or convert it
from JSON format if specified.
}
\author{
Ahmed El-Gabbas
}
