% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_GetPosts.R
\name{Mod_GetPosts}
\alias{Mod_GetPosts}
\title{Combines posteriors exported by \code{Hmsc-HPC} into an Hmsc object}
\usage{
Mod_GetPosts(Path_Post = NULL, FromJSON = FALSE)
}
\arguments{
\item{Path_Post}{Character vector. Path to the RDS files containing the
exported posterior files. This argument is mandatory and cannot be empty.}

\item{FromJSON}{Logical. Whether the loaded models should be converted from
\code{JSON} format. Defaults to \code{FALSE}, meaning the data will be read directly
from RDS files without conversion.}
}
\value{
Depending on the \code{FromJSON} parameter, returns an Hmsc object either
directly from the RDS files or after converting it from JSON format.
}
\description{
This function converts posterior files exported by \code{Hmsc-HPC} into an Hmsc
object. It can either read the data directly from RDS files or convert it
from JSON format if specified.
}
\author{
Ahmed El-Gabbas
}
