% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Prep4HPC.R, R/Mod_PrepData.R
\name{Mod_inputs}
\alias{Mod_inputs}
\alias{Mod_Prep4HPC}
\alias{Mod_PrepData}
\title{Prepare initial models for model fitting with Hmsc-HPC}
\usage{
Mod_Prep4HPC(
  Hab_Abb = NULL,
  Path_Model = NULL,
  MinEffortsSp = 100L,
  PresPerSpecies = 80L,
  EnvFile = ".env",
  GPP = TRUE,
  GPP_Dists = NULL,
  GPP_Save = TRUE,
  GPP_Plot = TRUE,
  MinLF = NULL,
  MaxLF = NULL,
  Alphapw = list(Prior = NULL, Min = 20, Max = 1200, Samples = 200),
  BioVars = c("bio3", "bio4", "bio11", "bio18", "bio19", "npp"),
  QuadraticVars = BioVars,
  EffortsAsPredictor = TRUE,
  RoadRailAsPredictor = TRUE,
  HabAsPredictor = TRUE,
  RiversAsPredictor = TRUE,
  NspPerGrid = 0L,
  ExcludeCult = TRUE,
  ExcludeZeroHabitat = TRUE,
  CV_NFolds = 4L,
  CV_NGrids = 20L,
  CV_NR = 2L,
  CV_NC = 2L,
  CV_Plot = TRUE,
  CV_SAC = FALSE,
  PhyloTree = TRUE,
  NoPhyloTree = FALSE,
  OverwriteRDS = TRUE,
  NCores = 8L,
  NChains = 4L,
  thin = NULL,
  samples = 1000L,
  transientFactor = 500L,
  verbose = 200L,
  SkipFitted = TRUE,
  NumArrayJobs = 210L,
  ModelCountry = NULL,
  VerboseProgress = TRUE,
  FromHPC = TRUE,
  PrepSLURM = TRUE,
  MemPerCpu = NULL,
  Time = NULL,
  JobName = NULL,
  Path_Hmsc = NULL,
  CheckPython = FALSE,
  ToJSON = FALSE,
  Precision = 64L,
  ...
)

Mod_PrepData(
  Hab_Abb = NULL,
  MinEffortsSp = 100L,
  ExcludeCult = TRUE,
  ExcludeZeroHabitat = TRUE,
  PresPerSpecies = 80L,
  EnvFile = ".env",
  Path_Model = NULL,
  VerboseProgress = TRUE,
  FromHPC = TRUE
)
}
\arguments{
\item{Hab_Abb}{Character. Abbreviation for the habitat type (based on
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab}) for which to prepare
data. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a}, \verb{4b}, \code{10}, \verb{12a}, \verb{12b}.
If \code{Hab_Abb} = \code{0}, data is prepared irrespective of the habitat type. For
more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{Path_Model}{Character. Directory path where all output files, including
the models to be fitted, will be saved.}

\item{MinEffortsSp}{Integer. Minimum number of vascular plant species per
grid cell (from GBIF data) required for inclusion in the models. This is to
exclude grid cells with very little sampling efforts. Defaults to \code{100}.}

\item{PresPerSpecies}{Integer. The minimum number of presence grid cells for
a species to be included in the analysis. The number of presence grid cells
per species is calculated after discarding grid cells with low sampling
efforts (\code{MinEffortsSp}) and zero percentage habitat coverage
\code{ExcludeZeroHabitat}. Defaults to \code{80}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{GPP}{Logical. Whether to fit spatial random effect using Gaussian
Predictive Process. Defaults to \code{TRUE}. If \code{FALSE}, non-spatial models will
be fitted.}

\item{GPP_Dists}{Integer. Spacing (in kilometers) between GPP knots, as well
as the minimum allowable distance between a knot and the nearest sampling
point. The knots are generated using the \link{Mod_PrepKnots} function, and this
value is used for both \code{knotDist} and \code{minKnotDist} in
\link[Hmsc:constructKnots]{Hmsc::constructKnots}.}

\item{GPP_Save}{Logical. Whether \code{RData} file. Default: \code{TRUE}.}

\item{GPP_Plot}{Logical. Whether to plot the coordinates of the sampling
units and the knots in a pdf file. Default: \code{TRUE}.}

\item{MinLF, MaxLF}{integer. Minimum and maximum number of latent factors to
be used. Both default to \code{NULL} which means that the number of latent
factors will be estimated from the data. If either is provided, the
respective values will be used as arguments to \link[Hmsc:setPriors]{Hmsc::setPriors}.}

\item{Alphapw}{Prior for the alpha parameter. Defaults to a list with \code{Prior = NULL}, \code{Min = 20}, \code{Max = 1200}, and \code{Samples = 200}. If \code{Alphapw} is
\code{NULL} or a list with all \code{NULL} list items, the default prior will be
used. If \code{Prior} is a matrix, it will be used as the prior. If \code{Prior = NULL}, the prior will be generated using \code{Min}, \code{Max}, and \code{Samples}. \code{Min}
and \code{Max} are the minimum and maximum values of the alpha parameter (in
kilometer). \code{Samples} is the number of samples to be used in the prior.}

\item{BioVars}{Character vector. Variables from CHELSA (bioclimatic variables
(bio1–bio19) and additional predictors (e.g., Net Primary Productivity,
npp)) to be used in the model. By default, six ecologically relevant and
minimally correlated variables are selected: c("bio3", "bio4", "bio11",
"bio18", "bio19", "npp").}

\item{QuadraticVars}{Character vector for variables for which quadratic terms
are used. Defaults to all variables of the \code{BioVars}. If \code{QuadraticVars} is
\code{NULL}, no quadratic terms will be used.}

\item{EffortsAsPredictor}{Logical. Whether to include the (log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}})
sampling efforts as predictor to the model. Default: \code{TRUE}.}

\item{RoadRailAsPredictor}{Logical. Whether to include the (log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}})
sum of road and railway intensity as predictor to the model. Default:
\code{TRUE}.}

\item{HabAsPredictor}{Logical. Whether to include the (log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}})
percentage coverage of respective habitat type per grid cell as predictor
to the model. Default: \code{TRUE}. Only valid if \code{Hab_Abb} not equals to \code{0}.}

\item{RiversAsPredictor}{Logical. Whether to include the (log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}})
total length of rivers per grid cell as predictor to the model. Default:
\code{TRUE}. See \link{River_Length} for more details.}

\item{NspPerGrid}{Integer. Minimum number of species required for a grid cell
to be included in the analysis. This filtering occurs after applying
\code{MinEffortsSp} (sampling effort thresholds), \code{PresPerSpecies} (minimum
species presence thresholds), and \code{ExcludeZeroHabitat} (exclude 0\% habitat
coverage). Default (0): Includes all grid cells. Positive value (>0):
Includes only grid cells where at least \code{NspPerGrid} species are present.}

\item{ExcludeCult}{Logical. Whether to exclude countries with cultivated or
casual observations per species. Defaults to \code{TRUE}.}

\item{ExcludeZeroHabitat}{Logical. Whether to exclude grid cells with zero
percentage habitat coverage. Defaults to \code{TRUE}.}

\item{CV_NFolds}{Integer. Number of cross-validation folds. Default: 4L.}

\item{CV_NGrids}{Integer. For \code{CV_Dist} cross-validation strategy (see
\link{Mod_GetCV}), this argument determines the size of the blocks (how many
grid cells in both directions).}

\item{CV_NR, CV_NC}{Integer. Number of rows and columns used in the \code{CV_Large}
cross-validation strategy  (see \link{Mod_GetCV}), in which the study area is
divided into large blocks given the provided \code{CV_NR} and \code{CV_NC} values.
Both default to 2 which means to split the study area into four large
blocks at the median latitude and longitude.}

\item{CV_Plot}{Logical. Indicating whether to plot the block cross-validation
folds.}

\item{CV_SAC}{Logical. Whether to use the spatial autocorrelation
to determine the block size. Defaults to \code{FALSE},}

\item{PhyloTree, NoPhyloTree}{Logical. Whether to fit models with (PhyloTree)
or without (NoPhyloTree) phylogenetic trees. Defaults are \code{PhyloTree = TRUE} and \code{NoPhyloTree = FALSE}, meaning only models with phylogenetic
trees are fitted by default. At least one of \code{PhyloTree} and \code{NoPhyloTree}
should be \code{TRUE}.}

\item{OverwriteRDS}{Logical. Whether to overwrite previously exported RDS
files for initial models. Default: \code{TRUE}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{NChains}{Integer. Number of model chains. Default: 4.}

\item{thin}{Integer vector. Thinning value(s) in MCMC sampling. If more than one
value is provided, a separate model will be fitted at each value of
thinning.}

\item{samples}{Integer vector. Value(s) for the number of MCMC samples. If more than
one value is provided, a separate model will be fitted at each value of
number of samples. Defaults to 1000.}

\item{transientFactor}{Integer. Transient multiplication factor. The value of
\code{transient} will equal the multiplication of \code{transientFactor} and \code{thin}.
Default: 500.}

\item{verbose}{Integer. Interval at which MCMC sampling progress is reported.
Default: \code{200}.}

\item{SkipFitted}{Logical. Whether to skip already fitted models. Default:
\code{TRUE}.}

\item{NumArrayJobs}{Integer. Number of jobs per SLURM script file. In LUMI
HPC, there is a limit of 210 submitted jobs per user for the \code{small-g}
partition. This argument is used to split the jobs into multiple SLURM
scripts if needed. Default: 210. See \href{https://docs.lumi-supercomputer.eu/runjobs/scheduled-jobs/partitions}{LUMI documentation}
for more details.}

\item{ModelCountry}{Character. Country or countries to filter observations
by. Default: \code{NULL}, which means prepare data for the whole Europe.}

\item{VerboseProgress}{Logical. Whether to print a message upon successful
saving of files. Defaults to \code{FALSE}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{PrepSLURM}{Logical. Whether to prepare SLURM command files. If \code{TRUE}
(default), the SLURM commands will be saved to disk using the
\link{Mod_SLURM} function.}

\item{MemPerCpu}{Character. Memory per CPU for the SLURM job. This value will
be assigned to the \verb{#SBATCH --mem-per-cpu=} SLURM argument. Example: "32G"
to request 32 gigabyte. Only effective if \code{PrepSLURM = TRUE}.}

\item{Time}{Character. Requested time for each job in the SLURM bash arrays.
Example: "01:00:00" to request an hour. Only effective if \code{PrepSLURM = TRUE}.}

\item{JobName}{Character. Name of the submitted job(s) for SLURM. If \code{NULL}
(Default), the job name will be prepared based on the folder path and the
\code{Hab_Abb} value. Only effective if \code{PrepSLURM = TRUE}.}

\item{Path_Hmsc}{Character. Directory path to \code{Hmsc-HPC} extension
installation. This will be provided as the \code{Path_Hmsc} argument of the
\link{Mod_SLURM} function.}

\item{CheckPython}{Logical. Whether to check if the Python executable exists.
Only valid if FromHPC = \code{FALSE}.}

\item{ToJSON}{Logical. Whether to convert unfitted models to JSON before
saving to RDS file. Default: \code{FALSE}.}

\item{Precision}{Integer (either 32 or 64). Defines the floating-point
precision mode for \code{Hmsc-HPC} sampling (--fp 32 or --fp 64). The default is
64, which is the default precision in \code{Hmsc-HPC}.}

\item{...}{Additional parameters provided to the \link{Mod_SLURM} function.}
}
\description{
The \strong{\code{Mod_Prep4HPC}} function prepares input data and initializes models
for fitting with \href{https://doi.org/10.1371/journal.pcbi.1011914}{Hmsc-HPC}. It
performs multiple tasks, including data preparation, defining spatial block
cross-validation folds, generating Gaussian Predictive Process (GPP) knots
(\href{https://doi.org/10.1002/ecy.2929}{Tikhonov et al.}), initializing models,
and creating HPC execution commands. The function supports parallel
processing and offers the option to include or exclude phylogenetic tree
data.\if{html}{\out{<br/>}}\if{html}{\out{<br/>}} The internal \strong{\code{Mod_PrepData}} function is used to prepare
habitat-specific data for Hmsc models. This function processes environmental
and species presence data, reads environment variables from a file, verifies
paths, loads and filters species data based on habitat type and minimum
presence grid cells per species, and merges various environmental layers
(e.g., CHELSA Bioclimatic variables, habitat coverage, road and railway
intensity, sampling efforts) into a single dataset. Processed data is saved
to disk as an \verb{*.RData} file.
}
\details{
The current models are fitted for 8 habitat types see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}:
\itemize{
\item \strong{1. Forests} -- closed vegetation dominated by deciduous or evergreen
trees
\item \strong{2. Open forests} -- woodlands with canopy openings created by
environmental stress or disturbance, including forest edges
\item \strong{3. Scrub} -- shrublands maintained by environmental stress (aridity) or
disturbance
\item \strong{4a. Natural grasslands} -- grasslands maintained by climate (aridity,
unevenly distributed precipitation), herbivores or environmental stress
(aridity, instability or toxicity of substrate)
\item \strong{4b. Human-maintained grasslands} -- grasslands dependent on regular
human-induced management (mowing, grazing by livestock, artificial burning)
\item \strong{10. Wetland} -- sites with the permanent or seasonal influence of
moisture, ranging from oligotrophic to eutrophic
\item \strong{12a. Ruderal habitats} -- anthropogenically disturbed or eutrophicated
sites, where the anthropogenic disturbance or fertilization is typically a
side-product and not the aim of the management
\item \strong{12b. Agricultural habitats} -- synanthropic habitats directly
associated with growing of agricultural products, thus dependent on
specific type of management (ploughing, fertilization)

\if{html}{\out{<br/>}}The following habitat types are excluded from the analysis:
\item \strong{5. Sandy} -- dunes and other habitats on unstable sandy substrate,
stressed by low nutrients, drought and disturbed by sand movement
\item \strong{6. Rocky} -- cliffs and rock outcrops with very shallow or no soil
\item \strong{7. Dryland} -- habitats in which drought stress limits vegetation
development
\item \strong{8. Saline} -- habitats stressed by high soil salinity
\item \strong{9. Riparian} -- a mosaic of wetlands, grasslands, tall-forb stands,
scrub and open forests in stream corridors
\item \strong{11. Aquatic} -- water bodies and streams with submerged and floating
plant species
}
}
\author{
Ahmed El-Gabbas
}
