% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_merge_chains.R
\name{mod_merge_chains}
\alias{mod_merge_chains}
\title{Merge model chains}
\usage{
mod_merge_chains(
  model_dir = NULL,
  n_cores = NULL,
  model_info_name = NULL,
  print_incomplete = TRUE,
  from_JSON = FALSE,
  out_extension = "qs2"
)
}
\arguments{
\item{model_dir}{Character. Path to the root directory of the fitted model.
Two folders will be created \code{Model_Fitted} and \code{Model_Coda} to store merged
model and coda objects, respectively.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.}

\item{model_info_name}{Character. Name of the file (without extension) where
the processed model information will be saved. If \code{NULL}, it overwrites the
\code{Model_Info.RData} file in the \code{model_dir} directory. If provided, a new
\code{.RData} file will be created with this name.}

\item{print_incomplete}{Logical. Whether to print the names of models that
were not successfully fitted to the console. Defaults to \code{TRUE}.}

\item{from_JSON}{Logical. Whether to convert loaded models from JSON format
before reading. Defaults to \code{FALSE}.}

\item{out_extension}{Character. File extension (without dot) for the files
containing the fitted model and coda objects. It can be either \code{qs2} or
\code{RData}. Defaults to \code{qs2}.}
}
\value{
The function does not return anything but saves the processed model
information to disk.
}
\description{
This function merges model chains, checks for missing or incomplete model
runs, and optionally prints information about incomplete models. It saves the
processed model objects to disk.
}
\author{
Ahmed El-Gabbas
}
