% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_LF.R
\name{Mod_Plot_LF}
\alias{Mod_Plot_LF}
\title{Plot spatial variation in site loadings of HMSC Models}
\usage{
Mod_Plot_LF(
  Path_Model = NULL,
  EnvFile = ".env",
  Plot_Width = 20,
  Plot_Height = 21
)
}
\arguments{
\item{Path_Model}{Path to the model file.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{Plot_Width, Plot_Height}{Numeric. The width and height of the
output plot in cm. Default is 20×21 cm.}
}
\description{
Generate and save spatial variation in site loadings of HMSC Models' latent
factors as a JPEG file.
}
\author{
Ahmed El-Gabbas
}
