% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Postprocess.R
\name{Mod_Postprocess_2_CPU}
\alias{Mod_Postprocess_2_CPU}
\title{Model pipeline for Hmsc analysis (using CPU)}
\usage{
Mod_Postprocess_2_CPU(
  ModelDir = NULL,
  Hab_Abb = NULL,
  NCores = 8L,
  FromHPC = TRUE,
  EnvFile = ".env",
  GPP_Dist = NULL,
  Tree = "Tree",
  Samples = 1000L,
  Thin = NULL,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  LF_NCores = NCores,
  LF_Check = FALSE,
  LF_Temp_Cleanup = TRUE,
  Temp_Cleanup = TRUE,
  N_Grid = 50L,
  CC_Models = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0",
    "UKESM1-0-LL"),
  CC_Scenario = c("ssp126", "ssp370", "ssp585"),
  RC_NCores = 8L,
  Pred_Clamp = TRUE,
  Fix_Efforts = "q90",
  Fix_Rivers = "q90",
  Pred_NewSites = TRUE
)
}
\arguments{
\item{ModelDir}{String. Path to the root directory of the fitted models
without the trailing slash. Two folders will be created \code{Model_Fitted} and
\code{Model_Coda} to store merged model and coda objects, respectively.}

\item{Hab_Abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type for
which data will be prepared. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a},
\verb{4b}, \code{10}, \verb{12a}, \verb{12b}. For more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{NCores}{Integer specifying the number of parallel cores for
parallelization. Default: 8 cores.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{GPP_Dist}{Integer specifying the distance in \emph{kilometers} between knots
for the selected model.}

\item{Tree}{Character string specifying if phylogenetic tree was used in the
selected model. Valid values are "Tree" or "NoTree". Default is "Tree".}

\item{Samples}{Integer specifying number of MCMC samples in the selected
model. Defaults to 1000.}

\item{Thin}{Integer specifying the value for thinning in the selected model.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{TF_use_single}{Logical indicating whether to use single precision for
the TF calculations. Defaults to \code{FALSE}.}

\item{LF_NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}

\item{LF_Check}{Logical. If TRUE, the function checks if the output files are
already created and valid. If FALSE, the function will only check if the
files exist without checking their integrity. Default is \code{FALSE}.}

\item{LF_Temp_Cleanup}{Logical indicating whether to delete temporary files
in the \code{Temp_Dir} after finishing the LF predictions.}

\item{Temp_Cleanup}{logical, indicating whether to clean up temporary files.
Defaults to \code{TRUE}.}

\item{N_Grid}{Integer specifying the number of points along the gradient for
continuous focal variables. Defaults to 50. See \link[Hmsc:constructGradient]{Hmsc::constructGradient}
for more details.}

\item{CC_Models}{Character vector. Specifies the climate models for future
predictions. Default: \code{c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL")}. Note: This parameter is temporary and may be
removed in future updates.}

\item{CC_Scenario}{Character vector. Specifies the climate scenarios for
future predictions. Default: \code{c("ssp126", "ssp370", "ssp585")}. Note: This
parameter is temporary and may be removed in future updates.}

\item{RC_NCores}{Integer specifying the number of cores to use for response
curve prediction. Defaults to \code{8}.}

\item{Pred_Clamp}{Logical indicating whether to clamp the sampling efforts at
a single value. Defaults to \code{TRUE}. If \code{TRUE}, the \code{Fix_Efforts} argument
must be provided.}

\item{Fix_Efforts}{Numeric or character. Defines the value to fix sampling
efforts less than the provided value. If numeric, the value is directly
used (log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be \code{median}, \code{mean},
\code{max}, or \code{q90} (q0\% Quantile). Using \code{max} can reflect extreme values
caused by rare, highly sampled locations (e.g., urban centers or popular
natural reserves). While using 90\% quantile avoid such extreme grid cells
while still capturing areas with high sampling effort. This argument is
mandatory when \code{Pred_Clamp} is set to \code{TRUE}.}

\item{Fix_Rivers}{Numeric or character. Similar to \code{Fix_Efforts}, but for
fixing the length of rivers. If numeric, the value is directly used
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be \code{median}, \code{mean}, \code{max},
\code{q90} (90\% quantile). It can be also \code{NULL} for not fixing the river length
predictor. Defaults to \code{q90}.}

\item{Pred_NewSites}{Logical indicating whether to predict habitat
suitability at new sites. Default: \code{TRUE}. Note: This parameter is
temporary and will be removed in future updates.}
}
\description{
This function continues running the analysis pipeline for postprocessing Hmsc
models. It is designed to automate the following:
\itemize{
\item process and visualize response curve
\item predict habitat suitability across different climate options
\item process and visualize variance partitioning
\item plot species & SR predictions as JPEG
\item compute and visualizing model internal evaluation (explanatory power)
\item postprocessing of fitted cross-validated models.
}
}
\details{
The function must be called after
\itemize{
\item completing \link{Mod_Postprocess_1_CPU} and \link{Mod_Prep_TF} on CPU,
\item executing \code{VP_SLURM.slurm} and \code{LF_SLURM.slurm} on GPU for making
processing response curves and LF predictions. Both SLURM scripts are
created using \link{Mod_Prep_TF},
\item fitting cross-validated models by submitting respective SLURM commands.
}
}
\author{
Ahmed El-Gabbas
}
