% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Pred2Array.R
\name{Mod_Pred2Array}
\alias{Mod_Pred2Array}
\title{Convert results from \code{Hmsc:::predict.Hmsc} function to array}
\usage{
Mod_Pred2Array(
  Model,
  Predict = TRUE,
  Preds = NULL,
  expected = TRUE,
  NCores = 1,
  predictEtaMean = TRUE
)
}
\arguments{
\item{Model}{Hmsc object or character for path to a saved model file. If
\code{Predict} is \code{True}, the model object will be used to make predictions
first.}

\item{Predict}{Logical; if \code{TRUE}, predictions are made using the Model
parameter before conversion. If \code{FALSE}, uses pre-computed predictions
provided in \code{Preds}.}

\item{Preds}{pre-computed predictions, resulted from \code{Hmsc:::predict.Hmsc},
to be converted into an array. Required if \code{Predict} is \code{FALSE}.}

\item{expected}{Logical; if \code{TRUE}, returns the location parameter of the
observation models. If \code{FALSE}, samples values from the observation models.
See \code{Hmsc:::predict.Hmsc} for more info.}

\item{NCores}{Integer; specifies the number of cores to use for parallel
computation of predicted values. Must be a positive integer.}

\item{predictEtaMean}{Logical; whether to predict the mean value of the
latent variable.}
}
\value{
An array of predictions.
}
\description{
This function converts the results from the (unexported)
\code{Hmsc:::predict.Hmsc} function into an array format. It allows for the direct
conversion of pre-computed predictions or the generation of predictions from
a model before conversion. It supports parallel computation for generating
predictions.
}
\examples{
Preds <- IASDT.R::Mod_Pred2Array(
    Predict = TRUE, Model = Hmsc::TD$m, expected = TRUE, NCores = 1)
str(Preds)

Preds <- IASDT.R::Mod_Pred2Array(
    Predict = TRUE, Model = Hmsc::TD$m, expected = FALSE, NCores = 1)
str(Preds)
}
\author{
Ahmed El-Gabbas
}
