% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_Plot.R
\name{Mod_Predict_Plot}
\alias{Mod_Predict_Plot}
\title{Plot species and level of invasion predictions as JPEG files using \code{ggplot2}}
\usage{
Mod_Predict_Plot(ModelDir, EnvFile = ".env", FromHPC = TRUE, NCores = 8)
}
\arguments{
\item{ModelDir}{Path to the model directory containing predictions.}

\item{EnvFile}{Path to the environment file (\code{.env}) for setting paths.}

\item{FromHPC}{Boolean indicating whether the environment is an HPC system.}

\item{NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}
}
\value{
Saves prediction plots as JPEG files in the specified output
directory.
}
\description{
Generate predictions for species and habitat models and saves the output as
JPEG files.
}
\author{
Ahmed El-Gabbas
}
