% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Postprocess.R
\name{Mod_Prep_TF}
\alias{Mod_Prep_TF}
\title{Prepare batch scripts for latent factor prediction on GPU}
\usage{
Mod_Prep_TF(
  Path_Models = "datasets/processed/model_fitting",
  NumFiles = 210,
  EnvFile = ".env",
  WD = NULL,
  Partition_Name = "small-g",
  LF_Time = "01:00:00",
  VP_Time = "01:30:00"
)
}
\arguments{
\item{Path_Models}{Character. Directory for fitted models. Default is
\code{datasets/processed/model_fitting}. A subdirectory \code{TF_postprocess} will
be created to store the batch scripts and log files.}

\item{NumFiles}{Integer. Number of output batch files to create. Must be less
than or equal to the maximum job limit of the HPC environment.}

\item{WD}{Character. Optional. Working directory to be set in batch scripts.
If If \code{NULL}, the working directory will not be changed.}

\item{Partition_Name}{Character. Name of the partition to submit the SLURM
jobs to. Default is \code{small-g}.}

\item{LF_Time}{Character. Time limit for LF prediction jobs. Default is
\code{01:00:00}.}

\item{VP_Time}{Character. Time limit for variance partitioning jobs. Default
is \code{01:00:00}.}
}
\value{
None. Writes batch files to \code{TF_postprocess} subdirectory.
}
\description{
This function automates the preparation of batch scripts for predicting the
latent factors (LF) of response curves and at new sites using GPU resources.
It reads input files that match a specified pattern, merges their contents,
sorts the commands, and distributes them into a user-defined number of output
files. Each batch script is designed to be compatible with an HPC
environment, such as LUMI, with TensorFlow setup included. The function
limits the number of output files to a specified maximum, defaults to 210 for
compatibility with LUMI's job limits.
}
\note{
This function is designed specifically for the LUMI HPC environment. It
assumes the \code{tensorflow} module is available and pre-configured with all
necessary Python packages. On other HPC systems, users may need to modify
the function to load a Python virtual environment or install the required
dependencies for TensorFlow and related packages.
}
\author{
Ahmed El-Gabbas
}
