% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepKnots.R
\name{Mod_PrepKnots}
\alias{Mod_PrepKnots}
\title{Prepare knot locations for Hmsc GPP models}
\usage{
Mod_PrepKnots(
  Coords = NULL,
  MinDist = NULL,
  JitterDist = 100,
  MinLF = NULL,
  MaxLF = NULL,
  Alphapw = list(Prior = NULL, Min = 20, Max = 1200, Samples = 200)
)
}
\arguments{
\item{Coords}{Numeric matrix or data frame containing the (x, y)
coordinates of sampling units.}

\item{MinDist}{Numeric. Minimum distance between knots
in meters. This distance is used for both \code{knotDist} and \code{minKnotDist}
parameters of the \link[Hmsc:constructKnots]{Hmsc::constructKnots} function.}

\item{JitterDist}{Numeric. The jitter distance applied to
overlapping coordinates to avoid exact duplicates. Defaults to 100 meters.}

\item{MinLF, MaxLF}{Integer. Minimum and maximum number of latent factors to
be used. Both default to \code{NULL} which means that the number of latent
factors will be estimated from the data. If either is provided, the
respective values will be used as arguments to \link[Hmsc:setPriors]{Hmsc::setPriors}.}

\item{Alphapw}{Prior for the alpha parameter. Defaults to a list with \code{Prior = NULL}, \code{Min = 20}, \code{Max = 1200}, and \code{Samples = 200}. If \code{Alphapw} is
\code{NULL} or a list with all \code{NULL} list items, the default prior will be
used. If \code{Prior} is a matrix, it will be used as the prior. If \code{Prior = NULL}, the prior will be generated using \code{Min}, \code{Max}, and \code{Samples}. \code{Min}
and \code{Max} are the minimum and maximum values of the alpha parameter (in
kilometer). \code{Samples} is the number of samples to be used in the prior.}
}
\value{
An object suitable for specifying the random level in HMSC GPP
models. This object contains the prepared knot locations.
}
\description{
Prepare the locations of knots for use in Gaussian Predictive Process (GPP)
models within the HMSC framework. It ensures that knots are spaced at a
minimum specified distance and applies jitter to any identical coordinates to
avoid overlap.
}
\author{
Ahmed El-Gabbas
}
