% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_SLURM.R
\name{Mod_SLURM}
\alias{Mod_SLURM}
\title{Prepare SLURM files for Hmsc-HPC model fitting}
\usage{
Mod_SLURM(
  ModelDir = NULL,
  JobName = NULL,
  CatJobInfo = TRUE,
  ntasks = 1L,
  CpusPerTask = 1L,
  GpusPerNode = 1L,
  MemPerCpu = NULL,
  Time = NULL,
  Partition = "small-g",
  EnvFile = ".env",
  FromHPC = TRUE,
  Path_Hmsc = NULL,
  Command_Prefix = "Commands2Fit",
  SLURM_Prefix = "Bash_Fit",
  Path_SLURM_Out = NULL
)
}
\arguments{
\item{ModelDir}{String. Path to the model files (without trailing slash).}

\item{JobName}{String. The name of the submitted job(s).}

\item{CatJobInfo}{Logical. Add bash lines to print information on the
submitted job. Default: \code{TRUE}.}

\item{ntasks}{Integer. The value for the \verb{#SBATCH --ntasks=} SLURM argument.
Default: 1.}

\item{CpusPerTask}{Integer. The value for the \verb{#SBATCH --cpus-per-task=}
SLURM argument. Default: 1.}

\item{GpusPerNode}{Integer. The value for the \verb{#SBATCH --gpus-per-node=}
SLURM argument. Default: 1.}

\item{MemPerCpu}{String. Memory per CPU allocation for the SLURM job.
Example: \verb{32G} for 32 gigabytes. Defaults to \code{NULL}. If not provided, the
function will throw an error.}

\item{Time}{String. Duration for which the job should run. Example:
\code{01:00:00} for one hour. If not provided, the function will throw an error.}

\item{Partition}{String. The name of the partition. Default: \code{small-g} for
running the array jobs on the GPU.}

\item{EnvFile}{String. Path to read the environment variables. Default value:
\code{.env}.}

\item{FromHPC}{Logical. Indicates if the operation is being performed from an
HPC environment. This adjusts file paths accordingly. Default: \code{TRUE}.}

\item{Path_Hmsc}{String. Path for the Hmsc-HPC.}

\item{Command_Prefix}{String. Prefix for the bash commands to be executed.
Default: \code{Commands2Fit}.}

\item{SLURM_Prefix}{String. Prefix for the exported SLURM file.}

\item{Path_SLURM_Out}{String indicating the directory where the SLURM file(s)
will be saved. Defaults to \code{NULL}, which means to identify the path from
\code{ModelDir}.}
}
\value{
The function does not return any value but writes SLURM script files
to the disk.
}
\description{
This function prepares and writes SLURM file(s) for submitting model fitting
jobs on an HPC environment. It dynamically generates bash scripts based on
the provided parameters, which are then used to submit jobs to the SLURM
workload manager.
}
\details{
The function reads the following environment variables:
\itemize{
\item \strong{\code{LUMI_Account_GPU}} for the BioDT LUMI project number.
\item \strong{\code{DP_R_Path_GPU_Check}} for the path of the python for reporting if
the GPU was used in the running SLURM job.
}
}
\author{
Ahmed El-Gabbas
}
