% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_SLURM_Refit.R
\name{Mod_SLURM_Refit}
\alias{Mod_SLURM_Refit}
\title{Prepare SLURM file for unfitted models on HPC}
\usage{
Mod_SLURM_Refit(
  Path_Model = NULL,
  MaxJobCounts = 210,
  JobName = NULL,
  MemPerCpu = NULL,
  Time = NULL,
  Partition = "small-g",
  Path_EnvFile = ".env",
  CatJobInfo = TRUE,
  ntasks = 1,
  CpusPerTask = 1,
  GpusPerNode = 1,
  CommandPrefix = "Commands_All",
  RefitPrefix = "Commands2Refit"
)
}
\arguments{
\item{Path_Model}{String. Path to the model files (without trailing slash)}

\item{MaxJobCounts}{Integer. Maximum number of jobs per slurm file}

\item{JobName}{String. The name of the submitted jobs}

\item{MemPerCpu}{String. The value for the \verb{#SBATCH --mem-per-cpu=} SLURM argument. Example: "32G" to request 32 gigabyte}

\item{Time}{String. The value for the requested time for each job in the bash arrays. Example: "01:00:00" to request an hour.}

\item{Partition}{String. The name of the partition. Default: \code{small-g}}

\item{Path_EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{CatJobInfo}{Logical. Add bash lines to report some job information. Default: \code{TRUE}}

\item{ntasks}{Integer. The value for the \verb{#SBATCH --ntasks=} SLURM argument. Default: 1}

\item{CpusPerTask}{Integer. The value for the \verb{#SBATCH --cpus-per-task=} SLURM argument. Default: 1}

\item{GpusPerNode}{Integer. The value for the \verb{#SBATCH --gpus-per-node=} SLURM argument. Default: 1}

\item{CommandPrefix}{String. The prefix for all model commands}

\item{RefitPrefix}{String. The prefix for commands to be re-fitted}
}
\description{
Prepare SLURM file for unfitted models on HPC
}
\author{
Ahmed El-Gabbas
}
