% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwf_naps_process.R, R/dwf_naps_distribution.R,
%   R/dwf_naps_plot.R, R/dwf_naps_standardisation.R
\name{naps_data}
\alias{naps_data}
\alias{naps_process}
\alias{naps_distribution}
\alias{naps_plot}
\alias{naps_standardisation}
\title{Process and map Naturalized Alien Plant Species (NAPS) data for the \code{IASDT}}
\usage{
naps_process(env_file = ".env", n_cores = 6L, strategy = "multisession")

naps_distribution(
  species = NULL,
  env_file = ".env",
  verbose = FALSE,
  dist_citizen = 100L
)

naps_plot(species = NULL, env_file = ".env")

naps_standardisation(env_file = ".env")
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{species}{Character. Species name for distribution mapping.}

\item{verbose}{Logical. If \code{TRUE}, prints progress messages. Default:
\code{FALSE}.}

\item{dist_citizen}{Numeric. Distance in km for spatial filtering of citizen
science data in GBIF or grid cells in countries at which species has not
yet recognized as "naturalized". Default is \code{100L} km.}
}
\description{
Processes and visualises Naturalized Alien Plant Species (NAPS) distribution
data from GBIF, EASIN, and eLTER for the Invasive Alien Species Digital Twin
(\code{IASDT}). Merges pre-processed data, creates presence-absence rasters,
summarises distributions, and generates maps using helper functions.
}
\section{Functions details}{

\itemize{
\item \strong{\code{naps_standardisation()}}: Load pre-prepared standardisation
information for NAPS verbatim names and prepare a unique \code{ias_id} for each
standardised NAPS. This function should be called \strong{only once} per
workflow version. The \code{ias_id} is determined based on sorting the
standardised species names alphabetically within their taxonomic hierarchy
(in the order of: class, order, family, and taxon_name). If a new
standardisation file is used, new \code{ias_id} values will be generated, which
will break reproducibility and prevent consistent data linkage across
analyses.
\item \strong{\code{naps_process()}}: Merges pre-processed GBIF (\link{gbif_process}), EASIN
(\link{easin_process}), and eLTER (\link{elter_process}) data (run these first).
Outputs \code{SpatRaster} distribution rasters, summary tables, and JPEG maps
using \code{naps_distribution()} and \code{naps_plot()}.
\item \strong{\code{naps_distribution()}}: Generates presence-absence maps (\code{.RData},
\code{.tif})  for a species, including all grid cells in the study area and a
set excluding cultivated/casual-only countries. Returns a file path to a
tibble with presence counts (total, by source) and summary statistics for
biogeographical regions
\item \strong{\code{naps_plot()}}: Creates JPEG distribution maps from GBIF, EASIN, and
eLTER data using \code{ggplot2}.
}
}

\author{
Ahmed El-Gabbas
}
