% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{NDecimals}
\alias{NDecimals}
\title{Number of decimal places in a vector}
\usage{
NDecimals(x)
}
\arguments{
\item{x}{vector}
}
\description{
Number of decimal places in a vector
}
\author{
Ahmed El-Gabbas
}
