% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_NormalizePath.R
\name{NormalizePath}
\alias{NormalizePath}
\title{Normalize and Quote File Paths}
\usage{
NormalizePath(Path, MustWork = FALSE)
}
\arguments{
\item{Path}{A character vector of file paths.}

\item{MustWork}{Logical; if \code{TRUE}, the function errors for non-existing
paths.}
}
\value{
A character vector of absolute, tidied, and shell-quoted paths.
}
\description{
This function ensures that file paths are expressed in a consistent and
canonical form. It first converts paths to absolute form using
\code{fs::path_abs()}, then tidies them with \code{fs::path_tidy()}, and finally quotes
them correctly based on the operating system. By default, \code{normalizePath()}
behaves differently on Windows and Linux when a file does not exist. On
Windows, it tries to construct an absolute path, while on Linux, it returns
the input path as-is (relative). To maintain consistency across platforms,
this function uses \code{fs::path_abs()} instead of \code{normalizePath()}.
}
\author{
Ahmed El-Gabbas
}
