% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{NUnique}
\alias{NUnique}
\title{Number of unique values for all columns of a data frame}
\usage{
NUnique(Data)
}
\arguments{
\item{Data}{input data}
}
\description{
Number of unique values for all columns of a data frame
}
\examples{
NUnique(mtcars)

NUnique(iris)
}
\references{
https://stackoverflow.com/questions/22196078/count-unique-values-for-every-column
}
