% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{Package_RemoteSha}
\alias{Package_RemoteSha}
\title{get remote sha of R packages}
\usage{
Package_RemoteSha(...)
}
\arguments{
\item{...}{name of one or more R packages}
}
\description{
get remote sha of R packages
}
\examples{
# Package_RemoteSha(IASDT.R, devtools)
}
