% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Path.R
\name{Path}
\alias{Path}
\title{Construct path to a file or directory}
\usage{
Path(..., ext = "")
}
\arguments{
\item{...}{character vectors, if any values are NA, the result will also be
NA. The paths follow the recycling rules used in the tibble package,
namely that only length 1 arguments are recycled.}

\item{ext}{An optional extension to append to the generated path.}
}
\value{
A character vector representing the constructed file path(s).
}
\description{
A wrapper around \code{fs::path()} that constructs file paths from input
components and returns them as a character string instead of an \code{"fs_path"}
object.
}
\examples{

# Basic usage
IASDT.R::Path("datasets", "processed", "model_fitting")

# Adding a file extension
IASDT.R::Path("results", "output", ext = "csv")

# Handling multiple components
IASDT.R::Path("home", "user", "documents", "report", ext = "pdf")

# Using vectorized input
IASDT.R::Path("folder", c("file1", "file2"), ext = "txt")
}
