% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Plot_Alpha.R
\name{Plot_Alpha}
\alias{Plot_Alpha}
\title{Plot convergence traceplots for the alpha parameter}
\usage{
Plot_Alpha(
  Post = NULL,
  Model = NULL,
  Title = NULL,
  NRC = NULL,
  AddFooter = TRUE,
  AddTitle = TRUE,
  Cols = c("red", "blue", "darkgreen", "darkgrey")
)
}
\arguments{
\item{Post}{Coda object or path to it}

\item{Model}{fitted model object or path to it}

\item{Title}{String. Plotting title}

\item{NRC}{Vector for the number of rows and columns per plot page}

\item{AddFooter}{Add footer to the plot for the page number}

\item{AddTitle}{Add main title to the plot}

\item{Cols}{Colours for lines for each chain}
}
\description{
Plot convergence traceplots for the alpha parameter
}
\author{
Ahmed El-Gabbas
}
