% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Plot_Convergence.R
\name{Plot_Convergence}
\alias{Plot_Convergence}
\title{Plot model convergence of multiple modelling alternatives}
\usage{
Plot_Convergence(
  Path_Model = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NChains = 4,
  maxOmega = 1000,
  NCores = NULL
)
}
\arguments{
\item{Path_Model}{String. Path to save all the output, including the to be fitted models (without trailing slash)}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{FromHPC}{Logical. Work from HPC? This is to adjust the file paths.}

\item{NChains}{Integer. Number of model chains}

\item{maxOmega}{Number of sample species interactions}

\item{NCores}{Number of parallel cores for parallelization}
}
\description{
Plot model convergence of multiple modelling alternatives
}
\author{
Ahmed El-Gabbas
}
