% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_evaluation.R
\name{plot_evaluation}
\alias{plot_evaluation}
\title{Generate plots for the explanatory power of Hmsc models}
\usage{
plot_evaluation(model_dir, env_file = ".env")
}
\arguments{
\item{model_dir}{Character. Path to the model directory containing
predictions.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}
}
\description{
This function generates four diagnostic plots (RMSE, AUC, Continuous Boyce
Index, and Tjur-R²) to evaluate the performance (explanatory power without
cross-validation) of Hmsc models.
}
\author{
Ahmed El-Gabbas
}
