% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_gelman.R
\name{plot_gelman}
\alias{plot_gelman}
\alias{plot_gelman_alpha}
\alias{plot_gelman_beta}
\alias{plot_gelman_omega}
\alias{plot_gelman_rho}
\title{Plot Gelman-Rubin-Brooks}
\usage{
plot_gelman(
  path_coda = NULL,
  alpha = TRUE,
  beta = TRUE,
  omega = TRUE,
  rho = TRUE,
  n_omega = 1000L,
  plotting_alpha = 0.25,
  env_file = ".env"
)

plot_gelman_alpha(coda_object, plotting_alpha = 0.25)

plot_gelman_beta(coda_object, env_file = ".env", plotting_alpha = 0.25)

plot_gelman_omega(coda_object, n_omega = 1000L, plotting_alpha = 0.25)

plot_gelman_rho(coda_object)
}
\arguments{
\item{path_coda}{Character or \code{mcmc.list}. Path to a file containing the coda
object, or an \code{mcmc.list} object representing MCMC samples.}

\item{alpha, beta, omega, rho}{Logical. If \code{TRUE}, plots the Gelman-Rubin
statistic for the respective model parameters (alpha, beta, omega, or rho).
Default: \code{TRUE} for all parameters.}

\item{n_omega}{Integer. Number of species sampled for the omega parameter.
Default: 1000L.}

\item{plotting_alpha}{Numeric. Transparency level (alpha) for plot lines (0 =
fully transparent, 1 = fully opaque). Default: 0.25.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{coda_object}{\code{mcmc.list}. An MCMC sample object containing posterior
distributions from an Hmsc model.}
}
\description{
The \verb{PlotGelman_*()} functions generate plots visualising the evolution of
the Gelman-Rubin-Brooks shrink factor for different model parameters as the
number of iterations increases. These plots help assess whether MCMC chains
have converged to a common distribution. Each plot includes: median (solid
line) and 97.5\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} percentile (dashed line) of the shrink factor and
a dashed horizontal line at 1.1, representing the common convergence
threshold. The primary function for users is \code{plot_gelman()}, which
internally calls:
\itemize{
\item \code{plot_gelman_alpha()}: Plots shrink factor for the \strong{Alpha} parameter
\item \code{plot_gelman_beta()}: Plots shrink factor for the \strong{Beta} parameters
\item \code{plot_gelman_omega()}: Plots shrink factor for the \strong{Omega} parameter
\item \code{plot_gelman_rho()}: Plots shrink factor for the \strong{Rho} parameter
}
}
\author{
Ahmed El-Gabbas
}
