% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_predict_latent_factor.R
\name{plot_latent_factor}
\alias{plot_latent_factor}
\title{Plot spatial variation in site loadings of HMSC models}
\usage{
plot_latent_factor(path_model = NULL, env_file = ".env")
}
\arguments{
\item{path_model}{Character. Path to fitted \code{Hmsc} model object.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}
}
\description{
Generate and save spatial variation in site loadings of HMSC models' latent
factors as a JPEG file.
}
\author{
Ahmed El-Gabbas
}
