% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Plot_RespCurv.R
\name{Plot_RespCurv}
\alias{Plot_RespCurv}
\title{Plot response curves}
\usage{
Plot_RespCurv(
  Model,
  ResCurvePath = getwd(),
  ResCurvePrefix = "ResCurv_",
  Verbose = TRUE,
  NCores = NULL
)
}
\arguments{
\item{Model}{Model}

\item{ResCurvePath}{ResCurvePath}

\item{ResCurvePrefix}{ResCurvePrefix}

\item{Verbose}{Verbose}

\item{NCores}{Number of cores to use}
}
\description{
Plot response curves
}
\author{
Ahmed El-Gabbas
}
