% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotBeta_bySp.R
\name{PlotBeta_bySp}
\alias{PlotBeta_bySp}
\title{PlotBeta_bySp}
\usage{
PlotBeta_bySp(
  Beta,
  NRC = c(3, 2),
  PdfPath = getwd(),
  FilePrefix = "Posterior_Sp",
  Verbose = TRUE
)
}
\arguments{
\item{Beta}{Beta}

\item{NRC}{NRC}

\item{PdfPath}{PdfPath}

\item{FilePrefix}{FilePrefix}

\item{Verbose}{Verbose}
}
\description{
PlotBeta_bySp
}
\author{
Ahmed El-Gabbas
}
