% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotBeta_byVar.R
\name{PlotBeta_byVar}
\alias{PlotBeta_byVar}
\title{PlotBeta_byVar}
\usage{
PlotBeta_byVar(
  Beta,
  NRC = c(3, 2),
  PdfPath = getwd(),
  FilePrefix = "Posterior_Sp",
  Verbose = TRUE,
  VerboseSp = FALSE
)
}
\arguments{
\item{Beta}{Beta}

\item{NRC}{NRC}

\item{PdfPath}{PdfPath}

\item{FilePrefix}{FilePrefix}

\item{Verbose}{Verbose}

\item{VerboseSp}{VerboseSp}
}
\description{
PlotBeta_byVar
}
\author{
Ahmed El-Gabbas
}
