% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotBetaGG.R
\name{PlotBetaGG}
\alias{PlotBetaGG}
\title{Heatmaps of parameter estimates or posterior support values of species' environmental responses (Beta parameters), i.e. how species in Y responds to covariates in X}
\usage{
PlotBetaGG(
  Path_Model = NULL,
  supportLevel = 0.95,
  PlotWidth = 18,
  PlotHeight = 20
)
}
\arguments{
\item{Path_Model}{String. Path to .RData file for selected model}

\item{supportLevel}{controls threshold posterior support for plotting. For more information, see \code{?Hmsc::plotBeta()}}

\item{PlotWidth, PlotHeight}{Integer. Plot size in cm}
}
\description{
Heatmaps of parameter estimates or posterior support values of species' environmental responses (Beta parameters), i.e. how species in Y responds to covariates in X. This function contain parts of the \code{?Hmsc::plotBeta()} function, with the main difference that this function uses \code{ggplot2} for plotting.
}
