% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotConvergence_AllModels.R
\name{PlotConvergence_AllModels}
\alias{PlotConvergence_AllModels}
\title{Plot model convergence of multiple modelling alternatives}
\usage{
PlotConvergence_AllModels(
  Path_Model = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NChains = 4,
  maxOmega = 1000,
  NCores = NULL
)
}
\arguments{
\item{Path_Model}{String. Path to save all the output, including the to be fitted models (without trailing slash). Must not be \code{NULL}.}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}.}

\item{FromHPC}{Logical. Indicates whether the function is being run on a High-Performance Computing (HPC) environment. Adjusts file paths accordingly.}

\item{NChains}{Integer. Number of MCMC chains used in the model fitting process.}

\item{maxOmega}{Integer. Maximum number of species interactions to sample for convergence diagnostics.}

\item{NCores}{Integer. Number of cores to use for parallel processing.}
}
\value{
The function does not return anything but saves a series of diagnostic plots in the specified path.
}
\description{
This function generates and saves a series of diagnostic plots to assess the convergence of Hmsc models across multiple modelling alternatives. It checks model convergence using trace plots and Gelman-Rubin diagnostics for key model parameters.
}
\author{
Ahmed El-Gabbas
}
