% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman_Alpha.R
\name{PlotGelman_Alpha}
\alias{PlotGelman_Alpha}
\title{Gelman-Rubin-Brooks plot \code{alpha} parameters}
\usage{
PlotGelman_Alpha(CodaObj = NULL, NCores = NULL, PlottingAlpha = 0.25)
}
\arguments{
\item{CodaObj}{an mcmc object}

\item{NCores}{Integer. Number of parallel processes.}

\item{PlottingAlpha}{Double. Plotting alpha for line transparency}
}
\description{
Gelman-Rubin-Brooks plot \code{alpha} parameters
}
\author{
Ahmed El-Gabbas
}
