% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman_Beta.R
\name{PlotGelman_Beta}
\alias{PlotGelman_Beta}
\title{Gelman-Rubin-Brooks plot for all beta parameters}
\usage{
PlotGelman_Beta(
  CodaObj = NULL,
  NCores = NULL,
  EnvFile = ".env",
  PlottingAlpha = 0.25
)
}
\arguments{
\item{CodaObj}{an mcmc object}

\item{NCores}{Integer. Number of parallel processes.}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{PlottingAlpha}{Double. Plotting alpha for line transparency}
}
\description{
Gelman-Rubin-Brooks plot for all beta parameters
}
\author{
Ahmed El-Gabbas
}
