% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman_Rho.R
\name{PlotGelman_Rho}
\alias{PlotGelman_Rho}
\title{Creates a Gelman-Rubin-Brooks plot for the \code{rho} parameter.}
\usage{
PlotGelman_Rho(CodaObj)
}
\arguments{
\item{CodaObj}{An object of class \code{mcmc.list}, representing the Hmsc samples.}
}
\value{
A ggplot object representing the Gelman-Rubin-Brooks plot for the
\code{rho} parameter.
}
\description{
This function generates a Gelman-Rubin-Brooks plot for the \code{rho} parameter of
a coda object. The plot includes lines for the median and the 97.5th
percentile of the shrink factor, with a dashed line at 1.1 indicating the
threshold for convergence. This function is not planned to be used in
isolation but rather within \link{PlotGelman}.
}
\author{
Ahmed El-Gabbas
}
