% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotOmegaGG.R
\name{PlotOmegaGG}
\alias{PlotOmegaGG}
\title{Heatmaps of parameter estimates or posterior support values of species' residual association (Omega parameters)}
\usage{
PlotOmegaGG(
  Path_Model = NULL,
  supportLevel = 0.95,
  PlotWidth = 22,
  PlotHeight = 20
)
}
\arguments{
\item{Path_Model}{String. Path to .RData file for selected model}

\item{supportLevel}{controls threshold posterior support for plotting.}

\item{PlotWidth, PlotHeight}{Integer. Plot size in cm}
}
\description{
Heatmaps of parameter estimates or posterior support values of species' residual association (Omega parameters)
}
