% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotRespCurv.R
\name{PlotRespCurv}
\alias{PlotRespCurv}
\title{PlotRespCurv}
\usage{
PlotRespCurv(
  Model,
  ResCurvePath = getwd(),
  ResCurvePrefix = "ResCurv_",
  Verbose = TRUE
)
}
\arguments{
\item{Model}{Model}

\item{ResCurvePath}{ResCurvePath}

\item{ResCurvePrefix}{ResCurvePrefix}

\item{Verbose}{Verbose}
}
\description{
PlotRespCurv
}
\author{
Ahmed El-Gabbas
}
