% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotRho.R
\name{PlotRho}
\alias{PlotRho}
\title{Plot convergence of the rho parameter}
\usage{
PlotRho(Post, Model, Title, Cols = c("red", "blue", "darkgreen", "darkgrey"))
}
\arguments{
\item{Post}{String. Path to Coda object}

\item{Model}{String. Path to the fitted model}

\item{Title}{String. Plotting title}

\item{Cols}{Colours for lines for each chain}
}
\description{
Plot convergence of the rho parameter
}
\author{
Ahmed El-Gabbas
}
