% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotVarPar.R
\name{PlotVarPar}
\alias{PlotVarPar}
\title{Plot variance partitioning of Hmsc model using ggplot2.}
\usage{
PlotVarPar(
  Model,
  Path_Plot,
  PlotTitlePrefix = NULL,
  ModelEval = NULL,
  NCores,
  VarPar = NULL,
  EnvFile = ".env",
  SaveVarPar = TRUE,
  SaveModelEval = TRUE,
  ReturnGG = FALSE,
  SaveGG = TRUE
)
}
\arguments{
\item{Model}{an object of class \code{Hmsc} or a path for the model file. Only needed if one of \code{ModelEval} and \code{VarPar} arguments are \code{NULL}.}

\item{Path_Plot}{String. Path where the output file(s) will be saved.}

\item{PlotTitlePrefix}{String (optional). Prefix to add to the title of the plot. Default: \code{NULL}, which means only 'Variance partitioning' will be used in the title.}

\item{ModelEval}{Result of the \code{Hmsc::evaluateModelFit} function. If \code{ModelEval = NULL} (default), \code{Hmsc::evaluateModelFit} will be executed on the model object to compute measures of model fit.}

\item{NCores}{Integer. Number of parallel computations for computing predicted values. This is used as the \code{nParallel} argument of the \code{Hmsc::computePredictedValues} function.}

\item{VarPar}{Variance partitioning object resulted from \code{Hmsc::computeVariancePartitioning}. If \code{VarPar = NULL} (default), \code{Hmsc::computeVariancePartitioning} will be executed on the model object.}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{SaveVarPar}{Logical. If \code{VarPar = NULL}, should the calculated variance partitioning be saved as an RData file?  Default: \code{TRUE}.}

\item{SaveModelEval}{Logical. If \code{ModelEval = NULL}, should the calculated model evaluation be saved as an RData file?  Default: \code{TRUE}.}

\item{ReturnGG}{Logical. Return the ggplot object. Default: \code{FALSE}.}

\item{SaveGG}{Logical. Should the plots be exported as an RData object?  Default: \code{TRUE}.}
}
\description{
This function generates and optionally saves plots visualizing the variance partitioning of a Hmsc model. It can calculate variance partitioning and model evaluation if not provided and supports parallel computation for model predictions.
}
\author{
Ahmed El-Gabbas
}
