% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_Polygon_Centroid.R
\name{Polygon_Centroid}
\alias{Polygon_Centroid}
\title{Replace the geometry of a polygon with its centroid point}
\usage{
Polygon_Centroid(x, Rename = FALSE, NewName = "")
}
\arguments{
\item{x}{simple feature object}

\item{Rename}{should the geometry field renamed}

\item{NewName}{the new name of geometry}
}
\description{
Replace the geometry of a polygon with its centroid point
}
\author{
Ahmed El-Gabbas
}
