% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Pred2Array.R
\name{Pred2Array}
\alias{Pred2Array}
\title{Convert results from predict function to array}
\usage{
Pred2Array(
  Predict = TRUE,
  Model = NULL,
  Preds = NULL,
  expected = TRUE,
  NCores = NULL
)
}
\arguments{
\item{Predict}{Logical. Make predictions first}

\item{Model}{Hmsc object or path of the saved model}

\item{Preds}{Pre-computed predictions}

\item{expected}{Logical. Whether to return the location parameter of the observation models or sample the values from those.}

\item{NCores}{Integer. Number of parallel computations for computing predicted values.}
}
\description{
Convert results from predict function to array, with possibility to make predictions first
}
\author{
Ahmed El-Gabbas
}
