% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_predict_hmsc.R
\name{predict_hmsc}
\alias{predict_hmsc}
\title{Calculates predicted values from a fitted Hmsc model}
\usage{
predict_hmsc(
  path_model,
  Loff = NULL,
  x_data = NULL,
  X = NULL,
  XRRRData = NULL,
  XRRR = NULL,
  gradient = NULL,
  Yc = NULL,
  mcmcStep = 1L,
  expected = TRUE,
  n_cores = 8L,
  strategy = "multisession",
  future_max_size = 1000L,
  model_name = "train",
  temp_dir = "temp_pred",
  temp_cleanup = TRUE,
  prediction_type = NULL,
  use_tf = TRUE,
  tf_environ = NULL,
  tf_use_single = FALSE,
  lf_out_file = NULL,
  lf_return = FALSE,
  lf_input_file = NULL,
  lf_only = FALSE,
  n_cores_lf = n_cores,
  lf_check = FALSE,
  lf_temp_cleanup = TRUE,
  lf_commands_only = FALSE,
  pred_directory = NULL,
  pred_pa = NULL,
  pred_xy = NULL,
  evaluate = FALSE,
  evaluation_name = NULL,
  evaluation_directory = "evaluation",
  verbose = TRUE,
  spatial_model = TRUE
)
}
\arguments{
\item{path_model}{Character. Path to the saved model object.}

\item{Loff}{See \link[Hmsc:predict.Hmsc]{Hmsc::predict.Hmsc} for more details.}

\item{x_data}{\code{data.frame}. The unpreprocessed covariates for the predictions
to be made. Works only if the \code{XFormula} argument was specified in the
\link[Hmsc:Hmsc]{Hmsc::Hmsc} model constructor call. Requirements are similar to those in
the \code{Hmsc} model constructor.}

\item{X}{\code{matrix}. Covariates for the predictions to be made. Only one of
\code{x_data} and \code{X} arguments may be provided.}

\item{XRRRData}{\code{data.frame}. Covariates for reduced-rank regression.}

\item{XRRR}{\code{matrix}. Covariates for reduced-rank regression.}

\item{gradient}{An object returned by \link[Hmsc:constructGradient]{Hmsc::constructGradient}. Providing
\code{gradient} is an alternative for providing \code{x_data}, \code{studyDesign} and
\code{ranLevels}. Cannot be used together with \code{Yc}.}

\item{Yc}{\code{matrix}. Outcomes that are assumed to be known for conditional
predictions. Cannot be used together with \code{gradient}.}

\item{mcmcStep}{Integer. Number of extra \code{mcmc} steps used for updating the
random effects. Defaults to 1L.}

\item{expected}{Logical. Whether to return the location parameter of the
observation models or sample the values from those. Defaults to \code{TRUE}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{future_max_size}{Numeric. Maximum allowed total size (in megabytes) of
global variables identified. See \code{future.globals.maxSize} argument of
\link[future:zzz-future.options]{future::future.options} for more details.}

\item{model_name}{Character. Prefix for temporary file names. Defaults to
\code{NULL}, in which case no prefix is used.}

\item{temp_dir}{Character. Path for temporary storage of intermediate files.}

\item{temp_cleanup}{Logical. Whether to clean up temporary files. Defaults to
\code{TRUE}.}

\item{prediction_type}{Character. Type of predictions to be made. If \code{NULL}
(default), predictions are made for the latent factors. If \code{c}, predictions
are made for response curves at mean coordinates. If \code{i}, predictions are
made for response curves at infinite coordinates.}

\item{use_tf}{Logical. Whether to use \code{TensorFlow} for calculations. Defaults
to \code{TRUE}.}

\item{tf_environ}{Character. Path to the Python environment. This argument is
required if \code{use_tf} is \code{TRUE} under Windows. Defaults to \code{NULL}.}

\item{tf_use_single}{Logical. Whether to use single precision for the
\code{TensorFlow} calculations. Defaults to \code{FALSE}.}

\item{lf_out_file}{Character. Path to save the outputs. If \code{NULL} (default),
the predicted latent factors are not saved to a file. This should end with
either \verb{*.qs2} or \verb{*.RData}.}

\item{lf_return}{Logical. Whether the output should be returned. Defaults to
\code{FALSE}. If \code{lf_out_file} is \code{NULL}, this parameter cannot be set to
\code{FALSE} because the function needs to return the result if it is not saved
to a file.}

\item{lf_input_file}{Character. File name where the latent factor predictions
are saved. If \code{NULL} (default), latent factor predictions will be computed.
If specified, latent factor predictions are read from this path. This
allows to predicting the latent factors for new sites only once.}

\item{lf_only}{Logical. Whether to return the latent factor predictions only.
Defaults to \code{FALSE}. This helps in predicting to new sites, allowing to
predicting the latent factors only once, then the output can be loaded in
other predictions when needed.}

\item{n_cores_lf}{Integer. Number of cores to use for parallel processing of
latent factor prediction. Defaults to 8L.}

\item{lf_check}{Logical. If \code{TRUE}, the function checks if the output files
are already created and valid. If \code{FALSE}, the function will only check if
the files exist without checking their integrity. Default is \code{FALSE}.}

\item{lf_temp_cleanup}{Logical. Whether to delete temporary files in the
\code{temp_dir} directory after finishing the LF predictions.}

\item{lf_commands_only}{Logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{pred_directory}{Character. Directory path indicating where the
predictions will be saved. Defaults to \code{NULL}, which saves model
predictions to "\code{model_prediction}" folder of the current working
directory.}

\item{pred_pa}{\code{matrix}. Presence-absence data for evaluation. If \code{NULL}
(default), the presence-absence data from the model object is used. This
argument is used only when \code{evaluate} is \code{TRUE}.}

\item{pred_xy}{\code{matrix}. Coordinates to be added to predicted values. If
\code{NULL} (default), the coordinates from the model object is used.}

\item{evaluate}{Logical. Whether to evaluate the model predictions. Defaults
to \code{FALSE}.}

\item{evaluation_name}{Character. Name of the evaluation results. If \code{NULL},
the default name is used (\verb{eval_[model_name].qs2}).}

\item{evaluation_directory}{Character. Directory where the evaluation results
will be saved. Defaults to \code{evaluation}.}

\item{verbose}{Logical. Whether to print a message upon successful saving of
files. Defaults to \code{FALSE}.}

\item{spatial_model}{Logical. Whether the fitted model is a spatial model.
Defaults to \code{TRUE}.}
}
\description{
This function modifies the \code{Hmsc:::predict.Hmsc} function.
}
