% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_LF.R
\name{Predict_LF}
\alias{Predict_LF}
\title{Draws samples from the conditional predictive distribution of latent factors.}
\usage{
Predict_LF(
  unitsPred,
  modelunits,
  postEta,
  postAlpha,
  LF_rL,
  LF_NCores = 8L,
  Temp_Dir = "TEMP_Pred",
  LF_Temp_Cleanup = TRUE,
  Model_Name = NULL,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  LF_OutFile = NULL,
  LF_Return = FALSE,
  LF_Check = FALSE,
  LF_Commands_Only = FALSE,
  solve_max_attempts = 5L,
  solve_chunk_size = 50L,
  Verbose = TRUE
)
}
\arguments{
\item{unitsPred}{a factor vector with random level units for which
predictions are to be made}

\item{modelunits}{a factor vector with random level units that are
conditioned on}

\item{postEta}{Character string specifying the path for postEta; a list
containing samples of random factors at conditioned units}

\item{postAlpha}{a list containing samples of range (lengthscale) parameters
for latent factors}

\item{LF_rL}{a HmscRandomLevel-class object that describes the random level
structure}

\item{LF_NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}

\item{Temp_Dir}{Character string specifying the path for temporary storage of
intermediate files.}

\item{LF_Temp_Cleanup}{Logical indicating whether to delete temporary files
in the \code{Temp_Dir} after finishing the LF predictions.}

\item{Model_Name}{Character string used as a prefix for temporary file names.
Defaults to NULL, in which case no prefix is used.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{TF_use_single}{Logical indicating whether to use single precision for
the TF calculations. Defaults to \code{FALSE}.}

\item{LF_OutFile}{Character string specifying the path to save the outputs.
If \code{NULL} (default), the predicted latent factors are not saved to a file.
This should end with either \verb{*.qs2} or \verb{*.RData}.}

\item{LF_Return}{Logical. Indicates if the output should be returned.
Defaults to \code{FALSE}. If \code{LF_OutFile} is \code{NULL}, this parameter cannot be
set to \code{FALSE} because the function needs to return the result if it is not
saved to a file.}

\item{LF_Check}{Logical. If TRUE, the function checks if the output files are
already created and valid. If FALSE, the function will only check if the
files exist without checking their integrity. Default is \code{FALSE}.}

\item{LF_Commands_Only}{logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{solve_max_attempts}{numeric (Optional). Maximum number of attempts to
run solve and crossprod functions. Default is 5.}

\item{solve_chunk_size}{numeric. Chunk size for solve_and_multiply python
function. Default is 50.}

\item{Verbose}{Logical. If TRUE, detailed output is printed. Default is
\code{FALSE}.}
}
\description{
This function is optimized for speed using parallel processing and optionally
TensorFlow for matrix operations. This function is adapted from
\link[Hmsc:predictLatentFactor]{Hmsc::predictLatentFactor} with equivalent results to the original function
when \code{predictMean = TRUE}.
}
\details{
The function is expected to be faster than the original function in
the \code{Hmsc} package, especially when using TensorFlow for calculations and
when working on parallel.

The main difference is that this function:
\itemize{
\item allow for parallel processing (\code{LF_NCores} argument);
\item it is possible to use TensorFlow (\code{UseTF} argument) to make matrix
calculations faster, particularly when used on GPU. The following modules
are needed: \code{numpy}, \code{os}, \code{tensorflow}, \code{rdata}, \code{xarray}, and \code{pandas}.
To use \code{TensorFlow}, the argument \code{TF_Environ} should be set to the path of
a Python environment with TensorFlow installed;
\item if \code{UseTF} is set to \code{FALSE}, the function uses R / CPP code in the
calculations;
\item calculates \code{D11} and \code{D12} matrices only once and save them to disk and
call them when needed.
}
}
\seealso{
\link[Hmsc:predictLatentFactor]{Hmsc::predictLatentFactor}
}
\author{
This script was adapted from the \link[Hmsc:predictLatentFactor]{Hmsc::predictLatentFactor} function
in the \code{Hmsc} package.
}
