% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PredictMaps.R
\name{Predict_Maps}
\alias{Predict_Maps}
\title{Predicts habitat suitability of \code{Hmsc} model at different climate scenarios}
\usage{
Predict_Maps(
  Path_Model = NULL,
  Hab_Abb = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NCores = 8,
  ChunkSize = 500,
  RemoveChunks = TRUE,
  TF_Environ = NULL,
  UseTF = TRUE
)
}
\arguments{
\item{Path_Model}{Character. Path to fitted \code{Hmsc} model object.}

\item{Hab_Abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type to
prepare data for. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a}, \verb{4b}, \code{10},
\verb{12a}, \verb{12b}. For more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{NCores}{Integer specifying the number of parallel cores for
parallelization. Default: 8 cores.}

\item{ChunkSize}{Integer. The size of chunks used to split the input raster
for parallel processing. Predictor rasters are converted into data frame
(tibble) then split into small chunks for processing on parallel. Default:
\code{500}.}

\item{RemoveChunks}{Logical. Whether to remove intermediate chunk files after
prediction.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}
}
\value{
Returns a tibble with the prediction summary and file paths for
output tif files.
}
\description{
This function generates prediction maps of \code{Hmsc} models under current and
future climates. It also predicts an ensemble predictions for different
climate models used. For each species and for species richness, the function
exports three maps representing the mean, standard deviation (sd), and
coefficient of variation (cov).
}
\author{
Ahmed El-Gabbas
}
