% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_predict_maps_CV.R
\name{predict_maps_CV}
\alias{predict_maps_CV}
\title{Predict and evaluate cross-validated \code{Hmsc} models}
\usage{
predict_maps_CV(
  model_dir = NULL,
  CV_name = NULL,
  CV_fold = NULL,
  n_cores = 8L,
  use_TF = TRUE,
  TF_environ = NULL,
  TF_use_single = FALSE,
  LF_n_cores = n_cores,
  LF_check = FALSE,
  LF_temp_cleanup = TRUE,
  LF_only = FALSE,
  LF_commands_only = FALSE,
  temp_cleanup = TRUE
)
}
\arguments{
\item{model_dir}{Character. Path to the root directory where the
cross-validated models were fitted.}

\item{CV_name}{Character. Name of the cross-validation strategy used. Valid
values are \code{CV_Dist}, \code{CV_Large}, or \code{CV_SAC}.}

\item{CV_fold}{Integer. The cross-validation fold number.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{use_TF}{Logical. Whether to use TensorFlow for calculations. Defaults
to \code{TRUE}.}

\item{TF_environ}{Character. Path to the Python environment. This argument is
required if \code{use_TF} is \code{TRUE} under Windows. Defaults to \code{NULL}.}

\item{TF_use_single}{Logical. Whether to use single precision for the
TensorFlow calculations. Defaults to \code{FALSE}.}

\item{LF_n_cores}{Integer. Number of cores to use for parallel processing of
latent factor prediction. Defaults to 8L.}

\item{LF_check}{Logical. If \code{TRUE}, the function checks if the output files
are already created and valid. If \code{FALSE}, the function will only check if
the files exist without checking their integrity. Default is \code{FALSE}.}

\item{LF_temp_cleanup}{Logical. Whether to delete temporary files in the
\code{temp_dir} directory after finishing the LF predictions.}

\item{LF_only}{Logical. Whether to return the latent factor predictions only.
Defaults to \code{FALSE}. This helps in predicting to new sites, allowing to
predicting the latent factors only once, then the output can be loaded in
other predictions when needed.}

\item{LF_commands_only}{Logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{temp_cleanup}{Logical. Whether to clean up temporary files. Defaults to
\code{TRUE}.}
}
\value{
A tibble containing the prediction summary and file paths for output
\verb{*.tif} files.
}
\description{
This function computes predicted values of cross-validated \code{Hmsc} models in
the testing cross-validation folders and evaluates the model performance
using different metrics. In contrast to \link{predict_maps}, this function
predicts only under current climate conditions and does not use clamping.
}
\seealso{
\link{predict_maps}
}
\author{
Ahmed El-Gabbas
}
