% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_predictHmsc.R
\name{predictHmsc}
\alias{predictHmsc}
\title{predictHmsc}
\usage{
predictHmsc(
  object,
  Loff = NULL,
  XData = NULL,
  X = NULL,
  XRRRData = NULL,
  XRRR = NULL,
  Gradient = NULL,
  Yc = NULL,
  mcmcStep = 1,
  expected = TRUE,
  NCores = 1,
  TempDir = "TEMP2Pred",
  PredDir = "Model_Predictions",
  Evaluate = FALSE,
  Evaluate_Name = NULL,
  EvalDir = "Eval",
  ModelName = "Train",
  RC = NULL,
  Pred_PA = NULL,
  Pred_XY = NULL,
  UseTF = TRUE,
  PythonScript = NULL,
  EnvPath = NULL,
  use_single = FALSE,
  Path_postEtaPred = NULL,
  Return_postEtaPred = TRUE,
  File_LF = NULL,
  nthreads = 5,
  ...
)
}
\arguments{
\item{object}{a fitted \code{Hmsc} model object}

\item{XData}{a dataframe specifying the unpreprocessed covariates for the
predictions to be made. Works only if the \code{XFormula} argument was
specified in the \code{Hmsc()} model constructor call. Requirements are
similar to those in the \code{Hmsc} model constructor.}

\item{X}{a matrix specifying the covariates for the predictions to be made.
Only one of XData and X arguments may be provided.}

\item{XRRRData}{a dataframe of covariates for reduced-rank regression}

\item{XRRR}{a matrix of covariates for reduced-rank regression}

\item{Gradient}{an object returned by \code{\link{constructGradient}}.
Providing \code{Gradient} is an alternative for providing \code{XData},
\code{studyDesign} and \code{ranLevels}. Cannot be used together with
\code{Yc}.}

\item{Yc}{a matrix of the outcomes that are assumed to be known for
conditional predictions. Cannot be used together with \code{Gradient}.}

\item{mcmcStep}{the number of extra mcmc steps used for updating the random
effects}

\item{expected}{boolean flag indicating whether to return the location
parameter of the observation models or sample the values from those.}

\item{NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}

\item{TempDir}{Character string specifying the path for temporary storage of
intermediate files.}

\item{ModelName}{Character string used as a prefix for temporary file names.
Defaults to NULL, in which case no prefix is used.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{PythonScript}{Character string specifying the path to the Python script
file containing the \code{crossprod_solve} function. Defaults to the path of the
file included in the package.}

\item{EnvPath}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{use_single}{Logical indicating whether to use single precision for the
TF calculations. Defaults to \code{FALSE}.}

\item{Path_postEtaPred}{Character string specifying the path to save the
outputs. If \code{NULL} (default), the predicted latent factors are not saved to
a file.}

\item{Return_postEtaPred}{Logical. Indicates if the output should be
returned. Defaults to \code{TRUE}. If \code{Path_postEtaPred} is \code{NULL}, this
parameter cannot be set to \code{FALSE} because the function needs to return the
result if it is not saved to a file.}

\item{nthreads}{Number of threads to use when reading \code{qs} files. Default 5;
see \link[qs:qread]{qs::qread}.}

\item{\dots}{other arguments passed to functions.}
}
\description{
Calculates predicted values from a fitted \code{Hmsc} model. This function
modifies the \code{Hmsc:::predict.Hmsc} function.
}
