% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PredictLatentFactor.R
\name{predictLatentFactor}
\alias{predictLatentFactor}
\title{predictLatentFactor}
\usage{
predictLatentFactor(
  unitsPred,
  modelunits,
  postEta,
  postAlpha,
  rL,
  NCores = 8,
  TempDir = "TEMP2Pred",
  ModelName = NULL
)
}
\arguments{
\item{unitsPred}{a factor vector with random level units for which
predictions are to be made}

\item{modelunits}{a factor vector with random level units that are
conditioned on}

\item{postEta}{a list containing samples of random factors at conditioned
units}

\item{postAlpha}{a list containing samples of range (lengthscale) parameters
for latent factors}

\item{rL}{a HmscRandomLevel-class object that describes the random level
structure}

\item{NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}

\item{TempDir}{Path for temporary directory.}

\item{ModelName}{Prefix}
}
\description{
Draws samples from the conditional predictive distribution of latent factors.
This function is adapted from \link[Hmsc:predictLatentFactor]{Hmsc::predictLatentFactor}.
}
