% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prepare_knots.R
\name{prepare_knots}
\alias{prepare_knots}
\title{Prepare knot locations for Hmsc GPP models}
\usage{
prepare_knots(
  coordinates = NULL,
  min_distance = NULL,
  jitter_distance = 100,
  min_lf = NULL,
  max_lf = NULL,
  alphapw = list(Prior = NULL, Min = 10, Max = 1500, Samples = 101)
)
}
\arguments{
\item{coordinates}{Numeric matrix or data frame containing the (x, y)
coordinates of sampling units.}

\item{min_distance}{Numeric. Minimum distance between knots in meters. This
distance is used for both \code{knotDist} and \code{minKnotDist} parameters of the
\link[Hmsc:constructKnots]{Hmsc::constructKnots} function.}

\item{jitter_distance}{Numeric. The jitter distance applied to overlapping
coordinates to avoid exact duplicates. Defaults to 100 meters.}

\item{min_lf, max_lf}{Integer. Minimum and maximum number of latent factors to
be used. Both default to \code{NULL} which means that the number of latent
factors will be estimated from the data. If either is provided, the
respective values will be used as arguments to \link[Hmsc:setPriors]{Hmsc::setPriors}.}

\item{alphapw}{Prior for the alpha parameter. Defaults to a list with \code{Prior = NULL}, \code{Min = 10}, \code{Max = 1500}, and \code{Samples = 101}. If \code{alphapw} is
\code{NULL} or a list with all \code{NULL} list items, the default prior will be
used. If \code{Prior} is a matrix, it will be used as the prior. If \code{Prior = NULL}, the prior will be generated using the minimum and maximum values of
the alpha parameter (\code{min} and \code{max}, respectively; in kilometre) and  the
number of samples (\code{Samples}). Defaults to a prior with 101 samples ranging
from 10 to 1500 km, with the first value in the second column set to 0.5.}
}
\value{
An object of class \code{HmscRandomLevel}, suitable for specifying the
random level in HMSC GPP models. This object contains the prepared knot
locations as a data frame with columns \code{var_1} and \code{var_2} (numeric
coordinates), after possible jittering and conversion to avoid overlap.
}
\description{
Prepare the locations of knots for use in Gaussian Predictive Process (GPP)
models within the HMSC framework. It ensures that knots are spaced at a
minimum specified distance and applies jitter to any identical coordinates to
avoid overlap.
}
\author{
Ahmed El-Gabbas
}
