% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepKnots.R
\name{PrepKnots}
\alias{PrepKnots}
\title{Prepare knot locations for Hmsc GPP models}
\usage{
PrepKnots(Coords = NULL, MinDist = NULL, JitterDist = 100)
}
\arguments{
\item{Coords}{A matrix or data frame containing the coordinates (x, y) of the
sampling units.}

\item{MinDist}{A numeric value specifying the minimum distance (in the same
units as Coords; here in meters) between knots. This distance is used for
both \code{knotDist} and \code{minKnotDist} parameters of the \link[Hmsc:constructKnots]{Hmsc::constructKnots}
function.}

\item{JitterDist}{A numeric value specifying the distance (in the same units
as Coords; here in meters) for jittering identical coordinates. Defaults to
100.}
}
\value{
An object suitable for specifying the random level in HMSC GPP
models. This object contains the prepared knot locations.
}
\description{
This function prepares the locations of knots for use in GPP models within
the HMSC framework. It ensures that knots are spaced at a minimum specified
distance and applies jitter to any identical coordinates to avoid overlap.
}
\author{
Ahmed El-Gabbas
}
