% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepMod4HPC.R
\name{PrepMod4HPC}
\alias{PrepMod4HPC}
\title{PrepMod4HPC}
\usage{
PrepMod4HPC(
  Hab_Abb = NULL,
  Path_Data = NULL,
  Path_Model = NULL,
  Path_EnvFile = ".env",
  GPP_Dists = c(20, 30, 40, 50, 60),
  GPP_Save = TRUE,
  GPP_Plot = TRUE,
  XVars = NULL,
  PhyloTree = TRUE,
  NoPhyloTree = TRUE,
  Npararallel = 8,
  nChains = 4,
  thin = c(5, 10, 20),
  samples = c(1000, 2000, 3000),
  transientFactor = 300,
  verbose = 1000,
  ModelCountry = NULL,
  MinPresPerCountry = 50
)
}
\arguments{
\item{Hab_Abb}{Hab_Abb}

\item{Path_Data}{Path_Data}

\item{Path_Model}{Path_Model}

\item{Path_EnvFile}{Path_EnvFile}

\item{GPP_Dists}{GPP_Dists}

\item{GPP_Save}{GPP_Save}

\item{GPP_Plot}{GPP_Plot}

\item{XVars}{XVars}

\item{PhyloTree}{PhyloTree}

\item{NoPhyloTree}{NoPhyloTree}

\item{Npararallel}{Npararallel}

\item{nChains}{nChains}

\item{thin}{thin}

\item{samples}{samples}

\item{transientFactor}{transientFactor}

\item{verbose}{verbose}

\item{ModelCountry}{ModelCountry}

\item{MinPresPerCountry}{MinPresPerCountry}
}
\description{
PrepMod4HPC
}
\author{
Ahmed El-Gabbas
}
