% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepModData.R
\name{PrepModData}
\alias{PrepModData}
\title{Prepare habitat-specific data for the models}
\usage{
PrepModData(
  Hab_Abb = NULL,
  MinPresGrids = 50,
  Path_EnvFile = ".env",
  BioVars = c("bio4", "bio6", "bio8", "bio12", "bio15", "bio18"),
  ReturnData = FALSE,
  OutputPath = NULL
)
}
\arguments{
\item{Hab_Abb}{Habitat type. This has to be one of the following: c("0", "1", "2", "3", "4a", "4b", "5", "6", "8", "10", "12a", "12b"). "0" means prepare data irrespective of the habitat type}

\item{MinPresGrids}{Minimum number of presence grid cells per species. Only species with >= this number will be considered}

\item{Path_EnvFile}{Path to read the environment variables. Default value: \code{.env}}

\item{BioVars}{The bioclimatic variables to get from CHELSA. Default value: \code{c("bio4", "bio6", "bio8", "bio12", "bio15", "bio18")}}

\item{ReturnData}{Should the resulted data be returned as an R object? Default: \code{FALSE}}

\item{OutputPath}{Path for the directory to save the output file.}
}
\description{
Prepare habitat-specific data for the models
}
\author{
Ahmed El-Gabbas
}
