% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepSLURM.R
\name{PrepSLURM}
\alias{PrepSLURM}
\title{Prepare SLURM file for model fitting on HPC}
\usage{
PrepSLURM(
  Path_Model = NULL,
  CommandFile = "Commands_All.txt",
  JobName = NULL,
  CatJobInfo = TRUE,
  ntasks = 1,
  CpusPerTask = 1,
  GpusPerNode = 1,
  MemPerCpu = NULL,
  Time = NULL,
  Partition = "small-g",
  Path_EnvFile = ".env"
)
}
\arguments{
\item{Path_Model}{String. Path to the model files (without trailing slash)}

\item{CommandFile}{CommandFile}

\item{JobName}{JobName}

\item{CatJobInfo}{CatJobInfo}

\item{ntasks}{ntasks}

\item{CpusPerTask}{CpusPerTask}

\item{GpusPerNode}{GpusPerNode}

\item{MemPerCpu}{MemPerCpu}

\item{Time}{Time}

\item{Partition}{Partition}

\item{Path_EnvFile}{Path_EnvFile}
}
\description{
Prepare SLURM file for model fitting on HPC
}
\author{
Ahmed El-Gabbas
}
