% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepKnots.R
\name{PriorAlpha}
\alias{PriorAlpha}
\title{Generate a prior distribution using a mix of truncated normal and uniform
samples}
\usage{
PriorAlpha(
  n_samples = 300,
  prob_norm = 0.8,
  tr_mean = 200,
  tr_sd = 50,
  tr_min = 5,
  tr_max = 400,
  dist_max = NULL,
  seed = NULL
)
}
\arguments{
\item{n_samples}{Integer; total number of samples to generate. Default: 300.}

\item{prob_norm}{Numeric; proportion of samples drawn from the truncated
normal distribution. Default: 0.8.}

\item{tr_mean, tr_sd, tr_min, tr_max}{Numeric; parameters for the truncated
normal distribution in the kilometer unit. See \link[truncnorm:dtruncnorm]{truncnorm::rtruncnorm} for
more details.}

\item{dist_max}{Numeric; maximum value for the uniform distribution samples.}

\item{seed}{Integer; an optional random seed for reproducibility.}
}
\description{
This function generates a prior alpha distribution using a mix of samples
from a truncated normal distribution and a uniform distribution.
}
\author{
Ahmed El-Gabbas
}
