% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_railway_intensity.R
\name{railway_intensity}
\alias{railway_intensity}
\title{Calculate railway intensity based on \code{OpenStreetMap} data}
\usage{
railway_intensity(
  env_file = ".env",
  n_cores = 6L,
  strategy = "multisession",
  delete_processed = TRUE
)
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{delete_processed}{Logical indicating whether to delete the raw
downloaded railways files after processing them. This helps to free large
unnecessary file space (> 55 GB). Defaults to \code{TRUE}.}
}
\value{
\code{NULL}. Outputs processed files to the directories specified in the
environment file.
}
\description{
This function downloads, processes, and analyses railway data extracted from
\href{https://www.openrailwaymap.org}{OpenRailwayMap} available from
\href{https://download.geofabrik.de/}{OpenStreetMap Data Extracts}. It supports
parallel processing for faster execution and can calculate the total length
of railways and distance to the nearest railway for each grid cell in Europe.
}
\author{
Ahmed El-Gabbas
}
