% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_RequireMultiple.R
\name{RequireMultiple}
\alias{RequireMultiple}
\title{Load (or install) multiple packages at once}
\usage{
RequireMultiple(..., Reload = FALSE, InstallMissing = TRUE)
}
\arguments{
\item{...}{character. Names of the packages to be loaded or installed.}

\item{Reload}{logical. Determines whether already loaded packages should be reloaded. Defaults to \code{FALSE}.}

\item{InstallMissing}{logical. Determines whether missing packages should be automatically installed and then loaded. Defaults to \code{TRUE}.}
}
\value{
Invisible NULL. The function is used for its side effects (loading/installing packages) and does not return any value.
}
\description{
This function attempts to load multiple R packages specified by the user. If a package is not installed, the function can optionally install it before loading. It also provides an option to reload already loaded packages.
}
\examples{
# Currently loaded packages
(P1 <- LoadedPackages())

RequireMultiple(dismo, MASS, mgcv)

# Loaded packages after implementing the function
(P2 <- LoadedPackages())

# Which packages were loaded?
setdiff(P2, P1)
}
\author{
Ahmed El-Gabbas
}
