% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PrepData.R
\name{RespCurv_PrepData}
\alias{RespCurv_PrepData}
\title{Prepare response curve data}
\usage{
RespCurv_PrepData(
  Path_Model = NULL,
  ngrid = 50,
  NCores = 15,
  ShowProgress = FALSE
)
}
\arguments{
\item{Path_Model}{String. Path to .RData file for selected model}

\item{ngrid}{Integer. Number of points along the gradient (for continuous focal variables)}

\item{NCores}{Integer. Number of parallel cores.}

\item{ShowProgress}{Logical. Show progress bar. Default: \code{FALSE}}
}
\description{
Prepare response curve data
}
