% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_river_length.R
\name{river_length}
\alias{river_length}
\title{Calculate the length of rivers in each Strahler order per grid cell}
\usage{
river_length(env_file = ".env", cleanup = FALSE)
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{cleanup}{Logical indicating whether to clean up temporary files from
the Interim directory after finishing calculations. Default: \code{FALSE}.}
}
\value{
\code{NULL}. The function outputs processed files to the specified
directories.
}
\description{
This function processes EU-Hydro River Network Database to calculate the
length of rivers in each Strahler number. The Strahler number is used as an
index for river network classification, with higher numbers representing
larger, more significant river segments. The function reads and processes
zip-compressed geographic data (GPKG files), extracts relevant information
about river segments, computes the length of rivers for each Strahler order
per grid cell, and outputs the results both as raster files and RData
objects. The calculated length represents the total length of rivers in each
Strahler number or larger (e.g., for STRAHLER_5, the length of rivers with
Strahler values of 5 or higher).
}
\details{
The data provides at pan-European level a photo-interpreted river
network, consistent of surface interpretation of water bodies (lakes and
wide rivers), and a drainage model (also called Drainage Network), derived
from EU-DEM, with catchments and drainage lines and nodes.
\itemize{
\item \strong{Data source}: EU-Hydro River Network Database v013 |
\item \strong{Temporal extent}: 2006-2012; \strong{Format}: Vector (GPKG); \strong{Size}: 4 GB
}
}
\references{
\itemize{
\item DOI: \url{https://doi.org/10.2909/393359a7-7ebd-4a52-80ac-1a18d5f3db9c}
\item Download link: \url{https://land.copernicus.eu/en/products/eu-hydro/eu-hydro-river-network-database}
}
}
\author{
Ahmed El-Gabbas
}
