% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Road_Intensity.R
\name{Road_Intensity}
\alias{Road_Intensity}
\title{Calculate road intensity}
\usage{
Road_Intensity(FromHPC = TRUE, EnvFile = ".env")
}
\arguments{
\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}
}
\value{
\code{NULL}. The function outputs processed files to the specified
directories.
}
\description{
This function downloads, processes, and analyzes \href{https://www.globio.info/download-grip-dataset}{GRIP global roads data} (\href{https://iopscience.iop.org/article/10.1088/1748-9326/aabd42/meta}{Meijer et al. 2018}). The
function calculates the total road lengths and the distance to the nearest
road per grid cell (for any road type and per road type).
}
\note{
\itemize{
\item The function downloads the most recent version of Global Roads Inventory
Project (\code{GRIP}) data from the URL specified in the environment variable
\code{DP_R_Roads_URL}. Original data format is a zipped file containing global
road data in the form of \code{fgdb} (\code{EPSG:3246}).
\item On LUMI HPC, loading the \code{libarchive} module is necessary to use the
\code{archive} R package: \verb{module load libarchive/3.6.2-cpeGNU-23.09}
\item The distance to roads is calculated by determining the distance from each
grid
cell to the nearest grid cell that overlaps with a road (not to the nearest
road line). Note that this is different from calculating the actual
distance to the nearest road line, which is computationally intensive and
not performed in this function.
}
}
\author{
Ahmed El-Gabbas
}
