% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_predict_latent_factor.R
\name{run_crossprod_solve}
\alias{run_crossprod_solve}
\title{run_crossprod_solve}
\usage{
run_crossprod_solve(
  tf_environ,
  s1,
  s2,
  post_eta,
  path_out,
  denom,
  chunk_size = 1000L,
  threshold_mb = 2000L,
  tf_use_single = TRUE,
  verbose = TRUE,
  solve_chunk_size = 50L,
  solve_max_attempts = 5L,
  lf_commands_only = FALSE
)
}
\arguments{
\item{tf_environ}{Character. Path to the Python environment. This argument is
required if \code{use_tf} is \code{TRUE} under Windows. Defaults to \code{NULL}.}

\item{s1}{Character. Path to the input file containing s1 coordinates.}

\item{s2}{Character Path to the input file containing s2 coordinates.}

\item{post_eta}{Character. Path to the file containing the \code{post_eta} matrix
data.}

\item{path_out}{Character. Path to rds file where the output results will be
saved.}

\item{denom}{Numeric. The denominator value used in the computation.}

\item{chunk_size}{Numeric (Optional). Size of chunks to process at a time.
Default is 1000.}

\item{threshold_mb}{Numeric (Optional). Memory threshold (in MB) to manage
processing. Default is 2000.}

\item{tf_use_single}{Logical. Whether to use single precision for the
\code{TensorFlow} calculations. Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE}, logs detailed information during
execution. Default is \code{TRUE}.}

\item{solve_chunk_size}{Integer. Chunk size for \code{solve_and_multiply} Python
function. Default is 50L.}

\item{solve_max_attempts}{Integer. Maximum number of attempts to run solve
and crossprod internal function \link{run_crossprod_solve}. Default is 5L.}

\item{lf_commands_only}{Logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}
}
\value{
Returns the \code{path_out} if successful. Returns \code{NULL} if all attempts
fail.
}
\description{
Internal function to executes a Python script that performs matrix
computations using \code{TensorFlow} with provided inputs. Retries up to three
times if the output file validation fails.
}
\details{
\itemize{
\item The function checks for the existence of required input files and the
Python executable in the specified virtual environment.
\item Executes the Python script using \code{system2}.
\item Verifies the output file validity using \code{ecokit::check_data}. Retries up
to 3 times if the output is invalid.
\item Generates detailed logs if \code{verbose} is set to \code{TRUE}.
}
}
\author{
Ahmed El-Gabbas
}
\keyword{internal}
