% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Sampling_Efforts.R
\name{Sampling_Efforts}
\alias{Sampling_Efforts}
\title{Download and Process sampling efforts data}
\usage{
Sampling_Efforts(
  FromHPC = TRUE,
  EnvFile = ".env",
  Renviron = ".Renviron",
  RequestData = TRUE,
  DownloadData = TRUE,
  NCores = 6,
  StartYear = 1980,
  Boundaries = c(-30, 50, 25, 75),
  ChunkSize = 10000
)
}
\arguments{
\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default: \code{.env}.}

\item{Renviron}{Character. The path to the \code{.Renviron} file containing GBIF
login credentials (email, user, password). Default: \code{.Renviron}.}

\item{RequestData}{Logical. If \code{TRUE}, the function requests data from GBIF.
If \code{FALSE}, previously requested data is loaded from disk. Defaults to
\code{TRUE}.}

\item{DownloadData}{Logical. If \code{TRUE}, the function downloads data and
stores it on disk. If \code{FALSE}, it skips the download step. Defaults to
\code{TRUE}.}

\item{NCores}{Integer. The number of cores to use for parallel processing.}

\item{StartYear}{Numeric. The starting year for the occurrence data. Only
records from this year onward will be requested from GBIF. Default: \code{1980}.}

\item{Boundaries}{Numeric vector of length 4. Specifies geographical
boundaries for the requested GBIF data in the order: Left, Right, Bottom,
Top. Default: \code{c(-30, 50, 25, 75)}.}

\item{ChunkSize}{Integer. The number of rows per chunk file. Default:
\verb{100,000}. See \link{Efforts_Split} and \link{Efforts_Process} for more details.}
}
\value{
Returns \code{NULL} invisibly. The function generates various output
files, maps, and logs, and it is designed to be used for its side effects.
}
\description{
This function downloads GBIF data for all vascular plants within a specified
geographical area (Europe), grouped by order, and converts the data to raster
format to represent the number of vascular plant observations and species per
grid cell.
}
\note{
This function is expected to take a substantial amount of time (>9
hours on a Windows PC with 6 cores). The data request from GBIF may take
around 5 hours to be ready. The function requests GBIF data for each
vascular plant order and waits for the data to be ready before processing
them.
}
\author{
Ahmed El-Gabbas
}
