% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{sapply_}
\alias{sapply_}
\title{Apply a Function over a List or Vector (with no return value)}
\usage{
sapply_(X, FUN, simplify = TRUE, ...)
}
\arguments{
\item{X}{vector}

\item{FUN}{function}

\item{simplify}{should be the output be simplified}

\item{...}{additional arguments to sapply}
}
\description{
Apply a Function over a List or Vector (with no return value)
}
\examples{
par(mfrow = c(1,2), oma = c(0.25, 0.25, 0.25, 0.25), mar = c(3,3,3,1))
sapply(list(x = 100:110, y = 110:120), function(V){plot(V, las = 1, main = "sapply")})

# -------------------------------------------

# nothing returned or printed, only the plotting
par(mfrow = c(1,2), oma = c(0.25, 0.25, 0.25, 0.25), mar = c(3,3,3,1))
sapply_(list(x = 100:110, y = 110:120), function(V){plot(V, las = 1, main = "sapply_")})
}
\author{
Ahmed El-Gabbas
}
