% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SaveAs}
\alias{SaveAs}
\title{Save an object with a different name}
\usage{
SaveAs(InObj, OutObj, OutPath)
}
\arguments{
\item{InObj}{input object}

\item{OutObj}{output object}

\item{OutPath}{save path}
}
\description{
Save an object with a different name
}
\author{
Ahmed El-Gabbas
}
