% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_SaveMultiple.R
\name{SaveMultiple}
\alias{SaveMultiple}
\title{Save multiple objects to their respective \code{.RData} files}
\usage{
SaveMultiple(
  Vars = NULL,
  OutFolder = getwd(),
  Overwrite = FALSE,
  Prefix = "",
  Verbose = FALSE
)
}
\arguments{
\item{Vars}{Character vector. Names of the variables to be saved. If \code{NULL}
or any specified variable does not exist in the global environment, the
function will stop with an error.}

\item{OutFolder}{Character. Path to the output folder where the \code{.RData}
files will be saved. Defaults to the current working directory.}

\item{Overwrite}{Logical. Whether existing \code{.RData} files should be
overwritten. If \code{FALSE} (Default) and files exist, the function will stop
with an error message.}

\item{Prefix}{Character. Prefix of each output file name. Useful for
organizing saved files or avoiding name conflicts. Defaults to an empty
string.}

\item{Verbose}{Logical. Whether to print a message upon successful saving of
files. Defaults to \code{FALSE}.}
}
\value{
The function is used for its side effect of saving files and does not
return a value.
}
\description{
This function saves specified variables from the global environment to
separate \code{.RData} files. It allows for optional file prefixing and
overwriting of existing files.
}
\examples{
\dontrun{
  TMP_Folder <- IASDT.R::Path(tempdir(), stringi::stri_rand_strings(1, 5))
  fs::dir_create(TMP_Folder)

  # ----------------------------------------------
  # Save x1 and x2 to disk
  # ----------------------------------------------
  x1 = 10; x2 = 20

  SaveMultiple(Vars = c("x1", "x2"), OutFolder = TMP_Folder)

  list.files(path = TMP_Folder, pattern = "^.+.RData")

  (x1Contents <- IASDT.R::LoadAs(IASDT.R::Path(TMP_Folder, "x1.RData")))

  (x2Contents <- IASDT.R::LoadAs(IASDT.R::Path(TMP_Folder, "x2.RData")))

  # ----------------------------------------------
  # Use Prefix
  # ----------------------------------------------

  SaveMultiple(Vars = c("x1", "x2"), OutFolder = TMP_Folder, Prefix = "A_")

  list.files(path = TMP_Folder, pattern = "^.+.RData")

  # ----------------------------------------------
  # File exists, no save
  # ----------------------------------------------
  try(SaveMultiple(Vars = c("x1", "x2"), OutFolder = TMP_Folder))

  # ----------------------------------------------
  # overwrite existing file
  # ----------------------------------------------
  x1 = 100; x2 = 200; x3 = 300

  SaveMultiple(Vars = c("x1", "x2", "x3"),
     OutFolder = TMP_Folder, Overwrite = TRUE)

  (x1Contents <- IASDT.R::LoadAs(IASDT.R::Path(TMP_Folder, "x1.RData")))

  (x2Contents <- IASDT.R::LoadAs(IASDT.R::Path(TMP_Folder, "x2.RData")))

     (x3Contents <- IASDT.R::LoadAs(IASDT.R::Path(TMP_Folder, "x3.RData")))
}
}
\author{
Ahmed El-Gabbas
}
