% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SaveXToDisk}
\alias{SaveXToDisk}
\title{Save many R objects at once to their respective \code{.RData} files}
\usage{
SaveXToDisk(Vars = NULL, OutFolder = getwd(), Overwrite = FALSE)
}
\arguments{
\item{Vars}{The name of variables to be saved}

\item{OutFolder}{The output directory. Default: the working directory}

\item{Overwrite}{overwrite available .RData files. Default: FALSE}
}
\description{
Save many R objects at once to their respective \code{.RData} files
}
\examples{
\dontrun{
x1 = 10; x2 = 20
TMP_Folder <- tempdir()
print(TMP_Folder)

SaveXToDisk(Vars = c("x1", "x2"), OutFolder = TMP_Folder)
list.files(path = TMP_Folder, pattern = "^.+.RData")

SaveXToDisk(Vars = c("x1", "x2"), OutFolder = TMP_Folder, Overwrite = TRUE)
}
}
\author{
Ahmed El-Gabbas
}
