% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_Scale_0_1.R
\name{Scale_0_1}
\alias{Scale_0_1}
\title{Scale a spatRaster object between 0 and 1}
\usage{
Scale_0_1(x)
}
\arguments{
\item{x}{x}
}
\description{
Scale a spatRaster object between 0 and 1
}
\author{
Ahmed El-Gabbas
}
