% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{scraplinks}
\alias{scraplinks}
\title{Extract link texts and urls from a web page}
\usage{
scraplinks(url)
}
\arguments{
\item{url}{the url}
}
\description{
source: https://gist.github.com/paulrougieux/e1ee769577b40cd9ed9db7f75e9a2cc2
}
\details{
Extract link texts and urls from a web page
}
\examples{
head(scraplinks("https://github.com/"))

}
\author{
Ahmed El-Gabbas
}
