% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_ScriptLocation.R
\name{ScriptLocation}
\alias{ScriptLocation}
\title{Retrieve the location of the current R script.}
\usage{
ScriptLocation()
}
\value{
A character string representing the file path of the current R script, or NA if the path cannot be determined.
}
\description{
This function attempts to find the location of the currently running R script. It first tries to identify the script's location based on the command line arguments used to start the script. If the script is being run in an interactive session within RStudio, it falls back to using the \code{rstudioapi} to find the file path of the script in the source editor. If the location cannot be determined, it returns NA.
}
\examples{
 \dontrun{
ScriptLocation()
}
}
\references{
\href{https://stackoverflow.com/questions/47044068/}{Click here}
}
