% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_SetChainName.R
\name{SetChainName}
\alias{SetChainName}
\title{Set name of the chain to a vector of strings}
\usage{
SetChainName(Obj, Chain)
}
\arguments{
\item{Obj}{Input object name}

\item{Chain}{Integer. Chain number}
}
\description{
Set name of the chain to a vector of strings
}
\author{
Ahmed El-Gabbas
}
