% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_setRastVals.R
\name{setRastVals}
\alias{setRastVals}
\title{Set values for \code{SpatRaster} Objects}
\usage{
setRastVals(R)
}
\arguments{
\item{R}{A \code{SpatRaster} object, a file path to a raster file, or an object
from the \code{raster} package (e.g., \code{RasterLayer}, \code{RasterStack},
\code{RasterBrick}).}
}
\value{
A \code{SpatRaster} object with values loaded into memory.
}
\description{
This function processes a \code{SpatRaster} object or a file path to it, ensuring
the raster is loaded and unpacked if necessary. It converts raster objects
from the \code{raster} package to \code{SpatRaster} objects. The function ensures that
the \code{SpatRaster} object is read from memory, not from file.
}
\details{
The function handles various types of input:
\itemize{
\item If a file path is provided, it attempts to load the raster using
\code{\link[terra:rast]{terra::rast()}}.
\item If the input is a packed \code{SpatRaster}, it unpacks the raster using
\code{\link[terra:wrap]{terra::unwrap()}}.
\item If the input is a raster object from the \code{raster} package, it is
converted to \code{SpatRaster}.
}
}
\author{
Ahmed El-Gabbas
}
