% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_spatial.R
\name{sf_add_coords}
\alias{sf_add_coords}
\title{Add coordinates of the current sf object as columns}
\usage{
sf_add_coords(Sf_Obj, NameX = "Long", NameY = "Lat", Overwrite = FALSE)
}
\arguments{
\item{Sf_Obj}{input sf object}

\item{NameX}{Name of the longitude column to be added}

\item{NameY}{Name of the latitude column to be added}

\item{Overwrite}{Should columns \code{NameX} or \code{NameY} be overwritten if these columns exist in the original data}
}
\description{
Add coordinates of the current sf object as columns
}
\examples{
pt1 = sf::st_point(c(0,1))
pt2 = sf::st_point(c(1,1))
d = data.frame(a = 1:2)
d$geom = sf::st_sfc(pt1, pt2)
df = sf::st_as_sf(d)
df
(df <- sf_add_coords(df))

(sf_add_coords(df))

(sf_add_coords(df, Overwrite = TRUE))
}
\author{
Ahmed El-Gabbas
}
