% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwf_soil.R
\name{soil_density_process}
\alias{soil_density_process}
\title{Retrieve and project soil bulk density data}
\usage{
soil_density_process(depths = NULL, env_file = ".env", n_cores = 6L)
}
\arguments{
\item{depths}{Character vector of depth intervals (e.g. \code{c("0-5","5-15")}).
If \code{NULL} (default) all valid depths are processed. Valid \code{depths} are the
SoilGrids BDOD standard horizons (cm): "0-5", "5-15", "15-30", "30-60",
"60-100", and "100-200". Each depth string should omit the trailing \code{cm}.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}
}
\value{
(Invisibly) the path to the \code{RData} file containing the processed
soil bulk density data.
}
\description{
Downloads and projects \href{https://soilgrids.org}{SoilGrids}' bulk density of
the fine earth fraction (bdod) layers for specified depth intervals. See
\href{https://www.isric.org/explore/soilgrids/faq-soilgrids}{here} for more
details. This function project the original data at different depths
intervals to the modelling reference grid.
}
\references{
\itemize{
\item SoilGrids: \url{https://soilgrids.org}
\item SoilGrids250m 2.0 - \href{https://data.isric.org/geonetwork/srv/eng/catalog.search#/metadata/713396f9-1687-11ea-a7c0-a0481ca9e724}{Bulk density}
\item Poggio \emph{et al.} (2021): SoilGrids 2.0: producing soil information for the
globe with quantified spatial uncertainty. Soil.
\href{https://doi.org/10.5194/soil-7-217-2021}{10.5194/soil-7-217-2021}
}
}
\author{
Ahmed El-Gabbas
}
