% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Solve1}
\alias{Solve1}
\title{Solve1: Invert a matrix}
\usage{
Solve1(x)
}
\arguments{
\item{x}{A square matrix to be inverted.}
}
\value{
The inverted matrix of the same dimensions as \code{x}.
//' @export
}
\description{
Computes the inverse of a square matrix using Armadillo's \code{inv} function.
}
