% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve2vect}
\alias{solve2vect}
\title{Solve a system of linear equations for a vector}
\usage{
solve2vect(A, B)
}
\arguments{
\item{A}{A square matrix of coefficients.}

\item{B}{A vector of constants.}
}
\value{
The solution vector \code{x}.
}
\description{
Solves a system of linear equations \code{A * x = B} where \code{A} is a square
matrix and \code{B} is a vector.
}
