% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{sort_}
\alias{sort_}
\title{Sort alphanumeric strings}
\usage{
sort_(
  x,
  decreasing = FALSE,
  na.last = TRUE,
  blank.last = FALSE,
  numeric.type = c("decimal", "roman"),
  roman.case = c("upper", "lower", "both")
)
}
\arguments{
\item{x}{Vector to be sorted.}

\item{decreasing}{logical. Should the sort be increasing or decreasing? Note that descending=TRUE reverses the meanings of na.last and blanks.last. Default: \code{FALSE}}

\item{na.last}{for controlling the treatment of NA values. If TRUE, missing values in the data are put last; if FALSE, they are put first; if NA, they are removed. Default: \code{TRUE}}

\item{blank.last}{for controlling the treatment of blank values. If TRUE, blank values in the data are put last; if FALSE, they are put first; if NA, they are removed. Default: \code{FALSE}}

\item{numeric.type}{either "decimal" (default) or "roman". Are numeric values represented as decimal numbers (numeric.type="decimal") or as Roman numerals (numeric.type="roman")?}

\item{roman.case}{one of "upper", "lower", or "both". Are roman numerals represented using only capital letters ('IX') or lower-case letters ('ix') or both?}
}
\description{
Sort alphanumeric strings
}
\author{
Ahmed El-Gabbas
}
