% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_SourceSilent.R
\name{SourceSilent}
\alias{SourceSilent}
\title{Silently source R script with optional message and warning suppression}
\usage{
SourceSilent(File, Messages = TRUE, Warnings = TRUE, ...)
}
\arguments{
\item{File}{Character. Path of the file to be sourced.}

\item{Messages}{Logical. If \code{TRUE} (default), messages generated by the
script are shown. If \code{FALSE}, messages are suppressed.}

\item{Warnings}{Logical. If \code{TRUE} (default), warnings generated by the
script are shown. If \code{FALSE}, warnings are suppressed.}

\item{...}{Additional arguments passed to the \link{source} function.}
}
\value{
It is used for its side effect of sourcing a file with optional
suppression of messages and warnings.
}
\description{
This function sources an R script file with options to suppress messages
and/or warnings. It is useful for running scripts that generate a lot of
console output that is not needed.
}
\author{
Ahmed El-Gabbas
}
