% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_SplitDF2Chunks.R
\name{SplitDF2Chunks}
\alias{SplitDF2Chunks}
\title{Split data.frame into smaller chunks}
\usage{
SplitDF2Chunks(DF = NULL, ChunkSize = NULL, NChunks = NULL, Prefix = "Chunk")
}
\arguments{
\item{DF}{The data frame to split}

\item{ChunkSize}{Integer. Number of rows per chunk}

\item{NChunks}{Integer. Number of chunks (only if \code{ChunkSize} not provided)}

\item{Prefix}{Prefix}
}
\description{
Split data.frame into smaller chunks
}
\examples{
SplitDF2Chunks(mtcars, ChunkSize = 16)

# -------------------------------------------

SplitDF2Chunks(mtcars, NChunks = 3)

# -------------------------------------------

SplitDF2Chunks(mtcars, NChunks = 3, Prefix = "T")
}
\author{
Ahmed El-Gabbas
}
