% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SplitVector}
\alias{SplitVector}
\title{Split a vector into smaller chunks}
\usage{
SplitVector(Vector = NULL, NSplit = NULL)
}
\arguments{
\item{Vector}{vector to split}

\item{NSplit}{number of splits}
}
\description{
Split a vector into smaller chunks
}
\author{
Ahmed El-Gabbas
}
