% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{System}
\alias{System}
\title{Run bash script depending on the operating system}
\usage{
System(command, RObj = TRUE, ...)
}
\arguments{
\item{command}{bash command to implement}

\item{RObj}{whether to make the output of the command an R object}

\item{...}{additional arguments to \code{shell} pr \code{system} functions}
}
\description{
Run bash script depending on the operating system
}
\examples{
# print working directory
System("pwd")

# first 5 files on the working directory
(A <- System("ls | head -n 5"))

(A <- System("ls | head -n 5", RObj = FALSE))
}
\author{
Ahmed El-Gabbas
}
