% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_Text2Coords.R
\name{Text2Coords}
\alias{Text2Coords}
\title{Extract longitude and latitude from string}
\usage{
Text2Coords(String = NULL, Long_Name = "Longitude", Lat_Name = "Latitude")
}
\arguments{
\item{String}{A string containing the coordinates in the format "POINT
(longitude latitude)". This parameter is required and cannot be NULL.}

\item{Long_Name, Lat_Name}{A string specifying the name to be used for the
longitude and Longitude columns in the output tibble. Defaults to
"Longitude" and "Latitude".}
}
\value{
A tibble with two columns containing the longitude and latitude
values extracted from the input string. The names of these columns are
determined by the \code{Long_Name} and \code{Lat_Name} parameters. If no names are
provided, the default names ("Longitude" and "Latitude") are used.

two column tibble for Longitude & Latitude
}
\description{
Extract longitude and latitude from string representing a geographical point
in the format "POINT (longitude latitude) and converts it into a two-column
tibble containing the longitude and latitude as numeric values. The names of
the columns in the resulting tibble can be customized. The default names for
the longitude and latitude columns are "Longitude" and "Latitude",
respectively.
}
\examples{
c("POINT (11.761 46.286)", "POINT (14.8336 42.0422)",
  "POINT (16.179999 38.427214)") \%>\%
 lapply(Text2Coords)

c("POINT (11.761 46.286)", "POINT (14.8336 42.0422)",
  "POINT (16.179999 38.427214)") \%>\%
 lapply(Text2Coords, Long_Name = "Long", Lat_Name = "Lat")
}
\author{
Ahmed El-Gabbas
}
