% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{valid_url}
\alias{valid_url}
\title{Check the validity of a URL}
\usage{
valid_url(url_in, t = 2)
}
\arguments{
\item{url_in}{URL path}

\item{t}{timeout}
}
\description{
Check the validity of a URL
}
\author{
Ahmed El-Gabbas
}
