% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_VarPar_Compute.R
\name{VarPar_Compute}
\alias{VarPar_Compute}
\title{Computes variance partitioning of Hmsc models}
\usage{
VarPar_Compute(
  Path_Model,
  group = NULL,
  groupnames = NULL,
  start = 1L,
  na.ignore = FALSE,
  NCores = 8L,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  Temp_Cleanup = TRUE,
  Chunk_size = 50L,
  Verbose = TRUE,
  VarParFile = "VarPar",
  VP_Commands_Only = FALSE
)
}
\arguments{
\item{Path_Model}{Character. Path to fitted \code{Hmsc} model object.}

\item{group}{vector of numeric values corresponding to group identifiers in
groupnames. If the model was defined with \code{XData} and \code{XFormula}, the
default is to use model terms.}

\item{groupnames}{vector of names for each group of fixed effect. Should
match \code{group}. If the model was defined with \code{XData} and \code{XFormula}, the
default is to use the labels of model terms.}

\item{start}{index of first MCMC sample included}

\item{na.ignore}{logical. If TRUE, covariates are ignored for sites where the
focal species is NA when computing variance-covariance matrices for each
species}

\item{NCores}{Integer specifying the number of parallel cores for
parallelization. This is only used when \code{UseTF} is TRUE.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{TF_use_single}{Logical indicating whether to use single precision for
the TF calculations. Defaults to \code{FALSE}.}

\item{Temp_Cleanup}{Logical. Whether to delete temporary files after
processing. Default: \code{TRUE}.}

\item{Chunk_size}{Integer. Size of each chunk of samples to process in
parallel. Only relevant for TensorFlow. Default: \code{50}.}

\item{Verbose}{Logical. Indicates whether progress messages should be
displayed. Defaults to \code{TRUE}.}

\item{VarParFile}{Character. Name of the output file to save the results.}

\item{VP_Commands_Only}{logical. If \code{TRUE}, returns the commands to run the
Python script. Default is \code{FALSE}. Only relevant when \code{UseTF} is \code{TRUE}.}
}
\description{
Computes variance components with respect to given grouping of fixed effects
and levels of random effects. This function is a wrapper around the
\code{Hmsc::computeVariancePartitioning}, but with the added functionality of
parallel computation using TensorFlow.
}
\author{
Ahmed El-Gabbas
}
