% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_VarPar_Plot.R
\name{VarPar_Plot}
\alias{VarPar_Plot}
\title{Plot variance partitioning of Hmsc model using ggplot2.}
\usage{
VarPar_Plot(
  Path_Model,
  EnvFile = ".env",
  FromHPC = TRUE,
  UseTF = TRUE,
  TF_Environ = NULL,
  NCores = 1
)
}
\arguments{
\item{Path_Model}{Character path for the model file.}

\item{EnvFile}{String. Path to read the environment variables. Default value:
\code{.env}}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{NCores}{Integer. Number of parallel computations for computing variance
partitioning using TensorFlow. See \link{VarPar_Compute} for more details.
Default: \code{1}.}
}
\description{
This function generates and optionally saves plots visualizing the variance
partitioning of a Hmsc model. It can calculate variance partitioning and
model evaluation if not provided and supports parallel computation for model
predictions.
}
\details{
The function reads the following environment variables:
\itemize{
\item \strong{\code{DP_R_TaxaInfo_RData}} (if \code{FromHPC} = \code{TRUE}) or
\strong{\code{DP_R_TaxaInfo_RData_Local}} (if \code{FromHPC} = \code{FALSE}) for the
location of the \code{TaxaList.RData} file containing species information.
}
}
\author{
Ahmed El-Gabbas
}
