% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwf_wetness_index.R
\name{wetness_index_process}
\alias{wetness_index_process}
\title{Download and Process Topographic Wetness Index Data}
\usage{
wetness_index_process(env_file = ".env")
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}
}
\value{
(Invisibly) The path to the processed wetness index \code{RData} file.
}
\description{
Downloads, extracts, and processes the global topographic wetness index data
at 30 arc-second resolution (Title & Bemmels, 2018). The function checks for
existing processed data, downloads the required dataset if necessary,
extracts the relevant TIFF file, re-projects and masks it to match a
reference grid, and saves the processed raster in both TIFF and \code{RData}
formats.
}
\references{
\itemize{
\item Title & Bemmels (2018): ENVIREM: an expanded set of bioclimatic and
topographic variables increases flexibility and improves performance of
ecological niche modeling. Ecography.
\href{https://doi.org/10.1111/ecog.02880}{10.1111/ecog.02880}
\item \href{https://deepblue.lib.umich.edu/data/concern/generic_works/gt54kn05f}{ENVIREM}: ENVIronmental Rasters for Ecological Modeling version 1.0
}
}
\author{
Ahmed El-Gabbas
}
