% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_AddImg2Plot.R
\name{AddImg2Plot}
\alias{AddImg2Plot}
\title{Add image to plot}
\usage{
AddImg2Plot(obj, x = NULL, y = NULL, width = NULL, interpolate = TRUE)
}
\arguments{
\item{obj}{an image file imported as an array (e.g. \code{png::readPNG}, \code{jpeg::readJPEG})}

\item{x}{mid x coordinate for image}

\item{y}{mid y coordinate for image}

\item{width}{width of image (in x coordinate units)}

\item{interpolate}{(passed to \code{graphics::rasterImage}) A logical vector (or scalar) indicating whether to apply linear interpolation to the image when drawing. Default = \code{TRUE}.}
}
\description{
Add image to plot
}
\examples{
LoadPackages(png)
myurl <- "https://upload.wikimedia.org/wikipedia/commons/e/e1/Jupiter_\%28transparent\%29.png"
z <- tempfile()
download.file(myurl, z, mode="wb")
pic <- readPNG(z)
file.remove(z) # cleanup

image(volcano)
AddImg2Plot(pic, x = 0.3, y = 0.5, width = 0.2)
AddImg2Plot(pic, x = 0.7, y = 0.7, width = 0.2)
AddImg2Plot(pic, x = 0.7, y = 0.2, width = 0.1)
}
\references{
\href{https://stackoverflow.com/questions/27800307/}{Click here}
}
