% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_AddMissingCols.R
\name{AddMissingCols}
\alias{AddMissingCols}
\title{Add missing columns to a data frame with specified fill values}
\usage{
AddMissingCols(DT, FillVal = NA_character_, ...)
}
\arguments{
\item{DT}{A data frame to which missing columns will be added. This parameter
cannot be NULL.}

\item{FillVal}{The value to fill the missing columns with. This parameter
defaults to \code{NA_character_}, but can be changed to any scalar value as
required.}

\item{...}{Column names as character strings.}
}
\value{
a data frame with the missing columns added, if any were missing.
}
\description{
This function checks a data frame for missing columns specified by the user.
If any are missing, it adds these columns to the data frame, filling them
with a specified value.
}
\examples{
mtcars \%>\%
 dplyr::select(1:3) \%>\%
 AddMissingCols(FillVal = NA_character_, A, B, C) \%>\%
 AddMissingCols(FillVal = as.integer(10), D)

# -------------------------------------------

AddCols <- c("Add1", "Add2")
mtcars \%>\%
 dplyr::select(1:3) \%>\%
 AddMissingCols(FillVal = NA_real_, AddCols)
}
\author{
Ahmed El-Gabbas
}
