% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{AddMissingCols}
\alias{AddMissingCols}
\title{Add missing columns to data frame}
\usage{
AddMissingCols(..., DT, FillVal = NA_character_)
}
\arguments{
\item{...}{list of column names to add}

\item{FillVal}{value to be used: default: NA_character}

\item{DF}{data frame}
}
\description{
Add missing columns to data frame
}
\examples{
AddMissingCols(DT = mtcars, FillVal = NA_real_, A, B, C)
}
\author{
Ahmed El-Gabbas
}
