% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{AddMissingCols}
\alias{AddMissingCols}
\title{Add missing columns to data frame}
\usage{
AddMissingCols(..., DT, FillVal = NA_character_)
}
\arguments{
\item{...}{list of column names to add}

\item{DT}{data frame}

\item{FillVal}{value to be used: default: NA_character}
}
\description{
Add missing columns to data frame
}
\examples{
mtcars \%>\%
 dplyr::select(1:3) \%>\%
 AddMissingCols(A, B, C, DT = ., FillVal = NA_character_) \%>\%
 AddMissingCols(D, DT = ., FillVal = as.integer(10))

# -------------------------------------------

AddCols <- c("Add1", "Add2")
mtcars \%>\%
 dplyr::select(1:3) \%>\%
 AddMissingCols(AddCols, DT = ., FillVal = NA_real_)
}
\author{
Ahmed El-Gabbas
}
