% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{AllObjSizes}
\alias{AllObjSizes}
\title{Size of objects in memory}
\usage{
AllObjSizes(GreaterThan = 0)
}
\arguments{
\item{GreaterThan}{Only show objects with size > specified value in MB. Default: 0, which means show all variables size}
}
\description{
This function print the size allocated by any of loaded objects into \code{R}
}
\examples{
AA1 <<- rep(1:1000, 10000)
AA2 <<- rep(1:1000, 100)

AllObjSizes()

AllObjSizes(GreaterThan = 1)

AllObjSizes(GreaterThan = 50)
}
\author{
Ahmed El-Gabbas
}
