% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{AssignIfNotExist}
\alias{AssignIfNotExist}
\title{Assign a value to a variable, only if not existing in the global environment}
\usage{
AssignIfNotExist(Variable, Value, Env = globalenv())
}
\arguments{
\item{Variable}{Variable name}

\item{Value}{Value to be assigned}

\item{Env}{environment to assign value to}
}
\description{
Assign a value to a variable, only if not existing in the global environment
}
\examples{
AssignIfNotExist(x, TRUE)
print(x)

# --------------------------------------------------

y <- 10

# y exists and thus its value was not changed
AssignIfNotExist(y, TRUE)
print(y)
}
\author{
Ahmed El-Gabbas
}
