% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Bioreg.R
\name{BioReg_Process}
\alias{BioReg_Process}
\title{Process biogeographical regions}
\usage{
BioReg_Process(FromHPC = TRUE, EnvFile = ".env")
}
\arguments{
\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}
}
\value{
Invisible \code{NULL}. Processed data is saved to disk as raster, vector,
and RData files.
}
\description{
Downloads and processes the Biogeographical Regions dataset (Europe 2016,
version 1) from the \href{https://www.eea.europa.eu/en/datahub/datahubitem-view/11db8d14-f167-4cd5-9205-95638dfd9618}{European Environment Agency}.
The dataset is processed to extract region names per reference grid cell in
EPSG:3035 projection. Outputs include raster (.tif), spatial vector (.shp),
and RData files. The outputs of this function are used to count presence grid
cells for each species per biogeographical region.
}
\details{
\itemize{
\item Temporal coverage: 2011-2015
\item Spatial coverage: 28째E-81째E, 31.27째W-62째E
\item CRS: EPSG:3035; read as: ETRS_1989_LAEA_L52_M10
\item file format: shapefile (compressed in zip file)
\item Requires external tools: \code{curl} (for downloading) and \code{unzip}
(for extraction)
}
}
\author{
Ahmed El-Gabbas
}
