% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{CatDiff}
\alias{CatDiff}
\title{Print time difference}
\usage{
CatDiff(InitTime, Prefix = "Completed in ", CatInfo = TRUE)
}
\arguments{
\item{InitTime}{day and time of the time event}

\item{Prefix}{Prefix character to print}

\item{CatInfo}{logical; also print \code{IASDT.R::InfoChunk("Session summary")}}
}
\description{
Print time difference
}
\author{
Ahmed El-Gabbas
}
