% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CatDiff.R
\name{CatDiff}
\alias{CatDiff}
\title{Print time difference}
\usage{
CatDiff(
  InitTime,
  ChunkText = "Session summary",
  Prefix = "Completed in ",
  CatInfo = FALSE,
  Level = 0,
  ...
)
}
\arguments{
\item{InitTime}{POSIXct; The initial time from which the difference is
calculated.}

\item{ChunkText}{character. The message printed as chunk info. Default
value: \verb{Session summary}. See: \link{InfoChunk} for more information.}

\item{Prefix}{character; A prefix string to prepend to the printed time
difference. Defaults to "Completed in ".}

\item{CatInfo}{logical; If \code{TRUE}, prints a session summary using
\link{InfoChunk} ("Session summary"). Defaults to \code{FALSE}.}

\item{Level}{integer; the level at which the message will be printed. If e.g.
Level = 1, the following string will be printed at the beginning of the
message: "   >>>   ". Default is \code{0}.}

\item{...}{Additional arguments for \link{CatTime}.}
}
\value{
The function is used for its side effect of printing to the console
and does not return any value.
}
\description{
This function calculates the time difference from a given initial time to the
current time and prints it with a specified prefix. Optionally, it can also
print a session summary.
}
\examples{
RefTime <- (lubridate::now() - lubridate::seconds(45))
CatDiff(RefTime)
# Completed in 00:00:45 hours

RefTime <- (lubridate::now() -
    (lubridate::minutes(50) + lubridate::seconds(45)))
CatDiff(RefTime)
# Completed in 00:50:46 hours

RefTime <- (lubridate::now() - lubridate::minutes(50))
CatDiff(RefTime)
# Completed in 00:50:01 hours

RefTime <- (lubridate::now() - lubridate::minutes(70))
CatDiff(RefTime)
# Completed in 01:10:00 hours

RefTime <- (lubridate::now() - lubridate::hours(4))
CatDiff(RefTime)
# Completed in 04:00:01 hours

RefTime <- lubridate::now() -
  (lubridate::hours(4) + lubridate::minutes(50) + lubridate::seconds(45))
  CatDiff(RefTime)
# Completed in 04:50:45 hours
}
\author{
Ahmed El-Gabbas
}
