% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{catSep}
\alias{catSep}
\title{Print separator(s) to the console}
\usage{
catSep(Rep = 1, Extra1 = 0, Extra2 = 0, Char = "-", CharReps = 50)
}
\arguments{
\item{Rep}{number of separator lines; default 1 row}

\item{Extra1}{number of extra empty lines before the separator; default: 0}

\item{Extra2}{number of extra empty lines after the separator; default: 0}

\item{Char}{The character to be used as a separator; default "-"}

\item{CharReps}{Number of times the character is repeated; default: 50}
}
\description{
Print separator(s) to the console
}
\examples{
catSep()
catSep(2)
catSep(2,2,3)
}
\author{
Ahmed El-Gabbas
}
