% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CatSep.R
\name{CatSep}
\alias{CatSep}
\title{Print separator(s) to the console}
\usage{
CatSep(
  Rep = 1L,
  Extra1 = 0L,
  Extra2 = 1L,
  Char = "-",
  CharReps = 50L,
  Bold = FALSE,
  Red = FALSE,
  ...
)
}
\arguments{
\item{Rep}{integer; the number of separator lines to print. Default is \code{1}.}

\item{Extra1, Extra2}{integer; the number of extra empty lines to print before
and after the separator lines. Default is \code{0} and \code{1}, respectively.}

\item{Char}{character; the character used to construct the separator line.
Default is \code{"-"}.}

\item{CharReps}{integer; the number of times the character is repeated to
form a separator line. Default is \code{50}.}

\item{Bold}{logical; whether to print the text in bold. Default is \code{FALSE}.}

\item{Red}{logical; whether to print the text in red. Default is \code{FALSE}.}

\item{...}{additional arguments to be passed to \code{cat()}.}
}
\value{
The function is called for its side effect (printing to the console)
and does not return a meaningful value.
}
\description{
This function prints customizable separator lines to the console, optionally
preceded and followed by empty lines. It is useful for improving the
readability of console output in R scripts or during interactive sessions.
}
\examples{
CatSep()

CatSep(2)

CatSep(2,2,3)

CatSep(2,2,3, Char = "*")

CatSep(2,2,3, Char = "*", CharReps = 20)
}
\author{
Ahmed El-Gabbas
}
