% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CatSep.R
\name{CatSep}
\alias{CatSep}
\title{Print separator(s) to the console}
\usage{
CatSep(Rep = 1, Extra1 = 0, Extra2 = 0, Char = "-", CharReps = 50)
}
\arguments{
\item{Rep}{integer; number of separator lines; default \code{1} row}

\item{Extra1}{integer; number of extra empty lines before the separator; default: \code{0}}

\item{Extra2}{integer; number of extra empty lines after the separator; default: \code{0}}

\item{Char}{character; the character to be used as a separator; default "-"}

\item{CharReps}{integer; number of times the character is repeated; default: 50}
}
\description{
Print separator(s) to the console
}
\examples{
CatSep()

CatSep(2)

CatSep(2,2,3)

CatSep(2,2,3, Char = "*")

CatSep(2,2,3, Char = "*", CharReps = 20)
}
\author{
Ahmed El-Gabbas
}
