% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{CatTime}
\alias{CatTime}
\title{Print text and a time stamp}
\usage{
CatTime(Text = NULL, NLines = 1, ...)
}
\arguments{
\item{Text}{the text to print}

\item{NLines}{number of empty lines after the printing; default: 1}

\item{...}{other arguments passed to \code{cat}}
}
\description{
Print text and a time stamp
}
\examples{
CatTime()
CatTime("Time now")
}
\author{
Ahmed El-Gabbas
}
