% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CatTime.R
\name{CatTime}
\alias{CatTime}
\title{Print text with time stamp}
\usage{
CatTime(Text = "", NLines = 1, Date = FALSE, TZ = "CET", ...)
}
\arguments{
\item{Text}{character; the text to print before the timestamp. If empty
(default), only the timestamp is printed.}

\item{NLines}{integer; the number of newline characters to print after the
message. Default is 1.}

\item{Date}{logical; whether to include the date in the timestamp. Default is
\code{FALSE}, meaning only the time is printed.}

\item{TZ}{character; the time zone to use for the timestamp. Default is
\code{CET}.}

\item{...}{additional arguments passed to \code{cat}.}
}
\value{
The function is called for its side effect of printing to the
console.
}
\description{
This function prints a given text followed by the current time (and
optionally the date) to the console. It allows for customization of the time
zone, the inclusion of the date, and the number of newline characters to
print after the message.
}
\examples{
CatTime()

CatTime(Date = TRUE)

CatTime("Time now")

CatTime("Time now", Date = TRUE)
}
\author{
Ahmed El-Gabbas
}
