% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckArgs.R
\name{CheckArgs}
\alias{CheckArgs}
\title{Check function arguments for specific types}
\usage{
CheckArgs(AllArgs, Args, Type)
}
\arguments{
\item{AllArgs}{Character vector. Input parameters of the function.
Usually as a result of \code{formals()} function}

\item{Args}{Character vector. Names of the arguments to be checked.}

\item{Type}{Character. The expected type of the arguments. Must be
one of "character", "logical", or "numeric".}
}
\value{
The function does not return a value but will stop execution and
throw an error if any of the specified arguments do not match the expected
type.
}
\description{
This function checks if the specified arguments of a function match the
expected type. It is useful for validating function inputs.
}
\author{
Ahmed El-Gabbas
}
