% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckCommands.R
\name{CheckCommands}
\alias{CheckCommands}
\title{Check system commands availability}
\usage{
CheckCommands(Commands)
}
\arguments{
\item{Commands}{A character vector of system command names to check (e.g.,
\code{c("git", "Rscript", "unzip")}).}
}
\value{
The function returns \code{TRUE} if all specified commands are available
on the system, \code{FALSE} if any is not available.
}
\description{
This function checks if a list of system commands are available on the user's
PATH. If any commands are missing, it stops execution and returns an
informative error message.
}
\examples{
CheckCommands(c("unzip", "head", "curl"))
}
\author{
Ahmed El-Gabbas
}
