% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckQuartoVersion.R
\name{CheckQuartoVersion}
\alias{CheckQuartoVersion}
\title{Check if \code{Quarto} should be updated}
\usage{
CheckQuartoVersion()
}
\description{
Check if \code{Quarto} should be updated
}
\examples{
CheckQuartoVersion()
}
\author{
Ahmed El-Gabbas
}
