% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckRData.R
\name{CheckRData}
\alias{CheckRData}
\title{Check if a file is a valid RData file}
\usage{
CheckRData(File)
}
\arguments{
\item{File}{character. The file path of the file to be checked. This can not
be empty.}
}
\value{
A logical value indicating if the file is a valid RData file
}
\description{
This function checks if a given file is a valid RData file by verifying its
extension and attempting to load it. It uses the \link{LoadAs} to load the file.
If the file loads successfully and the object exists, it returns \code{TRUE}. If
the file cannot be loaded or the object does not exist, it returns \code{FALSE}.
}
\author{
Ahmed El-Gabbas
}
