% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckRDS.R
\name{CheckRDS}
\alias{CheckRDS}
\title{Check if an RDS file can be successfully read}
\usage{
CheckRDS(File)
}
\arguments{
\item{File}{A character string specifying the path to the file to be checked.}
}
\value{
A logical value: \code{TRUE} if the file is an RDS file and contains a
non-null object, \code{FALSE} otherwise.
}
\description{
This function checks if the specified file is an RDS file and attempts to
read it. It returns \code{TRUE} if the file is an RDS file and contains a non-null
object, otherwise \code{FALSE.} If the file is not an RDS file, the function stops
with an error message.
}
\author{
Ahmed El-Gabbas
}
