% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckRStudioVersion.R
\name{CheckRStudioVersion}
\alias{CheckRStudioVersion}
\title{Check if the installed RStudio version is up to date}
\usage{
CheckRStudioVersion()
}
\value{
Side effects include printing messages to the console regarding the status of RStudio version.
}
\description{
This function checks the current installed version of RStudio against the latest version available online. If the versions do not match, it suggests updating RStudio.
}
\note{
This function requires internet access to check the latest version of RStudio online. If called outside of RStudio, it will only fetch and display the latest version without comparing.
}
\examples{
try(CheckRStudioVersion())
}
\author{
Ahmed El-Gabbas
}
