% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_CheckZip.R
\name{CheckZip}
\alias{CheckZip}
\title{Check the Integrity of a ZIP File}
\usage{
CheckZip(File)
}
\arguments{
\item{File}{Character. The path to the ZIP file that needs to be checked.}
}
\value{
Logical. Returns \code{TRUE} if the file exists and passes the integrity
check, and \code{FALSE} otherwise.
}
\description{
This function checks if a given ZIP file exists and whether it is intact
without any errors. The function uses the system's \code{unzip} command to test
the file's integrity.
}
\author{
Ahmed El-Gabbas
}
