% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_Chelsa_Extract_Matching.R
\name{Chelsa_Extract_Matching}
\alias{Chelsa_Extract_Matching}
\title{Extract time / climate models & scenarios from Chelsa URL}
\usage{
Chelsa_Extract_Matching(String, Time, Matches)
}
\arguments{
\item{String}{URL string}

\item{Time}{Time period to match}

\item{Matches}{further string to match}
}
\description{
A function to extract which climate model and scenario each file represent
}
\details{
A function to extract which climate model and scenario each file represent
}
\author{
Ahmed El-Gabbas
}
