% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_Chelsa_Info.R
\name{Chelsa_Info}
\alias{Chelsa_Info}
\title{Extracts and summarizes information from CHELSA climate model file names}
\usage{
Chelsa_Info(FileName)
}
\arguments{
\item{FileName}{character vector; a vector of CHELSA file names or paths.
Each element should be a string representing the file name or path of a
CHELSA climate model output file.}
}
\value{
A data frame with columns for the original file name (\code{FileName}),
the climate variable (\code{Variable}), the time period (\code{TimePeriod}), the
climate model (\code{ClimModel}), and the climate scenario (\code{ClimScenario}).
Additional variable information is joined from an external source
([IASDT.R::Chelsa_Vars`).
}
\description{
This function parses CHELSA climate model file names to extract and summarize
information such as the time period, climate model, climate scenario, and
variable represented in the file. It returns a data frame with these details
for each file name provided.
}
\examples{
BaseURL <- "https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/"
LFiles <- paste0(
 BaseURL,
 c(
   #'"2041-2070/MPI-ESM1-2-HR/ssp126/bio/CHELSA_bio14_2041-2070_mpi-esm1-2-hr_ssp126_V.2.1.tif",
   "2011-2040/MRI-ESM2-0/ssp126/bio/CHELSA_bio1_2011-2040_mri-esm2-0_ssp126_V.2.1.tif",
   "2011-2040/IPSL-CM6A-LR/ssp370/bio/CHELSA_lgd_2011-2040_ipsl-cm6a-lr_ssp370_V.2.1.tif",
   "2041-2070/MRI-ESM2-0/ssp370/bio/CHELSA_bio5_2041-2070_mri-esm2-0_ssp370_V.2.1.tif",
   "2011-2040/MRI-ESM2-0/ssp585/bio/CHELSA_scd_2011-2040_mri-esm2-0_ssp585_V.2.1.tif"))

Chelsa_Info(FileName = LFiles[1])

Chelsa_Info(FileName = LFiles)
}
\author{
Ahmed El-Gabbas
}
