% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_CHELSA_Prepare.R
\name{CHELSA_Prepare}
\alias{CHELSA_Prepare}
\title{Prepare and download CHELSA climate data}
\usage{
CHELSA_Prepare(
  EnvFile = ".env",
  FromHPC = TRUE,
  Download = FALSE,
  NCores = 4L,
  Overwrite = FALSE,
  Download_Attempts = 10L,
  Sleep = 5L,
  OtherVars = "npp",
  BaseURL = paste0("https://os.zhdk.cloud.switch.ch/envicloud/",
    "chelsa/chelsa_V2/GLOBAL/")
)
}
\arguments{
\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{Download}{Logical. Whether to download the CHELSA files. Defaults to
\code{FALSE}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 4.}

\item{Overwrite}{Logical. Whether to re-download files that already exist.
Defaults to \code{FALSE}.}

\item{Download_Attempts}{Integer. The maximum number of download attempts.
Defaults to 10.}

\item{Sleep}{Integer. Time in seconds to wait after each download attempt.
Defaults to 5.}

\item{OtherVars}{Character. First letters of variables other than bioclimatic
variables to be processed. Defaults to "npp", which means processing
bioclimatic variables (bio1-19) and Net Primary Productivity will be
processed. If \code{OtherVars} is set to "", only bioclimatic variables will be
processed. See \link{CHELSA_Vars} for a detailed information on CHELSA climate
variables.}

\item{BaseURL}{Character. The base URL for downloading CHELSA climate data.}
}
\description{
This function reads the contents of local text files containing URLs for
\code{CHELSA} climate data and extracts and saves their metadata (e.g., variable
name, climate model & scenario, time, and download URLs). It can optionally
download these links to desk, if requested.
}
\details{
The function returns information on 874 tif files, representing 19
bioclimatic variables (+ additional variables starting with the string in
\code{OtherVars}) at 46 climate options (current and 45 future scenarios)
}
\author{
Ahmed El-Gabbas
}
