% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Chelsa_Prepare_List}
\alias{Chelsa_Prepare_List}
\title{Prepare list of potential variables}
\usage{
Chelsa_Prepare_List(
  Down = FALSE,
  DownParallel = TRUE,
  DwnPath = NULL,
  OutPath = NULL,
  UpdateExisting = FALSE,
  Path_Chelsa = "Data/Chelsa"
)
}
\arguments{
\item{Down}{Should the Chelsa files be downloaded}

\item{DownParallel}{if \code{Down} was set as \code{TRUE}, should the download be on parallel}

\item{DwnPath}{Path for download}

\item{OutPath}{Out path}

\item{UpdateExisting}{reprocess existing file}

\item{Path_Chelsa}{Path for Chelsa files}
}
\description{
Prepare list of potential variables
}
\details{
list of variables exist under current and future climates.

46 variables available at 46 options (current and 45 future scenarios)
}
\author{
Ahmed El-Gabbas
}
