% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Chelsa_Prepare_List}
\alias{Chelsa_Prepare_List}
\title{Prepare list of potential variables}
\usage{
Chelsa_Prepare_List(
  Down = FALSE,
  DownParallel = TRUE,
  DwnPath = NULL,
  OutPath = NULL,
  UpdateExisting = FALSE,
  Path_Chelsa = "Data/Chelsa"
)
}
\arguments{
\item{Down}{logical; Download Chelsa files?}

\item{DownParallel}{logical; Download input files on parallel (if \code{Down} = \code{TRUE})}

\item{DwnPath}{Download path}

\item{OutPath}{Output path}

\item{UpdateExisting}{logical; Reprocess existing file?}

\item{Path_Chelsa}{Path for Chelsa analyses (including \code{Chelsa_Vars.txt} file)}
}
\description{
Prepare list of potential variables
}
\details{
list of variables exist under current and future climates.

46 variables available at 46 options (current and 45 future scenarios)
}
\author{
Ahmed El-Gabbas
}
