% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Chelsa_Process}
\alias{Chelsa_Process}
\title{Processing Chelsa data}
\usage{
Chelsa_Process(
  InputFile = NULL,
  OutFile = NULL,
  GridFile = NULL,
  ReturnMap = FALSE,
  Verbose = FALSE
)
}
\arguments{
\item{InputFile}{Input tif file}

\item{OutFile}{Output tif file}

\item{GridFile}{Path for the \verb{*.RData} file containing the reference grid. This grid will be used as reference grid for projection and the resulted file will be masked to it}

\item{ReturnMap}{logical; should the processed map be returned by the end of the function?}

\item{Verbose}{should the name of the processed file be printed to the console}
}
\value{
A map projected to \code{EPSG:3035} and cropped to the study area (masked by the \code{GridFile})
}
\description{
Processing Chelsa data
}
\author{
Ahmed El-Gabbas
}
