% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Chelsa_Process}
\alias{Chelsa_Process}
\title{Processing Chelsa data}
\usage{
Chelsa_Process(InputPath, OutPath, GridFile, Verbose = FALSE)
}
\arguments{
\item{InputPath}{Input path (all \verb{*.tif} files in this path will be processed)}

\item{OutPath}{Output path. Processed objects will be saved to this path (with the same original file name)}

\item{GridFile}{Path for the \verb{*.RData} file containing the reference grid. This grid will be used as reference grid for projection and the resulted file will be masked to it}

\item{Verbose}{should the name of the processed file be printed to the console}
}
\description{
Processing Chelsa data
}
\author{
Ahmed El-Gabbas
}
