% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_CHELSA_Process.R
\name{CHELSA_Process}
\alias{CHELSA_Process}
\title{Process CHELSA Climate Data}
\usage{
CHELSA_Process(
  FromHPC = TRUE,
  EnvFile = ".env",
  NCores = 8,
  Download = FALSE,
  Overwrite = FALSE,
  Download_Attempts = 10,
  Sleep = 5,
  OtherVars = "",
  BaseURL = paste0("https://os.zhdk.cloud.switch.ch/envicloud/",
    "chelsa/chelsa_V2/GLOBAL/"),
  Download_NCores = 4,
  CompressLevel = 5,
  OverwriteProcessed = FALSE
)
}
\arguments{
\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Defaults to 8.}

\item{Download}{Logical, whether to download the CHELSA files. Defaults to
\code{FALSE}.}

\item{Overwrite}{Logical, whether to re-download files that already exist.
Defaults to \code{FALSE}.}

\item{Download_Attempts}{Integer. The maximum number of download attempts.
Defaults to 10.}

\item{Sleep}{Integer. Time in seconds to wait after each download attempt.
Defaults to 5.}

\item{OtherVars}{Character. First letters of variables other than bioclimatic
variables to be processed. Defaults to "npp", which means processing
bioclimatic variables (bio1-19) and Net Primary Productivity will be
processed. If \code{OtherVars} is set to "", only bioclimatic variables will be
processed. See \link{CHELSA_Vars} for a detailed information on CHELSA climate
variables.}

\item{BaseURL}{String, the base URL for downloading CHELSA climate data.}

\item{Download_NCores}{integer. Number of CPU cores to use for parallel
downloading of CHELSA data. Only valid if Download = \code{TRUE}. Defaults to 4.}

\item{CompressLevel}{Integer; the level of compression to use when saving
NetCDF files. Default is 5.}

\item{OverwriteProcessed}{Logical; whether to overwrite already processed
files. Default is \code{FALSE}.}
}
\description{
This function processes CHELSA climate data, with an option to download them
first. It processes each variable to the European scale and reference grid,
and outputs in TIFF and NetCDF (NC) formats. It also saves grouped data for
each of the 46 climate scenarios.
}
\author{
Ahmed El-Gabbas
}
