% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Chelsa_Project}
\alias{Chelsa_Project}
\title{Project Chelsa data to the study area}
\usage{
Chelsa_Project(
  InputFile = NULL,
  OutFile = NULL,
  GridFile = NULL,
  ReturnMap = FALSE,
  DownPath = NULL,
  KeepDownloaded = TRUE,
  SaveTiff = FALSE
)
}
\arguments{
\item{InputFile}{character; file path or URL for input tif file}

\item{OutFile}{character; file path for output nc/tif file. File extension will be replaced when needed.}

\item{GridFile}{reference grid; either as \code{RasterLayer}, \code{SpatRaster}, or \code{PackedSpatRaster}. The reference grid will be used for projection and the resulted maps will be masked to it}

\item{ReturnMap}{logical; should the processed map be returned by the end of the function? Default: \code{FALSE}}

\item{DownPath}{character; Where to save downloaded files (if the \code{InputFile} is a URL)}

\item{KeepDownloaded}{logical; if URL is provided as input file, the file will be downloaded to disk first before processing. Should the downloaded file be kept in disk. Default: \code{FALSE}}

\item{SaveTiff}{logical; also save output map as *.tif file. Default: \code{FALSE}}
}
\value{
If \code{ReturnMap = TRUE}, a wrapped SpatRaster object (\code{PackedSpatRaster}) is returned; otherwise nothing is returned. By default the function exports a NetCDF file. If \code{SaveTiff} is set as \code{TRUE}, additional tiff file will be saved to disk.
}
\description{
Project Chelsa data to the study area
}
\author{
Ahmed El-Gabbas
}
