% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Chelsa_Project}
\alias{Chelsa_Project}
\title{Project Chelsa data}
\usage{
Chelsa_Project(
  InputFile = NULL,
  OutFile = NULL,
  GridFile = NULL,
  ReturnMap = FALSE,
  Verbose = FALSE,
  KeepDownloaded = TRUE
)
}
\arguments{
\item{InputFile}{Path or URL for input tif file}

\item{OutFile}{Path for output tif file}

\item{GridFile}{\code{raster} or \code{SpatRaster} for the reference grid. This grid will be used as reference grid for projection and the resulted file will be masked to it}

\item{ReturnMap}{logical; should the processed map be returned by the end of the function?}

\item{Verbose}{should the name of the processed file be printed to the console}

\item{KeepDownloaded}{if URL is provided as input file, the file will be downloaded to disk first before processing. Should the downloaded file be kept in disk.}
}
\value{
if \code{ReturnMap = TRUE}, the output raster object is returned; otherwise nothing is returned. The \verb{*.tif} files are saved to disk anyway.
}
\description{
Project Chelsa data
}
\author{
Ahmed El-Gabbas
}
