% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Chelsa_Project}
\alias{Chelsa_Project}
\title{Project Chelsa data to the study area}
\usage{
Chelsa_Project(
  InputFile = NULL,
  OutFile = NULL,
  GridFile = NULL,
  ReturnMap = FALSE,
  DownPath = NULL,
  KeepDownloaded = TRUE,
  SaveTiff = FALSE
)
}
\arguments{
\item{InputFile}{Path or URL for input tif file}

\item{OutFile}{Path for output tif file}

\item{GridFile}{\code{raster} or \code{SpatRaster} for the reference grid. This grid will be used as reference grid for projection and the resulted file will be masked to it}

\item{ReturnMap}{logical; should the processed map be returned by the end of the function?}

\item{DownPath}{Where to save downloaded files}

\item{KeepDownloaded}{if URL is provided as input file, the file will be downloaded to disk first before processing. Should the downloaded file be kept in disk.}

\item{SaveTiff}{Also save output map as *.tif file}
}
\value{
if \code{ReturnMap = TRUE}, the a wrapped SpatRaster object is returned (\code{PackedSpatRaster}); otherwise nothing is returned. By default the function exports a NetCDF file. If \code{SaveTiff} is set as \code{TRUE}, additional tiff file will be saved to disk.
}
\description{
Project Chelsa data to the study area
}
\author{
Ahmed El-Gabbas
}
