% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_CHELSA_variables.R
\docType{data}
\name{CHELSA_variables}
\alias{CHELSA_variables}
\title{Detailed information on CHELSA climate variables}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 6 columns.
}
\usage{
CHELSA_variables
}
\value{
A tibble containing detailed information about various climate
variables in \code{CHELSA} climate data. Each row in the tibble represents a
different climate variable, with columns providing additional details
such as the long name, unit, scale, offset, and an explanation of the
variable
}
\description{
Detailed information on CHELSA climate variables
}
\examples{
library(dplyr)
data(CHELSA_variables)
print(CHELSA_variables, n = Inf)

}
\keyword{datasets}
