% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwf_clc_process.R
\name{clc_process}
\alias{clc_process}
\title{Process Corine Land Cover (CLC) data for the \code{IASDT}}
\usage{
clc_process(env_file = ".env", min_land_percent = 15L, plot_clc = TRUE)
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{min_land_percent}{Numeric. Minimum land percentage per grid cell for
the reference grid. Default: \code{15}.}

\item{plot_clc}{Logical. If \code{TRUE}, plots percentage coverage for CLC levels
and habitat types. Default: \code{TRUE}.}
}
\value{
Returns \code{invisible(NULL)}; saves processed data and optional plots to
disk.
}
\description{
Processes \href{https://land.copernicus.eu/pan-european/corine-land-cover/clc2018}{Corine Land Cover (CLC)}
data for the Invasive Alien Species Digital Twin (\code{IASDT}). Calculates
percentage coverage and most common classes per grid cell at three CLC
levels, plus \code{eunis19} and \code{SynHab} habitat types. Prepares a reference grid
and optionally generates percentage coverage maps as JPEG.
}
\references{
\itemize{
\item Data source:
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc2018}
\item Data citation:
\url{https://doi.org/10.2909/960998c1-1870-4e82-8051-6485205ebbac}
}
}
\author{
Ahmed El-Gabbas
}
