% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_ClearConsole.R
\name{ClearConsole}
\alias{ClearConsole}
\title{Clear the console}
\usage{
ClearConsole()
}
\description{
Clear the console. This function is equivalent to \code{cat("\\014")}
}
\examples{
\dontrun{
ClearConsole()
}
}
