% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_ClipRasterByPolygon.R
\name{ClipRasterByPolygon}
\alias{ClipRasterByPolygon}
\title{Clip raster by a spatial polygon}
\usage{
ClipRasterByPolygon(raster = NULL, shape = NULL)
}
\arguments{
\item{raster}{raster layer}

\item{shape}{Polygon}
}
\description{
Clip raster by a spatial polygon
}
\examples{
LoadPackages(sp)
LoadPackages(raster)
LoadPackages(rworldmap)

# Example Polygon
SPDF <- getMap(resolution = "low") \%>\%
   subset(NAME == "Germany")

# Example RasterLayer
r <- raster::raster(nrow = 1e3, ncol = 1e3, crs = proj4string(SPDF))
r[] <- 1:length(r)
plot(r)
plot(SPDF, add = TRUE)

# ----------------------------------

SPDF_DE <- ClipRasterByPolygon(r, SPDF)
plot(raster::extent(SPDF_DE), axes = FALSE, xlab = "", ylab = "")
plot(SPDF_DE, add = TRUE)
plot(SPDF, add = TRUE)
}
