% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Coda_to_tibble.R
\name{Coda_to_tibble}
\alias{Coda_to_tibble}
\title{Convert \code{mcmc.list} to tibble}
\usage{
Coda_to_tibble(CodaObj = NULL, Type = NULL, EnvFile = ".env", NOmega = 1000)
}
\arguments{
\item{CodaObj}{\code{mcmc.list} object}

\item{Type}{String. Parameter type. It has to be one of \code{rho}, \code{alpha}, \code{omega}, or \code{beta}.}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{NOmega}{Integer. Number of species to be sampled for the Omega parameter}
}
\description{
Convert \code{mcmc.list} to tibble
}
\author{
Ahmed El-Gabbas
}
