% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_coda_to_tibble.R
\name{coda_to_tibble}
\alias{coda_to_tibble}
\title{Convert a Coda object to a tibble with specified parameter transformations}
\usage{
coda_to_tibble(
  coda_object = NULL,
  posterior_type = NULL,
  env_file = ".env",
  n_omega = 100
)
}
\arguments{
\item{coda_object}{An object of class \code{mcmc.list} or \code{mcmc}, representing the
MCMC output.}

\item{posterior_type}{Character. The parameter type to transform and extract.
Must be one of \code{rho}, \code{alpha}, \code{omega}, or \code{beta}.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_omega}{Integer. The number of species to be sampled for the \code{Omega}
parameter transformation. Defaults to 100.}
}
\value{
A tibble containing the transformed parameters based on the specified
\code{posterior_type}. The structure of the returned tibble varies depending on
the \code{posterior_type} parameter.
}
\description{
This function converts a Coda object (\code{mcmc.list} or \code{mcmc}) into a tibble
format, facilitating further analysis and visualisation. It supports
transformation for specific parameter types: \code{rho}, \code{alpha}, \code{omega}, and
\code{beta}.
}
\examples{
ecokit::load_packages(Hmsc, coda, dplyr)

coda_object <- Hmsc::convertToCodaObject(Hmsc::TD$m)
IASDT.R::coda_to_tibble(
   coda_object = coda_object$Alpha[[1]], posterior_type = "Alpha")
}
\author{
Ahmed El-Gabbas
}
