% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Convergence_Plot_All.R
\name{Convergence_Plot_All}
\alias{Convergence_Plot_All}
\title{Plot model convergence of multiple modelling alternatives}
\usage{
Convergence_Plot_All(
  ModelDir = NULL,
  maxOmega = 1000L,
  NCores = NULL,
  FromHPC = TRUE,
  MarginType = "histogram"
)
}
\arguments{
\item{ModelDir}{Character. Path to the root directory of the fitted models.
The convergence outputs will be saved to the \code{Model_Convergence_All}
subfolder.}

\item{maxOmega}{Integer. Maximum number of species interactions to sample for
convergence diagnostics.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{MarginType}{Character. The type of marginal plot to add to
the main plot. Valid options are "histogram" (default) or "density".}
}
\value{
The function does not return anything but saves a series of
diagnostic plots in the specified path.
}
\description{
This function generates and saves a series of diagnostic plots to assess the
convergence of Hmsc models across multiple modelling alternatives. It checks
model convergence using trace plots and Gelman-Rubin diagnostics for key
model parameters.
}
\author{
Ahmed El-Gabbas
}
