% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convergence_plot_all.R
\name{convergence_plot_all}
\alias{convergence_plot_all}
\title{Plot model convergence of multiple modelling alternatives}
\usage{
convergence_plot_all(
  model_dir = NULL,
  n_omega = 1000L,
  n_cores = NULL,
  margin_type = "histogram"
)
}
\arguments{
\item{model_dir}{Character. Path to the root directory of the fitted model.
The convergence outputs will be saved to the
\code{Model_Convergence_All} subfolder.}

\item{n_omega}{Integer. Number of species interactions sampled for Omega
parameter diagnostics. Default: 1000L}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.}

\item{margin_type}{Character. The type of marginal plot to add to the main
plot. Valid options are "histogram" (default) or "density".}
}
\value{
The function does not return anything but saves a series of
diagnostic plots in the specified path.
}
\description{
This function generates and saves a series of diagnostic plots to assess the
convergence of Hmsc models across multiple modelling alternatives. It checks
model convergence using trace plots and Gelman-Rubin diagnostics for key
model parameters.
}
\author{
Ahmed El-Gabbas
}
