% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convergence_plot_all.R
\name{convergence_plot_all}
\alias{convergence_plot_all}
\title{Plot model convergence of multiple modelling alternatives}
\usage{
convergence_plot_all(
  model_dir = NULL,
  n_omega = 1000L,
  margin_type = "histogram",
  spatial_model = TRUE,
  n_cores = NULL,
  strategy = "multisession"
)
}
\arguments{
\item{model_dir}{Character. Path to the root directory of the fitted model.
The convergence outputs will be saved to the \code{Model_Convergence_All}
subdirectory.}

\item{n_omega}{Integer. Number of species interactions sampled for Omega
parameter diagnostics. Default: 1000L}

\item{margin_type}{Character. The type of marginal plot to add to the main
plot. Valid options are "histogram" (default) or "density".}

\item{spatial_model}{Logical. Whether the model is a spatial model. If \code{TRUE}
(default), the function will generate additional plots for the model's
\code{Alpha} parameter.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}
}
\value{
The function returns \code{invisible(NULL)} and does not return any value,
but saves a series of diagnostic plots in the specified path.
}
\description{
This function generates and saves a series of diagnostic plots to assess the
convergence of Hmsc models across multiple modelling alternatives. It checks
model convergence using trace plots and Gelman-Rubin diagnostics for key
model parameters.
}
\author{
Ahmed El-Gabbas
}
