% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convergence_plot.R,
%   R/mod_convergence_alpha.R, R/mod_convergence_rho.R
\name{convergence_plots}
\alias{convergence_plots}
\alias{convergence_plot}
\alias{convergence_alpha}
\alias{convergence_rho}
\alias{convergence_beta_ranges}
\title{Plot model convergence of a selected model}
\usage{
convergence_plot(
  path_coda = NULL,
  env_file = ".env",
  title = " ",
  n_omega = 1000L,
  n_cores = 8L,
  strategy = "multisession",
  future_max_size = 2000L,
  n_RC = c(2L, 2L),
  beta_n_RC = c(3L, 3L),
  pages_per_file = 20L,
  chain_colors = NULL,
  margin_type = "histogram",
  spatial_model = TRUE
)

convergence_alpha(
  posterior = NULL,
  title = NULL,
  n_RC = NULL,
  add_footer = TRUE,
  add_title = TRUE,
  chain_colors = NULL,
  margin_type = "histogram",
  n_chains = NULL,
  n_samples = NULL
)

convergence_rho(
  posterior = NULL,
  title = NULL,
  chain_colors = NULL,
  margin_type = "histogram",
  n_chains = NULL,
  n_samples = NULL
)

convergence_beta_ranges(model_dir = NULL, beta_data = NULL, n_chains = NULL)
}
\arguments{
\item{path_coda}{Character. Path to a saved coda object containing MCMC
samples.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{title}{Character. title for \strong{rho} and \strong{alpha} convergence plots.
Default: " "}

\item{n_omega}{Integer. Number of species interactions sampled for Omega
parameter diagnostics. Default: 1000L}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{future_max_size}{Numeric. Maximum allowed total size (in megabytes) of
global variables identified. See \code{future.globals.maxSize} argument of
\link[future:zzz-future.options]{future::future.options} for more details.}

\item{n_RC}{Numeric vector. Grid layout (rows×columns) for arranging
alpha parameter plots. Default: \code{c(2, 2)}. If \code{NULL}, the layout is
automatically determined based on the number of alpha levels.}

\item{beta_n_RC}{Numeric vector. The grid layout (rows×columns) for
arranging beta parameter plots. Default: \code{c(3, 3)}.}

\item{pages_per_file}{Integer. Number of plots per page in the Omega
parameter output. Default: 20L.}

\item{chain_colors}{Character vector. MCMC chain colours (optional). Default:
\code{NULL}.}

\item{margin_type}{Character. The type of marginal plot to add to the main
plot. Valid options are "histogram" (default) or "density".}

\item{spatial_model}{Logical. Whether the model is a spatial model. If \code{TRUE}
(default), the function will generate additional plots for the model's
\code{Alpha} parameter.}

\item{posterior}{\code{mcmc.list} or character. Either an MCMC object
(\code{mcmc.list}) containing posterior samples, or a file path to a saved coda
object.}

\item{add_footer}{Logical. If \code{TRUE} (default), adds a footer with page
numbers to each plot.}

\item{add_title}{Logical. If \code{TRUE} (default), adds the main title (\code{title})
to the plot.}

\item{n_chains}{Integer. Number of MCMC chains.}

\item{n_samples}{Integer. Number of MCMC samples.}

\item{model_dir}{Character. A path to the model directory.}

\item{beta_data}{Data frame. Beta parameter summary data frame.}
}
\description{
The \code{convergence_plot()} function generates and saves convergence diagnostics
plots for the \code{rho}, \code{alpha}, \code{omega}, and \code{beta} parameters in an Hmsc
model. These plots help assess whether the MCMC chains have reached
stationarity. It supports parallel processing and can work with models fitted
on HPC environments.
}
\details{
\code{convergence_alpha()}, \code{convergence_rho()}, and
\code{convergence_beta_ranges} are internal functions and should not be called
directly. The \code{convergence_beta_ranges} plots the convergence range of the
each species beta parameters. It can be used to check if any of the chains
show convergence issues; i.e., showing exceptionally high or low beta
values.
}
\author{
Ahmed El-Gabbas
}
