% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convergence_plot.R,
%   R/mod_convergence_alpha.R, R/mod_convergence_rho.R
\name{convergence_plots}
\alias{convergence_plots}
\alias{convergence_plot}
\alias{convergence_alpha}
\alias{convergence_rho}
\alias{convergence_Beta_ranges}
\title{Plot model convergence of a selected model}
\usage{
convergence_plot(
  path_coda = NULL,
  path_model = NULL,
  env_file = ".env",
  title = " ",
  n_omega = 1000L,
  n_cores = 8L,
  strategy = "future::multicore",
  n_RC = c(2L, 2L),
  beta_n_RC = c(3L, 3L),
  save_plotting_data = TRUE,
  pages_per_file = 20L,
  chain_colors = NULL,
  margin_type = "histogram"
)

convergence_alpha(
  posterior = NULL,
  model_object = NULL,
  title = NULL,
  n_RC = NULL,
  add_footer = TRUE,
  add_title = TRUE,
  chain_colors = NULL,
  margin_type = "histogram"
)

convergence_rho(
  posterior = NULL,
  model_object = NULL,
  title = NULL,
  chain_colors = NULL,
  margin_type = "histogram"
)

convergence_Beta_ranges(model_dir)
}
\arguments{
\item{path_coda}{Character. Path to a saved coda object containing MCMC
samples.}

\item{path_model}{Character. Path to a saved Hmsc model object.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{title}{Character. title for \strong{rho} and \strong{alpha} convergence plots.
Default: " "}

\item{n_omega}{Integer. Number of species interactions sampled for Omega
parameter diagnostics. Default: 1000L}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "future::sequential", "future::multisession",
"future::multicore", and "future::cluster". Defaults to
\code{"future::multicore"} (\code{"future::multisession"} on Windows). See
\code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{n_RC}{Numeric vector. Grid layout (rows×columns) for arranging
alpha parameter plots. Default: \code{c(2, 2)}. If \code{NULL}, the layout is
automatically determined based on the number of alpha levels.}

\item{beta_n_RC}{Numeric vector. The grid layout (rows×columns) for
arranging beta parameter plots. Default: \code{c(3, 3)}.}

\item{save_plotting_data}{Logical. If \code{TRUE} (default), saves the plot data
as \code{.RData} files.}

\item{pages_per_file}{Integer. Number of plots per page in the Omega
parameter output. Default: 20L.}

\item{chain_colors}{Character vector. MCMC chain colours (optional). Default:
\code{NULL}.}

\item{margin_type}{Character. The type of marginal plot to add to the main
plot. Valid options are "histogram" (default) or "density".}

\item{posterior}{\code{mcmc.list} or character. Either an MCMC object
(\code{mcmc.list}) containing posterior samples, or a file path to a saved coda
object.}

\item{model_object}{\code{Hmsc} object or character. Either a fitted Hmsc model
object, or a file path to a saved Hmsc model.}

\item{add_footer}{Logical. If \code{TRUE} (default), adds a footer with page
numbers to each plot.}

\item{add_title}{Logical. If \code{TRUE} (default), adds the main title (\code{title})
to the plot.}

\item{model_dir}{Character. A path to the model directory.}
}
\description{
The \code{convergence_plot()} function generates and saves convergence diagnostics
plots for the \code{rho}, \code{alpha}, \code{omega}, and \code{beta} parameters in an Hmsc
model. These plots help assess whether the MCMC chains have reached
stationarity. It supports parallel processing and can work with models fitted
on HPC environments.
}
\details{
\code{convergence_alpha()}, \code{convergence_rho()}, and
\code{convergence_Beta_ranges} are internal functions and should not be called
directly. The \code{convergence_Beta_ranges} plots the convergence range of the
each species beta parameters. It can be used to check if any of the chains
show convergence issues; i.e., showing exceptionally high or low beta
values.
}
\author{
Ahmed El-Gabbas
}
