% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{CurrOS}
\alias{CurrOS}
\title{Current operating system}
\usage{
CurrOS()
}
\value{
String for the current operating system
}
\description{
Current operating system
}
\examples{
CurrOS()

}
\author{
Ahmed El-Gabbas
}
