% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{DirCreate}
\alias{DirCreate}
\title{Create directory if not existed}
\usage{
DirCreate(Path)
}
\arguments{
\item{Path}{Path of the folder}
}
\description{
Create directory if not existed
}
\author{
Ahmed El-Gabbas
}
