% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{DirCreate}
\alias{DirCreate}
\title{Create directory if not existed}
\usage{
DirCreate(Path, Verbose = TRUE)
}
\arguments{
\item{Path}{Path of the folder}

\item{Verbose}{logical; print a message whether the folder was created or already available. Default: \code{TRUE}}
}
\description{
Create directory if not existed
}
\examples{
# create new folder (random name) in the temporary folder
LoadPackages(stringi)
Path2Create <- file.path(tempdir(), stri_rand_strings(1, 5))
file.exists(Path2Create)

DirCreate(Path2Create)
DirCreate(Path2Create)
DirCreate(Path2Create, Verbose = FALSE)
file.exists(Path2Create)

}
\author{
Ahmed El-Gabbas
}
