% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_DownBoundary.R
\name{DownBoundary}
\alias{DownBoundary}
\title{Determine the boundaries of the requested GBIF data}
\usage{
DownBoundary(L, R, B, T)
}
\arguments{
\item{L}{left boundary}

\item{R}{right boundary}

\item{B}{bottom boundary}

\item{T}{top boundary}
}
\value{
WKT string
}
\description{
\code{rgbif::pred_within()} function used to download GBIF data only accepts a WKT string. This function takes the values of the boundary and converts it to a WKT string. Default values are determined by the variables: Bound_L, R = Bound_R, Bound_B, Bound_T...
}
\details{
Determine the boundaries of the requested GBIF data
}
\examples{
IASDT.R::DownBoundary(20, 30, 40, 50)
}
\author{
Ahmed El-Gabbas
}
