% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_DownBoundary.R
\name{DownBoundary}
\alias{DownBoundary}
\title{Determine the boundaries of the requested GBIF data}
\usage{
DownBoundary(L = NULL, R = NULL, B = NULL, T = NULL)
}
\arguments{
\item{L, R, B, T}{Numeric, the left, right, bottom, and top boundary of the area.}
}
\value{
A character string representing the WKT of the polygon that outlines the specified boundaries.
}
\description{
\code{rgbif::pred_within()} function used to download GBIF data only accepts a WKT string. This function takes the values of the boundary and converts it to a WKT string.
}
\details{
This function constructs a Well-Known Text (WKT) string representing a polygon that outlines the specified boundaries. It is used to define the area of interest for downloading GBIF data through the \code{rgbif::pred_within()} function.
}
\examples{
IASDT.R::DownBoundary(Left = 20, Right = 30, Bottom = 40, Top = 50)
}
\author{
Ahmed El-Gabbas
}
