% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_EASIN_process.R, R/DWF_EASIN_taxonomy.R,
%   R/DWF_EASIN_download.R, R/DWF_EASIN_plot.R
\name{EASIN_data}
\alias{EASIN_data}
\alias{EASIN_process}
\alias{EASIN_taxonomy}
\alias{EASIN_download}
\alias{EASIN_plot}
\title{Process EASIN data for the \code{IASDT}}
\usage{
EASIN_process(
  extract_taxa = TRUE,
  extract_data = TRUE,
  n_download_attempts = 10L,
  n_cores = 6L,
  strategy = "multisession",
  sleep_time = 10L,
  n_search = 1000L,
  env_file = ".env",
  delete_chunks = TRUE,
  start_year = 1981L,
  plot = TRUE
)

EASIN_taxonomy(
  env_file = ".env",
  kingdom = "Plantae",
  phylum = "Tracheophyta",
  n_search = 100L
)

EASIN_download(
  species_key,
  timeout = 200,
  verbose = FALSE,
  env_file = ".env",
  n_search = 1000,
  n_attempts = 10,
  sleep_time = 5,
  delete_chunks = TRUE,
  return_data = FALSE
)

EASIN_plot(env_file = ".env")
}
\arguments{
\item{extract_taxa}{Logical. If \code{TRUE}, extracts taxonomy using
\code{EASIN_taxonomy()}. Default: \code{TRUE}.}

\item{extract_data}{Logical.If \code{TRUE}, downloads occurrence data with
\code{EASIN_download()}. Default: \code{TRUE}.}

\item{n_download_attempts}{Integer. Retry attempts for downloads. Default:
\code{10}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6. The maximum number of allowed cores are 8.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{sleep_time}{Integer. Number of seconds to pause between each data
retrieval request to prevent overloading the server. Default: 5 second.}

\item{n_search}{Integer. Number of records to attempt to retrieve per
request. Default: 1000, which is the current maximum allowed by the API.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{delete_chunks}{Logical. Whether to delete temporary files for data
chunks from the \code{file_parts} subdirectory. Defaults to \code{TRUE}.}

\item{start_year}{Integer. Earliest year for occurrence data (excludes
earlier records). Default: \code{1981} (aligned with CHELSA climate data).}

\item{plot}{Logical. If \code{TRUE}, generates plots via \code{EASIN_plot()}. Default:
\code{TRUE}.}

\item{kingdom}{Character. Taxonomic kingdom to query. Default: \code{"Plantae"}.}

\item{phylum}{Character. Taxonomic phylum within kingdom. Default:
\code{"Tracheophyta"}}

\item{species_key}{Character. EASIN taxon ID for which data is to be
retrieved. This parameter cannot be \code{NULL}.}

\item{timeout}{Integer. Download timeout in seconds. Default: \code{200}.}

\item{verbose}{Logical. If \code{TRUE}, prints progress messages. Default:
\code{FALSE}.}

\item{n_attempts}{Integer. Max download attempts per chunk. Default: \code{10}.}

\item{return_data}{Logical. If \code{TRUE}, returns data as a dataframe;
otherwise, saves to disk and returns \code{invisible(NULL)}. Default: \code{FALSE}.}
}
\description{
Extracts, processes, and visualises data from the \href{https://easin.jrc.ec.europa.eu/}{European Alien Species Information Network (EASIN)} for the
Invasive Alien Species Digital Twin (\code{IASDT}). Manages taxonomy, occurrence
data, and plots, handling API pagination and server limits. Orchestrated by
\code{EASIN_process()} with helpers \code{EASIN_taxonomy()}, \code{EASIN_download()}, and
\code{EASIN_plot()}.
}
\note{
Uses a static RDS file with EASIN-GBIF taxonomic standardization,
prepared by Marina Golivets (Feb 2024).
}
\section{Functions details}{

\itemize{
\item \strong{\code{EASIN_process()}}: Orchestrates taxonomy extraction, data downloads,
and plotting for EASIN species data.
\item \strong{\code{EASIN_taxonomy()}}: Fetches taxonomy data in chunks via the EASIN API,
filtered by kingdom and phylum. Returns a tibble.
\item \strong{\code{EASIN_download()}}: Downloads occurrence data for a given EASIN ID,
handling pagination and pauses. Returns a dataframe if \code{return_data = TRUE}, else \code{invisible(NULL)}.
\item \strong{\code{EASIN_plot()}}: Creates summary plots (observations count, species
count, distribution by partner) as JPEGs. Returns \code{invisible(NULL)}.
}
}

\author{
Ahmed El-Gabbas
}
