% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Sampling_Efforts.R
\name{Efforts_Download}
\alias{Efforts_Download}
\title{Download and Manage GBIF Data for Vascular Plant Orders}
\usage{
Efforts_Download(NCores = 6, Path_Raw, Path_Interim, Path_Efforts)
}
\arguments{
\item{NCores}{Integer. Number of cores to use for parallel processing.  Must
be a positive integer. This directory must exist or be created beforehand.}

\item{Path_Raw}{Character. Path where the raw downloaded data will be saved.
This directory must exist or be created beforehand.}

\item{Path_Interim}{Character. Path where the interim CSV files will be
saved. This directory must exist or be created beforehand.}

\item{Path_Efforts}{Character. Path where the final processed data will be
saved. This directory must exist or be created beforehand.}
}
\value{
A data frame (\code{Efforts_AllRequests}) with updated download paths and
interim file paths.
}
\description{
This function handles the downloading of GBIF data in parallel, checks the
validity of downloaded files, and stores the data in specified directories.
If data has already been downloaded, it validates the files instead of
downloading them again.
}
\author{
Ahmed El-Gabbas
}
