% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Sampling_Efforts.R
\name{Efforts_Plot}
\alias{Efforts_Plot}
\title{Plot the output of efforts maps}
\usage{
Efforts_Plot(Path_Efforts, EU_Bound)
}
\arguments{
\item{Path_Efforts}{Character. Path where the final processed data will be
saved.}

\item{EU_Bound}{Character. Path to file containing country boundaries.}
}
\value{
The function saves generated plots as JPEG files in the specified
directory and returns NULL invisibly.
}
\description{
This function generates and saves multiple plots of plant observation
efforts, both in raw and log10 scales, using provided spatial boundary and
summary data.
}
\details{
This function generates and saves effort maps visualizing the number
of plant observations and species, including both native and non-native
species, within Europe. It produces both standard and log10-scaled plots.
}
\author{
Ahmed El-Gabbas
}
