% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Sampling_Efforts.R
\name{Efforts_Process}
\alias{Efforts_Process}
\title{Process and Summarize GBIF Data for vascular plants}
\usage{
Efforts_Process(
  NCores,
  Path_Efforts,
  Path_Interim,
  Path_Data,
  Path_Grid,
  IAS_List,
  Efforts_AllRequests
)
}
\arguments{
\item{NCores}{Numeric. The number of cores to use for parallel processing.}

\item{Path_Efforts}{Character. Path where the final processed data will be
saved.}

\item{Path_Interim}{Character. The directory path to save interim data.}

\item{Path_Data}{Character. The directory path to save detailed effort data
as \code{RData}.}

\item{Path_Grid}{Character. The directory path to load the grid data.}

\item{IAS_List}{A list of invasive alien species keys.}

\item{Efforts_AllRequests}{A data frame containing the details of the GBIF
download, including paths to CSV files, zip files, order, class, and total
records.}
}
\value{
The function returns and saves the GBIF data summary.
}
\description{
This function processes GBIF data for vascular plants by extracting,
summarizing, and saving the data. It also creates summary maps and saves them
in both \code{RData} and \code{TIFF} formats.
}
\author{
Ahmed El-Gabbas
}
