% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_Extract_BB.R
\name{Extract_BB}
\alias{Extract_BB}
\title{Extract a specific column from the output of GBIF standardization}
\usage{
Extract_BB(x, var)
}
\arguments{
\item{x}{A list or data frame containing the results of GBIF standardization (\code{rgbif::name_backbone}).}

\item{var}{A character string specifying the name of the column to extract from \code{x}.}
}
\value{
Returns the content of the specified column \code{var} from the input \code{x}. If \code{var} is not found within \code{x}, the function returns \code{NA_character_}.
}
\description{
This function is designed to extract a specific column from the output of GBIF standardization, typically obtained via the \code{rgbif::name_backbone} function. It is useful for retrieving specific information such as the scientific name of the accepted taxa.
}
\details{
Extract a specific column from the output of GBIF
}
\examples{

\dontrun{
rgbif::name_backbone(name = "Helianthus annuus", kingdom = "plants") \%>\%
   Extract_BB(status)

# ------------------------

c("Helianthus annuus", "Tagetes patula L.") \%>\%
   tibble::tibble(Taxa = .) \%>\%
   dplyr::mutate(
      BB = purrr::map(Taxa, rgbif::name_backbone),
      status = purrr::map_chr(BB, Extract_BB, status),
      SpKey = purrr::map_int(BB, Extract_BB, speciesKey))
}
}
\author{
Ahmed El-Gabbas
}
