% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{FunctionsInPackage}
\alias{FunctionsInPackage}
\title{List of functions in a package}
\usage{
FunctionsInPackage(Package)
}
\arguments{
\item{Package}{name of the package}
}
\description{
List of functions in a package
}
\examples{
FunctionsInPackage("raster")
}
\author{
Ahmed El-Gabbas
}
