% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GBIF_process.R, R/DWF_GBIF_download.R,
%   R/DWF_GBIF_read_chunk.R, R/DWF_GBIF_species_data.R
\name{GBIF_data}
\alias{GBIF_data}
\alias{GBIF_process}
\alias{GBIF_download}
\alias{GBIF_read_chunk}
\alias{GBIF_species_data}
\title{Process GBIF occurrence data for the \code{IASDT}}
\usage{
GBIF_process(
  env_file = ".env",
  r_environ = ".Renviron",
  n_cores = 6L,
  strategy = "multisession",
  request = TRUE,
  download = TRUE,
  split_chunks = TRUE,
  overwrite = FALSE,
  delete_chunks = TRUE,
  chunk_size = 50000L,
  boundaries = c(-30, 50, 25, 75),
  start_year = 1981L
)

GBIF_download(
  env_file = ".env",
  r_environ = ".Renviron",
  request = TRUE,
  download = TRUE,
  split_chunks = TRUE,
  chunk_size = 50000L,
  boundaries = c(-30, 50, 25, 75),
  start_year = 1981L
)

GBIF_read_chunk(
  chunk_file,
  env_file = ".env",
  max_uncertainty = 10L,
  start_year = 1981L,
  save_RData = TRUE,
  return_data = FALSE,
  overwrite = FALSE
)

GBIF_species_data(
  species = NULL,
  env_file = ".env",
  verbose = TRUE,
  plot_tag = NULL
)
}
\arguments{
\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{r_environ}{Character. Path to \code{.Renviron} file with GBIF credentials
(\code{GBIF_EMAIL}, \code{GBIF_USER}, \code{GBIF_PWD}). Default: \code{".Renviron"}. The
credentials must be in the format:
\itemize{
\item \code{GBIF_EMAIL=your_email}
\item \code{GBIF_USER=your_username}
\item \code{GBIF_PWD=your_password}
}}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{request}{Logical. If \code{TRUE} (default), requests GBIF data; otherwise,
loads from disk.}

\item{download}{Logical. If \code{TRUE} (default), downloads and saves GBIF data.}

\item{split_chunks}{Logical. If \code{TRUE} (default), splits data into chunks for
easier processing.}

\item{overwrite}{Logical. If \code{TRUE}, reprocesses existing \code{.RData} chunks.
Default: \code{FALSE}. This helps to continue working on previously processed
chunks if the previous try failed, e.g. due to memory issue.}

\item{delete_chunks}{Logical. If \code{TRUE} (default), deletes chunk files.}

\item{chunk_size}{Integer. Records per data chunk. Default: \code{50000}.}

\item{boundaries}{Numeric vector (length 4). GBIF data bounds (Left, Right,
Bottom, Top). Default: \code{c(-30, 50, 25, 75)}.}

\item{start_year}{Integer. Earliest collection year to be included. Default
is 1981.}

\item{chunk_file}{Character. Path of chunk file for processing.}

\item{max_uncertainty}{Numeric. Maximum spatial uncertainty in kilometres.
Default: \code{10}.}

\item{save_RData}{Logical. If \code{TRUE} (default), saves chunk data as \code{.RData}.}

\item{return_data}{If \code{TRUE}, returns chunk data; otherwise,
\code{invisible(NULL)}. Default: \code{FALSE}.}

\item{species}{Character. Species name for processing.}

\item{verbose}{Logical. If \code{TRUE} (default), prints progress messages.}

\item{plot_tag}{Character. Tag for plot titles.}
}
\description{
Extracts, processes, and visualises occurrence data from the \href{https://www.gbif.org}{Global Biodiversity Information Facility (GBIF)} for the
Invasive Alien Species Digital Twin (\code{IASDT}). Orchestrated by
\code{GBIF_process()}, it requests, downloads, cleans, chunks, and maps species
data using helper functions.
}
\note{
Relies on a static RDS file listing IAS species, GBIF keys, and
metadata, standardized by Marina Golivets (Feb 2024).
}
\section{Functions details}{

\itemize{
\item \strong{\code{GBIF_process()}}: Orchestrates GBIF data requests, downloads,
processing, and mapping. Saves \code{RData}, Excel, and JPEG summary files.
\item \strong{\code{GBIF_download()}}: Requests and downloads GBIF data (if \code{download = TRUE}), using the specified criteria (taxa, coordinates, time period, and
boundaries), splits into small chunks (if \code{split_chunks = TRUE}), and saves
metadata. Returns \code{invisible(NULL)}.
\item \strong{\code{GBIF_read_chunk()}}: Filters chunk data (spatial/temporal, e.g.,
spatial uncertainty, collection year, coordinate precision, and taxonomic
rank), select relevant columns, and saves as \code{.RData} (if \code{save_RData = TRUE}) or returns it (if \code{return_data = TRUE}). Skips if \code{.RData} exists
and \code{overwrite = FALSE}.
\item \strong{\code{GBIF_species_data()}}: Converts species-specific data to \code{sf} and
raster formats, generating distribution maps.
}
}

\author{
Ahmed El-Gabbas
}
