% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GBIF_ReadChunk.R
\name{GBIF_ReadChunk}
\alias{GBIF_ReadChunk}
\title{Read GBIF chunk data}
\usage{
GBIF_ReadChunk(
  ChunkFile,
  EnvFile = ".env",
  FromHPC = TRUE,
  MaxUncert = 10,
  MinYear = 1981,
  SaveRData = TRUE,
  ReturnData = FALSE,
  Overwrite = FALSE
)
}
\arguments{
\item{ChunkFile}{A string specifying the path of the chunk file to be
processed.}

\item{EnvFile}{A string specifying the path of the environment file. Default
is ".env".}

\item{FromHPC}{A logical value indicating whether the environment variables
should be read from an HPC (High-Performance Computing) environment.
Default is TRUE.}

\item{MaxUncert}{A numeric value specifying the maximum accepted spatial
uncertainty in kilometers. Default is 10 km.}

\item{MinYear}{An integer specifying the earliest collection year to be
included. Default is 1981.}

\item{SaveRData}{Logical; indicating whether to save the cleaned data for the
current chunk as \verb{*.RData} file.}

\item{ReturnData}{Logical; indicating whether to return the cleaned data for
the current chunk. Defaults to \code{FALSE}.}

\item{Overwrite}{Logical; indicating whether to process the current chunk
file if it has already processed and saved as \verb{*.RData} file. This helps to
continue working on previously processed chunks if the previous try failed,
e.g. due to memory issue.}
}
\value{
The output of the function is a tibble (sf) object containing the
processed and filtered GBIF data for the input chunk file. Whether the
tibble is exported or saved as \verb{*.RData} file depends on the values of the
\code{SaveRData}, \code{ReturnData}, and \code{Overwrite} parameters.

If \code{SaveRData = TRUE} (default), the processed data will be saved as
\code{RData} file with the same base name of the chunk file and at the same
directory. By default, the function does not return any value, unless
\code{ReturnData} is set to \code{TRUE}. \code{SaveRData} and \code{ReturnData} can not both
set to \code{FALSE}. The function can optionally skip processing the current
chunk file if the \verb{*.RData} file for this chunk already exist and
\code{Overwrite} is set as \code{FALSE} (default). If \code{Overwrite} is set to \code{TRUE},
the data for the current chunk will be re-processed irrespective of the
existence of the RData file. This helps to continue working on previously
processed chunks if a previous run of \link{GBIF_Processing} failed, e.g. due to
memory issue.
}
\description{
This function reads and processes a chunk of GBIF (Global Biodiversity
Information Facility) data. It filters the data based on several criteria
including spatial uncertainty, collection year, coordinate precision, and
taxonomic rank and returns only selected columns.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only called from the \link{GBIF_Processing} function.
}
\author{
Ahmed El-Gabbas
}
