% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GBIF_SpData.R
\name{GBIF_SpData}
\alias{GBIF_SpData}
\title{Extract species data as sf and Raster and plot species distribution}
\usage{
GBIF_SpData(
  Species = NULL,
  FromHPC = TRUE,
  EnvFile = ".env",
  Verbose = TRUE,
  LastUpdate = NULL
)
}
\arguments{
\item{Species}{Character. Species name.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{Verbose}{Logical. Whether to print progress messages. Default is
\code{TRUE}.}

\item{LastUpdate}{Character. Last update date to be displayed on the plots.}
}
\description{
Extracts and processes species data from processed GBIF data, generating
grids, rasters, and visual maps.
}
\note{
This function is not intended to be used directly by the user or in the
IAS-pDT, but only used inside the \link{GBIF_Process} function.
}
\author{
Ahmed El-Gabbas
}
