% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Gelman_Alpha.R
\name{Gelman_Alpha}
\alias{Gelman_Alpha}
\title{Gelman-Rubin-Brooks plot for alpha}
\usage{
Gelman_Alpha(CodaObj = NULL, NCores = NULL, PlottingAlpha = 0.25)
}
\arguments{
\item{CodaObj}{an mcmc object}

\item{NCores}{Integer. Number of parallel processes.}

\item{PlottingAlpha}{Double. Plotting alpha for line transparency}
}
\description{
Gelman-Rubin-Brooks plot for alpha
}
\author{
Ahmed El-Gabbas
}
