% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Gelman_Plot.R
\name{Gelman_Plot}
\alias{Gelman_Plot}
\title{Plot Gelman-Rubin-Brooks plots}
\usage{
Gelman_Plot(
  InputCoda = NULL,
  Beta = TRUE,
  Rho = TRUE,
  Omega = TRUE,
  Alpha = TRUE,
  NCores = NULL,
  NOmega = 1000,
  PlotAlpha = 0.25,
  SavePlot = TRUE,
  ReturnPlots = FALSE,
  OutPath = NULL
)
}
\arguments{
\item{InputCoda}{Path to RData file containing the coda object}

\item{Beta}{Logical. Run \code{IASDT.R::Gelman_Beta}?}

\item{Rho}{Logical. Run \code{IASDT.R::Gelman_Rho}?}

\item{Omega}{Logical. Run \code{IASDT.R::Gelman_Omega}?}

\item{Alpha}{Logical. Run \code{IASDT.R::Gelman_Alpha} (not yet available)?}

\item{NCores}{Integer. Number of parallel processes.}

\item{NOmega}{Integer. Number of species to be sampled for the Omega parameter}

\item{PlotAlpha}{Double. Plotting alpha for line transparency}

\item{SavePlot}{Logical. Save the outputs as PDF}

\item{ReturnPlots}{Logical. Return ggplot objects}

\item{OutPath}{String. Folder path to save the output figures}
}
\description{
Plot Gelman-Rubin-Brooks plots. This plot shows the evolution of Gelman and Rubin's shrink factor as the number of iterations increases. For more information, see \code{coda:::gelman.plot}
}
\author{
Ahmed El-Gabbas
}
