% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Gelman_Rho.R
\name{Gelman_Rho}
\alias{Gelman_Rho}
\title{Gelman-Rubin-Brooks plot for all beta parameters}
\usage{
Gelman_Rho(CodaObj = NULL)
}
\arguments{
\item{CodaObj}{an mcmc object}
}
\description{
Gelman-Rubin-Brooks plot for all beta parameters
}
\author{
Ahmed El-Gabbas
}
