% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{Get_EASIN_Data}
\alias{Get_EASIN_Data}
\title{Extract EASIN data using EASIN ID}
\usage{
Get_EASIN_Data(
  SpKey,
  NSearch = 1000,
  SleepPart = 1,
  Path_Raw = "Data/EASIN/Raw",
  ReturnVal = FALSE,
  SleepAfter = 5
)
}
\arguments{
\item{SpKey}{character; EASIN taxon ID}

\item{NSearch}{number of observations to extract each time; default: 1000 (the current maximum per request)}

\item{SleepPart}{integer; time in seconds to wait between each request. Default: 1.}

\item{Path_Raw}{character; a valid path for the location to store the output data. A subfolder named \code{FileParts} will be created for temporary files for individual requests.}

\item{ReturnVal}{logical; should the data be returned or only saved to disk. Default: \code{FALSE}}

\item{SleepAfter}{integer; time in seconds to wait after finishing with downloading taxon data. Default: 10.}
}
\description{
Extract EASIN data using EASIN ID
}
\details{
A function to extract EASIN data for a given EASIN_ID
}
\author{
Ahmed El-Gabbas
}
