% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_IAS.R
\name{GetAcceptedName}
\alias{GetAcceptedName}
\title{Get accepted name of a taxa}
\usage{
GetAcceptedName(ID)
}
\arguments{
\item{ID}{GBIF id}
}
\value{
Scientific name of the accepted taxa
}
\description{
Get accepted name of a taxa
}
\details{
Get accepted name of a taxa
}
\author{
Ahmed El-Gabbas
}
