% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_GetCV.R
\name{GetCV}
\alias{GetCV}
\title{Prepare cross-validation data}
\usage{
GetCV(DT, NR = 4, NC = 3, Path_Grid)
}
\arguments{
\item{DT}{Input data set}

\item{NR}{Integer. Number of rows}

\item{NC}{Integer. Number of columns}

\item{Path_Grid}{String. Path of the reference grid.}
}
\description{
Prepare cross-validation data
}
\author{
Ahmed El-Gabbas
}
