% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_GetCV.R
\name{GetCV}
\alias{GetCV}
\title{Prepare spatial-block cross-validation folds for spatial analysis}
\usage{
GetCV(
  DT,
  EnvFile = ".env",
  XVars,
  CV_NFolds = 4,
  CV_NGrids = 20,
  CV_NR = 2,
  CV_NC = 2,
  OutPath = NULL,
  FromHPC = TRUE,
  CV_Plot = TRUE
)
}
\arguments{
\item{DT}{A data frame or tibble containing the input dataset. This data
frame should include two columns for \code{x} and \code{y} coordinates as long as
other columns matching the names of predictors listed in \code{XVars} argument.}

\item{EnvFile}{String specifying the path to read environment variables from,
with a default value of \code{.env}.}

\item{XVars}{Vector of strings specifying variables to be used in the model.
This argument is mandatory and can not be empty.}

\item{CV_NFolds}{Number of cross-validation folds. Default: 4.}

\item{CV_NGrids}{For \code{CV_Dist} cross-validation strategy (see below), this
argument determines the size of the blocks (how many grid cells in both
directions).}

\item{CV_NR, CV_NC}{Integer, the number of rows and columns used in the
\code{CV_Large} cross-validation strategy (see below), in which the study area
is divided into large blocks given the provided \code{CV_NR} and \code{CV_NC}
values. Both default to 2 which means to split the study area into four
large blocks at the median latitude and longitude.}

\item{OutPath}{String specifying the folder path to save the cross-validation
results. Default: \code{NULL}.}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}

\item{CV_Plot}{Logical. Indicating whether to plot the block cross-validation
folds.}
}
\value{
The function returns a modified version of the input dataset \code{DT}
with 3 additional numeric columns (integer) indicating the cross-validation
folds:
\enumerate{
\item \code{CV_SAC} in which the size of the blocks is determined by the median
spatial autocorrelation range in the predictor data (estimated using
\link[blockCV:cv_spatial_autocor]{blockCV::cv_spatial_autocor}). This requires the availability of the
\code{automap} R package.
\item \code{CV_Dist} in which the size of spatial cross-validation blocks is
determined by the \code{CV_NGrids} argument. The default \code{CV_NGrids} value is 20,
which means blocks of 20x20 grid cell each.
\item \code{CV_Large} which splits the study area into large blocks, as determined
by the  \code{CV_NR} and \code{CV_NC} arguments. if \code{CV_NR = CV_NC} = 2 (default), four large
blocks will be used, split the study area at the median coordinates..
}
}
\description{
This function assign modelling input data into spatial-block cross-validation
folds. The function uses \link[blockCV:cv_spatial]{blockCV::cv_spatial} to assign grid cells to
cross-validation folds using three strategies (see below). The function is
planned to be used inside the \link{Mod_Prep4HPC} function.
}
\details{
The function reads the following environment variable:
\itemize{
\item \strong{\code{DP_R_Grid}} (if \code{FromHPC = TRUE}) or
\strong{\code{DP_R_Grid_Local}} (if \code{FromHPC = FALSE}). The function reads
the content of the \code{Grid_10_Land_Crop.RData} file from this path.
\item \strong{\code{DP_R_EUBound_sf}} (if \code{FromHPC} = \code{TRUE}) or
\strong{\code{DP_R_EUBound_sf_Local}} (if \code{FromHPC} = \code{FALSE}): path for the
\code{RData} file containing the country boundaries (\code{sf} object).
}
}
\author{
Ahmed El-Gabbas
}
