% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_GetMode.R
\name{GetMode}
\alias{GetMode}
\title{Calculate the mode of numbers}
\usage{
GetMode(v)
}
\arguments{
\item{v}{vector}
}
\description{
Calculate the mode of numbers
}
\examples{
GetMode(c(1:10,1,1,3,3,3,3))
}
\references{
https://www.tutorialspoint.com/r/r_mean_median_mode.htm
}
