% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_GetPosts.R
\name{GetPosts}
\alias{GetPosts}
\title{Combining posteriors exported by \code{Hmsc-HPC} into Hmsc object}
\usage{
GetPosts(FilePath)
}
\arguments{
\item{FilePath}{String. Vector of the file paths of the \code{rds} files created by \code{Hmsc-HPC}}
}
\description{
Combining posteriors exported by \code{Hmsc-HPC} into Hmsc object
}
\author{
Ahmed El-Gabbas
}
