% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_GetPosts.R
\name{GetPosts}
\alias{GetPosts}
\title{Combining posteriors into Hmsc object}
\usage{
GetPosts(FilePath)
}
\arguments{
\item{FilePath}{vector of \code{rds} files created by Hmsc-HPC}
}
\description{
Combining posteriors into Hmsc object
}
\author{
Ahmed El-Gabbas
}
