% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS_GetSpeciesName.R
\name{GetSpeciesName}
\alias{GetSpeciesName}
\title{Get species information for specific species number}
\usage{
GetSpeciesName(EnvFile = ".env", SpID = NULL)
}
\arguments{
\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{SpID}{Species ID}
}
\description{
Get species information for specific species number
}
\author{
Ahmed El-Gabbas
}
