% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GetSpeciesName.R
\name{GetSpeciesName}
\alias{GetSpeciesName}
\title{Get Species Name or Information Based on IASDT Species ID}
\usage{
GetSpeciesName(EnvFile = ".env", SpID = NULL, FromHPC = TRUE)
}
\arguments{
\item{EnvFile}{EnvFile A string specifying the path to the environment
variables file. Default is ".env". This file must exist in the specified
path; otherwise, the function will stop with an error message.}

\item{SpID}{optional IASDT species ID for which detailed information is
required. If not provided, the function returns the entire species list.}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}
}
\value{
A data frame containing species information. If a species ID \code{SpID}
is provided, it only returns species information for the listed species,
otherwise return the full list of IAS.
}
\description{
This function retrieves detailed information on IASDT Species, optionally
filtered by a specific IASDT species ID. It first loads environment variables
from a given file, then reads a species list from a path specified in the
environment variables. If a species ID \code{SpID} is provided, it only returns
species information for the listed species, otherwise return the full list of
IAS.
}
\details{
The function reads the following environment variables:
\itemize{
\item \strong{\code{DP_R_TaxaInfo}} (if \code{FromHPC} = \code{TRUE}) or
\strong{\code{DP_R_TaxaInfo_Local}} (if \code{FromHPC} = \code{FALSE}) for the location of the \code{Species_List_ID.txt} file containing species information.
\item \strong{\code{DP_R_PA}} (if \code{FromHPC} = \code{TRUE}) or \strong{\code{DP_R_PA_Local}} (if \code{FromHPC} = \code{FALSE}). The function reads the contents of the \code{Sp_PA_Summary_DF.RData} file from this path.
}
}
\author{
Ahmed El-Gabbas
}
