% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_GetSpeciesName.R
\name{GetSpeciesName}
\alias{GetSpeciesName}
\title{Get species name or information of an \code{IAS-pDT} species ID}
\usage{
GetSpeciesName(SpeciesID = NULL, EnvFile = ".env", FromHPC = TRUE)
}
\arguments{
\item{SpeciesID}{optional IASDT species ID for which detailed information is
required. If not provided, the function returns the entire species list.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}
}
\value{
A data frame containing species information. If a species ID
\code{SpeciesID} is provided, it only returns species information for the listed
species, otherwise return the full list of IAS.
}
\description{
This function retrieves detailed information on \code{IAS-pDT} species list,
optionally filtered by a specific IAS-pDT species ID (\code{SpeciesID}).
}
\author{
Ahmed El-Gabbas
}
