% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_ht.R
\name{ht}
\alias{ht}
\title{Print head and tail of data frame}
\usage{
ht(DF = NULL, NRows = 5L)
}
\arguments{
\item{DF}{\verb{data.frame. A data frame to print. This parameter cannot be }NULL`.}

\item{NRows}{Integer. Number of rows to print from both the head and tail of
the data frame. Defaults to 5.}
}
\value{
The function is used for its side effect (printing) and does not
return any value.
}
\description{
This function takes a data frame and an optional number of rows to print from
both the top (head) and bottom (tail) of the data frame. It is useful for
quickly inspecting the first few and last few rows of a large data frame to
understand its structure and contents.
}
\examples{
ht(mtcars)

# -------------------------------------------

ht(DF = mtcars, 2)

# -------------------------------------------

ht(DF = mtcars, 6)
}
\author{
Ahmed El-Gabbas
}
