% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_IAS_Process.R, R/DWF_IAS_Distribution.R,
%   R/DWF_IAS_Plot.R
\name{IAS_data}
\alias{IAS_data}
\alias{IAS_Process}
\alias{IAS_Distribution}
\alias{IAS_Plot}
\title{Process and map Invasive Alien Species (IAS) data for the \code{IAS-pDT}}
\usage{
IAS_Process(EnvFile = ".env", NCores = 6L, Overwrite = TRUE)

IAS_Distribution(
  Species = NULL,
  EnvFile = ".env",
  Verbose = FALSE,
  Overwrite = FALSE
)

IAS_Plot(Species = NULL, EnvFile = ".env", Overwrite = FALSE)
}
\arguments{
\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 6.}

\item{Overwrite}{Logical. If \code{TRUE}, overwrites existing maps; otherwise,
skips if maps exist. Default: \code{FALSE}.}

\item{Species}{Character. Species name for distribution mapping.}

\item{Verbose}{Logical. If \code{TRUE}, prints progress messages. Default:
\code{FALSE}.}
}
\description{
Processes and visualizes Invasive Alien Species (IAS) distribution data from
GBIF, EASIN, and eLTER for the Invasive Alien Species prototype Digital Twin
(\code{IAS-pDT}). Merges pre-processed data, creates presence-absence rasters,
summarizes distributions, and generates maps using helper functions.
}
\section{Functions details}{

\itemize{
\item \strong{\code{IAS_Process()}}: Merges pre-processed GBIF (\code{GBIF_Process()}), EASIN
(\code{EASIN_Process()}), and eLTER (\code{eLTER_Process()}) data (run these first).
Outputs \code{SpatRaster} distribution rasters, summary tables, and JPEG maps
using \code{IAS_Distribution()} and \code{IAS_Plot()}.
\item \strong{\code{IAS_Distribution()}}: Generates presence-absence maps (\code{.RData},
\code{.tif})  for a species, including all grid cells in the study area and a set
excluding cultivated/casual-only countries. Returns a tibble with presence
counts (total, by source) and summary statistics for biogeographical
regions
\item \strong{\code{IAS_Plot()}}: Creates JPEG distribution maps from GBIF, EASIN, and
eLTER data using \code{ggplot2}.
}
}

\author{
Ahmed El-Gabbas
}
