% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAS.DT-package.R
\docType{package}
\name{IASDT.R-package}
\alias{IASDT.R}
\alias{IASDT.R-package}
\title{IASDT.R: Helper R functions for IAS pDT project}
\description{
Helper functions to be used in the Invasive Alien Species (IAS) prototype Digital Twin (IAS pDT).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biodt.eu}
  \item \url{https://www.ufz.de/index.php?en=36737}
  \item Report bugs at \url{https://git.ufz.de/biodt/IASDT.R/-/issues}
}

}
\author{
\strong{Maintainer}: Ahmed El-Gabbas \email{ahmed.el-gabbas@ufz.de} (\href{https://orcid.org/0000-0003-2225-088X}{ORCID})

}
\keyword{internal}
