% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{InfoChunk}
\alias{InfoChunk}
\title{Print Information chunk}
\usage{
InfoChunk(Message = "", ...)
}
\arguments{
\item{Message}{String passed to the \code{IASDT.R::CatTime} function}

\item{...}{additional arguments for the \code{IASDT.R::CatSep} function}
}
\description{
Print Information chunk
}
\examples{
InfoChunk(Message = "Started")

InfoChunk(Message = "finished", Char = "*", CharReps = 60)
}
\author{
Ahmed El-Gabbas
}
