% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_InfoChunk.R
\name{InfoChunk}
\alias{InfoChunk}
\title{Print Information chunk with time stamp}
\usage{
InfoChunk(Message = "", ...)
}
\arguments{
\item{Message}{A character string representing the main message to be timestamped.  parameter is mandatory and cannot be NULL or empty.}

\item{...}{Additional arguments passed to \link{CatSep} for customizing the separators.}
}
\value{
The function does not return any value but prints the message and separators to the console.
}
\description{
This function prints a formatted message with a timestamp, surrounded by separators for better readability in console outputs or logs.
}
\examples{
InfoChunk(Message = "Started")

InfoChunk(Message = "finished", Char = "*", CharReps = 60)
}
\author{
Ahmed El-Gabbas
}
