% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_InfoChunk.R
\name{InfoChunk}
\alias{InfoChunk}
\title{Print Information chunk with time stamp}
\usage{
InfoChunk(
  Message = "",
  Date = TRUE,
  Extra1 = 1L,
  Extra2 = 1L,
  Bold = FALSE,
  Red = FALSE,
  Time = FALSE,
  Level = 0L,
  NLines = 1L,
  ...
)
}
\arguments{
\item{Message}{A character string representing the main message to be
timestamped. This parameter is mandatory and cannot be \code{NULL} or empty.}

\item{Date}{logical; whether to include the date in the timestamp. Default is
\code{FALSE}, meaning only the time is printed. See \link{CatTime}.}

\item{Extra1, Extra2}{integer; the number of extra empty lines to print before
and after the separator lines. See \link{CatSep} for more details.}

\item{Bold}{logical; whether to print the text in bold. Default is \code{FALSE}.}

\item{Red}{logical; whether to print the text in red. Default is \code{FALSE}.}

\item{Time}{logical; whether to include the time in the timestamp. Default is
\code{FALSE}.}

\item{Level}{integer; the level at which the message will be printed. If e.g.
Level = 1, the following string will be printed at the beginning of the
message: "   >>>   ". Default is \code{0}.}

\item{NLines}{integer; the number of newline characters to print after the
message. Default is 1.}

\item{...}{Additional arguments passed to \link{CatSep} for customizing
the separators.}
}
\value{
The function does not return any value but prints the message and
separators to the console.
}
\description{
This function prints a formatted message with a timestamp, surrounded by
separators for better readability in console outputs or logs.
}
\examples{
InfoChunk(Message = "Started")
InfoChunk(Message = "finished", Char = "*", CharReps = 60)
}
\author{
Ahmed El-Gabbas
}
