% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_InfoChunk.R
\name{InfoChunk}
\alias{InfoChunk}
\title{Print Information chunk with time stamp}
\usage{
InfoChunk(Message = "", ...)
}
\arguments{
\item{Message}{A character string passed to \code{IASDT.R::CatTime}}

\item{...}{additional arguments for the \code{IASDT.R::CatSep}}
}
\description{
Print Information chunk with time stamp
}
\examples{
InfoChunk(Message = "Started")

InfoChunk(Message = "finished", Char = "*", CharReps = 60)
}
\author{
Ahmed El-Gabbas
}
