% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_lapply_.R
\name{lapply_}
\alias{lapply_}
\title{Apply a Function over a List or Vector with Optional Silence}
\usage{
lapply_(X, FUN, Silent = TRUE, ...)
}
\arguments{
\item{X}{A list or vector. The input data over which \code{FUN} is to be applied.}

\item{FUN}{A function to be applied to each element of \code{X}.}

\item{Silent}{A logical value. If \code{TRUE}, the function suppresses the return value of \code{FUN} and returns \code{NULL} invisibly. If \code{FALSE}, the function returns the result of applying \code{FUN} over \code{X}.}

\item{...}{Additional arguments to be passed to \code{FUN}.}
}
\value{
If \code{Silent} is \code{TRUE}, returns \code{NULL} invisibly, otherwise returns a list of the same length as \code{X}, where each element is the result of applying \code{FUN} to the corresponding element of \code{X}.
}
\description{
This function is a wrapper around the base \link[base:lapply]{base::lapply} function that allows for the application of a function over a list or vector. It optionally allows for the suppression of the function's return value, making it useful for operations where the user is only interested in the side effects of the function.
}
\examples{
par(mfrow = c(1,2), oma = c(0.25, 0.25, 0.25, 0.25), mar = c(3,3,3,1))
lapply(list(x = 100:110, y = 110:120), function(V) {
    plot(V, las = 1, main = "lapply")
})

# -------------------------------------------

par(mfrow = c(1,2), oma = c(0.25, 0.25, 0.25, 0.25), mar = c(3,3,3,1))
lapply_(list(x = 100:110, y = 110:120), function(V) {
    plot(V, las = 1, main = "lapply_")
})
}
\author{
Ahmed El-Gabbas
}
