% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{lapply_}
\alias{lapply_}
\title{Apply a Function over a List or Vector (with no return value)}
\usage{
lapply_(X, FUN, Silent = TRUE, ...)
}
\arguments{
\item{X}{vector}

\item{FUN}{function}

\item{Silent}{should the output be silenced?}

\item{...}{additional arguments to lapply}
}
\description{
Apply a Function over a List or Vector (with no return value)
}
\author{
Ahmed El-Gabbas
}
