% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{List2RData}
\alias{List2RData}
\title{Split list items to separate RData files}
\usage{
List2RData(List, Prefix = "", Dir = getwd(), Overwrite = FALSE)
}
\arguments{
\item{List}{list object}

\item{Prefix}{prefix string}

\item{Dir}{output directory}

\item{Overwrite}{overwrite existing files?}
}
\description{
Split list items to separate RData files
}
\examples{
# split iris data by species name
iris2 <- iris \%>\%
  tibble::tibble() \%>\%
  split(~Species)

str(iris2, 1)

(TMP_Folder <- file.path(tempdir(), stringi::stri_rand_strings(1, 5)))
list.files(TMP_Folder)

List2RData(iris2, Dir = TMP_Folder)
list.files(TMP_Folder)
}
\author{
Ahmed El-Gabbas
}
