% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{List2RData}
\alias{List2RData}
\title{Split list items to separate RData files}
\usage{
List2RData(List, Prefix = "", Dir = getwd(), Overwrite = FALSE)
}
\arguments{
\item{List}{list object}

\item{Prefix}{prefix string}

\item{Dir}{output directory}

\item{Overwrite}{overwrite existing files?}
}
\description{
Split list items to separate RData files
}
\author{
Ahmed El-Gabbas
}
