% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_LoadAs.R
\name{LoadAs}
\alias{LoadAs}
\title{Load RData file and return its contents as a list or a single object}
\usage{
LoadAs(File, nthreads = 5)
}
\arguments{
\item{File}{character; the file path of the \code{.RData} file to be loaded.}

\item{nthreads}{Number of threads to use when reading \code{qs} files. Default 5;
see \link[qs:qread]{qs::qread}.}
}
\value{
Depending on the contents of the \code{RData} file, this function returns
either a single R object or a named list of R objects. The names of the
list elements (if a list is returned) correspond to the names of the
objects stored within the RData file.
}
\description{
This function loads an \code{RData} file specified by the \code{File} parameter. If the
\code{RData} file contains a single object, that object is returned directly. If
the file contains multiple objects, they are returned as a list with each
object accessible by its name. This allows for flexible handling of loaded
data without needing to know the names of the objects stored within the RData
file ahead of time. The function also supports loading \code{qs} and \code{rds} files.
}
\examples{

File <- system.file("testdata", "culcita_dat.RData", package = "lme4")

# ---------------------------------------------------------
# loading RData using base library
# ---------------------------------------------------------
(load(File))

ls()

tibble::tibble(culcita_dat)

# ---------------------------------------------------------
# Loading as custom object name
# ---------------------------------------------------------
NewObj <- LoadAs(File = File)

ls()

print(tibble::tibble(NewObj))

# ---------------------------------------------------------
# Loading multiple objects stored in single RData file
# ---------------------------------------------------------
# store three objects to single RData file
mtcars2 <- mtcars3 <- mtcars
TempFile <- tempfile(pattern = "mtcars_", fileext = ".RData")

save(mtcars2, mtcars3, mtcars, file = TempFile)
mtcars_all <- LoadAs(TempFile)

# overwrite the file with different order of objects
save(mtcars, mtcars2, mtcars3, file = TempFile)
mtcars_all2 <- LoadAs(TempFile)

# single list object with 3 items, keeping original object names and order
names(mtcars_all)
names(mtcars_all2)
}
\author{
Ahmed El-Gabbas
}
