% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{LoadAs}
\alias{LoadAs}
\title{Load RData file as specific object; i.e. rename loaded object}
\usage{
LoadAs(File = NA)
}
\arguments{
\item{File}{path of file}
}
\description{
Load RData file as specific object; i.e. rename loaded object
}
\examples{
File <- system.file("testdata", "culcita_dat.RData", package = "lme4")

# ---------------------------------------------------------
# loading RData using base library
# ---------------------------------------------------------
(load(File))

ls()

tibble::tibble(culcita_dat)

rm(culcita_dat)

# ---------------------------------------------------------
# Loading as custom object name
# ---------------------------------------------------------
NewObj <- LoadAs(File = File)

ls()

print(tibble::tibble(NewObj))
}
\author{
Ahmed El-Gabbas
}
