% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_LoadMultiple.R
\name{LoadMultiple}
\alias{LoadMultiple}
\title{Load multiple RData files together}
\usage{
LoadMultiple(
  Files = NULL,
  Verbose = TRUE,
  OneObject = TRUE,
  ReturnNames = TRUE
)
}
\arguments{
\item{Files}{A character vector specifying the paths of \code{.RData} files to be loaded.}

\item{Verbose}{A logical flag indicating whether to print messages about the loading process. Only effective when \code{OneObject} is \code{FALSE}. Defaults to \code{TRUE}.}

\item{OneObject}{A logical flag indicating whether to load all objects into a single list (\code{TRUE}) or directly into the global environment (\code{FALSE}). Defaults to \code{TRUE}.}

\item{ReturnNames}{A logical flag indicating whether to return the names of the loaded objects. Only effective when \code{OneObject} is \code{FALSE}. Defaults to \code{TRUE}.}
}
\value{
If \code{OneObject} is \code{TRUE}, returns a named list of all objects loaded from the specified files. If \code{OneObject} is \code{FALSE} and \code{ReturnNames} is \code{TRUE}, returns a character vector of the names of the objects loaded into the global environment. Otherwise, returns \code{NULL}.
}
\description{
This function loads multiple \code{.RData} files either into a single list object or directly into the global environment. It provides options for verbosity, returning object names, and handling of non-existent files.
}
\examples{
(Files <- system.file("testdata", c("culcita_dat.RData", "gopherdat2.RData"), package = "lme4"))

ls()

# ---------------------------------------------------
# Load multiple *.RData files to one list object
# `OneObject = TRUE`
# ---------------------------------------------------
MultiObj <- LoadMultiple(Files = Files, OneObject = TRUE)
ls()

str(MultiObj, 1)

# ---------------------------------------------------
# Load multiple *.RData files separately to the current environment
# `OneObject = FALSE`
# ---------------------------------------------------
ls()
rm("MultiObj")
ls()

LoadMultiple(Files = Files, OneObject = FALSE)

str(Gdat, 1)

str(culcita_dat, 1)

# ---------------------------------------------------
# Load multiple *.RData files, one object already exists
# ---------------------------------------------------
ls()
rm("culcita_dat")
ls()

try(LoadMultiple(Files = Files, OneObject = FALSE))

}
\author{
Ahmed El-Gabbas
}
