% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{LoadPackages}
\alias{LoadPackages}
\title{Load package silently and print version}
\usage{
LoadPackages(Package)
}
\arguments{
\item{Package}{package name}
}
\description{
Load package silently and print version
}
\author{
Ahmed El-Gabbas
}
