% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_LoadPackages.R
\name{LoadPackages}
\alias{LoadPackages}
\title{Load package silently (+ install missing packages)}
\usage{
LoadPackages(..., List = NULL, Verbose = FALSE)
}
\arguments{
\item{...}{packages to load / install}

\item{List}{vector for the name of packages to be loaded}

\item{Verbose}{print a message of the package name and version}
}
\description{
Load package silently (+ install missing packages)
}
\examples{
LoadPackages(raster)

LoadPackages(terra, Verbose = TRUE)
}
\author{
Ahmed El-Gabbas
}
