% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CV_evaluate.R
\name{mod_CV_evaluate}
\alias{mod_CV_evaluate}
\title{Evaluate cross-validated Hmsc Model}
\usage{
mod_CV_evaluate(path_CV = NULL, predict_eta_mean = TRUE, n_cores = 8L)
}
\arguments{
\item{path_CV}{Character. The directory path where cross-validation models
and outputs are stored.}

\item{predict_eta_mean}{boolean flag indicating whether to use the estimated
mean values of posterior predictive distribution for random effects
corresponding for the new units. See \code{Hmsc:::predict.Hmsc} for more
details.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}
}
\description{
This function evaluates Hmsc models trained via cross-validation using. It
merges posterior chains, makes predictions, and computes 4 evaluation
metrics: AUC, RMSE, Tjur's R2, and Boyce Index.
}
\author{
Ahmed El-Gabbas
}
