% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CV_fit.R
\name{mod_CV_fit}
\alias{mod_CV_fit}
\title{Prepare cross-validated Hmsc models for HPC fitting}
\usage{
mod_CV_fit(
  path_model = NULL,
  CV_name = c("CV_Dist", "CV_Large"),
  Partitions = NULL,
  env_file = ".env",
  init_par = NULL,
  job_name = "CV_Models",
  updater = list(Gamma2 = FALSE, GammaEta = FALSE),
  align_posterior = TRUE,
  to_JSON = FALSE,
  SLURM_prepare = TRUE,
  memory_per_cpu = NULL,
  job_runtime = NULL,
  path_Hmsc = NULL,
  precision = 64,
  ...
)
}
\arguments{
\item{path_model}{Character. Path to a saved model file (\verb{*.qs2}).}

\item{CV_name}{Character vector. Column name(s) in the model input data to be
used to cross-validate the models (see \link{mod_prepare_data} and
\link{mod_CV_prepare}). The function allows the possibility of using more than
one way of assigning grid cells into cross-validation folders. If multiple
names are provided, separate cross-validation models will be fitted for
each cross-validation type. Currently, there are three cross-validation
strategies: \code{CV_SAC}, \code{CV_Dist}, and \code{CV_Large}. Defaults to \code{c("CV_Dist", "CV_Large")}.}

\item{Partitions}{A vector for cross-validation created by
\link[Hmsc:createPartition]{Hmsc::createPartition} or similar. Defaults to \code{NULL}, which means to use
column name(s) provided in the \code{CV_name} argument. If the \code{Partitions}
vector is provided, the label used in the output files will be \code{CV_Custom}.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{init_par}{a named list of parameter values used for initialisation of
MCMC states. See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information.
Default: \code{NULL}.}

\item{job_name}{Character. Name of the submitted job(s) for SLURM. Default:
\code{CV_Models}.}

\item{updater}{named \code{list}. Which conditional updaters should be omitted?
See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information. Defaults to
\code{list(Gamma2 = FALSE, GammaEta = FALSE)} to disable the following warnings:
\verb{setting updater$Gamma2=FALSE due to specified phylogeny matrix} and
\verb{setting updater$GammaEta=FALSE: not implemented for spatial methods 'GPP' and 'NNGP'}.}

\item{align_posterior}{Logical. Whether the posterior of each chains should
be aligned. See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information.
Default: \code{TRUE}.}

\item{to_JSON}{Logical. Whether to convert unfitted models to JSON before
saving to RDS file. Default: \code{FALSE}.}

\item{SLURM_prepare}{Logical. Whether to prepare SLURM command files. If
\code{TRUE} (default), the SLURM commands will be saved to disk using the
\link{mod_SLURM} function.}

\item{memory_per_cpu}{Character. Memory allocation per CPU core. Example:
"32G" for 32 gigabytes. Defaults to "64G".}

\item{job_runtime}{Character. Maximum allowed runtime for the job. Example:
"01:00:00" for one hour. Required --- if not provided, the function throws
an error.}

\item{path_Hmsc}{Character. Path to the Hmsc-HPC installation.}

\item{precision}{Integer (either 32 or 64). Defines the floating-point
precision mode for \code{Hmsc-HPC} sampling (--fp 32 or --fp 64). The default is
64, which is the default precision in \code{Hmsc-HPC}.}

\item{...}{Additional arguments passed to the \link{mod_SLURM} function.}
}
\description{
This function prepares cross-validated Hmsc models for fitting using HPC. It
handles data preparation, model initialisation, and generation of SLURM
commands.
}
\details{
The function copies part of the \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues}
function, which currently does not support performing cross-validation
using Hmsc-HPC.
}
\author{
Ahmed El-Gabbas
}
