% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_CV_Fit.R
\name{Mod_CV_Fit}
\alias{Mod_CV_Fit}
\title{Prepare cross-validated Hmsc models for HPC fitting}
\usage{
Mod_CV_Fit(
  Path_Model = NULL,
  CVName = c("CV_Dist", "CV_Large"),
  Partitions = NULL,
  EnvFile = ".env",
  InitPar = NULL,
  JobName = "CV_Models",
  Updater = list(Gamma2 = FALSE, GammaEta = FALSE),
  AlignPost = TRUE,
  ToJSON = FALSE,
  PrepSLURM = TRUE,
  MemPerCpu = NULL,
  Time = NULL,
  Path_Hmsc = NULL,
  Precision = 64,
  ...
)
}
\arguments{
\item{Path_Model}{Character. Path to a saved model file (\verb{*.qs2}).}

\item{CVName}{Character vector. Column name(s) in the model input data to be
used to cross-validate the models (see \link{Mod_PrepData} and \link{Mod_GetCV}).
The function allows the possibility of using more than one way of
assigning grid cells into cross-validation folders. If multiple names are
provided, separate cross-validation models will be fitted for each
cross-validation type. Currently, there are three cross-validation
strategies: \code{CV_SAC}, \code{CV_Dist}, and \code{CV_Large}. Defaults to \code{c("CV_Dist", "CV_Large")}.}

\item{Partitions}{A vector for cross-validation created by
\link[Hmsc:createPartition]{Hmsc::createPartition} or similar. Defaults to \code{NULL}, which means to use
column name(s) provided in the \code{CVName} argument. If the \code{Partitions}
vector is provided, the label used in the output files will be \code{CV_Custom}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{InitPar}{a named list of parameter values used for initialization of
MCMC states. See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information.
Default: \code{NULL}.}

\item{JobName}{Character. Name of the submitted job(s) for SLURM. Default:
\code{CV_Models}.}

\item{Updater}{named \code{list}. Which conditional updaters should be
omitted? See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information. Defaults
to \code{list(Gamma2 = FALSE, GammaEta = FALSE)} to disable the following
warnings: \verb{setting updater$Gamma2=FALSE due to specified phylogeny matrix}
and \verb{setting updater$GammaEta=FALSE: not implemented for spatial methods 'GPP' and 'NNGP'}.}

\item{AlignPost}{Logical. Whether the posterior of each chains
should be aligned. See \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} for more information.
Default: \code{TRUE}.}

\item{ToJSON}{Logical. Whether to convert unfitted models to JSON before
saving to RDS file. Default: \code{FALSE}.}

\item{PrepSLURM}{Logical. Whether to prepare SLURM command files. If \code{TRUE}
(default), the SLURM commands will be saved to disk using the \link{Mod_SLURM}
function.}

\item{MemPerCpu}{Character. Memory allocation per CPU core. Example: "32G"
for 32 gigabytes. Required --- if not provided, the function throws an
error.}

\item{Time}{Character. Maximum allowed runtime for the job. Example:
"01:00:00" for one hour. Required --- if not provided, the function throws
an error.}

\item{Path_Hmsc}{Character. Path to the Hmsc-HPC installation.}

\item{Precision}{Integer (either 32 or 64). Defines the floating-point
precision mode for \code{Hmsc-HPC} sampling (--fp 32 or --fp 64). The default is
64, which is the default precision in \code{Hmsc-HPC}.}

\item{...}{Additional arguments passed to the \link{Mod_SLURM} function.}
}
\description{
This function prepares cross-validated Hmsc models for fitting using HPC. It
handles data preparation, model initialization, and generation of SLURM
commands.
}
\details{
The function copies part of the \link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues}
function, which currently does not support performing cross-validation
using Hmsc-HPC.
}
\author{
Ahmed El-Gabbas
}
