% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Evaluation_Plot.R
\name{Mod_Eval_Plot}
\alias{Mod_Eval_Plot}
\title{Generate Plots for the Explanatory Power of Hmsc models}
\usage{
Mod_Eval_Plot(ModelDir, EnvFile = ".env", FromHPC = TRUE)
}
\arguments{
\item{ModelDir}{Path to the model directory containing predictions.}

\item{EnvFile}{Path to the environment file (\code{.env}) for setting paths.}

\item{FromHPC}{Boolean indicating whether the environment is an HPC system.}
}
\description{
This function generates four diagnostic plots (RMSE, AUC, Continuous Boyce
Index, and Tjur-R²) to evaluate the performance (explanatory power without
cross-validation) of Hmsc models.
}
\author{
Ahmed El-Gabbas
}
