% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Fit_WS.R
\name{Mod_Fit_WS}
\alias{Mod_Fit_WS}
\title{Fit Hmsc-HPC models on UFZ Windows Server}
\usage{
Mod_Fit_WS(Path_Model = NULL, EnvFile = ".env", NCores = NULL)
}
\arguments{
\item{Path_Model}{String. Path to the model files (without trailing slash)}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{NCores}{Integer. Number of parallel processes.}
}
\description{
Fit Hmsc-HPC models on UFZ Windows Server
}
\author{
Ahmed El-Gabbas
}
