% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Merge_Chains.R
\name{Mod_Merge_Chains}
\alias{Mod_Merge_Chains}
\title{Merge model chains}
\usage{
Mod_Merge_Chains(
  ModelDir = NULL,
  NCores = NULL,
  ModInfoName = NULL,
  PrintIncomplete = TRUE,
  FromHPC = TRUE,
  FromJSON = FALSE,
  Extension = "qs2"
)
}
\arguments{
\item{ModelDir}{Character. Path to the root directory of the fitted models.
Two folders will be created \code{Model_Fitted} and \code{Model_Coda} to store merged
model and coda objects, respectively.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.}

\item{ModInfoName}{Character. Name of the file (without extension) where the
processed model information will be saved. If \code{NULL}, it overwrites the
\code{Model_Info.RData} file in the \code{ModelDir} directory. If provided, a new
\code{.RData} file will be created with this name.}

\item{PrintIncomplete}{Logical. Whether to print the names of models that
were not successfully fitted to the console. Defaults to \code{TRUE}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{FromJSON}{Logical. Whether to convert loaded models from JSON format
before reading. Defaults to \code{FALSE}.}

\item{Extension}{Character. File extension (without dot) for the files
containing the fitted model and coda objects. It can be either \code{qs2} or
\code{RData}. Defaults to \code{qs2}.}
}
\value{
The function does not return anything but saves the processed model
information to disk.
}
\description{
This function merges model chains, checks for missing or incomplete model
runs, and optionally prints information about incomplete models. It saves the
processed model objects to disk.
}
\author{
Ahmed El-Gabbas
}
