% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_LF.R
\name{Mod_Plot_LF}
\alias{Mod_Plot_LF}
\title{Plot spatial variation in site loadings of HMSC Models}
\usage{
Mod_Plot_LF(
  Path_Model = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  Plot_Width = 20,
  Plot_Height = 21
)
}
\arguments{
\item{Path_Model}{Path to the model file.}

\item{EnvFile}{Path to the environment variables file. Default is ".env".}

\item{FromHPC}{Logical flag indicating whether to load environment variables
from HPC. Default is \code{TRUE}.}

\item{Plot_Width}{Numeric value specifying the width of the output plot in
cm. Default is 20.}

\item{Plot_Height}{Numeric value specifying the height of the output plot in
cm. Default is 21.}
}
\value{
Invisibly returns NULL after saving the plot.
}
\description{
Generate and save spatial variation in site loadings of HMSC Models' latent
factors as a JPEG file.
}
\author{
Ahmed El-Gabbas
}
