% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Prep4HPC.R
\name{Mod_Prep4HPC}
\alias{Mod_Prep4HPC}
\title{Prepare initial models in R to be used by the Hmsc-HPC}
\usage{
Mod_Prep4HPC(
  Hab_Abb = NULL,
  Path_Data = NULL,
  Path_Model = NULL,
  MinPresGrids = NULL,
  EnvFile = ".env",
  PrepareData = TRUE,
  GPP_Dists = NULL,
  GPP_Save = TRUE,
  GPP_Plot = TRUE,
  XVars = NULL,
  PhyloTree = TRUE,
  NoPhyloTree = TRUE,
  NParallel = 8,
  nChains = 4,
  thin = NULL,
  samples = NULL,
  transientFactor = 300,
  verbose = 1000,
  SkipFitted = TRUE,
  MaxJobCounts = 210,
  ModelCountry = NULL,
  MinPresPerCountry = 50,
  VerboseProgress = FALSE,
  FromHPC = TRUE,
  PrepSLURM = TRUE,
  MemPerCpu = NULL,
  Time = NULL,
  JobName = NULL,
  ToJSON = FALSE,
  ...
)
}
\arguments{
\item{Hab_Abb}{String. Habitat type. This has to be one of the following: c("0", "1", "2", "3", "4a", "4b", "5", "6", "8", "10", "12a", "12b"). "0" means prepare data irrespective of the habitat type}

\item{Path_Data}{String. Path to read modelling data (without trailing slash)}

\item{Path_Model}{String. Path to save all the output, including the to be fitted models (without trailing slash)}

\item{MinPresGrids}{Integer. Minimum number of presence grid cells per species. Only species with ≥ this number will be considered}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{PrepareData}{Logical. Should the input data be prepared or loaded from disk?}

\item{GPP_Dists}{Integer. Distance in kilometer for the distance between the knots and the minimum distance of a knot to the nearest data point. The same value will be used for the \code{knotDist}    and \code{minKnotDist}    arguments of the \code{Hmsc::constructKnots} function.}

\item{GPP_Save}{Logical. Save the resulted knots as RData. Default: \code{TRUE}}

\item{GPP_Plot}{Logical. Plot the coordinates of the sampling units and the knots. Default: \code{TRUE}}

\item{XVars}{String. Variables to be used in the model. Default value: \code{NULL}, which means to use the following variables: bio4, bio6, bio8, bio12, bio15, bio18, RoadRailLog, BiasLog}

\item{PhyloTree}{Logical. Fit model variants with phylogenetic trees. Default: \code{TRUE}. At least one of \code{PhyloTree} and \code{NoPhyloTree} can not be \code{FALSE}.}

\item{NoPhyloTree}{Logical. Fit model variants without phylogenetic trees. Default: \code{TRUE}. At least one of \code{PhyloTree} and \code{NoPhyloTree} can not be \code{FALSE}.}

\item{NParallel}{Number of parallel cores for parallelization}

\item{nChains}{Integer. Number of model chains}

\item{thin}{Integer. Value(s) for thinning}

\item{samples}{Integer. Value(s) for the number of MCMC samples}

\item{transientFactor}{Integer. Transient multiplication factor. The value of transient = \code{transientFactor * thin}.}

\item{verbose}{Integer. How often the results of the MCMC sampling be reported}

\item{SkipFitted}{Logic. Should the already fitted models be skipped.}

\item{MaxJobCounts}{Integer. Maximum number of jobs per slurm file}

\item{ModelCountry}{String. Filter observations to the following country list. Default: \code{NULL}, which means prepare data for the whole Europe}

\item{MinPresPerCountry}{Integer. Minimum of grid cells for the selected country/countries for species to be considered in the models. Effective only if a valid \code{ModelCountry} is provided.}

\item{VerboseProgress}{Logical. Show messages for the progress of creating files}

\item{FromHPC}{Logical. Work from HPC? This is to adjust the file paths.}

\item{PrepSLURM}{Logical. Prepare SLURM command? If \code{TRUE}, the SLURM commands will be exported using the \code{IASDT.R::Mod_SLURM} function.}

\item{MemPerCpu}{String. The value for the \verb{#SBATCH --mem-per-cpu=} SLURM argument. Example: "32G" to request 32 gigabyte. Only effective if \code{PrepSLURM = TRUE}.}

\item{Time}{String. The value for the requested time for each job in the bash arrays. Example: "01:00:00" to request an hour. Only effective if \code{PrepSLURM = TRUE}.}

\item{JobName}{String. The name of the submitted job(s). Only effective if \code{PrepSLURM = TRUE}.}

\item{ToJSON}{Logical. Convert unfitted models to JSON before saving to rds file}

\item{...}{additional parameters provided to the \code{IASDT.R::Mod_SLURM} function}
}
\description{
Prepare initial models in R to be used by the Hmsc-HPC
}
\author{
Ahmed El-Gabbas
}
