% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Prep4HPC.R
\name{Mod_Prep4HPC}
\alias{Mod_Prep4HPC}
\title{Prepare initial models in R for model fitting by Hmsc-HPC}
\usage{
Mod_Prep4HPC(
  Hab_Abb = NULL,
  Path_Model = NULL,
  MinEffortsSp = 100L,
  PresPerVar = 10L,
  EnvFile = ".env",
  GPP_Dists = NULL,
  GPP_Save = TRUE,
  GPP_Plot = TRUE,
  MinLF = NULL,
  MaxLF = NULL,
  BioVars = c("bio4", "bio6", "bio8", "bio12", "bio15", "bio18"),
  EffortsAsPredictor = FALSE,
  RoadRailAsPredictor = TRUE,
  HabAsPredictor = TRUE,
  NspPerGrid = 1L,
  ExcludeCult = TRUE,
  CV_NFolds = 4L,
  CV_NGrids = 20L,
  CV_NR = 2L,
  CV_NC = 2L,
  CV_Plot = TRUE,
  PhyloTree = TRUE,
  NoPhyloTree = TRUE,
  OverwriteRDS = TRUE,
  NCores = 8L,
  NChains = 4L,
  thin = NULL,
  samples = 1000L,
  transientFactor = 300L,
  verbose = 200L,
  SkipFitted = TRUE,
  NArrayJobs = 210L,
  ModelCountry = NULL,
  VerboseProgress = TRUE,
  FromHPC = TRUE,
  PrepSLURM = TRUE,
  MemPerCpu = NULL,
  Time = NULL,
  JobName = NULL,
  Path_Hmsc = NULL,
  ToJSON = FALSE,
  ...
)
}
\arguments{
\item{Hab_Abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type to
prepare data for. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a}, \verb{4b}, \code{10},
\verb{12a}, \verb{12b}. If \code{Hab_Abb} = \code{0}, data is prepared irrespective of the
habitat type. For more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{Path_Model}{String (without trailing slash) specifying the path where
all output, including models to be fitted, will be saved.}

\item{MinEffortsSp}{Minimum number of vascular plant species per grid cell in
GBIF for a grid cell to be included in the models. This is to exclude
grid cells with very little sampling efforts. Defaults to \code{100}.}

\item{PresPerVar}{Integer. Minimum number of presence grid cells per
predictor for a species to be included in the analysis. The number of
presence grid cells per species is calculated as the product of this factor
and the number of variables used in the models \code{NVars} and is calculated
after discarding grid cells with low sampling efforts (\code{MinEffortsSp}).
Defaults to \code{10}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{GPP_Dists}{Integer specifying the distance in kilometers for both the
distance between knots and the minimum distance of a knot to the nearest
data point. The GPP knots are prepared by the \link{PrepKnots}
function. The same value will be used for the \code{knotDist} and
\code{minKnotDist}    arguments of the \link[Hmsc:constructKnots]{Hmsc::constructKnots} function.}

\item{GPP_Save}{Logical indicating whether to save the resulted knots as
\code{RData} file. Default: \code{TRUE}.}

\item{GPP_Plot}{Logical indicating whether to plot the coordinates of the
sampling units and the knots in a pdf file. Default: \code{TRUE}.}

\item{MinLF, MaxLF}{integer. Minimum and maximum number of latent factors to
be used. Both default to \code{NULL} which means that the number of latent
factors will be estimated from the data. If either is provided, the
respective values will be used as arguments to \link[Hmsc:setPriors]{Hmsc::setPriors}.}

\item{BioVars}{Character vector. Specifies the Bioclimatic variables to be
included from CHELSA. Defaults to 6 ecologically meaningful and less
correlated Bioclimatic variables: \code{bio4}, \code{bio6}, \code{bio8}, \code{bio12},
\code{bio15}, and \code{bio18}.}

\item{EffortsAsPredictor}{Logical indicating whether to include the
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}}) sampling efforts as predictor to the model. Default:
\code{FALSE}.}

\item{RoadRailAsPredictor}{Logical indicating whether to include the
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}}) sum of road and railway intensity as predictor to the
model. Default: \code{TRUE}.}

\item{HabAsPredictor}{Logical indicating whether to include the
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}}) percentage coverage of respective habitat type per grid
cell as predictor to the model. Default: \code{TRUE}. Only valid if \code{Hab_Abb}
not equals to \code{0}.}

\item{NspPerGrid}{Integer. Indicating the minimum number of species per grid
cell for a grid cell to be include in the analysis. This is calculated
after filtering grid cells by sampling efforts (\code{MinEffortsSp}) and
filtering species by the number of presence grid cells per predictor
(\code{PresPerVar}). If \code{NspPerGrid} = \code{NULL} or 1 (default), grid cells with
at least one species presence will be considered in the models.}

\item{ExcludeCult}{Logical. Indicates whether to exclude countries with
cultivated or casual observations per species. Defaults to \code{TRUE}.}

\item{CV_NFolds}{Number of cross-validation folds. Default: 4.}

\item{CV_NGrids}{For \code{CV_Dist} cross-validation strategy (see below), this
argument determines the size of the blocks (how many grid cells in both
directions).}

\item{CV_NR, CV_NC}{Integer, the number of rows and columns used in the
\code{CV_Large} cross-validation strategy (see below), in which the study area
is divided into large blocks given the provided \code{CV_NR} and \code{CV_NC}
values. Both default to 2 which means to split the study area into four
large blocks at the median latitude and longitude.}

\item{CV_Plot}{Logical. Indicating whether to plot the block cross-validation
folds.}

\item{PhyloTree, NoPhyloTree}{Logical indicating whether to fit model variants
with or without phylogenetic trees, respectively. The default of both
arguments is \code{TRUE}, which means to fit a model variant with the respective
option. If both \code{PhyloTree} and \code{NoPhyloTree} are \code{TRUE} (Default), models
for both options will be fitted. At least one of \code{PhyloTree} and
\code{NoPhyloTree} should be \code{TRUE}.}

\item{OverwriteRDS}{Logical. Indicating whether to overwrite previously
exported RDS files for initial models. Default: \code{TRUE}.}

\item{NCores}{Integer specifying the number of parallel cores for
parallelization. Default: 8 cores.}

\item{NChains}{Integer specifying the number of model chains. Default: 4.}

\item{thin}{Integer specifying the value(s) for thinning in MCMC sampling. If
more than one value is provided, a separate model will be fitted at each
value of thinning.}

\item{samples}{Integer specifying the value(s) for the number of MCMC
samples. If more than one value is provided, a separate model will be
fitted at each value of number of samples. Defaults to 1000.}

\item{transientFactor}{Integer specifying the transient multiplication
factor. The value of \code{transient} will equal  the multiplication of
\code{transientFactor} and \code{thin}. Default: 300.}

\item{verbose}{Integer specifying how often the results of the MCMC sampling
should be reported. Default: \code{200}.}

\item{SkipFitted}{Logical indicating whether to skip already fitted models.
Default: \code{TRUE}.}

\item{NArrayJobs}{Integer specifying the maximum allowed number of array
jobs per SLURM file. Default: 210. See
\href{https://docs.lumi-supercomputer.eu/runjobs/scheduled-jobs/partitions}{LUMI documentation}
for more details.}

\item{ModelCountry}{String or vector of strings specifying the country or
countries to filter observations by. Default: \code{NULL}, which means prepare
data for the whole Europe.}

\item{VerboseProgress}{Logical. Indicates whether progress messages should be
displayed. Defaults to \code{TRUE}.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{PrepSLURM}{Logical indicating whether to prepare SLURM command files.
If \code{TRUE} (default), the SLURM commands will be saved to disk using the
\link{Mod_SLURM} function.}

\item{MemPerCpu}{String specifying the memory per CPU for the SLURM job. This
value will be assigned to the \verb{#SBATCH --mem-per-cpu=} SLURM argument.
Example: "32G" to request 32 gigabyte. Only effective if \code{PrepSLURM = TRUE}.}

\item{Time}{String specifying the requested time for each job in the SLURM
bash arrays. Example: "01:00:00" to request an hour. Only effective if
\code{PrepSLURM = TRUE}.}

\item{JobName}{String specifying the name of the submitted job(s) for SLURM.
If \code{NULL} (Default), the job name will be prepared based on the folder path
and the \code{Hab_Abb} value. Only effective if \code{PrepSLURM = TRUE}.}

\item{Path_Hmsc}{String specifying the path for the Hmsc-HPC. This will be
provided as the \code{Path_Hmsc} argument of the \link{Mod_SLURM} function.}

\item{ToJSON}{Logical indicating whether to convert unfitted models to JSON
before saving to RDS file. Default: \code{FALSE}.}

\item{...}{Additional parameters provided to the \link{Mod_SLURM}
function.}
}
\value{
The function is used for its side effects of preparing data and
models for HPC and does not return any value.
}
\description{
This function prepares initial models in R for model fitting by Hmsc-HPC. It
involves data preparation, define spatial block cross-validation folds, model
initialization, GPP knots, and generating commands for running models on HPC.
It supports parallel processing, options to include/not include phylogenetic
tree data. The models will be fitted using Gaussian Predictive Process (GPP;
see \href{https://doi.org/10.1002/ecy.2929}{Tikhonov et al.}) for more details)
via the \href{https://doi.org/10.1101/2024.02.13.580046}{Hmsc-HPC} extension.
}
\details{
The function provides options for:
\itemize{
\item for which habitat types the models will be fitted
\item excluding grid cells with very low sampling efforts (\code{MinEffortsSp})
\item selection of species based on minimum number of presence-grid cells
(\code{PresPerVar} * number of predictors)
\item optionally model fitting on specified list of countries: (\code{ModelCountry})
\item whether to exclude grid cells with few species (\code{NspPerGrid})
\item number of cross-validation folds
\item options for whether or not to include phylogenetic information to the model
\item different values for knot distance for GPP (\code{GPP_Dists})
\item which Bioclimatic variables to be uses in the models (\code{BioVars})
\item whether to include sampling efforts \code{EffortsAsPredictor}, percentage of
respective habitat type per grid cell \code{HabAsPredictor}, and railway and
road intensity per grid cell \code{RoadRailAsPredictor}
\item Hmsc options (\code{NChains}, \code{thin}, \code{samples}, \code{transientFactor}, and \code{verbose})
\item prepare SLURM commands (\code{PrepSLURM}) and some specifications (e.g.
\code{NArrayJobs}, \code{MemPerCpu}, \code{Time}, \code{JobName})

The function reads the following environment variables:
\itemize{
\item \strong{\code{DP_R_Grid}} (if \code{FromHPC = TRUE}) or
\strong{\code{DP_R_Grid_Local}} (if \code{FromHPC = FALSE}). The function reads
the content of the \code{Grid_10_Land_Crop.RData} file from this path.
\item \strong{\code{DP_R_Path_Python}}: Python path on LUMI
\item \strong{\code{DP_R_TaxaInfo}} or \strong{\code{DP_R_TaxaInfo_Local}} for the location of the
\code{Species_List_ID.txt} file representing species information.
\item \strong{\code{DP_R_EUBound_sf}} or \strong{\code{DP_R_EUBound_sf_Local}} for the path of the
\code{RData} file containing the country boundaries (\code{sf} object)
\item \strong{\code{DP_R_PA}} or \strong{\code{DP_R_PA_Local}}: The function reads the contents of
the \code{Sp_PA_Summary_DF.RData} file from this path
}
}
}
\author{
Ahmed El-Gabbas
}
