% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Prep4HPC.R
\name{Mod_Prep4HPC}
\alias{Mod_Prep4HPC}
\title{Prepare initial models in R for model fitting by Hmsc-HPC}
\usage{
Mod_Prep4HPC(
  Hab_Abb = NULL,
  Path_Data = NULL,
  Path_Model = NULL,
  MinPresGrids = 50,
  EnvFile = ".env",
  PrepareData = TRUE,
  GPP_Dists = NULL,
  GPP_Save = TRUE,
  GPP_Plot = TRUE,
  XVars = NULL,
  PhyloTree = TRUE,
  NoPhyloTree = TRUE,
  NParallel = 8,
  nChains = 4,
  thin = NULL,
  samples = NULL,
  transientFactor = 300,
  verbose = 1000,
  SkipFitted = TRUE,
  MaxJobCounts = 210,
  ModelCountry = NULL,
  MinPresPerCountry = 50,
  VerboseProgress = FALSE,
  FromHPC = TRUE,
  PrepSLURM = TRUE,
  MemPerCpu = NULL,
  Time = NULL,
  JobName = NULL,
  Path_Hmsc = NULL,
  ToJSON = FALSE,
  ...
)
}
\arguments{
\item{Hab_Abb}{String indicating the habitat type to be considered for model preparation. This has to be one of the following: c("0", "1", "2", "3", "4a", "4b", "5", "6", "8", "10", "12a", "12b"). "0" means prepare data irrespective of the habitat type}

\item{Path_Data}{String specifying the path where modeling data is read from.}

\item{Path_Model}{String (without trailing slash) specifying the path where all output, including models to be fitted, will be saved.}

\item{MinPresGrids}{Integer indicating the minimum number of presence grid cells per species for a species to be used in the model. Only species with ≥ this number will be considered. Default: 50.}

\item{EnvFile}{String specifying the path to read environment variables from, with a default value of ".env".}

\item{PrepareData}{Logical indicating whether to prepare input data or load it from disk. Defaults to \code{TRUE} which means the input data will be prepared using the \link{Mod_PrepData} function.}

\item{GPP_Dists}{Integer specifying the distance in kilometers for both the distance between knots and the minimum distance of a knot to the nearest data point. The same value will be used for the \code{knotDist} and \code{minKnotDist}    arguments of the \link[Hmsc:constructKnots]{Hmsc::constructKnots} function.}

\item{GPP_Save}{Logical indicating whether to save the resulted knots as RData Default: \code{TRUE}.}

\item{GPP_Plot}{Logical indicating whether to plot the coordinates of the sampling units and the knots in a pdf file. Default: \code{TRUE}}

\item{XVars}{Vector of strings specifying variables to be used in the model. Default value: \code{NULL}, which means to use the following variables: bio4, bio6, bio8, bio12, bio15, bio18, RoadRailLog, BiasLog.}

\item{PhyloTree, NoPhyloTree}{Logical indicating whether to fit model variants  with or without phylogenetic trees, respectively. Default: \code{TRUE}, which means to fit a model variant with the respective option. If both \code{PhyloTree} and \code{NoPhyloTree} are \code{TRUE} (Default), models for both options will be fitted. At least one of \code{PhyloTree} and \code{NoPhyloTree} should be \code{TRUE}.}

\item{NParallel}{Integer specifying the number of parallel cores for parallelization. Default: 8 cores.}

\item{nChains}{Integer specifying the number of model chains. Default: 4.}

\item{thin}{Integer specifying the value(s) for thinning in MCMC sampling. If more than one value is provided, a separate model will be fitted at each value of thinning.}

\item{samples}{Integer specifying the value(s) for the number of MCMC samples. If more than one value is provided, a separate model will be fitted at each value of number of samples.}

\item{transientFactor}{Integer specifying the transient multiplication factor. The value of \code{transient} will equal  the multiplication of \verb{transientFactor and thin}. Default: 300.}

\item{verbose}{Integer specifying how often the results of the MCMC sampling should be reported. Default: 1000.}

\item{SkipFitted}{Logical indicating whether to skip already fitted models. Default: \code{TRUE}.}

\item{MaxJobCounts}{Integer specifying the maximum allowed number of array jobs per SLURM file. Default: 210. See \href{https://docs.lumi-supercomputer.eu/runjobs/scheduled-jobs/partitions/}{here} for more details.}

\item{ModelCountry}{String or vector of strings specifying the country or countries to filter observations by. Default: \code{NULL}, which means prepare data for the whole Europe.}

\item{MinPresPerCountry}{Integer specifying the minimum number of grid cells for the selected country/countries for species to be considered in the models. Effective only if a valid \code{ModelCountry} is provided. Default: 50.}

\item{VerboseProgress}{Logical indicating whether to show messages for the progress of creating files. Default: \code{FALSE}.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to adjust file paths accordingly. Default: \code{TRUE}.}

\item{PrepSLURM}{Logical indicating whether to prepare SLURM command files. If \code{TRUE} (default), the SLURM commands will be saved to disk using the \link{Mod_SLURM} function.}

\item{MemPerCpu}{String specifying the memory per CPU for the SLURM job. This value will be assigned to the \verb{#SBATCH --mem-per-cpu=} SLURM argument. Example: "32G" to request 32 gigabyte. Only effective if \code{PrepSLURM = TRUE}.}

\item{Time}{String specifying the requested time for each job in the SLURM bash arrays. Example: "01:00:00" to request an hour. Only effective if \code{PrepSLURM = TRUE}.}

\item{JobName}{String specifying the name of the submitted job(s) for SLURM. If \code{NULL} (Default), the job name will be prepared based on the folder path and the \code{Hab_Abb} value. Only effective if \code{PrepSLURM = TRUE}.}

\item{Path_Hmsc}{String specifying the path for the Hmsc-HPC. This will be provided as the \code{Path_Hmsc} argument of the \link{Mod_SLURM} function.}

\item{ToJSON}{Logical indicating whether to convert unfitted models to JSON before saving to RDS file. Default: \code{FALSE}.}

\item{...}{Additional parameters provided to the \link{Mod_SLURM} function.}
}
\value{
The function is used for its side effects of preparing data and models for HPC and does not return any value.
}
\description{
This function prepares initial models in R for model fitting by Hmsc-HPC. It involves data preparation, model initialization, and generating commands for running models on HPC. It supports parallel processing, options to include/not include phylogenetic tree data.
}
\author{
Ahmed El-Gabbas
}
