% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Postprocess.R
\name{Mod_Prep_TF}
\alias{Mod_Prep_TF}
\title{Prepare Batch Scripts for LF Prediction on GPU}
\usage{
Mod_Prep_TF(
  Path = "datasets/processed/model_fitting",
  NumFiles = 210,
  Pattern = "LF_.+_Commands_.+.txt|LF_.+_Commands_.+txt|VP_.+Command.txt",
  WD = NULL,
  Path_Out = "TF_BatchFiles"
)
}
\arguments{
\item{Path}{Character. Directory containing input files with commands.}

\item{NumFiles}{Integer. Number of output batch files to create. Must be less
than or equal to the maximum job limit of the HPC environment.}

\item{Pattern}{Character. Regex pattern to match input files.}

\item{WD}{Character. Working directory for batch files. If \code{NULL}, defaults
to the current directory.}

\item{Path_Out}{Character. Directory to save output files. Default is
\code{TF_BatchFiles}.}
}
\value{
None. Writes batch files to \code{Path_Out}.
}
\description{
This function automates the preparation of batch scripts for predicting the
latent factors (LF) of response curves and at new sites using GPU resources.
It reads input files that match a specified pattern, merges their contents,
sorts the commands, and distributes them into a user-defined number of output
files. Each batch script is designed to be compatible with an HPC
environment, such as LUMI, with TensorFlow setup included. The function
limits the number of output files to a specified maximum, defaults to 210 for
compatibility with LUMI's job limits.
}
\note{
This function is designed specifically for the LUMI HPC environment. It
assumes the \code{tensorflow} module is available and pre-configured with all
necessary Python packages. On other HPC systems, users may need to modify
the function to load a Python virtual environment or install the required
dependencies for TensorFlow and related packages.
}
\author{
Ahmed El-Gabbas
}
