% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepData.R
\name{Mod_PrepData}
\alias{Mod_PrepData}
\title{Prepare habitat-specific data for Hmsc models}
\usage{
Mod_PrepData(
  Hab_Abb = NULL,
  MinEffortsSp = 100L,
  ExcludeCult = TRUE,
  ExcludeZeroHabitat = TRUE,
  PresPerSpecies = 80L,
  EnvFile = ".env",
  Path_Model = NULL,
  VerboseProgress = TRUE,
  FromHPC = TRUE
)
}
\arguments{
\item{Hab_Abb}{Character. Abbreviation for the habitat type (based on
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab}) for which to prepare
data. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a}, \verb{4b}, \code{10}, \verb{12a}, \verb{12b}.
If \code{Hab_Abb} = \code{0}, data is prepared irrespective of the habitat type. For
more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{MinEffortsSp}{Integer specifying the minimum number of vascular plant
species per grid cell (from GBIF data) required for inclusion in the
models. This is to exclude grid cells with very little sampling efforts.
Defaults to \code{100}.}

\item{ExcludeCult}{Logical. Indicates whether to exclude countries with
cultivated or casual observations per species. Defaults to \code{TRUE}.}

\item{ExcludeZeroHabitat}{Logical. Indicates whether to exclude grid cells
with zero habitat coverage. Defaults to \code{TRUE}.}

\item{PresPerSpecies}{Integer. The minimum number of presence grid cells for
a species to be included in the analysis. The number of presence grid cells
per species is calculated after discarding grid cells with low sampling
efforts (\code{MinEffortsSp}). Defaults to \code{80}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{Path_Model}{Character. Path where the output file should be saved.}

\item{VerboseProgress}{Logical. Indicates whether progress messages should be
displayed. Defaults to \code{TRUE}.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}
}
\value{
a tibble containing modelling data.
}
\description{
This function processes environmental and species presence data to prepare
habitat-specific datasets for use in Hmsc models. It checks input arguments,
reads environment variables from a file, verifies paths, loads and filters
species data based on habitat type and minimum presence grid cells per
species, and merges various environmental layers (e.g., CHELSA Bioclimatic
variables, habitat coverage, road and railway intensity, sampling efforts)
into a single dataset. Processed data is saved to disk as an \verb{*.RData}
file.
}
\details{
The function reads the following environment variables:
\itemize{
\item \strong{\code{DP_R_Grid}} (if \code{FromHPC} = \code{TRUE}) or \strong{\code{DP_R_Grid_Local}} (if
\code{FromHPC} = \code{FALSE}). The function reads the content of the
\code{Grid_10_Land_Crop.RData} and \code{Grid_10_Land_Crop_sf_Country.RData} files
\item \strong{\code{DP_R_Grid_Ref}} or \strong{\code{DP_R_Grid_Ref_Local}}: The function reads the
content of \code{Grid_10_sf.RData} file from this path.
\item \strong{\code{DP_R_PA}} or \strong{\code{DP_R_PA_Local}}: The function reads the contents
of the \code{Sp_PA_Summary_DF.RData} file from this path.
\item \strong{\code{DP_R_CLC_Summary}} / \strong{\code{DP_R_CLC_Summary_Local}}: Path containing
the \code{PercCov_SynHab_Crop.RData} file. This file contains maps for the
percentage coverage of each SynHab habitat type per grid cell.
\item \strong{\code{DP_R_CHELSA_Output}} / \strong{\code{DP_R_CHELSA_Output_Local}}: Path
for processed CHELSA data.
\item \strong{\code{DP_R_Roads}} / \strong{\code{DP_R_Roads_Local}}: Path for processed road data.
The function reads the contents of: \code{Road_Length.RData} for the total length
of any road type per grid cell.
\item \strong{\code{DP_R_Railway}} / \strong{\code{DP_R_Railway_Local}}: Path for processed
railway data. The function reads the contents of: \code{Railway_Length.RData} for
the total length of any railway type per grid cell.
\item \strong{\code{DP_R_Efforts}} / \strong{\code{DP_R_Efforts_Local}}: Path for processed
sampling efforts analysis. The function reads the content of
\code{Bias_GBIF_SummaryR.RData} file containing the total number of GBIF vascular
plant observations per grid cell.
}

The current models are fitted for 8 habitat types see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}:
\itemize{
\item \strong{1. Forests} -- closed vegetation dominated by deciduous or evergreen
trees
\item \strong{2. Open forests} -- woodlands with canopy openings created by
environmental stress or disturbance, including forest edges
\item \strong{3. Scrub} -- shrublands maintained by environmental stress (aridity) or
disturbance
\item \strong{4a. Natural grasslands} -- grasslands mantained by climate (aridity,
unevenly distributed precipitation), herbivores or environmental stress
(aridity, instability or toxicity of substrate)
\item \strong{4b. Human-maintained grasslands} -- grasslands dependent on regular
human-induced management (mowing, grazing by livestock, artificial burning)
\item \strong{10. Wetland} -- sites with the permanent or seasonal influence of
moisture, ranging from oligotrophic to eutrophic
\item \strong{12a. Ruderal habitats} -- anthropogenically disturbed or eutrophicated
sites, where the anthropogenic disturbance or fertilization is typically a
side-product and not the aim of the management
\item \strong{12b. Agricultural habitats} -- synanthropic habitats directly
associated with growing of agricultural products, thus dependent on specific
type of management (ploughing, fertilization)
}

The following habitat types are excluded from the analysis:
\itemize{
\item \strong{5. Sandy} -- dunes and other habitats on unstable sandy substrate, stressed
by low nutrients, drought and disturbed by sand movement
\item \strong{6. Rocky} -- cliffs and rock outcrops with very shallow or no soil
\item \strong{7. Dryland} -- habitats in which drought stress limits vegetation
development
\item \strong{8. Saline} -- habitats stressed by high soil salinity
\item \strong{9. Riparian} -- a mosaic of wetlands, grasslands, tall-forb stands,
scrub and open forests in stream corridors
\item \strong{11. Aquatic} -- water bodies and streams with submerged and floating
plant species
}
}
\author{
Ahmed El-Gabbas
}
