% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_SLURM.R
\name{Mod_SLURM}
\alias{Mod_SLURM}
\title{Prepare SLURM file for model fitting on HPC}
\usage{
Mod_SLURM(
  Path_Model = NULL,
  JobName = NULL,
  CatJobInfo = TRUE,
  ntasks = 1,
  CpusPerTask = 1,
  GpusPerNode = 1,
  MemPerCpu = NULL,
  Time = NULL,
  Partition = "small-g",
  EnvFile = ".env",
  FromHPC = TRUE,
  Path_Hmsc = NULL,
  Command_Prefix = "Commands_All",
  SLURM_Prefix = "Bash_Fit",
  Path_SLURM_Out = NULL
)
}
\arguments{
\item{Path_Model}{String. Path to the model files (without trailing slash).}

\item{JobName}{String. The name of the submitted job(s).}

\item{CatJobInfo}{Logical. Add bash lines to print information on the submitted job. Default: \code{TRUE}.}

\item{ntasks}{Integer. The value for the \verb{#SBATCH --ntasks=} SLURM argument. Default: 1.}

\item{CpusPerTask}{Integer. The value for the \verb{#SBATCH --cpus-per-task=} SLURM argument. Default: 1.}

\item{GpusPerNode}{Integer. The value for the \verb{#SBATCH --gpus-per-node=} SLURM argument. Default: 1.}

\item{MemPerCpu}{String. The value for the \verb{#SBATCH --mem-per-cpu=} SLURM argument. Example: "32G" to request 32 gigabytes.}

\item{Time}{String. The value for the requested time for each job in the bash arrays. Example: "01:00:00" to request an hour.}

\item{Partition}{String. The name of the partition. Default: \code{small-g}.}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}.}

\item{FromHPC}{Logical. Indicates if the operation is being performed from an HPC environment. This adjusts file paths accordingly. Default: \code{TRUE}.}

\item{Path_Hmsc}{String. Path for the Hmsc-HPC.}

\item{Command_Prefix}{String. Prefix for the bash commands to be executed.}

\item{SLURM_Prefix}{String. Prefix for the exported SLURM file.}

\item{Path_SLURM_Out}{String indicating the directory where the SLURM file(s) will be saved. Defaults to \code{NULL}, which means to identify the path from \code{Path_Model}.}
}
\value{
The function does not return any value but writes SLURM script files to the disk.
}
\description{
This function prepares and writes a SLURM file(s) for submitting model fitting jobs on an HPC environment. It dynamically generates bash scripts based on the provided parameters, which are then used to submit jobs to the SLURM workload manager.
}
\author{
Ahmed El-Gabbas
}
