% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_SLURM_Refit.R
\name{Mod_SLURM_Refit}
\alias{Mod_SLURM_Refit}
\title{Prepare SLURM file for failed / Unfitted Models}
\usage{
Mod_SLURM_Refit(
  ModelDir = NULL,
  NumArrayJobs = 210,
  JobName = NULL,
  MemPerCpu = NULL,
  Time = NULL,
  Partition = "small-g",
  EnvFile = ".env",
  CatJobInfo = TRUE,
  ntasks = 1,
  CpusPerTask = 1,
  GpusPerNode = 1,
  FromHPC = TRUE,
  PrepSLURM = TRUE,
  Path_Hmsc = NULL,
  Refit_Prefix = "Commands2Refit",
  SLURM_Prefix = "Bash_Refit"
)
}
\arguments{
\item{ModelDir}{String. Path to the model files directory. Must not end
with a slash.}

\item{NumArrayJobs}{Integer specifying the maximum allowed number of array
jobs per SLURM file. Default: 210. See \href{https://docs.lumi-supercomputer.eu/runjobs/scheduled-jobs/partitions}{LUMI documentation}
for more details.}

\item{JobName}{String. The name of the submitted job(s).}

\item{MemPerCpu}{String. Memory per CPU allocation for the SLURM job.
Example: \verb{32G} for 32 gigabytes. Defaults to \code{NULL}. If not provided, the
function will throw an error.}

\item{Time}{String. Duration for which the job should run. Example:
\code{01:00:00} for one hour. If not provided, the function will throw an error.}

\item{Partition}{String. The name of the partition. Default: \code{small-g} for
running the array jobs on the GPU.}

\item{EnvFile}{String. Path to read the environment variables. Default value:
\code{.env}.}

\item{CatJobInfo}{Logical. Add bash lines to print information on the
submitted job. Default: \code{TRUE}.}

\item{ntasks}{Integer. The value for the \verb{#SBATCH --ntasks=} SLURM argument.
Default: 1.}

\item{CpusPerTask}{Integer. The value for the \verb{#SBATCH --cpus-per-task=}
SLURM argument. Default: 1.}

\item{GpusPerNode}{Integer. The value for the \verb{#SBATCH --gpus-per-node=}
SLURM argument. Default: 1.}

\item{FromHPC}{Logical. Indicates if the operation is being performed from an
HPC environment. This adjusts file paths accordingly. Default: \code{TRUE}.}

\item{PrepSLURM}{Logical indicating whether to prepare SLURM command files.
If \code{TRUE} (default), the SLURM commands will be saved to disk using the
\link{Mod_SLURM} function.}

\item{Path_Hmsc}{String. Path for the Hmsc-HPC.}

\item{Refit_Prefix}{String. Prefix for the files that contain the commands to
refit the models.}

\item{SLURM_Prefix}{String. Prefix for the exported SLURM file.}
}
\value{
The function does not return any value but creates command and SLURM
batch files for refitting models.
}
\description{
This function prepares SLURM files for models that have not been successfully
fitted yet. It generates command files for refitting and creates
corresponding SLURM batch files to execute these commands on a HPC
environment.
}
\author{
Ahmed El-Gabbas
}
