% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{NDecimals}
\alias{NDecimals}
\title{Number of decimal places in a vector}
\usage{
NDecimals(x)
}
\arguments{
\item{x}{vector}
}
\value{
integer; number of decimal places
}
\description{
Number of decimal places in a vector
}
\examples{
NDecimals("13.45554545")

# -------------------------------------------

NDecimals(15.01500)

NDecimals('15.01500')

# -------------------------------------------

NDecimals(13.45554545)
}
\author{
Ahmed El-Gabbas
}
