% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_NetCDFGlobalAttrs.R
\name{NetCDFGlobalAttrs}
\alias{NetCDFGlobalAttrs}
\title{Get global attributes for NetCDF files}
\usage{
NetCDFGlobalAttrs(nc)
}
\arguments{
\item{nc}{path to the NetCDF file}
}
\description{
https://github.com/rspatial/terra/issues/1443
}
\details{
Get global attributes for NetCDF files
}
