% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Package_RemoteSha.R
\name{Package_RemoteSha}
\alias{Package_RemoteSha}
\title{Get remote sha of R packages}
\usage{
Package_RemoteSha(...)
}
\arguments{
\item{...}{name of one or more R packages}
}
\description{
Get remote sha of R packages
}
\examples{
# Package_RemoteSha(IASDT.R, devtools)
}
