% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Package_RemoteSha.R
\name{Package_RemoteSha}
\alias{Package_RemoteSha}
\title{Get the remote SHA of an R packages}
\usage{
Package_RemoteSha(...)
}
\arguments{
\item{...}{Character. Names of one or more R packages for which to retrieve
the remote SHA.}
}
\value{
A named character vector where names are the package names and values
are the corresponding remote SHAs.
}
\description{
This function retrieves the remote SHA (Secure Hash Algorithm) reference for
one or more specified R packages. It is useful for tracking the exact version
of a package's source code.
}
\examples{
Package_RemoteSha(IASDT.R, devtools)
}
\author{
Ahmed El-Gabbas
}
