% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameter_heatmap.R
\name{parameter_heatmap}
\alias{parameter_heatmap}
\alias{mod_heatmap_beta}
\alias{mod_heatmap_omega}
\title{Heatmaps for the \code{beta} and \code{omega} parameters of the Hmsc model}
\usage{
mod_heatmap_beta(
  path_model = NULL,
  support_level = 0.95,
  width = 25,
  height = 35
)

mod_heatmap_omega(
  path_model = NULL,
  support_level = 0.95,
  width = 26,
  height = 22.5
)
}
\arguments{
\item{path_model}{Character. Path to the fitted \code{Hmsc} model object.}

\item{support_level}{Numeric. The posterior support threshold for determining
which values are considered significant in the heatmap. Defaults to 0.95,
indicating 95\% posterior support. Values above this threshold (or below 1 -
threshold for negative associations) are considered significant and will be
plotted (see \link[Hmsc:plotBeta]{Hmsc::plotBeta}).}

\item{width, height}{Integer. The width and height of the generated
heatmaps in centimetres. Defaults to 26×22.5 for \code{omega}; 25×35
for \code{beta}.}
}
\value{
Both functions do not return a value but saves heatmap plots as JPEG
files in the \code{model_postprocessing/parameters_summary} subdirectory.
}
\description{
The \code{mod_heatmap_beta()} and \code{mod_heatmap_omega()} functions generate
heatmaps using \code{ggplot2} to visualise parameter estimates or posterior
support values for species' environmental responses (\code{beta} parameters, which
describes how species (\emph{Y}) respond to various covariates (\emph{X}); see
\link[Hmsc:plotBeta]{Hmsc::plotBeta}) and residual associations (\code{omega} parameter),
respectively.
}
\details{
The functions exports three types of visualisations (see
\link[Hmsc:plotBeta]{Hmsc::plotBeta}):
\itemize{
\item \code{Mean}: posterior mean estimate,
\item \code{support}: statistical support level, measured by the posterior
probability for a positive or negative response,
\item \code{Sign}: indicates whether the response is positive, negative, or neither
of these based on the chosen \code{support_level}.
}

For the \code{omega} parameter, the \code{mod_heatmap_omega()} function generates two
JPEG files: signs and mean values. While for the \code{beta} parameter, the
\code{mod_heatmap_beta()} function generates four JPEG files : support, signs,
mean values (including and excluding the intercept).
}
\author{
Ahmed El-Gabbas. The \code{mod_heatmap_beta()} function is adapted from
\link[Hmsc:plotBeta]{Hmsc::plotBeta}
}
