% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_prediction.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot species and level of invasion predictions as JPEG files using \code{ggplot2}}
\usage{
plot_prediction(model_dir = NULL, env_file = ".env", n_cores = 8L)
}
\arguments{
\item{model_dir}{Character. Path to the model directory containing
predictions.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}
}
\value{
Saves prediction plots as JPEG files in the specified output
directory.
}
\description{
Generate predictions for species and habitat models and saves the output as
JPEG files.
}
\author{
Ahmed El-Gabbas
}
