% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotAlpha.R
\name{PlotAlpha}
\alias{PlotAlpha}
\title{Plot convergence traceplots for the alpha parameter}
\usage{
PlotAlpha(
  Post = NULL,
  Model = NULL,
  Title = NULL,
  NRC = NULL,
  AddFooter = TRUE,
  AddTitle = TRUE,
  Cols = NULL,
  FromHPC = TRUE
)
}
\arguments{
\item{Post}{A \code{coda} object or a string path to a saved \code{coda} object
containing the posterior samples for the alpha parameter.}

\item{Model}{A fitted model object or a string path to a saved model object
containing the samples and metadata necessary for plotting.}

\item{Title}{A string specifying the main title of the plot.}

\item{NRC}{numeric vector of length 2 specifying the number of rows and
columns per page of plots. If \code{NULL}, it is automatically determined based
on the number of levels of the alpha parameter.}

\item{AddFooter}{A logical value indicating whether to add a footer with page
numbers to each plot page. Defaults to \code{TRUE}.}

\item{AddTitle}{A logical value indicating whether to add the main title
(specified by \code{Title}) to the plot. Defaults to \code{TRUE}.}

\item{Cols}{Character vector for chain colours (optional). Default: \code{NULL}.}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}
}
\value{
A grid of \code{ggplot2} objects representing the traceplots for the alpha
parameter, arranged according to the specified number of rows and columns
per page. If \code{AddTitle} or \code{AddFooter} is \code{TRUE}, additional text elements
are included as specified.
}
\description{
This function generates and plots convergence traceplots for the alpha
parameter from an Hmsc model, using various diagnostics and visualizations to
assess convergence. It supports customization of plot aesthetics and layout.
}
\author{
Ahmed El-Gabbas
}
