% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotBetaGG.R
\name{PlotBetaGG}
\alias{PlotBetaGG}
\title{Heatmaps of parameter estimates or posterior support values of species'
environmental responses (Beta parameters)}
\usage{
PlotBetaGG(
  Path_Model = NULL,
  supportLevel = 0.95,
  PlotWidth = 26,
  PlotHeight = 25
)
}
\arguments{
\item{Path_Model}{String. Path to the fitted Hmsc model object.}

\item{supportLevel}{Numeric. The threshold for posterior support used in
plotting. Values above this threshold (and below 1 - threshold) are
considered significant and will be plotted. The default value is 0.95,
indicating 95\% posterior support. For more information, see
\link[Hmsc:plotBeta]{Hmsc::plotBeta}}

\item{PlotWidth, PlotHeight}{Numeric. The width and height of the plot in
centimeters. Default is \code{26} cm x \code{25} cm.}
}
\value{
The function does not return a value but saves heatmap plots as JPEG
files in a directory related to the model's path.
}
\description{
This function generates heatmaps of parameter estimates or posterior support
values for species' environmental responses, represented by Beta parameters.
It is designed to visualize how species (Y) respond to various covariates (X)
using \code{ggplot2} for plotting. The function is an adaptation of
\link[Hmsc:plotBeta]{Hmsc::plotBeta}, focusing on \code{ggplot2}-based visualizations.
}
