% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotConvergence.R
\name{PlotConvergence}
\alias{PlotConvergence}
\title{Plot model convergence of a selected model}
\usage{
PlotConvergence(
  Path_Model = NULL,
  Path_Coda = NULL,
  Path_FittedModel = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NChains = 4,
  Title = " ",
  NOmega = 1000,
  NCores = NULL,
  NRC = c(2, 3),
  SavePlotData = FALSE,
  Cols = c("red", "blue", "darkgreen", "darkgrey")
)
}
\arguments{
\item{Path_Model}{String. Path to save all the output, including the to be fitted models (without trailing slash)}

\item{Path_Coda}{String. Path to the coda object.}

\item{Path_FittedModel}{String. Path to the model object}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{FromHPC}{Logical. Work from HPC? This is to adjust the file paths.}

\item{NChains}{Integer. Number of model chains}

\item{Title}{String. Title of the rho and alpha plots}

\item{NOmega}{Number of sample species interactions}

\item{NCores}{Number of parallel cores for parallelization}

\item{NRC}{Vector. Number of rows and columns}

\item{SavePlotData}{Logical. Save plots as RData file}

\item{Cols}{Colours for lines for each chain}
}
\description{
Plot model convergence of a selected model
}
\author{
Ahmed El-Gabbas
}
