% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotConvergence_All.R
\name{PlotConvergence_All}
\alias{PlotConvergence_All}
\title{Plot model convergence of multiple modelling alternatives}
\usage{
PlotConvergence_All(
  Path_Model = NULL,
  maxOmega = 1000,
  NCores = NULL,
  FromHPC = TRUE
)
}
\arguments{
\item{Path_Model}{String. Path to the root directory of the fitted models
without the trailing slash. The convergence outputs will be saved to the
\code{Model_Convergence_All} subfolder.}

\item{maxOmega}{Integer. Maximum number of species interactions to sample for
convergence diagnostics.}

\item{NCores}{Integer. Number of cores to use for parallel processing.}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}
}
\value{
The function does not return anything but saves a series of
diagnostic plots in the specified path.
}
\description{
This function generates and saves a series of diagnostic plots to assess the
convergence of Hmsc models across multiple modelling alternatives. It checks
model convergence using trace plots and Gelman-Rubin diagnostics for key
model parameters.
}
\author{
Ahmed El-Gabbas
}
