% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman.R
\name{PlotGelman}
\alias{PlotGelman}
\title{Plot Gelman-Rubin-Brooks}
\usage{
PlotGelman(
  InputCoda = NULL,
  Beta = TRUE,
  Rho = TRUE,
  Omega = TRUE,
  Alpha = TRUE,
  NCores = NULL,
  NOmega = 1000,
  PlottingAlpha = 0.25,
  EnvFile = ".env",
  SavePlot = TRUE,
  ReturnPlots = FALSE,
  OutPath = NULL
)
}
\arguments{
\item{InputCoda}{Path to RData file containing the coda object}

\item{Beta}{Logical. Run \code{IASDT.R::PlotGelman_Beta}?}

\item{Rho}{Logical. Run \code{IASDT.R::PlotGelman_Rho}?}

\item{Omega}{Logical. Run \code{IASDT.R::PlotGelman_Omega}?}

\item{Alpha}{Logical. Run \code{IASDT.R::PlotGelman_Alpha}}

\item{NCores}{Integer. Number of parallel processes.}

\item{NOmega}{Integer. Number of species to be sampled for the Omega parameter}

\item{PlottingAlpha}{Double. Plotting alpha for line transparency}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{SavePlot}{Logical. Save the outputs as PDF}

\item{ReturnPlots}{Logical. Return ggplot objects}

\item{OutPath}{String. Folder path to save the output figures}
}
\description{
Plot Gelman-Rubin-Brooks. This plot shows the evolution of Gelman and Rubin's shrink factor as the number of iterations increases. For more information, see \code{coda:::gelman.plot}
}
\author{
Ahmed El-Gabbas
}
