% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman.R
\name{PlotGelman}
\alias{PlotGelman}
\alias{PlotGelman_Alpha}
\alias{PlotGelman_Beta}
\alias{PlotGelman_Omega}
\alias{PlotGelman_Rho}
\title{Plot Gelman-Rubin-Brooks}
\usage{
PlotGelman(
  Path_Coda = NULL,
  Alpha = TRUE,
  Beta = TRUE,
  Omega = TRUE,
  Rho = TRUE,
  NOmega = 1000L,
  PlottingAlpha = 0.25,
  EnvFile = ".env",
  ReturnPlots = FALSE
)

PlotGelman_Alpha(CodaObj, PlottingAlpha = 0.25)

PlotGelman_Beta(CodaObj, EnvFile = ".env", PlottingAlpha = 0.25)

PlotGelman_Omega(CodaObj, NOmega = 1000L, PlottingAlpha = 0.25)

PlotGelman_Rho(CodaObj)
}
\arguments{
\item{Path_Coda}{Character or \code{mcmc.list}. Path to a file containing the coda
object, or an \code{mcmc.list} object representing MCMC samples.}

\item{Alpha, Beta, Omega, Rho}{Logical. If \code{TRUE}, plots the Gelman-Rubin
statistic for the respective model parameters (Alpha, Beta, Omega, or Rho).
Default: \code{TRUE} for all parameters.}

\item{NOmega}{Integer. Number of species sampled for the Omega parameter.
Default: 1000L.}

\item{PlottingAlpha}{Numeric. Transparency level (alpha) for plot lines (0 =
fully transparent, 1 = fully opaque). Default: 0.25.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{ReturnPlots}{Character. Path to the folder where the output plots will
be saved.}

\item{CodaObj}{\code{mcmc.list}. An MCMC sample object containing posterior
distributions from an Hmsc model.}
}
\description{
The \verb{PlotGelman_*()} functions generate plots visualizing the evolution of
the Gelman-Rubin-Brooks shrink factor for different model parameters as the
number of iterations increases. These plots help assess whether MCMC chains
have converged to a common distribution. Each plot includes: median (solid
line) and 97.5\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} percentile (dashed line) of the shrink factor and
a dashed horizontal line at 1.1, representing the common convergence
threshold. The primary function for users is \code{PlotGelman()}, which internally
calls:
\itemize{
\item \code{PlotGelman_Alpha()}: Plots shrink factor for the \strong{Alpha} parameter
\item \code{PlotGelman_Beta()}: Plots shrink factor for the \strong{Beta} parameters
\item \code{PlotGelman_Omega()}: Plots shrink factor for the \strong{Omega} parameter
\item \code{PlotGelman_Rho()}: Plots shrink factor for the \strong{Rho} parameter
}
}
\author{
Ahmed El-Gabbas
}
