% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman_Alpha.R
\name{PlotGelman_Alpha}
\alias{PlotGelman_Alpha}
\title{Creates a Gelman-Rubin-Brooks plot for \code{alpha} parameters from an Hmsc
object.}
\usage{
PlotGelman_Alpha(CodaObj, PlottingAlpha = 0.25)
}
\arguments{
\item{CodaObj}{An object of class \code{mcmc.list}, representing the Hmsc samples.}

\item{PlottingAlpha}{A double specifying the alpha (transparency) level for
the plot lines. Default is 0.25.}
}
\value{
A ggplot object(s) representing the Gelman-Rubin-Brooks plot for the
\code{beta} parameters.
}
\description{
This function generates a Gelman-Rubin-Brooks plot for \code{alpha} parameters. It
uses parallel processing to speed up the computation. The plot includes lines
for the median and the 97.5th percentile of the shrink factor, with a dashed
line at 1.1 indicating the threshold for convergence. This function is not
planned to be used in isolation but rather within \link{PlotGelman}.
}
\author{
Ahmed El-Gabbas
}
