% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman_Omega.R
\name{PlotGelman_Omega}
\alias{PlotGelman_Omega}
\title{Gelman-Rubin-Brooks plot for all omega parameters}
\usage{
PlotGelman_Omega(
  CodaObj = NULL,
  NCores = NULL,
  NOmega = 1000,
  PlottingAlpha = 0.25
)
}
\arguments{
\item{CodaObj}{an mcmc object}

\item{NCores}{Integer. Number of parallel processes.}

\item{NOmega}{Integer. Number of species to be sampled for the Omega parameter}

\item{PlottingAlpha}{Double. Plotting alpha for line transparency}
}
\description{
Gelman-Rubin-Brooks plot for all omega parameters
}
\author{
Ahmed El-Gabbas
}
