% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotGelman_Rho.R
\name{PlotGelman_Rho}
\alias{PlotGelman_Rho}
\title{Gelman-Rubin-Brooks plot for the rho parameter}
\usage{
PlotGelman_Rho(CodaObj = NULL)
}
\arguments{
\item{CodaObj}{an mcmc object}
}
\description{
Gelman-Rubin-Brooks plot for the rho parameter
}
\author{
Ahmed El-Gabbas
}
