% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotRespCurv.R
\name{PlotRespCurv}
\alias{PlotRespCurv}
\title{Plot response curves}
\usage{
PlotRespCurv(
  Model,
  ResCurvePath = getwd(),
  ResCurvePrefix = "ResCurv_",
  Verbose = TRUE,
  NCores = NULL
)
}
\arguments{
\item{Model}{Model}

\item{ResCurvePath}{ResCurvePath}

\item{ResCurvePrefix}{ResCurvePrefix}

\item{Verbose}{Verbose}

\item{NCores}{Number of cores to use}
}
\description{
Plot response curves
}
\author{
Ahmed El-Gabbas
}
