% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotAlpha.R, R/Mod_PlotRho.R
\name{PlotRho}
\alias{PlotRho}
\alias{PlotAlpha}
\title{Plot convergence traceplots for the alpha parameter}
\usage{
PlotAlpha(
  Post = NULL,
  Model = NULL,
  Title = NULL,
  NRC = NULL,
  Cols = c("red", "blue", "darkgreen", "darkgrey")
)

PlotRho(
  Post = NULL,
  Model = NULL,
  Title = NULL,
  Cols = c("red", "blue", "darkgreen", "darkgrey")
)
}
\arguments{
\item{Post}{Coda object or path to it}

\item{Model}{Fitted model object or path to it}

\item{Title}{String. Plotting title}

\item{NRC}{Vector for the number of rows and columns per plot page}

\item{Cols}{Colours for lines for each chain}
}
\description{
Plot convergence traceplots for the alpha parameter

Plot convergence traceplots for the rho parameter
}
\author{
Ahmed El-Gabbas
}
