% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_VarParPlot.R
\name{PlotVarPar}
\alias{PlotVarPar}
\title{Plot variance partitioning of Hmsc model using ggplot2.}
\usage{
PlotVarPar(
  Path_Model,
  EnvFile = ".env",
  FromHPC = TRUE,
  UseTF = TRUE,
  TF_Environ = NULL,
  NCores = 1
)
}
\arguments{
\item{Path_Model}{Character path for the model file. Only needed if one of
\code{Model_Eval} and \code{VarPar} arguments are \code{NULL}.}

\item{EnvFile}{String. Path to read the environment variables. Default value:
\code{.env}}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}

\item{NCores}{Integer. Number of parallel computations for computing
predicted values. This is used as the \code{nParallel} argument of the
\link[Hmsc:computePredictedValues]{Hmsc::computePredictedValues} function.}

\item{PlotTitlePrefix}{String (optional). Prefix to add to the title of the
plot. Default: \code{NULL}, which means only 'Variance partitioning' will be
used in the title.}

\item{Model_Eval}{Result of the \link[Hmsc:evaluateModelFit]{Hmsc::evaluateModelFit} function. If
\code{Model_Eval = NULL} (default), \link[Hmsc:evaluateModelFit]{Hmsc::evaluateModelFit} will be executed on
the model object to compute measures of model fit.}

\item{VarPar}{Variance partitioning object resulted from
\link[Hmsc:computeVariancePartitioning]{Hmsc::computeVariancePartitioning}. If \code{VarPar = NULL} (default),
\link[Hmsc:computeVariancePartitioning]{Hmsc::computeVariancePartitioning} will be executed on the model object.}
}
\description{
This function generates and optionally saves plots visualizing the variance
partitioning of a Hmsc model. It can calculate variance partitioning and
model evaluation if not provided and supports parallel computation for model
predictions.
}
\details{
The function reads the following environment variables:
\itemize{
\item \strong{\code{DP_R_TaxaInfo_RData}} (if \code{FromHPC} = \code{TRUE}) or
\strong{\code{DP_R_TaxaInfo_RData_Local}} (if \code{FromHPC} = \code{FALSE}) for the
location of the \code{TaxaList.RData} file containing species information.
}
}
\author{
Ahmed El-Gabbas
}
