% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PlotVarPar.R
\name{PlotVarPar}
\alias{PlotVarPar}
\title{Plot variance partitioning of the model}
\usage{
PlotVarPar(
  Model,
  PlotPath = NULL,
  ModelName = NULL,
  ModelEval = NULL,
  VarPar = NULL,
  EnvFile = ".env"
)
}
\arguments{
\item{Model}{an object of class \code{Hmsc} or a path for the model. Only needed if one of \code{ModelEval} and \code{VarPar} arguments are \code{NULL}.}

\item{PlotPath}{String. Path to save the output file.}

\item{ModelName}{String. Prefix to add to the title of the plot. Default: \code{NULL}, which means only use 'Variance paritioning' in the title.}

\item{ModelEval}{Result of the \code{Hmsc::evaluateModelFit} function. If \code{ModelEval = NULL} (default), \code{Hmsc::evaluateModelFit} will be executed on the model object to compute measures of model fit.}

\item{VarPar}{Variance partitioning. An object resulted from \code{Hmsc::computeVariancePartitioning}.}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}
}
\description{
Plot variance partitioning of the model
}
\author{
Ahmed El-Gabbas
}
