% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_predict_maps.R, R/mod_predict_maps_cv.R
\name{predict_maps}
\alias{predict_maps}
\alias{predict_maps_cv}
\title{Predict habitat suitability of \code{Hmsc} models}
\usage{
predict_maps(
  path_model = NULL,
  hab_abb = NULL,
  env_file = ".env",
  n_cores = 8L,
  strategy = "multisession",
  clamp_pred = TRUE,
  fix_efforts = "q90",
  fix_rivers = "q90",
  pred_new_sites = TRUE,
  use_tf = TRUE,
  tf_environ = NULL,
  tf_use_single = FALSE,
  n_cores_lf = n_cores,
  lf_check = FALSE,
  lf_temp_cleanup = TRUE,
  lf_only = FALSE,
  lf_commands_only = FALSE,
  temp_dir = "temp_pred",
  temp_cleanup = TRUE,
  tar_predictions = TRUE,
  spatial_model = TRUE,
  n_cores_pred = n_cores,
  climate_models = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0",
    "UKESM1-0-LL"),
  climate_scenario = c("ssp126", "ssp370", "ssp585")
)

predict_maps_cv(
  model_dir = NULL,
  cv_name = NULL,
  cv_fold = NULL,
  n_cores = 8L,
  strategy = "multisession",
  env_file = ".env",
  use_tf = TRUE,
  tf_environ = NULL,
  tf_use_single = FALSE,
  n_cores_lf = n_cores,
  lf_check = FALSE,
  lf_temp_cleanup = TRUE,
  lf_only = FALSE,
  lf_commands_only = FALSE,
  temp_cleanup = TRUE
)
}
\arguments{
\item{path_model}{Character. Path to fitted \code{Hmsc} model object.}

\item{hab_abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type. Valid
values: \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a}, \verb{4b}, \code{10}, \verb{12a}, \verb{12b}. See \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.} for details.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{clamp_pred}{Logical indicating whether to clamp the sampling efforts at
a single value. If \code{TRUE} (default), the \code{fix_efforts} argument must be
provided.}

\item{fix_efforts}{Numeric or character. When \code{clamp_pred = TRUE}, fixes the
sampling efforts predictor at this value during predictions. If numeric,
uses the value directly (on log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, must be
one of \code{identity} (i.e., do not fix), \code{median}, \code{mean}, \code{max}, or \code{q90}
(90\% quantile). Using \code{max} may reflect extreme sampling efforts from
highly sampled locations, while \code{q90} captures high sampling areas without
extremes. Required if \code{clamp_pred = TRUE}.}

\item{fix_rivers}{Numeric, character, or \code{NULL}. Similar to \code{fix_efforts},
but for the river length predictor. If \code{NULL}, the river length is not
fixed. Default: \code{q90}.}

\item{pred_new_sites}{Logical. Whether to predict suitability at new sites.
Default: \code{TRUE}.}

\item{use_tf}{Logical. Whether to use \code{TensorFlow} for calculations. Defaults
to \code{TRUE}.}

\item{tf_environ}{Character. Path to the Python environment. This argument is
required if \code{use_tf} is \code{TRUE} under Windows. Defaults to \code{NULL}.}

\item{tf_use_single}{Logical. Whether to use single precision for the
\code{TensorFlow} calculations. Defaults to \code{FALSE}.}

\item{n_cores_lf}{Integer. Number of cores to use for parallel processing of
latent factor prediction. Defaults to 8L.}

\item{lf_check}{Logical. If \code{TRUE}, the function checks if the output files
are already created and valid. If \code{FALSE}, the function will only check if
the files exist without checking their integrity. Default is \code{FALSE}.}

\item{lf_temp_cleanup}{Logical. Whether to delete temporary files in the
\code{temp_dir} directory after finishing the LF predictions.}

\item{lf_only}{Logical. Whether to predict only the latent factor. This is
useful for distributing processing load between GPU and CPU. When \code{lf_only = TRUE}, latent factor prediction needs to be computed separately on GPU.
When computations are finished on GPU, the function can later be rerun with
\code{lf_only = FALSE} (default) to predict habitat suitability using the
already-computed latent factor predictions.}

\item{lf_commands_only}{Logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{temp_dir}{Character. Path for temporary storage of intermediate files.}

\item{temp_cleanup}{Logical. Whether to clean up temporary files. Defaults to
\code{TRUE}.}

\item{tar_predictions}{Logical. Whether to compress tiff files for predicted
habitat suitability into a single \verb{*.tar} file (without compression).
Default: \code{TRUE}.}

\item{spatial_model}{Logical. Whether the fitted model is a spatial model.
Defaults to \code{TRUE}.}

\item{n_cores_pred}{Integer. Number of cores to use for predicting species'
habitat suitability.}

\item{climate_models}{Character vector. Climate models for future
predictions. Available options are \code{c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL")} (default).}

\item{climate_scenario}{Character vector. Climate scenarios for future
predictions. Available options are: \code{c("ssp126", "ssp370", "ssp585")}
(default).}

\item{model_dir}{Character. Path to the directory containing cross-validated
models.}

\item{cv_name}{Character. Cross-validation strategy. Valid values are
\code{cv_dist}, \code{cv_large}, or \code{cv_sac}.}

\item{cv_fold}{Integer. The cross-validation fold number.}
}
\description{
This package provides two functions for predicting habitat suitability of
\code{Hmsc} models in the \code{IASDT} framework. \link{predict_maps} generates current and
future habitat suitability maps (mean, sd, cov) from a full Hmsc model fit.
\link{predict_maps_cv} predicts and evaluates cross-validated Hmsc models for
current climate conditions. For more details, see the respective function
documentation and the details section below.
}
\details{
\itemize{
\item \strong{\code{predict_maps}}: Generates habitat suitability maps for \code{Hmsc} models
fitted on the full dataset, for both current and future climate options. It
produces maps for mean, standard deviation (sd), and coefficient of variation
(cov) of suitability for each species and overall species richness. It
evaluate model's explanatory power using various metrics. For future
predictions, it also generates anomaly maps (future - current). The function
supports ensemble predictions across multiple climate models and prepares
data for upload to the \href{http://opendap.biodt.eu/ias-pdt/}{OPeNDAP} server for
use in the \code{IASDT} \href{https://app.biodt.eu/}{Shiny App}.
\item \strong{\code{predict_maps_cv}}: Computes predictions for cross-validated \code{Hmsc}
models using only the testing folds. It evaluates model performance
(predictive power) with various metrics and plots evaluation results for
predictive and explanatory power. Unlike \code{predict_maps}, this function does
not perform clamping and does not generate future climate predictions.
}
}
\seealso{
predict_hmsc
}
\author{
Ahmed El-Gabbas
}
