% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_Maps.R
\name{Predict_Maps}
\alias{Predict_Maps}
\title{Predicts habitat suitability of \code{Hmsc} model across different climate options}
\usage{
Predict_Maps(
  Path_Model = NULL,
  Hab_Abb = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NCores = 8L,
  Pred_Clamp = TRUE,
  Fix_Efforts = "q90",
  Fix_Rivers = "q90",
  Pred_NewSites = TRUE,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  LF_NCores = NCores,
  LF_Check = FALSE,
  LF_Temp_Cleanup = TRUE,
  LF_Only = FALSE,
  LF_Commands_Only = FALSE,
  Temp_Dir = "TEMP_Pred",
  Temp_Cleanup = TRUE,
  CC_Models = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0",
    "UKESM1-0-LL"),
  CC_Scenario = c("ssp126", "ssp370", "ssp585")
)
}
\arguments{
\item{Path_Model}{Character. Path to fitted \code{Hmsc} model object.}

\item{Hab_Abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type for
which data will be prepared. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a},
\verb{4b}, \code{10}, \verb{12a}, \verb{12b}. For more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{NCores}{Integer specifying the number of parallel cores for
parallelization. Default: 8 cores.}

\item{Pred_Clamp}{Logical indicating whether to clamp the sampling efforts at
a single value. Defaults to \code{TRUE}. If \code{TRUE}, the \code{Fix_Efforts} argument
must be provided.}

\item{Fix_Efforts}{Numeric or character. Defines the value to fix sampling
efforts less than the provided value. If numeric, the value is directly
used (log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be \code{median}, \code{mean},
\code{max}, or \code{q90} (90\% Quantile). Using \code{max} can reflect extreme values
caused by rare, highly sampled locations (e.g., urban centers or popular
natural reserves). While using 90\% Quantile avoid such extreme grid cells
while still capturing areas with high sampling effort. This argument is
mandatory when \code{Pred_Clamp} is set to \code{TRUE}.}

\item{Fix_Rivers}{Numeric or character. Similar to \code{Fix_Efforts}, but for
fixing the length of rivers. If numeric, the value is directly used
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be \code{median}, \code{mean}, \code{max},
\code{q90} (90\% Quantile). It can be also \code{NULL} for not fixing the river length
predictor. Defaults to \code{q90}.}

\item{Pred_NewSites}{Logical indicating whether to predict habitat
suitability at new sites. Default: \code{TRUE}. Note: This parameter is
temporary and will be removed in future updates.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{TF_use_single}{Logical indicating whether to use single precision for
the TF calculations. Defaults to \code{FALSE}.}

\item{LF_NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}

\item{LF_Check}{Logical. If TRUE, the function checks if the output files are
already created and valid. If FALSE, the function will only check if the
files exist without checking their integrity. Default is \code{FALSE}.}

\item{LF_Temp_Cleanup}{Logical indicating whether to delete temporary files
in the \code{Temp_Dir} after finishing the LF predictions.}

\item{LF_Only}{Logical. Indicates whether to predict only the latent factor.
Useful for distributing processing load between GPU and CPU. When \code{LF_Only = TRUE}, the latent factor prediction can be computed on the GPU. The
function can then be rerun with \code{LF_Only = FALSE} to predict habitat
suitability using the predicted latent factor on the CPU. Default: \code{FALSE}.}

\item{LF_Commands_Only}{logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{Temp_Dir}{Character string specifying the path for temporary storage of
intermediate files.}

\item{Temp_Cleanup}{logical, indicating whether to clean up temporary files.
Defaults to \code{TRUE}.}

\item{CC_Models}{Character vector. Specifies the climate models for future
predictions. Default: \code{c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL")}. Note: This parameter is temporary and may be
removed in future updates.}

\item{CC_Scenario}{Character vector. Specifies the climate scenarios for
future predictions. Default: \code{c("ssp126", "ssp370", "ssp585")}. Note: This
parameter is temporary and may be removed in future updates.}
}
\value{
A tibble containing the prediction summary and file paths for output
\verb{*.tif} files.
}
\description{
This function generates prediction maps of \code{Hmsc} models for current and
future climate scenarios. It also predicts an ensemble predictions for
different climate models. For each species and overall species richness, the
function exports three maps: mean, standard deviation (sd), and coefficient
of variation (cov).
}
\author{
Ahmed El-Gabbas
}
