% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_predict_maps.R
\name{predict_maps}
\alias{predict_maps}
\title{Predict habitat suitability of \code{Hmsc} model across different climate options}
\usage{
predict_maps(
  path_model = NULL,
  hab_abb = NULL,
  env_file = ".env",
  n_cores = 8L,
  clamp_pred = TRUE,
  fix_efforts = "q90",
  fix_rivers = "q90",
  pred_new_sites = TRUE,
  use_TF = TRUE,
  TF_environ = NULL,
  TF_use_single = FALSE,
  LF_n_cores = n_cores,
  LF_check = FALSE,
  LF_temp_cleanup = TRUE,
  LF_only = FALSE,
  LF_commands_only = FALSE,
  temp_dir = "TEMP_Pred",
  temp_cleanup = TRUE,
  tar_predictions = TRUE,
  CC_models = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0",
    "UKESM1-0-LL"),
  CC_scenario = c("ssp126", "ssp370", "ssp585")
)
}
\arguments{
\item{path_model}{Character. Path to fitted \code{Hmsc} model object.}

\item{hab_abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type for
which data will be prepared. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a},
\verb{4b}, \code{10}, \verb{12a}, \verb{12b}. For more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{clamp_pred}{Logical indicating whether to clamp the sampling efforts at
a single value. If \code{TRUE} (default), the \code{fix_efforts} argument must be
provided.}

\item{fix_efforts}{Numeric or character. If \code{clamp_pred = TRUE}, the sampling
efforts predictor with values U+02264 \code{fix_efforts} is fixed at
\code{fix_efforts} during predictions. If numeric, the value is directly used
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be one of \code{median}, \code{mean},
\code{max}, or \code{q90} (90\% Quantile). Using \code{max} can reflect extreme values
caused by rare, highly sampled locations (e.g., urban centers or popular
natural reserves). While using 90\% quantile avoid such extreme grid cells
while still capturing areas with high sampling effort. This argument is
mandatory when \code{clamp_pred} is set to \code{TRUE}.}

\item{fix_rivers}{Numeric or character. Similar to \code{fix_efforts}, but for
fixing the length of rivers. If numeric, the value is directly used
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be one of \code{median}, \code{mean},
\code{max}, \code{q90} (90\% quantile). It can be also \code{NULL} for not fixing the river
length predictor. Defaults to \code{q90}.}

\item{pred_new_sites}{Logical. Whether to predict habitat suitability at new
sites. Default: \code{TRUE}. Note: This parameter is temporary and will be
removed in future updates.}

\item{use_TF}{Logical. Whether to use TensorFlow for calculations. Defaults
to \code{TRUE}.}

\item{TF_environ}{Character. Path to the Python environment. This argument is
required if \code{use_TF} is \code{TRUE} under Windows. Defaults to \code{NULL}.}

\item{TF_use_single}{Logical. Whether to use single precision for the
TensorFlow calculations. Defaults to \code{FALSE}.}

\item{LF_n_cores}{Integer. Number of cores to use for parallel processing of
latent factor prediction. Defaults to 8L.}

\item{LF_check}{Logical. If \code{TRUE}, the function checks if the output files
are already created and valid. If \code{FALSE}, the function will only check if
the files exist without checking their integrity. Default is \code{FALSE}.}

\item{LF_temp_cleanup}{Logical. Whether to delete temporary files in the
\code{temp_dir} directory after finishing the LF predictions.}

\item{LF_only}{Logical. Whether to predict only the latent factor. This is
useful for distributing processing load between GPU and CPU. When \code{LF_only = TRUE}, latent factor prediction needs to be computed separately on GPU.
When computations are finished on GPU, the function can later be rerun with
\code{LF_only = FALSE} (default) to predict habitat suitability using the
already-computed latent factor predictions.}

\item{LF_commands_only}{Logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{temp_dir}{Character. Path for temporary storage of intermediate files.}

\item{temp_cleanup}{Logical. Whether to clean up temporary files. Defaults to
\code{TRUE}.}

\item{tar_predictions}{Logical. Whether to compress the add files into a
single \verb{*.tar} file (without compression). Default: \code{TRUE}.}

\item{CC_models}{Character vector. Climate models for future predictions.
Available options are \code{c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL")} (default).}

\item{CC_scenario}{Character vector. Climate scenarios for future
predictions. Available options are: \code{c("ssp126", "ssp370", "ssp585")}
(default).}
}
\value{
A tibble containing the prediction summary and file paths for output
\verb{*.tif} files.
}
\description{
This function generates prediction maps of \code{Hmsc} models for current and
future climate options. It also predicts an ensemble predictions for
different climate models. For each species and overall species richness, the
function exports three maps: mean, standard deviation (sd), and coefficient
of variation (cov). For future predictions, the function generates maps for
prediction anomaly (future - current). The function prepares data to be
uploaded to the \href{http://opendap.biodt.eu/ias-pdt/}{OPeNDAP} data server for
the use of the IAS-pDT \href{https://app.biodt.eu/}{Shiny App}.
}
\author{
Ahmed El-Gabbas
}
