% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_Maps.R
\name{Predict_Maps}
\alias{Predict_Maps}
\title{Predict habitat suitability of \code{Hmsc} model across different climate options}
\usage{
Predict_Maps(
  Path_Model = NULL,
  Hab_Abb = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NCores = 8L,
  Pred_Clamp = TRUE,
  Fix_Efforts = "q90",
  Fix_Rivers = "q90",
  Pred_NewSites = TRUE,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  LF_NCores = NCores,
  LF_Check = FALSE,
  LF_Temp_Cleanup = TRUE,
  LF_Only = FALSE,
  LF_Commands_Only = FALSE,
  Temp_Dir = "TEMP_Pred",
  Temp_Cleanup = TRUE,
  CC_Models = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0",
    "UKESM1-0-LL"),
  CC_Scenario = c("ssp126", "ssp370", "ssp585")
)
}
\arguments{
\item{Path_Model}{Character. Path to fitted \code{Hmsc} model object.}

\item{Hab_Abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type for
which data will be prepared. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a},
\verb{4b}, \code{10}, \verb{12a}, \verb{12b}. For more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical. Whether the processing is being done on an
High-Performance Computing (HPC) environment, to adjust file paths
accordingly. Default: \code{TRUE}.}

\item{NCores}{Integer. Number of CPU cores to use for parallel processing.
Default: 8.}

\item{Pred_Clamp}{Logical indicating whether to clamp the sampling efforts at
a single value. If \code{TRUE} (default), the \code{Fix_Efforts} argument must be
provided.}

\item{Fix_Efforts}{Numeric or character. If \code{Pred_Clamp = TRUE}, the sampling
efforts predictor with values U+02264 \code{Fix_Efforts} is fixed at
\code{Fix_Efforts} during predictions. If numeric, the value is directly used
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be one of \code{median}, \code{mean},
\code{max}, or \code{q90} (90\% Quantile). Using \code{max} can reflect extreme values
caused by rare, highly sampled locations (e.g., urban centers or popular
natural reserves). While using 90\% quantile avoid such extreme grid cells
while still capturing areas with high sampling effort. This argument is
mandatory when \code{Pred_Clamp} is set to \code{TRUE}.}

\item{Fix_Rivers}{Numeric or character. Similar to \code{Fix_Efforts}, but for
fixing the length of rivers. If numeric, the value is directly used
(log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} scale). If character, it can be one of \code{median}, \code{mean},
\code{max}, \code{q90} (90\% quantile). It can be also \code{NULL} for not fixing the river
length predictor. Defaults to \code{q90}.}

\item{Pred_NewSites}{Logical. Whether to predict habitat suitability at new
sites. Default: \code{TRUE}. Note: This parameter is temporary and will be
removed in future updates.}

\item{UseTF}{Logical. Whether to use TensorFlow for calculations. Defaults to
\code{TRUE}.}

\item{TF_Environ}{Character. Path to the Python environment. This argument is
required if \code{UseTF} is \code{TRUE}.}

\item{TF_use_single}{Logical. Whether to use single precision for the
TensorFlow calculations. Defaults to \code{FALSE}.}

\item{LF_NCores}{Integer. Number of cores to use for parallel processing of
latent factor prediction. Defaults to 8L.}

\item{LF_Check}{Logical. If \code{TRUE}, the function checks if the output files
are already created and valid. If \code{FALSE}, the function will only check if
the files exist without checking their integrity. Default is \code{FALSE}.}

\item{LF_Temp_Cleanup}{Logical. Whether to delete temporary files in the
\code{Temp_Dir} directory after finishing the LF predictions.}

\item{LF_Only}{Logical. Whether to predict only the latent factor. This is
useful for distributing processing load between GPU and CPU. When \code{LF_Only = TRUE}, latent factor prediction needs to be computed separately on GPU.
When computations are finished on GPU, the function can later be rerun with
\code{LF_Only = FALSE} (default) to predict habitat suitability using the
already-computed latent factor predictions.}

\item{LF_Commands_Only}{Logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{Temp_Dir}{Character. Path for temporary storage of intermediate files.}

\item{Temp_Cleanup}{Logical. Whether to clean up temporary files.
Defaults to \code{TRUE}.}

\item{CC_Models}{Character vector. Climate models for future predictions.
Available options are \code{c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL")} (default).}

\item{CC_Scenario}{Character vector. Climate scenarios for future
predictions. Available options are: \code{c("ssp126", "ssp370", "ssp585")}
(default).}
}
\value{
A tibble containing the prediction summary and file paths for output
\verb{*.tif} files.
}
\description{
This function generates prediction maps of \code{Hmsc} models for current and
future climate options. It also predicts an ensemble predictions for
different climate models. For each species and overall species richness, the
function exports three maps: mean, standard deviation (sd), and coefficient
of variation (cov). For future predictions, the function generates maps for
prediction anomaly (future - current). The function prepares data to be
uploaded to the \href{http://opendap.biodt.eu/ias-pdt/}{OPeNDAP} data server for
the use of the IAS-pDT \href{https://app.biodt.eu/}{Shiny App}.
}
\author{
Ahmed El-Gabbas
}
