% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_Maps.R
\name{Predict_Maps}
\alias{Predict_Maps}
\title{Predicts habitat suitability of \code{Hmsc} model at different climate scenarios}
\usage{
Predict_Maps(
  Path_Model = NULL,
  Hab_Abb = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NCores = 8,
  Pred_Clamp = TRUE,
  Fix_Efforts = "mean",
  Pred_NewSites = TRUE,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  Temp_Dir = "TEMP2Pred",
  Temp_Cleanup = FALSE,
  CC_Models = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0",
    "UKESM1-0-LL"),
  CC_Scenario = c("ssp126", "ssp370", "ssp585")
)
}
\arguments{
\item{Path_Model}{Character. Path to fitted \code{Hmsc} model object.}

\item{Hab_Abb}{Character. Habitat abbreviation indicating the specific
\href{https://www.preslia.cz/article/pdf?id=11548}{SynHab} habitat type to
prepare data for. Valid values are \code{0}, \code{1}, \code{2}, \code{3}, \verb{4a}, \verb{4b}, \code{10},
\verb{12a}, \verb{12b}. For more details, see \href{https://doi.org/10.23855/preslia.2022.447}{Pysek et al.}.}

\item{EnvFile}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{NCores}{Integer specifying the number of parallel cores for
parallelization. Default: 8 cores.}

\item{Pred_Clamp}{Logical indicating whether to clamp the sampling efforts at
a single value. Defaults to \code{TRUE}. If \code{TRUE}, the \code{Fix_Efforts} argument
must be provided.}

\item{Fix_Efforts}{Numeric or character. Value to fix the sampling efforts
at. If numeric, the value will be used as is. If character, it can be one
of the following: \code{median}, \code{mean}, or \code{max}. Default: \code{mean}. This
argument is required when \code{Pred_Clamp} is \code{TRUE}.}

\item{Pred_NewSites}{Logical indicating whether to predict habitat
suitability at new sites. Default: \code{TRUE}. This argument is only temporary
to make predictions for the workflow and will be removed in the future.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{TF_use_single}{Logical indicating whether to use single precision for
the TF calculations. Defaults to \code{FALSE}.}

\item{Temp_Dir}{Character string specifying the path for temporary storage of
intermediate files.}

\item{Temp_Cleanup}{Logical indicating whether to delete temporary files in
the \code{Temp_Dir} after finishing the calculations. Defaults to \code{FALSE}; i.e.,
the temporary files will be kept, enabling continue working on unfinished
calculations, if the function failed, e.g. due to OOM.}

\item{CC_Models}{Character vector specifying the climate models to use for
future predictions. Default: \code{c("GFDL-ESM4", "IPSL-CM6A-LR" , "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL")}. This argument is only
temporary to make predictions for the workflow and will be removed in the
future.}

\item{CC_Scenario}{Character vector specifying the climate scenarios to use
for future predictions. Default: \code{c("ssp126", "ssp370", "ssp585")}. This
argument is only temporary to make predictions for the workflow and will be
removed in the future.}
}
\value{
Returns a tibble with the prediction summary and file paths for
output tif files.
}
\description{
This function generates prediction maps of \code{Hmsc} models under current and
future climates. It also predicts an ensemble predictions for different
climate models used. For each species and for species richness, the function
exports three maps representing the mean, standard deviation (sd), and
coefficient of variation (cov).
}
\author{
Ahmed El-Gabbas
}
