% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_predictHmsc.R
\name{predictHmsc}
\alias{predictHmsc}
\title{predictHmsc}
\usage{
predictHmsc(
  object,
  Loff = NULL,
  XData = NULL,
  X = NULL,
  XRRRData = NULL,
  XRRR = NULL,
  Gradient = NULL,
  Yc = NULL,
  mcmcStep = 1,
  expected = TRUE,
  NCores = 8,
  Model_Name = "Train",
  Temp_Dir = "TEMP2Pred",
  RC = NULL,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  LF_OutFile = NULL,
  LF_Return = TRUE,
  LF_InputFile = NULL,
  Pred_Dir = "Model_Predictions",
  Pred_PA = NULL,
  Pred_XY = NULL,
  Evaluate = FALSE,
  Eval_Name = NULL,
  Eval_Dir = "Evaluation",
  Verbose = TRUE
)
}
\arguments{
\item{object}{a fitted \code{Hmsc} model object or a character string specifying a
file name where the model object is saved.}

\item{XData}{a dataframe specifying the unpreprocessed covariates for the
predictions to be made. Works only if the \code{XFormula} argument was specified
in the \link[Hmsc:Hmsc]{Hmsc::Hmsc} model constructor call. Requirements are similar to
those in the \code{Hmsc} model constructor.}

\item{X}{a matrix specifying the covariates for the predictions to be made.
Only one of \code{XData} and \code{X} arguments may be provided.}

\item{XRRRData}{a dataframe of covariates for reduced-rank regression.}

\item{XRRR}{a matrix of covariates for reduced-rank regression.}

\item{Gradient}{an object returned by \link[Hmsc:constructGradient]{Hmsc::constructGradient}. Providing
\code{Gradient} is an alternative for providing \code{XData}, \code{studyDesign} and
\code{ranLevels}. Cannot be used together with \code{Yc}.}

\item{Yc}{a matrix of the outcomes that are assumed to be known for
conditional predictions. Cannot be used together with \code{Gradient}.}

\item{mcmcStep}{the number of extra mcmc steps used for updating the random
effects}

\item{expected}{boolean flag indicating whether to return the location
parameter of the observation models or sample the values from those.}

\item{NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}

\item{Model_Name}{Character string used as a prefix for temporary file names.
Defaults to NULL, in which case no prefix is used.}

\item{Temp_Dir}{Character string specifying the path for temporary storage of
intermediate files.}

\item{RC}{a character string specifying the type of predictions to be made.
If \code{NULL} (default), predictions are made for the latent factors. If \code{c},
predictions are made for response curves at mean coordinates. If \code{i},
predictions are made for response curves at infinite coordinates.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{TF_use_single}{Logical indicating whether to use single precision for the
TF calculations. Defaults to \code{FALSE}.}

\item{LF_OutFile}{Character string specifying the path to save the outputs.
If \code{NULL} (default), the predicted latent factors are not saved to a
file. This should end with either \code{qs} or \code{RData}.}

\item{LF_Return}{Logical. Indicates if the output should be
returned. Defaults to \code{TRUE}. If \code{LF_OutFile} is \code{NULL}, this parameter
cannot be set to \code{FALSE} because the function needs to return the result if
it is not saved to a file.}

\item{LF_InputFile}{a character string specifying the file name where the
latent factor predictions are saved. If \code{NULL}, the predictions are
saved to a temporary file. This argument is used only when \code{RC} is \code{NULL}.}

\item{Pred_Dir}{a character string specifying the directory where the
predictions will be saved.}

\item{Pred_PA}{a matrix of presence-absence data for evaluation. If \code{NULL}
(default), the presence-absence data from the model object is used. This
argument is used only when \code{Evaluate} is \code{TRUE}.}

\item{Pred_XY}{a matrix of coordinates to be added to predicted values. If
\code{NULL} (default), the coordinates from the model object is used.}

\item{Evaluate}{a logical flag indicating whether to evaluate the model
predictions. Defaults to \code{FALSE}.}

\item{Eval_Name}{a character string specifying the name of the evaluation
results. If \code{NULL}, the default name is used (\code{Eval_Model_Name.RData}).}

\item{Eval_Dir}{a character string specifying the directory where the
evaluation results will be saved. Defaults to \code{Evaluation}.}

\item{Verbose}{Logical. If TRUE, detailed output is printed. Default is
\code{FALSE}.}
}
\description{
Calculates predicted values from a fitted \code{Hmsc} model. This function
modifies the \code{Hmsc:::predict.Hmsc} function.
}
