% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepKnots.R
\name{PrepKnots}
\alias{PrepKnots}
\title{Prepare knot locations for GPP models}
\usage{
PrepKnots(Coords = NULL, MinDist = NULL, JitterDist = 100)
}
\arguments{
\item{Coords}{A matrix or data frame for the coordinates of the sampling units.}

\item{MinDist}{Numeric. Minimum distance in meters for the \code{knotDist} and \code{minKnotDist} arguments of the \code{Hmsc::constructKnots} function.}

\item{JitterDist}{Numeric. Distance in meter to be used in the \code{sf::st_jitter} function.}
}
\description{
Prepare knot locations for GPP models
}
\author{
Ahmed El-Gabbas
}
