% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepareKnots.R
\name{PrepKnots}
\alias{PrepKnots}
\title{PrepKnots}
\usage{
PrepKnots(Coords = DT_xy, MinDist, JitterDist = 100)
}
\arguments{
\item{Coords}{Coords}

\item{MinDist}{MinDist}

\item{JitterDist}{JitterDist}
}
\description{
PrepKnots
}
\author{
Ahmed El-Gabbas
}
