% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepKnots.R
\name{PrepKnots}
\alias{PrepKnots}
\title{Prepare knot locations for Hmsc GPP models}
\usage{
PrepKnots(
  Coords = NULL,
  MinDist = NULL,
  JitterDist = 100,
  MinLF = NULL,
  MaxLF = NULL,
  SetAlphapw = TRUE,
  AlphapwPar = list(n_samples = 300, prob_norm = 0.8, tr_mean = 200, tr_sd = 50, tr_min =
    5, tr_max = 400, seed = NULL, Alphapw = NULL)
)
}
\arguments{
\item{Coords}{A matrix or data frame containing the coordinates (x, y) of the
sampling units.}

\item{MinDist}{A numeric value specifying the minimum distance (in the same
units as Coords; here in meters) between knots. This distance is used for
both \code{knotDist} and \code{minKnotDist} parameters of the \link[Hmsc:constructKnots]{Hmsc::constructKnots}
function.}

\item{JitterDist}{A numeric value specifying the distance (in the same units
as Coords; here in meters) for jittering identical coordinates. Defaults to
100.}

\item{MinLF, MaxLF}{integer. Minimum and maximum number of latent factors to
be used. Both default to \code{NULL} which means that the number of latent
factors will be estimated from the data. If either is provided, the
respective values will be used as arguments to \link[Hmsc:setPriors]{Hmsc::setPriors}.}

\item{SetAlphapw}{Logical; whether to set the prior for alphapw using
\link{PriorAlpha}. Default: \code{TRUE}.}

\item{AlphapwPar}{List; parameters for generating prior distribution. If the
list item \code{Alphapw} is provided, it will be used as the prior for alphapw.
Otherwise, the prior will be generated using \code{PriorAlpha} function. Default
values are as follows:  \code{list(n_samples = 300, prob_norm = 0.8, tr_mean = 200, tr_sd = 50, tr_min = 5, tr_max = 400, seed = NULL, Alphapw = NULL)}.
See \link{PriorAlpha} for more details.}
}
\value{
An object suitable for specifying the random level in HMSC GPP
models. This object contains the prepared knot locations.
}
\description{
This function prepares the locations of knots for use in GPP models within
the HMSC framework. It ensures that knots are spaced at a minimum specified
distance and applies jitter to any identical coordinates to avoid overlap.
}
\author{
Ahmed El-Gabbas
}
