% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PrepModelData.R
\name{PrepModData}
\alias{PrepModData}
\title{Prepare habitat-specific data for the models}
\usage{
PrepModData(
  Hab_Abb = NULL,
  MinPresGrids = 50,
  Path_EnvFile = ".env",
  ReturnData = FALSE,
  OutputPath = "Data/Models"
)
}
\arguments{
\item{Hab_Abb}{Habitat type}

\item{MinPresGrids}{Minimum number of presence grid cells per species. Only species with >= this number will be considered}

\item{Path_EnvFile}{Path to read the environment variables}

\item{ReturnData}{return data object}

\item{OutputPath}{Dir path for the output file}
}
\description{
Prepare habitat-specific data for the models
}
\author{
Ahmed El-Gabbas
}
