% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_SLURM.R
\name{PrepSLURM}
\alias{PrepSLURM}
\title{Prepare SLURM file for model fitting on HPC}
\usage{
PrepSLURM(
  Path_Model = NULL,
  JobName = NULL,
  CatJobInfo = TRUE,
  ntasks = 1,
  CpusPerTask = 1,
  GpusPerNode = 1,
  MemPerCpu = NULL,
  Time = NULL,
  Partition = "small-g",
  Path_EnvFile = ".env",
  CommandPrefix = "Commands_All",
  SLURM_Prefix = "Commands_All"
)
}
\arguments{
\item{Path_Model}{String. Path to the model files (without trailing slash)}

\item{JobName}{String. The name of the submitted jobs}

\item{CatJobInfo}{Logical. Add bash lines to report some job information. Default: \code{TRUE}}

\item{ntasks}{Integer. The value for the \verb{#SBATCH --ntasks=} SLURM argument. Default: 1}

\item{CpusPerTask}{Integer. The value for the \verb{#SBATCH --cpus-per-task=} SLURM argument. Default: 1}

\item{GpusPerNode}{Integer. The value for the \verb{#SBATCH --gpus-per-node=} SLURM argument. Default: 1}

\item{MemPerCpu}{String. The value for the \verb{#SBATCH --mem-per-cpu=} SLURM argument. Example: "32G" to request 32 gigabyte}

\item{Time}{String. The value for the requested time for each job in the bash arrays. Example: "01:00:00" to request an hour.}

\item{Partition}{String. The name of the partition. Default: \code{small-g}}

\item{Path_EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{CommandPrefix}{String. Prefix for the bash commands to be executed}

\item{SLURM_Prefix}{String. Prefix for the exported SLURM file.}
}
\description{
Prepare SLURM file for model fitting on HPC
}
\author{
Ahmed El-Gabbas
}
