% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Railway_Intensity.R
\name{Railway_Intensity}
\alias{Railway_Intensity}
\title{Calculate Railway Intensity Based on \code{OpenStreetMap} Data}
\usage{
Railway_Intensity(
  FromHPC = TRUE,
  EnvFile = ".env",
  Download = TRUE,
  Extract = TRUE,
  GetDownLinks = TRUE,
  CheckZip = TRUE,
  NCores = 6
)
}
\arguments{
\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default is ".env".}

\item{Download}{Logical. If \code{TRUE} (default), downloads railway data from the
specified source.}

\item{Extract}{Logical. If \code{TRUE} (default), extracts railway data.}

\item{GetDownLinks}{Logical. If \code{TRUE}, fetches download links for railway
data. Default is \code{TRUE}.}

\item{CheckZip}{Logical. If \code{TRUE} (default), validates the integrity of
downloaded ZIP files.}

\item{NCores}{Numeric. Number of CPU cores to use for parallel processing.
Default is 6.}
}
\value{
\code{NULL}. Outputs processed files to the directories specified in the
environment file.
}
\description{
This function downloads, processes, and analyzes railway data extracted from
\href{https://www.openrailwaymap.org}{OpenRailwayMap} available from
\href{https://download.geofabrik.de/}{OpenStreetMap Data Extracts}. It supports
parallel processing for faster execution and can calculate the total length
of railways and distance to the nearest railway for each grid cell in Europe.
}
\author{
Ahmed El-Gabbas
}
