% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Railway_Intensity.R
\name{Railway_Intensity}
\alias{Railway_Intensity}
\title{Calculate railway intensity based on \code{OpenStreetMap} data}
\usage{
Railway_Intensity(
  FromHPC = TRUE,
  EnvFile = ".env",
  NCores = 6,
  DeleteProcessed = TRUE
)
}
\arguments{
\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default is ".env".}

\item{NCores}{Numeric. Number of CPU cores to use for parallel processing.
Default is 6.}

\item{DeleteProcessed}{Logical indicating whether to delete the raw
downloaded railways files after processing them. This helps to free large
unnecessary file space (> 55 GB). Defaults to \code{TRUE}.}
}
\value{
\code{NULL}. Outputs processed files to the directories specified in the
environment file.
}
\description{
This function downloads, processes, and analyzes railway data extracted from
\href{https://www.openrailwaymap.org}{OpenRailwayMap} available from
\href{https://download.geofabrik.de/}{OpenStreetMap Data Extracts}. It supports
parallel processing for faster execution and can calculate the total length
of railways and distance to the nearest railway for each grid cell in Europe.
}
\author{
Ahmed El-Gabbas
}
