% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Range2NewVal.R
\name{Range2NewVal}
\alias{Range2NewVal}
\title{Changes values within a specified range or larger than or equals to / less than or equals to specific value to a new value in a vector, data.frame, or raster.}
\usage{
Range2NewVal(
  x,
  Between = NULL,
  MoreThan = NULL,
  LessThan = NULL,
  NewVal,
  InvertSelection = FALSE
)
}
\arguments{
\item{x}{A vector, data.frame, or RasterLayer object whose values are to be modified.}

\item{Between}{A numeric vector of length 2 specifying the range of values to be changed or kept. If specified, \code{MoreThan} and \code{LessThan} are ignored.}

\item{MoreThan, LessThan}{A numeric value specifying the threshold above/below which values in \code{x} will be changed to \code{NewVal}. Only applied if \code{Between} is not specified.}

\item{NewVal}{The new value to assign to the selected elements in \code{x}.}

\item{InvertSelection}{A logical value indicating whether to invert the selection specified by \code{Between}. If \code{TRUE}, values outside the specified range are changed to \code{NewVal}. Default is \code{FALSE}.}
}
\value{
The modified object \code{x} with values changed according to the specified conditions.
}
\description{
This function modifies values in the input object \code{x} based on the specified conditions. It can operate on vectors, data.frames, or RasterLayer objects. The function allows for changing values within a specified range (\code{Between}), greater than or equals to a specified value (\code{MoreThan}), or less than or equals to a specified value (\code{LessThan}) to a new value (\code{NewVal}). An option to invert the selection is also available for ranges.
}
\examples{
# Vector

Range2NewVal(x =  1:10, Between = c(5, 8), NewVal = NA)


Range2NewVal(x =  1:10, Between = c(5, 8), NewVal = NA, InvertSelection = TRUE)

Range2NewVal(x =  1:10, Between = c(5, 8), NewVal = NA, MoreThan = 4)

# ---------------------------------------------

# tibble

iris \%>\%
 tibble::as_tibble() \%>\%
 dplyr::slice_head(n = 50) \%>\%
 dplyr::select(-Sepal.Length, -Petal.Length, -Petal.Width) \%>\%
 dplyr::mutate(
   Sepal.Width.New = Range2NewVal(
       x = Sepal.Width, Between = c(3, 3.5), NewVal = NA, InvertSelection = FALSE),
   Sepal.Width.Rev = Range2NewVal(
       x = Sepal.Width, Between = c(3, 3.5), NewVal = NA, InvertSelection = TRUE)) \%>\%
 dplyr::arrange(-Sepal.Width) \%>\%
 print(n = 50)

# ---------------------------------------------

# raster

library(raster)

RRR <- system.file("external/test.grd", package = "raster") \%>\%
    raster::raster()

RRR2 <- Range2NewVal(x = RRR, LessThan = 500, NewVal = NA)
RRR3 <- Range2NewVal(x = RRR, MoreThan = 500, NewVal = NA)
par(mar = c(0.5, 0.5, 3, 3))
plot(raster::stack(RRR, RRR2, RRR3), nr = 1, main = c("Original", "<500 to NA", ">500 to NA"))

RRR2 <- Range2NewVal(x = RRR, Between = c(1000, 1800), NewVal = 1800, InvertSelection = FALSE)
RRR3 <- Range2NewVal(x = RRR, Between = c(1000, 1800), NewVal = 1800, InvertSelection = TRUE)
plot(stack(RRR>=1000, RRR2, RRR3), nr = 1, main = c(">1000 ?", "<500 to NA", ">500 to NA"))
}
\author{
Ahmed El-Gabbas
}
