% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{Range2NewVal}
\alias{Range2NewVal}
\title{Change values of \code{vector}/\code{data.frame}/\code{raster} within/outside a certain range to another value}
\usage{
Range2NewVal(
  x = NULL,
  Between = NULL,
  MoreThan = NULL,
  LessThan = NULL,
  NewVal = NA,
  InvertSelection = FALSE
)
}
\arguments{
\item{x}{vector / data.frame / raster object.}

\item{Between}{Range of values to be changed or kept.}

\item{MoreThan}{The value above which the original values will be changed. Only applied if \code{Between} is not used.}

\item{LessThan}{The value below which the original values will be changed. Only applied if \code{Between} and \code{MoreThan} are not used.}

\item{NewVal}{The new value to be assigned. Default: \code{NA}.}

\item{InvertSelection}{Should the selection be inverted? Only valid if \code{Between} used. If \code{InvertSelection = TRUE}, then the values not in the range of \code{Between} argument will be changed. Default: \code{FALSE}.}
}
\description{
Change values of \code{vector}/\code{data.frame}/\code{raster} within/outside a certain range to another value
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(raster, warn.conflicts = FALSE)

# ---------------------------------------------

# Vector

Range2NewVal(x =  1:10, Between = c(5, 8), NewVal = NA)

Range2NewVal(x =  1:10, Between = c(5, 8), NewVal = NA, InvertSelection = TRUE)

# ---------------------------------------------

# tibble

iris \%>\% as_tibble() \%>\%
 slice_head(n = 50) \%>\%
 dplyr::select(-Sepal.Length, -Petal.Length, -Petal.Width) \%>\%
 mutate(
   Sepal.Width.New = Range2NewVal(
       x = Sepal.Width, Between = c(3, 3.5),
       NewVal = NA, InvertSelection = FALSE),
   Sepal.Width.Rev = Range2NewVal(
       x = Sepal.Width, Between = c(3, 3.5),
       NewVal = NA, InvertSelection = TRUE)) \%>\%
 arrange(-Sepal.Width) \%>\%
 print(n = 50)

# ---------------------------------------------

# raster

require(raster)

RRR <- system.file("external/test.grd", package = "raster") \%>\%
    raster::raster()

RRR2 <- Range2NewVal(x = RRR, LessThan = 500, NewVal = NA)
RRR3 <- Range2NewVal(x = RRR, MoreThan = 500, NewVal = NA)
par(mar = c(0.5, 0.5, 3, 3))
plot(raster::stack(RRR, RRR2, RRR3), nr = 1, main = c("Original", "<500 to NA", ">500 to NA"))

RRR2 <- Range2NewVal(x = RRR, Between = c(1000, 1800), NewVal = 1800, InvertSelection = FALSE)
RRR3 <- Range2NewVal(x = RRR, Between = c(1000, 1800), NewVal = 1800, InvertSelection = TRUE)
plot(raster::stack(RRR>=1000, RRR2, RRR3), nr = 1, main = c(">1000 ?", "<500 to NA", ">500 to NA"))
}
\author{
Ahmed El-Gabbas
}
