% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_spatial.R
\name{RastPA}
\alias{RastPA}
\title{convert raster map into binary (1/0)}
\usage{
RastPA(x, NA_to_0 = TRUE, Zero_to_NA = FALSE)
}
\arguments{
\item{x}{input map; either \code{PackedSpatRaster}, \code{RasterLayer}, or \code{SpatRaster}}

\item{NA_to_0}{logical (default: \code{TRUE}); should NA be replaced with 0}

\item{Zero_to_NA}{logical (default: \code{FALSE}); should 0 be replaced with NA}
}
\description{
convert raster map into binary (1/0)
}
\examples{
IASDT.R::LoadPackages(dplyr, raster, ggplot2, tidyterra)

r <- raster::raster(system.file("external/test.grd", package = "raster"))

ggplot2::ggplot() +
  tidyterra::geom_spatraster(data = terra::rast(r), maxcell = Inf) +
  ggplot2::theme_minimal()

R2 <- raster::stack(
  RastPA(r),                            # NA replaced with 0
  RastPA(r, NA_to_0 = FALSE),           # NA is kept as NA
  RastPA(RastPA(r), Zero_to_NA = TRUE)) # 0 replaced with NA in the second map

ggplot2::ggplot() +
  tidyterra::geom_spatraster(data = terra::as.factor(terra::rast(R2)), maxcell = Inf) +
  ggplot2::facet_wrap(~lyr) +
  ggplot2::scale_fill_manual(values = c("grey30", "red"), na.value = "transparent") +
  ggplot2::theme_minimal()
}
\author{
Ahmed El-Gabbas
}
