% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{ReloadPackage}
\alias{ReloadPackage}
\title{Reload an R package}
\usage{
ReloadPackage(Package = NULL)
}
\arguments{
\item{Package}{The name of packages to reload}
}
\description{
Reload an R package
}
\examples{
library(sf)
ReloadPackage("sf")
}
