% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_spatial.R
\name{Rename_geometry}
\alias{Rename_geometry}
\title{Rename active geometry column of an sf object}
\usage{
Rename_geometry(g, name)
}
\arguments{
\item{g}{simple feature object}

\item{name}{the new name of geometry}
}
\description{
Rename active geometry column of an sf object
}
\author{
Ahmed El-Gabbas
}
