% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_ReplaceSpace.R
\name{ReplaceSpace}
\alias{ReplaceSpace}
\title{Replace space with underscore}
\usage{
ReplaceSpace(x)
}
\arguments{
\item{x}{string}
}
\description{
Replace space with underscore
}
\examples{
ReplaceSpace("Genus species")

ReplaceSpace("Genus species subspecies")
}
\author{
Ahmed El-Gabbas
}
