% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_RequireMultiple.R
\name{RequireMultiple}
\alias{RequireMultiple}
\title{Load (or install) multiple packages at once}
\usage{
RequireMultiple(..., Reload = FALSE, InstallMissing = TRUE)
}
\arguments{
\item{...}{Packages to load or install}

\item{Reload}{Should the already loaded packages re-loaded? Default: \code{FALSE}}

\item{InstallMissing}{Should the missing packages be installed and loaded? Default: \code{TRUE}}
}
\description{
Load (or install) multiple packages at once
}
\examples{
# Currently loaded packages
(P1 <- LoadedPackages())

RequireMultiple(dismo, MASS, mgcv)

# Loaded packages after implementing the function
(P2 <- LoadedPackages())

# Which pacakages were loaded?
setdiff(P2, P1)
}
\author{
Ahmed El-Gabbas
}
