% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PlotSp.R
\name{RespCurv_PlotSp}
\alias{RespCurv_PlotSp}
\title{Plot species response curves}
\usage{
RespCurv_PlotSp(
  ModelDir = NULL,
  NCores = 20,
  EnvFile = ".env",
  FromHPC = TRUE,
  ReturnData = FALSE
)
}
\arguments{
\item{ModelDir}{String. Path to the root directory of the fitted models
without the trailing slash. The function reads data from \code{RespCurv_DT}
subdirectory created by \link{RespCurv_PrepData}.}

\item{NCores}{Integer. Number of cores to use for parallel processing.
Defaults to 20.}

\item{EnvFile}{String. Path to the environment variables file. Defaults to
".env".}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}

\item{ReturnData}{Logical. Indicates whether the output data be returned as
an R object.}
}
\description{
Generates and saves response curve plots for species based on environmental
variables and other factors. Plots show predicted habitat suitability across
different values of each variable.
}
\author{
Ahmed El-Gabbas
}
