% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PlotSp.R
\name{RespCurv_PlotSp}
\alias{RespCurv_PlotSp}
\title{Plot species response curves}
\usage{
RespCurv_PlotSp(
  Path_RC = NULL,
  NCores = 20,
  EnvFile = ".env",
  SaveGG = FALSE,
  ShowProgress = FALSE
)
}
\arguments{
\item{Path_RC}{String. Path to response curve folder.}

\item{NCores}{Integer. Number of parallel cores}

\item{EnvFile}{String. Path to read the environment variables. Default value: \code{.env}}

\item{SaveGG}{Logical. Should the ggplot object be saved as RData file. Default: \code{FALSE}}

\item{ShowProgress}{Logical. Show progress bar. Default: \code{FALSE}}
}
\description{
Plot species response curves
}
