% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PlotSpAll.R
\name{RespCurv_PlotSpAll}
\alias{RespCurv_PlotSpAll}
\title{Plot all species response curves}
\usage{
RespCurv_PlotSpAll(
  ModelDir = NULL,
  NCores = 8L,
  ReturnData = FALSE,
  PlotAlpha = 0.3
)
}
\arguments{
\item{ModelDir}{String. Path to the root directory of the fitted models
without the trailing slash. The function reads data from \code{RespCurv_DT}
subdirectory created by \link{RespCurv_PrepData}.}

\item{NCores}{Integer. Number of cores to use for parallel processing.
Defaults to 20.}

\item{ReturnData}{Logical. Indicates whether the output data be returned as
an R object.}

\item{PlotAlpha}{Numeric. The alpha value (transparency) for the response
curve lines. Defaults to 0.3.}
}
\description{
Generates and saves response curve plots for all species together in a single
plot.
}
