% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PlotSR.R
\name{RespCurv_PlotSR}
\alias{RespCurv_PlotSR}
\title{Plot species richness response curves}
\usage{
RespCurv_PlotSR(Path_RC)
}
\arguments{
\item{Path_RC}{String. Path to response curve folder.}
}
\description{
Plot species richness response curves
}
