% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PlotSR.R
\name{RespCurv_PlotSR}
\alias{RespCurv_PlotSR}
\title{Plot species richness response curves}
\usage{
RespCurv_PlotSR(Path_Model)
}
\arguments{
\item{Path_Model}{String. The path to the directory containing the models.
The function reads data from the \code{Model_Postprocessing/RespCurv_SR}
sub-directory, resulted from \link{RespCurv_PrepData}.}
}
\value{
This function does not return a value but saves JPEG images of the
response curves in a subdirectory within the specified path.
}
\description{
This function generates and saves species richness response curves as JPEG
images for each variable in the dataset.
}
\seealso{
RespCurv_PlotSp RespCurv_PrepData
}
\author{
Ahmed El-Gabbas
}
