% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PlotSR.R
\name{RespCurv_PlotSR}
\alias{RespCurv_PlotSR}
\title{Plot species richness response curves}
\usage{
RespCurv_PlotSR(ModelDir, Verbose = TRUE, NCores = 8)
}
\arguments{
\item{ModelDir}{String. Path to the root directory of the fitted models
without the trailing slash. The function reads data from \code{RespCurv_DT}
subdirectory created by \link{RespCurv_PrepData}.}

\item{Verbose}{Logical. If TRUE, detailed output is printed. Default is
\code{FALSE}.}

\item{NCores}{Integer. Number of cores to use for parallel processing.
Defaults to 8.}
}
\value{
This function does not return a value but saves JPEG images of the
response curves in a subdirectory within the specified path.
}
\description{
Generate species richness response curves as JPEG images for each variable
used in the model.
}
\seealso{
RespCurv_PlotSp RespCurv_PrepData
}
\author{
Ahmed El-Gabbas
}
