% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PrepData.R
\name{RespCurv_PrepData}
\alias{RespCurv_PrepData}
\title{Prepare and Process Response Curve Data for Hmsc Models}
\usage{
RespCurv_PrepData(
  Path_Model = NULL,
  N_Grid = 50,
  NCores = 8,
  ReturnData = FALSE,
  Probabilities = c(0.025, 0.5, 0.975),
  UseTF = TRUE,
  TF_Environ = NULL,
  Temp_Dir = "TEMP2Pred",
  Verbose = TRUE
)
}
\arguments{
\item{Path_Model}{Character. Path file containing the model.}

\item{N_Grid}{Integer specifying the number of points along the gradient for
continuous focal variables. Defaults to 50. See \link[Hmsc:constructGradient]{Hmsc::constructGradient}
for more details.}

\item{NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 8.}

\item{ReturnData}{Logical. If \code{TRUE}, returns processed response curve data
as an R object. Default is \code{FALSE}.}

\item{Probabilities}{quantiles to be calculated. Defaults to \code{c(0.025, 0.5, 0.975)}. See \link[stats:quantile]{stats::quantile} for more details.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{Temp_Dir}{Character string specifying the path for temporary storage of
intermediate files.}

\item{Verbose}{Logical. If TRUE, detailed output is printed. Default is
\code{FALSE}.}
}
\value{
Depending on the value of \code{ReturnData}, either returns response curve
data or \code{NULL} invisibly.
}
\description{
This function prepares and processes data for generating response curves for
Hmsc models. It supports parallel processing and can return the processed
data.
}
\seealso{
RespCurv_PlotSp RespCurv_PlotSR
}
