% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_RespCurv_PrepData.R
\name{RespCurv_PrepData}
\alias{RespCurv_PrepData}
\title{Prepare and Process Response Curve Data for Hmsc Models}
\usage{
RespCurv_PrepData(
  Path_Model = NULL,
  ngrid = 50,
  NCores = 15,
  ReturnData = FALSE
)
}
\arguments{
\item{Path_Model}{String specifying the path to the .RData file containing
the model to be used.}

\item{ngrid}{Integer specifying the number of points along the gradient for
continuous focal variables. Defaults to 50.}

\item{NCores}{Integer specifying the number of cores to use for parallel
processing. Defaults to 15.}

\item{ReturnData}{Logical indicating whether the processed response curve
data should be returned as an R object. Defaults to \code{FALSE}.}
}
\value{
Depending on the value of \code{ReturnData}, either returns response curve
data or \code{NULL} invisibly.
}
\description{
This function prepares and processes data for generating response curves for
Hmsc models. It supports parallel processing and can return the processed
data.
}
\seealso{
RespCurv_PlotSp RespCurv_PlotSR
}
