% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resp_curv_prepare_data.R,
%   R/mod_resp_curv_plot_species.R, R/mod_resp_curv_plot_species_all.R,
%   R/mod_resp_curv_plot_SR.R
\name{response_curves}
\alias{response_curves}
\alias{resp_curv_prepare_data}
\alias{resp_curv_plot_species}
\alias{resp_curv_plot_species_all}
\alias{resp_curv_plot_SR}
\title{Prepare and plot response curve data for Hmsc models}
\usage{
resp_curv_prepare_data(
  path_model = NULL,
  n_grid = 50L,
  n_cores = 8L,
  strategy = "multisession",
  return_data = FALSE,
  probabilities = c(0.025, 0.5, 0.975),
  use_TF = TRUE,
  TF_environ = NULL,
  TF_use_single = FALSE,
  LF_n_cores = n_cores,
  LF_check = FALSE,
  LF_temp_cleanup = TRUE,
  LF_commands_only = FALSE,
  temp_dir = "TEMP_Pred",
  temp_cleanup = TRUE,
  verbose = TRUE
)

resp_curv_plot_species(
  model_dir = NULL,
  n_cores = 20,
  env_file = ".env",
  return_data = FALSE
)

resp_curv_plot_species_all(
  model_dir = NULL,
  n_cores = 8L,
  strategy = "multisession",
  return_data = FALSE,
  plotting_alpha = 0.3
)

resp_curv_plot_SR(
  model_dir,
  verbose = TRUE,
  n_cores = 8L,
  strategy = "multisession"
)
}
\arguments{
\item{path_model}{Character. Path to the file containing the fitted Hmsc
model.}

\item{n_grid}{Integer. Number of points along the gradient for continuous
focal variables. Higher values result in smoother curves. Default: 50. See
\link[Hmsc:constructGradient]{Hmsc::constructGradient} for details.}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing.
Defaults to 8L for all functions, except for \code{resp_curv_plot_species}, in
which it defaults to 20L.}

\item{strategy}{Character. The parallel processing strategy to use. Valid
options are "sequential", "multisession" (default), "multicore", and
"cluster". See \code{\link[future:plan]{future::plan()}} and \code{\link[ecokit:set_parallel]{ecokit::set_parallel()}} for details.}

\item{return_data}{Logical. If \code{TRUE}, the function returns processed data as
an R object. Default: \code{FALSE}.}

\item{probabilities}{Numeric vector. Quantiles to calculate in response curve
predictions. Default: \code{c(0.025, 0.5, 0.975)}. See \link[stats:quantile]{stats::quantile} for
details.}

\item{use_TF}{Logical. Whether to use \code{TensorFlow} for calculations. Defaults
to \code{TRUE}.}

\item{TF_environ}{Character. Path to the Python environment. This argument is
required if \code{use_TF} is \code{TRUE} under Windows. Defaults to \code{NULL}.}

\item{TF_use_single}{Logical. Whether to use single precision for the
\code{TensorFlow} calculations. Defaults to \code{FALSE}.}

\item{LF_n_cores}{Integer. Number of cores to use for parallel processing of
latent factor prediction. Defaults to 8L.}

\item{LF_check}{Logical. If \code{TRUE}, the function checks if the output files
are already created and valid. If \code{FALSE}, the function will only check if
the files exist without checking their integrity. Default is \code{FALSE}.}

\item{LF_temp_cleanup}{Logical. Whether to delete temporary files in the
\code{temp_dir} directory after finishing the LF predictions.}

\item{LF_commands_only}{Logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}

\item{temp_dir}{Character. Path for temporary storage of intermediate files.}

\item{temp_cleanup}{Logical. Whether to clean up temporary files. Defaults to
\code{TRUE}.}

\item{verbose}{Logical. Whether to print a message upon successful saving of
files. Defaults to \code{FALSE}.}

\item{model_dir}{Character. Path to the root directory containing fitted
models. The function reads data from the \code{RespCurv_DT} subdirectory, which
is created by \code{resp_curv_prepare_data}.}

\item{env_file}{Character. Path to the environment file containing paths to
data sources. Defaults to \code{.env}.}

\item{plotting_alpha}{Numeric. Opacity level for response curve lines (0 =
fully transparent, 1 = fully opaque). Default: 0.3.}
}
\description{
The \verb{RespCurv_*()} functions process and visualise response curves for Hmsc
models. They support parallel computation and optionally return processed
data. There are four functions in this group:
\itemize{
\item \code{resp_curv_prepare_data()}: Prepares response curve data for analysis
\item \code{resp_curv_plot_species()}: Generates response curve plots for
individual species
\item \code{resp_curv_plot_species_all()}: Generates response curves for all
species together in a single plot
\item \code{resp_curv_plot_SR()}: Plots response curves for species richness.
}
}
\author{
Ahmed El-Gabbas
}
