% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Predict_LF.R
\name{run_crossprod_solve}
\alias{run_crossprod_solve}
\title{run_crossprod_solve
Run Crossprod Solve}
\usage{
run_crossprod_solve(
  TF_Environ,
  s1,
  s2,
  postEta,
  path_out,
  denom,
  chunk_size = 1000L,
  threshold_mb = 2000L,
  use_single = TRUE,
  verbose = TRUE,
  solve_chunk_size = 50L,
  solve_max_attempts = 5L,
  LF_Commands_Only = FALSE
)
}
\arguments{
\item{TF_Environ}{character. Path to the virtual environment containing
Python and required dependencies.}

\item{s1}{character. Path to the input file containing s1 coordinates.}

\item{s2}{character Path to the input file containing s2 coordinates.}

\item{postEta}{character. Path to the file containing the \code{postEta} matrix
data.}

\item{path_out}{character. Path to rds file where the output results will be
saved.}

\item{denom}{numeric. A denominator value used in the computation.}

\item{chunk_size}{numeric (Optional). Size of chunks to process at a time.
Default is 1000.}

\item{threshold_mb}{numeric (Optional). Memory threshold (in MB) to manage
processing. Default is 2000.}

\item{use_single}{logical. Indicates whether to use single precision.
Defaults to \code{TRUE}.}

\item{verbose}{logical (Optional). If \code{TRUE}, logs detailed information
during execution. Default is \code{TRUE}.}

\item{solve_chunk_size}{numeric. Chunk size for solve_and_multiply python
function. Default is 50.}

\item{solve_max_attempts}{numeric (Optional). Maximum number of attempts to
run solve and crossprod functions. Default is 5.}

\item{LF_Commands_Only}{logical. If \code{TRUE}, returns the command to run the
Python script. Default is \code{FALSE}.}
}
\value{
Returns the \code{path_out} if successful. Returns \code{NULL} if all attempts
fail.
}
\description{
Internal function to executes a Python script that performs matrix
computations using TensorFlow with provided inputs. Retries up to three times
if the output file validation fails.
}
\details{
\itemize{
\item The function checks for the existence of required input files and the
Python executable in the specified virtual environment.
\item Executes the Python script using \code{system2}.
\item Verifies the output file validity using \code{IASDT.R::CheckData}. Retries up
to 3 times if the output is invalid.
\item Generates detailed logs if \code{verbose} is set to \code{TRUE}.
}
}
\author{
Ahmed El-Gabbas
}
\keyword{internal}
