% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWF_Sampling_Efforts.R
\name{Sampling_Efforts}
\alias{Sampling_Efforts}
\title{Download and Process sampling efforts data}
\usage{
Sampling_Efforts(
  FromHPC = TRUE,
  EnvFile = ".env",
  Renviron = ".Renviron",
  RequestData = TRUE,
  DownloadData = TRUE,
  NCores = 6,
  StartYear = 1980,
  Boundaries = c(-30, 50, 25, 75)
)
}
\arguments{
\item{FromHPC}{Logical indicating whether the work is being done from HPC, to
adjust file paths accordingly. Default: \code{TRUE}.}

\item{EnvFile}{Character. The path to the environment file containing
variables required by the function. Default: \code{.env}.}

\item{Renviron}{Character. The path to the \code{.Renviron} file containing GBIF
login credentials (email, user, password). Default: \code{.Renviron}.}

\item{RequestData}{Logical. If \code{TRUE}, the function requests data from GBIF.
If \code{FALSE}, previously requested data is loaded from disk. Defaults to
\code{TRUE}.}

\item{DownloadData}{Logical. If \code{TRUE}, the function downloads data and
stores it on disk. If \code{FALSE}, it skips the download step. Defaults to
\code{TRUE}.}

\item{NCores}{Integer. The number of cores to use for parallel processing.}

\item{StartYear}{Numeric. The starting year for the occurrence data. Only
records from this year onward will be requested from GBIF. Default: \code{1980}.}

\item{Boundaries}{Numeric vector of length 4. Specifies geographical
boundaries for the requested GBIF data in the order: Left, Right, Bottom,
Top. Default: \code{c(-30, 50, 25, 75)}.}
}
\description{
This script download GBIF data for all vascular plants (grouped by order) in
the study area. Data is converted to raster to represent the number of
vascular plant observation per grid
}
\note{
The function is expected to take substantial amount of time (> 9 hours
on a Windows PC running with 6 cores). The data request is expected to take
5 hours, as it includes waiting for the GBIF data to be ready.
}
\author{
Ahmed El-Gabbas
}
