% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{sapply_}
\alias{sapply_}
\title{Apply a Function over a List or Vector (with no return value)}
\usage{
sapply_(X, FUN, simplify = TRUE, ...)
}
\arguments{
\item{X}{vector}

\item{FUN}{function}

\item{simplify}{should be the output be simplified}

\item{...}{additional arguments to sapply}
}
\description{
Apply a Function over a List or Vector (with no return value)
}
\author{
Ahmed El-Gabbas
}
