% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_SaveAs.R
\name{SaveAs}
\alias{SaveAs}
\title{Save an object with a different name}
\usage{
SaveAs(InObj, OutObj, OutPath)
}
\arguments{
\item{InObj}{input object}

\item{OutObj}{output object}

\item{OutPath}{save path}
}
\description{
Save an object with a different name
}
\examples{
TMP_Folder <- file.path(tempdir(), stringi::stri_rand_strings(1, 5))
fs::dir_create(TMP_Folder)

# save iris data in `iris2.RData` with `iris2` object name
SaveAs(iris, "iris2", file.path(TMP_Folder, "iris2.RData"))

list.files(TMP_Folder, pattern = "^.+.RData")

(load(file.path(TMP_Folder, "iris2.RData")))

tibble::tibble(iris2)
}
\author{
Ahmed El-Gabbas
}
