% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SaveMultiple}
\alias{SaveMultiple}
\title{Save multiple objects to their respective \code{.RData} files}
\usage{
SaveMultiple(Vars = NULL, OutFolder = getwd(), Overwrite = FALSE)
}
\arguments{
\item{Vars}{variables to save}

\item{OutFolder}{output path}

\item{Overwrite}{overwrite existing files}
}
\description{
Save multiple objects to their respective \code{.RData} files
}
\author{
Ahmed El-Gabbas
}
