% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SaveSession}
\alias{SaveSession}
\title{Save all objects (except functions) of the global environment as list items}
\usage{
SaveSession(Path = getwd(), ExcludeObs = NULL)
}
\arguments{
\item{Path}{Path of where to save the output RData file}
}
\description{
Save all objects (except functions) of the global environment as list items
}
\author{
Ahmed El-Gabbas
}
