% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_SaveSession.R
\name{SaveSession}
\alias{SaveSession}
\title{Save all objects (except functions) of the global environment as list items}
\usage{
SaveSession(Path = getwd(), ExcludeObj = NULL, Prefix = "S")
}
\arguments{
\item{Path}{Path of where to save the output RData file}

\item{ExcludeObj}{objects not to save}

\item{Prefix}{file prefix}
}
\description{
Save all objects (except functions) of the global environment as list items
}
\author{
Ahmed El-Gabbas
}
