% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SaveSessionInfo}
\alias{SaveSessionInfo}
\title{Save all objects (except functions) of the global environment as list items}
\usage{
SaveSessionInfo(Path = getwd(), SessionObj = NULL)
}
\arguments{
\item{Path}{Path of where to save the output RData file}

\item{SessionObj}{List of objects and their sizes (typically a a result of \code{IASDT::SaveSession} function)}
}
\description{
Save all objects (except functions) of the global environment as list items
}
\author{
Ahmed El-Gabbas
}
