% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{ScrapLinks}
\alias{ScrapLinks}
\title{Extract link texts and urls from a web page}
\usage{
ScrapLinks(url)
}
\arguments{
\item{url}{the url}
}
\description{
Extract link texts and urls from a web page
}
\examples{
ScrapLinks("https://github.com/")
}
\references{
https://gist.github.com/paulrougieux/e1ee769577b40cd9ed9db7f75e9a2cc2
}
