% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_ScriptLocation.R
\name{ScriptLocation}
\alias{ScriptLocation}
\title{The location of current script}
\usage{
ScriptLocation()
}
\description{
The location of current script
}
\examples{
 \dontrun{
ScriptLocation()
}
}
\references{
\href{https://stackoverflow.com/questions/47044068/}{Click here}
}
