% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_Set_geometry.R
\name{Set_geometry}
\alias{Set_geometry}
\title{Set geometry of an sf object in the pipe line}
\usage{
Set_geometry(x, Name)
}
\arguments{
\item{x}{simple feature data frame}

\item{Name}{the name of the geometry column to be used}
}
\description{
Set geometry of an sf object in the pipe line
}
\author{
Ahmed El-Gabbas
}
