% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Set_parallel.R
\name{Set_parallel}
\alias{Set_parallel}
\title{Set up or stop parallel processing plan}
\usage{
Set_parallel(
  NCores = 1L,
  Strategy = "future::multisession",
  Stop = FALSE,
  Cat = TRUE,
  Future_maxSize = 8L,
  ...
)
}
\arguments{
\item{NCores}{Integer. Number of cores to use. If \code{NULL}, defaults to
sequential mode. Default is \code{1}.}

\item{Strategy}{Character. The parallel processing strategy to use. Valid
options are \code{future::sequential} (sequential), \code{future::multisession}
(default), \code{future::multicore} (not supported on Windows), and
\code{future::cluster}. If \code{Strategy} is not one of the valid options or if
\code{future::multicore} on Windows PC, it defaults to \code{future::multisession}.
See \code{future::plan()} for more details.}

\item{Stop}{Logical. If \code{TRUE}, stops any parallel cluster and resets to
sequential mode. If \code{FALSE} (default), sets up a new plan.}

\item{Cat}{Logical. If \code{TRUE} (default), logs messages via
\code{\link[=CatTime]{CatTime()}}.}

\item{Future_maxSize}{Numeric. Maximum allowed total size (in megabytes) of
global variables identified. See \code{future.globals.maxSize} argument of
\link[future:future.options]{future::future.options} for more details. Default is \code{8}.}

\item{...}{Additional arguments to pass to \link{CatTime}.}
}
\description{
Configures parallel processing with \code{future::plan()} or stops an existing
plan. When stopping, it resets to sequential mode.
}
\examples{
# Prepare working on parallel
IASDT.R::Set_parallel(NCores = 2)
future::plan()

# ---------------------------------------------

# Stopping parallel processing
IASDT.R::Set_parallel(Stop = TRUE)
future::plan()

# ---------------------------------------------

# Prepare working on parallel using `future::cluster`
IASDT.R::Set_parallel(NCores = 2, Strategy = "future::cluster")
future::plan()

# Stopping parallel processing
IASDT.R::Set_parallel(Stop = TRUE)
future::plan()
}
\author{
Ahmed El-Gabbas
}
