% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_SetChainName.R
\name{SetChainName}
\alias{SetChainName}
\title{SetChainName}
\usage{
SetChainName(Obj, Chain)
}
\arguments{
\item{Obj}{Obj}

\item{Chain}{Chain}
}
\description{
SetChainName
}
\author{
Ahmed El-Gabbas
}
