% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve2}
\alias{solve2}
\title{Solve a system of linear equations}
\usage{
solve2(A, B)
}
\arguments{
\item{A}{A square matrix representing the coefficients.}

\item{B}{A matrix or vector representing the constants.}
}
\value{
The solution matrix or vector \code{X}.
}
\description{
Solves a system of linear equations \code{A * X = B} for \code{X}, where \code{A} is a
square matrix and \code{B} is a matrix of compatible dimensions.
}
