% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_sort_.R
\name{sort_}
\alias{sort_}
\title{Sort alphanumeric strings with enhanced options}
\usage{
sort_(
  x,
  decreasing = FALSE,
  na.last = TRUE,
  blank.last = FALSE,
  numeric.type = c("decimal", "roman"),
  roman.case = c("upper", "lower", "both")
)
}
\arguments{
\item{x}{Vector to be sorted.}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
Note that \code{descending=TRUE} reverses the meanings of \code{na.last} and
\code{blanks.last}.}

\item{na.last}{for controlling the treatment of \code{NA} values.  If
\code{TRUE}, missing values in the data are put last; if \code{FALSE}, they
are put first; if \code{NA}, they are removed.}

\item{blank.last}{for controlling the treatment of blank values.  If
\code{TRUE}, blank values in the data are put last; if \code{FALSE}, they
are put first; if \code{NA}, they are removed.}

\item{numeric.type}{either "decimal" (default) or "roman".  Are numeric
values represented as decimal numbers (\code{numeric.type="decimal"}) or as
Roman numerals (\code{numeric.type="roman"})?}

\item{roman.case}{one of "upper", "lower", or "both".  Are roman numerals
represented using only capital letters ('IX') or lower-case letters ('ix')
or both?}
}
\value{
A vector of sorted alphanumeric strings.
}
\description{
This function extends the sorting capabilities for alphanumeric strings by
allowing for sorting of mixed numeric and character strings, with additional
control over sorting direction, treatment of \code{NA} and blank values, and
handling of numeric values represented as either decimal numbers or Roman
numerals. This function is a wrapper function for the \link[gtools:mixedsort]{gtools::mixedsort}
function.
}
\examples{
# example code
(AA <- paste0("V", 1:12))

sort(x = AA)

sort_(x = AA)

}
