% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SourceSilent}
\alias{SourceSilent}
\title{Silently source R script}
\usage{
SourceSilent(File, ...)
}
\arguments{
\item{File}{path of the file to be sourced}

\item{...}{additional arguments passed to \code{source} function}
}
\description{
Silently source R script
}
\author{
Ahmed El-Gabbas
}
