% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_SourceSilent.R
\name{SourceSilent}
\alias{SourceSilent}
\title{Silently source R script}
\usage{
SourceSilent(File, Messages = TRUE, Warnings = TRUE, ...)
}
\arguments{
\item{File}{path of the file to be sourced}

\item{Messages}{Show messages; default: \code{TRUE}}

\item{Warnings}{Show warnings; default: \code{TRUE}}

\item{...}{additional arguments passed to \code{source} function}
}
\description{
Silently source R script
}
\author{
Ahmed El-Gabbas
}
