% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SplitDF2Chunks}
\alias{SplitDF2Chunks}
\title{Split data.frame into smaller chunks}
\usage{
SplitDF2Chunks(DF = NULL, ChunkSize = NULL, NChunks = NULL, Prefix = "Chunk")
}
\arguments{
\item{DF}{dataframe to split}

\item{ChunkSize}{number of rows per chunk}

\item{NChunks}{number of chunks}

\item{Prefix}{prefix}
}
\description{
Split data.frame into smaller chunks
}
\author{
Ahmed El-Gabbas
}
