% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{SplitDF2Chunks}
\alias{SplitDF2Chunks}
\title{Split data.frame into smaller chunks}
\usage{
SplitDF2Chunks(DF = NULL, ChunkSize = NULL, NChunks = NULL, Prefix = "Chunk")
}
\arguments{
\item{DF}{the data frame to split}

\item{ChunkSize}{number of rows per chunk}

\item{NChunks}{number of chunks (only if \code{ChunkSize} not provided)}

\item{Prefix}{prefix}
}
\description{
Split data.frame into smaller chunks
}
\examples{
SplitDF2Chunks(mtcars, ChunkSize = 16)

# -------------------------------------------

SplitDF2Chunks(mtcars, NChunks = 3)

# -------------------------------------------

SplitDF2Chunks(mtcars, NChunks = 3, Prefix = "T")
}
\author{
Ahmed El-Gabbas
}
