% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_spatial.R
\name{SplitRaster}
\alias{SplitRaster}
\title{Split a raster object into a list of smaller rasters}
\usage{
SplitRaster(
  raster,
  Ncol = 4,
  NRow = 4,
  save = FALSE,
  SplitPath = "",
  plot = FALSE,
  Extent = FALSE
)
}
\arguments{
\item{raster}{raster object to split}

\item{Ncol}{number of columns}

\item{NRow}{number of rows}

\item{save}{save output to disk?}

\item{SplitPath}{file path}

\item{plot}{plot the results}

\item{Extent}{return only the extent of split raster}
}
\description{
Split a raster object into a list of smaller rasters
}
\details{
#' References:
https://stackoverflow.com/questions/29784829/
https://stackoverflow.com/questions/22109774/r-raster-mosaic-from-list-of-rasters
}
\author{
Ahmed El-Gabbas
}
