% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spat_SplitRaster.R
\name{SplitRaster}
\alias{SplitRaster}
\title{Split a raster object into a list of smaller rasters}
\usage{
SplitRaster(
  raster,
  Ncol = 4,
  Nrow = 4,
  save = FALSE,
  SplitPath = "",
  plot = FALSE,
  Extent = FALSE
)
}
\arguments{
\item{raster}{raster object to split}

\item{Ncol}{number of columns}

\item{Nrow}{number of rows}

\item{save}{save output to disk?}

\item{SplitPath}{file path}

\item{plot}{plot the results}

\item{Extent}{return only the extent of split raster}
}
\description{
Split a raster object into a list of smaller rasters
}
\details{
#' References:
https://stackoverflow.com/questions/29784829/
https://stackoverflow.com/questions/22109774/r-raster-mosaic-from-list-of-rasters
}
\examples{
LoadPackages(raster)
logo <- raster(system.file("external/rlogo.grd", package = "raster"))
plot(logo, axes = FALSE, legend = FALSE, bty = "n",
     box = FALSE, main = "Original raster layer")
# --------------------------------------------------

# Split into 3 rows and 3 columns
logoSplit <- SplitRaster(logo, Ncol = 3, Nrow = 3, plot = TRUE)

print(logoSplit) # a list object of 9 items

# --------------------------------------------------

# Merging split maps again
logoSplit$fun <- mean
logoSplit$na.rm <- TRUE
logoSplit2 <- do.call(mosaic, logoSplit)
par(mfrow = c(1, 1))
plot(logoSplit2, axes = FALSE, legend = FALSE, bty = "n",
     box = FALSE, main = "Merged raster layers")

print({logoSplit2 - logo}) # No value difference!

# --------------------------------------------------

logoSplit <- SplitRaster(logo, Ncol = 3, Nrow = 3, Extent = TRUE)
print(logoSplit)

}
\author{
Ahmed El-Gabbas
}
