% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_SplitVector.R
\name{SplitVector}
\alias{SplitVector}
\title{Split a vector into smaller chunks}
\usage{
SplitVector(Vector = NULL, NSplit = NULL, Prefix = "Chunk")
}
\arguments{
\item{Vector}{Vector to split}

\item{NSplit}{Integer. Number of splits}

\item{Prefix}{String. Prefix string for the name of the resulted list}
}
\description{
Split a vector into smaller chunks
}
\examples{
SplitVector(1:100, 3)

# -------------------------------------------

SplitVector(1:100, 2, "T")

# -------------------------------------------

\dontrun{
SplitVector(1:100)
}
}
\author{
Ahmed El-Gabbas
}
