% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_general.R
\name{System}
\alias{System}
\title{Run bash script depending on the operating system}
\usage{
System(command, intern = TRUE, ...)
}
\arguments{
\item{command}{bash command to implement}

\item{intern}{whether to make the output of the command an R object}

\item{...}{additional arguments to \code{shell} pr \code{system} functions}
}
\description{
Run bash script depending on the operating system
}
\author{
Ahmed El-Gabbas
}
