% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs_spatial.R
\name{Text2Coords}
\alias{Text2Coords}
\title{Extract longitude / latitude from text}
\usage{
Text2Coords(x)
}
\arguments{
\item{x}{string of coordinate}
}
\description{
Extract longitude / latitude from text
}
\examples{
c("POINT (11.761 46.286)",
 "POINT (14.8336 42.0422)",
 "POINT (16.179999 38.427214)") \%>\%
 lapply(Text2Coords)

}
\author{
Ahmed El-Gabbas
}
