% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Valid_URL.R
\name{Valid_URL}
\alias{Valid_URL}
\title{Check the validity of a URL}
\usage{
Valid_URL(url_in, t = 2)
}
\arguments{
\item{url_in}{URL path}

\item{t}{Timeout}
}
\description{
Check the validity of a URL
}
\examples{
urls <- c("http://www.amazon.com", "http://this.isafakelink.biz", "https://stackoverflow.com")
sapply(urls, Valid_URL)
}
\references{
https://stackoverflow.com/questions/52911812/check-if-url-exists-in-r
}
