% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_VarPar_Plot.R
\name{VarPar_Plot}
\alias{VarPar_Plot}
\title{Plot variance partitioning of Hmsc model}
\usage{
VarPar_Plot(
  Path_Model,
  EnvFile = ".env",
  VarParFile = NULL,
  FromHPC = TRUE,
  UseTF = TRUE,
  TF_Environ = NULL,
  NCores = 1,
  Fig_width = 30,
  Fig_height = 15,
  axis_text = 4
)
}
\arguments{
\item{Path_Model}{Character path for the model file.}

\item{EnvFile}{String. Path to read the environment variables. Default value:
\code{.env}}

\item{VarParFile}{Character. Name of the output file to save the results.}

\item{FromHPC}{Logical. Indicates whether the function is being run on an HPC
environment, affecting file path handling. Default: \code{TRUE}.}

\item{UseTF}{Logical indicating whether to use TensorFlow for calculations.
Defaults to TRUE.}

\item{TF_Environ}{Character string specifying the path to the Python
environment. Defaults to NULL. This argument is required if \code{UseTF} is
TRUE.}

\item{NCores}{Integer. Number of parallel computations for computing variance
partitioning using TensorFlow. See \link{VarPar_Compute} for more details.
Default: \code{1}.}

\item{Fig_width, Fig_height}{Numeric. Width and height of the output plot in
centimeters. Default: \code{30} and \code{15}, respectively.}

\item{axis_text}{Numeric. Size of the axis text. Default: \code{4}.}
}
\description{
This function generates plots for the variance partitioning of an Hmsc model.
It can optionally compute the variance partitioning if not available on disk,
supporting parallel computation and using TensorFlow; see \link{VarPar_Compute}.
It plots the relative variance partitioning sorted by the mean value per
predictor or by the original species order. It also plots the raw variance
partitioning (relative variance partitioning multiplied by the Tjur-R^2
value). The plots are saved as JPEG files.
}
\details{
The function reads the following environment variables:
\itemize{
\item \strong{\code{DP_R_TaxaInfo_RData}} (if \code{FromHPC} = \code{TRUE}) or
\strong{\code{DP_R_TaxaInfo_RData_Local}} (if \code{FromHPC} = \code{FALSE}) for the
location of the \code{TaxaList.RData} file containing species information.
}
}
\author{
Ahmed El-Gabbas
}
